Trait xcb::base::BaseError

pub trait BaseError {
    const EXTENSION: Option<Extension>;
    const NUMBER: u32;

    unsafe fn from_raw(raw: *const u8) -> Self;

    unsafe fn into_raw(self) -> *const u8;

    fn as_slice(&self) -> &[u8];
}

marker trait to designate base events (aka. non GE_GENERIC events)

Associated Constants

const EXTENSION: Option<Extension>

The extension associated to this event, or None for the main protocol

const NUMBER: u32

The number associated to this error

Required methods

unsafe fn from_raw(raw: *const u8) -> Self

Build an error from a raw pointer

Safety

raw must be a valid pointer to the error, and be allocated with libc::malloc

unsafe fn into_raw(self) -> *const u8

Convert the error into a raw pointer

Safety

returned value should be freed with libc::free to avoid memory leak, or used to build a new error

fn as_slice(&self) -> &[u8]

Access to the raw error data