Trait xcb::base::FromWireError

pub trait FromWireError {
    unsafe fn from_wire_error(
        first_error: u8, 
        error: *mut xcb_generic_error_t
    ) -> Self;
}

trait for converting wire data to an error Self is normally an enum of several errors subtypes

Required methods

unsafe fn from_wire_error(
first_error: u8,
error: *mut xcb_generic_error_t
) -> Self

Convert a pointer to xcb_generic_error_t to Self, inferring the correct subtype using response_type field and first_error

Panics

Panics if the error subtype cannot be resolved for self. That is, response_type field must be checked beforehand to be in range with first_error.

Safety

err must be a valid, non-null error obtained by xcb_wait_for_reply or similar function