Trait xcb::base::GeEvent

pub trait GeEvent {
    unsafe fn from_raw(raw: *const u8) -> Self;

    unsafe fn into_raw(self) -> *const u8;
}

marker trait to designate GE_GENERIC events

Required methods

unsafe fn from_raw(raw: *const u8) -> Self

Build an event from a raw pointer

Safety

raw must be a valid pointer to the event, and be allocated with libc::malloc

unsafe fn into_raw(self) -> *const u8

Convert the event into a raw pointer

Safety

returned value should be freed with libc::free to avoid memory leak, or used to build a new event