xcb

Module damage

The DAMAGE X extension.

Accessible with the damage cargo feature.

Structs

Add

The Add request.

BadDamageError

The BadDamageError error.

Create

The Create request.

Damage

Destroy

The Destroy request.

NotifyEvent

The NotifyEvent event.

QueryVersion

The QueryVersion request.

QueryVersionCookie

Cookie type for QueryVersion.

QueryVersionCookieUnchecked

Unchecked cookie type for QueryVersion.

QueryVersionReply

Reply type for QueryVersion.

Subtract

The Subtract request.

Enums

Error

Unified error type for the Damage extension

Event

Unified event type for the Damage extension

ReportLevel

Constants

MAJOR_VERSION

The major version of the Damage extension.

MINOR_VERSION

The minor version of the Damage extension.

VERSION_STRING

The version string of the Damage extension.

XNAME

The official identifier for the Damage extension.

Functions

get_extension_data

Fetch server runtime info data of the Damage extension.

prefetch_extension_data

Prefetch server runtime info data of the Damage extension.