xcb
Enum ConnError

pub enum ConnError {
    Connection,
    ClosedExtNotSupported,
    ClosedMemInsufficient,
    ClosedReqLenExceed,
    ClosedParseErr,
    ClosedInvalidScreen,
    ClosedFdPassingFailed,
    XOpenDisplay,
}

Error type that is returned by Connection::has_error.

Variants

Connection

xcb connection errors because of socket, pipe and other stream errors.

ClosedExtNotSupported

xcb connection shutdown because of extension not supported

ClosedMemInsufficient

malloc(), calloc() and realloc() error upon failure, for eg ENOMEM

ClosedReqLenExceed

Connection closed, exceeding request length that server accepts.

ClosedParseErr

Connection closed, error during parsing display string.

ClosedInvalidScreen

Connection closed because the server does not have a screen matching the display.

ClosedFdPassingFailed

Connection closed because some file descriptor passing operation failed.

XOpenDisplay

XOpenDisplay returned NULL