xcb

Module ffi

Module for Foreign Function Interface bindings.

Structs

xcb_ge_generic_event_t

FFI type for the Generic Event Extension.

xcb_generic_error_t

Generic error

xcb_generic_event_t

Generic event.

Enums

xcb_connection_t

XCB Connection structure.

Constants

XCBOwnsEventQueue

XCB owns the event queue

XCB_CONN_CLOSED_EXT_NOTSUPPORTED

xcb connection shutdown because of extension not supported

XCB_CONN_CLOSED_FDPASSING_FAILED

Connection closed because some FD passing operation failed

XCB_CONN_CLOSED_INVALID_SCREEN

Connection closed because the server does not have a screen matching the display.

XCB_CONN_CLOSED_MEM_INSUFFICIENT

malloc(), calloc() and realloc() error upon failure, for eg ENOMEM

XCB_CONN_CLOSED_PARSE_ERR

Connection closed, error during parsing display string.

XCB_CONN_CLOSED_REQ_LEN_EXCEED

Connection closed, exceeding request length that server accepts.

XCB_CONN_ERROR

xcb connection errors because of socket, pipe and other stream errors.

XCB_GE_GENERIC

response_type number corresponding to a xcb_ge_generic_event_t.

X_PROTOCOL

Current protocol version

X_PROTOCOL_REVISION

Current minor version

X_TCP_PORT

X_TCP_PORT + display number = server port for TCP transport

XlibOwnsEventQueue

Xlib owns the event queue

Functions

XGetXCBConnection⚠

Get an XCB connection from the xlib::Display.

XSetEventQueueOwner⚠

Set the owner of the X client event queue.

Type Aliases

XEventQueueOwner

Type for XSetEventQueueOwner owner parameter