Function xcb::ffi::base::xcb_connect_to_display_with_auth_info

pub unsafe extern "C" fn xcb_connect_to_display_with_auth_info(
    display: *const c_char, 
    auth: *mut xcb_auth_info_t, 
    screen: *mut c_int
) -> *mut xcb_connection_t

Connects to the X server, using an authorization information. display: The name of the display. auth: The authorization information. screen: A pointer to a preferred screen number. Returns A newly allocated xcb_connection_t structure.

Connects to the X server specified by displayname, using the authorization auth. If a particular screen on that server is preferred, the int pointed to by screenp (if not NULL) will be set to that screen; otherwise screenp will be set to 0.

Always returns a non-NULL pointer to a xcb_connection_t, even on failure. Callers need to use xcb_connection_has_error() to check for failure. When finished, use xcb_disconnect() to close the connection and free the structure.