xcb::ffi::base::xcb_connect_to_fd

pub unsafe extern fn xcb_connect_to_fd(fd: c_int, auth_info: *mut xcb_auth_info_t) -> *mut xcb_connection_t

Connects to the X server.

Connects to an X server, given the open socket @p fd and the xcb_auth_info_t @p auth_info. The file descriptor @p fd is bidirectionally connected to an X server. If the connection should be unauthenticated, @p auth_info must be @c NULL.

Always returns a non-NULL pointer to a xcb_connection_t, even on failure. Callers need to use xcb_connection_has_error() to check for failure. When finished, use xcb_disconnect() to close the connection and free the structure.