Function xcb::ffi::base::xcb_connection_has_error

pub unsafe extern "C" fn xcb_connection_has_error(
    c: *mut xcb_connection_t
) -> c_int

Test whether the connection has shut down due to a fatal error.

Some errors that occur in the context of an xcb_connection_t are unrecoverable. When such an error occurs, the connection is shut down and further operations on the xcb_connection_t have no effect, but memory will not be freed until xcb_disconnect() is called on the xcb_connection_t.

Returns XCB_CONN_ERROR, because of socket errors, pipe errors or other stream errors. Returns XCB_CONN_CLOSED_EXT_NOTSUPPORTED, when extension not supported. Returns XCB_CONN_CLOSED_MEM_INSUFFICIENT, when memory not available. Returns XCB_CONN_CLOSED_REQ_LEN_EXCEED, exceeding request length that server accepts. Returns XCB_CONN_CLOSED_PARSE_ERR, error during parsing display string. Returns XCB_CONN_CLOSED_INVALID_SCREEN, because the server does not have a screen matching the display.

Returns > 0 if the connection is in an error state; 0 otherwise.