Function xcb::ffi::base::xcb_get_extension_data

pub unsafe extern "C" fn xcb_get_extension_data(
    c: *mut xcb_connection_t, 
    ext: *mut xcb_extension_t
) -> *const xcb_query_extension_reply_t

Caches reply information from QueryExtension requests.

This function is the primary interface to the “extension cache”, which caches reply information from QueryExtension requests. Invoking this function may cause a call to xcb_query_extension to retrieve extension information from the server, and may block until extension data is received from the server.

The result must not be freed. This storage is managed by the cache itself.

Returns A pointer to the xcb_query_extension_reply_t for the extension.