Function xcb::ffi::base::xcb_get_maximum_request_length

pub unsafe extern "C" fn xcb_get_maximum_request_length(
    c: *mut xcb_connection_t
) -> u32

Returns the maximum request length that this server accepts.

In the absence of the BIG-REQUESTS extension, returns the maximum request length field from the connection setup data, which may be as much as 65535. If the server supports BIG-REQUESTS, then the maximum request length field from the reply to the BigRequestsEnable request will be returned instead.

Note that this length is measured in four-byte units, making the theoretical maximum lengths roughly 256kB without BIG-REQUESTS and 16GB with.

Returns The maximum request length field.