Function xcb::ffi::base::xcb_parse_display

pub unsafe extern "C" fn xcb_parse_display(
    name: *const c_char, 
    host: *mut *mut c_char, 
    display: *mut c_int, 
    screen: *mut c_int
) -> c_int

Parses a display string name in the form documented by X(7x). name: The name of the display. host: A pointer to a malloc'd copy of the hostname. display: A pointer to the display number. screen: A pointer to the screen number.

Parses the display string name display_name in the form documented by X(7x). Has no side effects on failure. If displayname is NULL or empty, it uses the environment variable DISPLAY. hostp is a pointer to a newly allocated string that contain the host name. displayp is set to the display number and screenp to the preferred screen number. screenp can be NULL. If displayname does not contain a screen number, it is set to 0.

Returns 0 on failure, non 0 otherwise.