xcb::ffi::base

Function xcb_request_check

pub unsafe extern "C" fn xcb_request_check(
    c: *mut xcb_connection_t,
    cookie: xcb_void_cookie_t,
) -> *mut xcb_generic_error_t

Return the error for a request, or NULL if none can ever arrive.

The xcb_void_cookie_t cookie supplied to this function must have resulted from a call to xcb_[request_name]_checked(). This function will block until one of two conditions happens. If an error is received, it will be returned. If a reply to a subsequent request has already arrived, no error can arrive for this request, so this function will return NULL.

Note that this function will perform a sync if needed to ensure that the sequence number will advance beyond that provided in cookie; this is a convenience to avoid races in determining whether the sync is needed.