Struct xcb::ffi::base::xcb_auth_info_t

#[repr(C)]

pub struct xcb_auth_info_t {
    pub namelen: c_int,
    pub name: *mut c_char,
    pub datalen: c_int,
    pub data: *mut c_char,
}

Container for authorization information. A container for authorization information to be sent to the X server

Fields
namelen: c_int

length of the string name (as returned by strlen)

name: *mut c_char

String containing the authentication protocol name, such as "MIT-MAGIC-COOKIE-1" or "XDM-AUTHORIZATION-1".

datalen: c_int

length of the data member

data: *mut c_char

data interpreted in a protocol specific manner