Module xcb::ffi::composite

Structs

xcb_composite_create_region_from_border_clip_request_t

xcb_composite_get_overlay_window_cookie_t

xcb_composite_get_overlay_window_reply_t

xcb_composite_get_overlay_window_request_t

xcb_composite_name_window_pixmap_request_t

xcb_composite_query_version_cookie_t

xcb_composite_query_version_reply_t

xcb_composite_query_version_request_t

xcb_composite_redirect_subwindows_request_t

xcb_composite_redirect_window_request_t

xcb_composite_release_overlay_window_request_t

xcb_composite_unredirect_subwindows_request_t

xcb_composite_unredirect_window_request_t

Constants

XCB_COMPOSITE_CREATE_REGION_FROM_BORDER_CLIP

XCB_COMPOSITE_GET_OVERLAY_WINDOW

XCB_COMPOSITE_MAJOR_VERSION

XCB_COMPOSITE_MINOR_VERSION

XCB_COMPOSITE_NAME_WINDOW_PIXMAP

XCB_COMPOSITE_QUERY_VERSION

XCB_COMPOSITE_REDIRECT_AUTOMATIC

XCB_COMPOSITE_REDIRECT_MANUAL

XCB_COMPOSITE_REDIRECT_SUBWINDOWS

XCB_COMPOSITE_REDIRECT_WINDOW

XCB_COMPOSITE_RELEASE_OVERLAY_WINDOW

XCB_COMPOSITE_UNREDIRECT_SUBWINDOWS

XCB_COMPOSITE_UNREDIRECT_WINDOW

Statics

xcb_composite_id

Functions

xcb_composite_create_region_from_border_clip⚠

xcb_composite_create_region_from_border_clip_checked⚠

xcb_composite_get_overlay_window⚠

xcb_composite_get_overlay_window_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_composite_get_overlay_window_unchecked⚠

xcb_composite_name_window_pixmap⚠

xcb_composite_name_window_pixmap_checked⚠

xcb_composite_query_version⚠

xcb_composite_query_version_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_composite_query_version_unchecked⚠

xcb_composite_redirect_subwindows⚠

xcb_composite_redirect_subwindows_checked⚠

xcb_composite_redirect_window⚠

xcb_composite_redirect_window_checked⚠

xcb_composite_release_overlay_window⚠

xcb_composite_release_overlay_window_checked⚠

xcb_composite_unredirect_subwindows⚠

xcb_composite_unredirect_subwindows_checked⚠

xcb_composite_unredirect_window⚠

xcb_composite_unredirect_window_checked⚠

Type Definitions

xcb_composite_redirect_t