Module xcb::ffi::dri2

Structs

xcb_dri2_attach_format_iterator_t

xcb_dri2_attach_format_t

xcb_dri2_authenticate_cookie_t

xcb_dri2_authenticate_reply_t

xcb_dri2_authenticate_request_t

xcb_dri2_buffer_swap_complete_event_t

xcb_dri2_connect_cookie_t

xcb_dri2_connect_reply_t

xcb_dri2_connect_request_t

xcb_dri2_copy_region_cookie_t

xcb_dri2_copy_region_reply_t

xcb_dri2_copy_region_request_t

xcb_dri2_create_drawable_request_t

xcb_dri2_destroy_drawable_request_t

xcb_dri2_dri2_buffer_iterator_t

xcb_dri2_dri2_buffer_t

xcb_dri2_get_buffers_cookie_t

xcb_dri2_get_buffers_reply_t

xcb_dri2_get_buffers_request_t

xcb_dri2_get_buffers_with_format_cookie_t

xcb_dri2_get_buffers_with_format_reply_t

xcb_dri2_get_buffers_with_format_request_t

xcb_dri2_get_msc_cookie_t

xcb_dri2_get_msc_reply_t

xcb_dri2_get_msc_request_t

xcb_dri2_get_param_cookie_t

xcb_dri2_get_param_reply_t

xcb_dri2_get_param_request_t

xcb_dri2_invalidate_buffers_event_t

xcb_dri2_query_version_cookie_t

xcb_dri2_query_version_reply_t

xcb_dri2_query_version_request_t

xcb_dri2_swap_buffers_cookie_t

xcb_dri2_swap_buffers_reply_t

xcb_dri2_swap_buffers_request_t

xcb_dri2_swap_interval_request_t

xcb_dri2_wait_msc_cookie_t

xcb_dri2_wait_msc_reply_t

xcb_dri2_wait_msc_request_t

xcb_dri2_wait_sbc_cookie_t

xcb_dri2_wait_sbc_reply_t

xcb_dri2_wait_sbc_request_t

Constants

XCB_DRI2_MAJOR_VERSION

XCB_DRI2_MINOR_VERSION

XCB_DRI_2_ATTACHMENT_BUFFER_ACCUM

XCB_DRI_2_ATTACHMENT_BUFFER_BACK_LEFT

XCB_DRI_2_ATTACHMENT_BUFFER_BACK_RIGHT

XCB_DRI_2_ATTACHMENT_BUFFER_DEPTH

XCB_DRI_2_ATTACHMENT_BUFFER_DEPTH_STENCIL

XCB_DRI_2_ATTACHMENT_BUFFER_FAKE_FRONT_LEFT

XCB_DRI_2_ATTACHMENT_BUFFER_FAKE_FRONT_RIGHT

XCB_DRI_2_ATTACHMENT_BUFFER_FRONT_LEFT

XCB_DRI_2_ATTACHMENT_BUFFER_FRONT_RIGHT

XCB_DRI_2_ATTACHMENT_BUFFER_HIZ

XCB_DRI_2_ATTACHMENT_BUFFER_STENCIL

XCB_DRI_2_AUTHENTICATE

XCB_DRI_2_BUFFER_SWAP_COMPLETE

XCB_DRI_2_CONNECT

XCB_DRI_2_COPY_REGION

XCB_DRI_2_CREATE_DRAWABLE

XCB_DRI_2_DESTROY_DRAWABLE

XCB_DRI_2_DRIVER_TYPE_DRI

XCB_DRI_2_DRIVER_TYPE_VDPAU

XCB_DRI_2_EVENT_TYPE_BLIT_COMPLETE

XCB_DRI_2_EVENT_TYPE_EXCHANGE_COMPLETE

XCB_DRI_2_EVENT_TYPE_FLIP_COMPLETE

XCB_DRI_2_GET_BUFFERS

XCB_DRI_2_GET_BUFFERS_WITH_FORMAT

XCB_DRI_2_GET_MSC

XCB_DRI_2_GET_PARAM

XCB_DRI_2_INVALIDATE_BUFFERS

XCB_DRI_2_QUERY_VERSION

XCB_DRI_2_SWAP_BUFFERS

XCB_DRI_2_SWAP_INTERVAL

XCB_DRI_2_WAIT_MSC

XCB_DRI_2_WAIT_SBC

Statics

xcb_dri2_id

Functions

xcb_dri2_attach_format_end

xcb_dri2_attach_format_next

xcb_dri2_authenticate

xcb_dri2_authenticate_reply
the returned value must be freed by the caller using libc::free().

xcb_dri2_authenticate_unchecked

xcb_dri2_connect

xcb_dri2_connect_alignment_pad

xcb_dri2_connect_alignment_pad_end

xcb_dri2_connect_alignment_pad_length

xcb_dri2_connect_device_name

xcb_dri2_connect_device_name_end

xcb_dri2_connect_device_name_length

xcb_dri2_connect_driver_name

xcb_dri2_connect_driver_name_end

xcb_dri2_connect_driver_name_length

xcb_dri2_connect_reply
the returned value must be freed by the caller using libc::free().

xcb_dri2_connect_unchecked

xcb_dri2_copy_region

xcb_dri2_copy_region_reply
the returned value must be freed by the caller using libc::free().

xcb_dri2_copy_region_unchecked

xcb_dri2_create_drawable

xcb_dri2_create_drawable_checked

xcb_dri2_destroy_drawable

xcb_dri2_destroy_drawable_checked

xcb_dri2_dri2_buffer_end

xcb_dri2_dri2_buffer_next

xcb_dri2_get_buffers

xcb_dri2_get_buffers_buffers

xcb_dri2_get_buffers_buffers_iterator

xcb_dri2_get_buffers_buffers_length

xcb_dri2_get_buffers_reply
the returned value must be freed by the caller using libc::free().

xcb_dri2_get_buffers_unchecked

xcb_dri2_get_buffers_with_format

xcb_dri2_get_buffers_with_format_buffers

xcb_dri2_get_buffers_with_format_buffers_iterator

xcb_dri2_get_buffers_with_format_buffers_length

xcb_dri2_get_buffers_with_format_reply
the returned value must be freed by the caller using libc::free().

xcb_dri2_get_buffers_with_format_unchecked

xcb_dri2_get_msc

xcb_dri2_get_msc_reply
the returned value must be freed by the caller using libc::free().

xcb_dri2_get_msc_unchecked

xcb_dri2_get_param

xcb_dri2_get_param_reply
the returned value must be freed by the caller using libc::free().

xcb_dri2_get_param_unchecked

xcb_dri2_query_version

xcb_dri2_query_version_reply
the returned value must be freed by the caller using libc::free().

xcb_dri2_query_version_unchecked

xcb_dri2_swap_buffers

xcb_dri2_swap_buffers_reply
the returned value must be freed by the caller using libc::free().

xcb_dri2_swap_buffers_unchecked

xcb_dri2_swap_interval

xcb_dri2_swap_interval_checked

xcb_dri2_wait_msc

xcb_dri2_wait_msc_reply
the returned value must be freed by the caller using libc::free().

xcb_dri2_wait_msc_unchecked

xcb_dri2_wait_sbc

xcb_dri2_wait_sbc_reply
the returned value must be freed by the caller using libc::free().

xcb_dri2_wait_sbc_unchecked

Type Definitions

xcb_dri2_attachment_t

xcb_dri2_driver_type_t

xcb_dri2_event_type_t