Module xcb::ffi::glx

Structs

xcb_glx_are_textures_resident_cookie_t

xcb_glx_are_textures_resident_reply_t

xcb_glx_are_textures_resident_request_t

xcb_glx_bool32_iterator_t

xcb_glx_buffer_swap_complete_event_t

xcb_glx_change_drawable_attributes_request_t

xcb_glx_client_info_request_t

xcb_glx_context_iterator_t

xcb_glx_context_tag_iterator_t

xcb_glx_copy_context_request_t

xcb_glx_create_context_attribs_arb_request_t

xcb_glx_create_context_request_t

xcb_glx_create_glx_pixmap_request_t

xcb_glx_create_new_context_request_t

xcb_glx_create_pbuffer_request_t

xcb_glx_create_pixmap_request_t

xcb_glx_create_window_request_t

xcb_glx_delete_lists_request_t

xcb_glx_delete_queries_arb_request_t

xcb_glx_delete_textures_request_t

xcb_glx_delete_window_request_t

xcb_glx_destroy_context_request_t

xcb_glx_destroy_glx_pixmap_request_t

xcb_glx_destroy_pbuffer_request_t

xcb_glx_destroy_pixmap_request_t

xcb_glx_drawable_iterator_t

xcb_glx_end_list_request_t

xcb_glx_fbconfig_iterator_t

xcb_glx_feedback_buffer_request_t

xcb_glx_finish_cookie_t

xcb_glx_finish_reply_t

xcb_glx_finish_request_t

xcb_glx_float32_iterator_t

xcb_glx_float64_iterator_t

xcb_glx_flush_request_t

xcb_glx_gen_lists_cookie_t

xcb_glx_gen_lists_reply_t

xcb_glx_gen_lists_request_t

xcb_glx_gen_queries_arb_cookie_t

xcb_glx_gen_queries_arb_reply_t

xcb_glx_gen_queries_arb_request_t

xcb_glx_gen_textures_cookie_t

xcb_glx_gen_textures_reply_t

xcb_glx_gen_textures_request_t

xcb_glx_generic_error_t

xcb_glx_get_booleanv_cookie_t

xcb_glx_get_booleanv_reply_t

xcb_glx_get_booleanv_request_t

xcb_glx_get_clip_plane_cookie_t

xcb_glx_get_clip_plane_reply_t

xcb_glx_get_clip_plane_request_t

xcb_glx_get_color_table_cookie_t

xcb_glx_get_color_table_parameterfv_cookie_t

xcb_glx_get_color_table_parameterfv_reply_t

xcb_glx_get_color_table_parameterfv_request_t

xcb_glx_get_color_table_parameteriv_cookie_t

xcb_glx_get_color_table_parameteriv_reply_t

xcb_glx_get_color_table_parameteriv_request_t

xcb_glx_get_color_table_reply_t

xcb_glx_get_color_table_request_t

xcb_glx_get_compressed_tex_image_arb_cookie_t

xcb_glx_get_compressed_tex_image_arb_reply_t

xcb_glx_get_compressed_tex_image_arb_request_t

xcb_glx_get_convolution_filter_cookie_t

xcb_glx_get_convolution_filter_reply_t

xcb_glx_get_convolution_filter_request_t

xcb_glx_get_convolution_parameterfv_cookie_t

xcb_glx_get_convolution_parameterfv_reply_t

xcb_glx_get_convolution_parameterfv_request_t

xcb_glx_get_convolution_parameteriv_cookie_t

xcb_glx_get_convolution_parameteriv_reply_t

xcb_glx_get_convolution_parameteriv_request_t

xcb_glx_get_doublev_cookie_t

xcb_glx_get_doublev_reply_t

xcb_glx_get_doublev_request_t

xcb_glx_get_drawable_attributes_cookie_t

xcb_glx_get_drawable_attributes_reply_t

xcb_glx_get_drawable_attributes_request_t

xcb_glx_get_error_cookie_t

xcb_glx_get_error_reply_t

xcb_glx_get_error_request_t

xcb_glx_get_fb_configs_cookie_t

xcb_glx_get_fb_configs_reply_t

xcb_glx_get_fb_configs_request_t

xcb_glx_get_floatv_cookie_t

xcb_glx_get_floatv_reply_t

xcb_glx_get_floatv_request_t

xcb_glx_get_histogram_cookie_t

xcb_glx_get_histogram_parameterfv_cookie_t

xcb_glx_get_histogram_parameterfv_reply_t

xcb_glx_get_histogram_parameterfv_request_t

xcb_glx_get_histogram_parameteriv_cookie_t

xcb_glx_get_histogram_parameteriv_reply_t

xcb_glx_get_histogram_parameteriv_request_t

xcb_glx_get_histogram_reply_t

xcb_glx_get_histogram_request_t

xcb_glx_get_integerv_cookie_t

xcb_glx_get_integerv_reply_t

xcb_glx_get_integerv_request_t

xcb_glx_get_lightfv_cookie_t

xcb_glx_get_lightfv_reply_t

xcb_glx_get_lightfv_request_t

xcb_glx_get_lightiv_cookie_t

xcb_glx_get_lightiv_reply_t

xcb_glx_get_lightiv_request_t

xcb_glx_get_mapdv_cookie_t

xcb_glx_get_mapdv_reply_t

xcb_glx_get_mapdv_request_t

xcb_glx_get_mapfv_cookie_t

xcb_glx_get_mapfv_reply_t

xcb_glx_get_mapfv_request_t

xcb_glx_get_mapiv_cookie_t

xcb_glx_get_mapiv_reply_t

xcb_glx_get_mapiv_request_t

xcb_glx_get_materialfv_cookie_t

xcb_glx_get_materialfv_reply_t

xcb_glx_get_materialfv_request_t

xcb_glx_get_materialiv_cookie_t

xcb_glx_get_materialiv_reply_t

xcb_glx_get_materialiv_request_t

xcb_glx_get_minmax_cookie_t

xcb_glx_get_minmax_parameterfv_cookie_t

xcb_glx_get_minmax_parameterfv_reply_t

xcb_glx_get_minmax_parameterfv_request_t

xcb_glx_get_minmax_parameteriv_cookie_t

xcb_glx_get_minmax_parameteriv_reply_t

xcb_glx_get_minmax_parameteriv_request_t

xcb_glx_get_minmax_reply_t

xcb_glx_get_minmax_request_t

xcb_glx_get_pixel_mapfv_cookie_t

xcb_glx_get_pixel_mapfv_reply_t

xcb_glx_get_pixel_mapfv_request_t

xcb_glx_get_pixel_mapuiv_cookie_t

xcb_glx_get_pixel_mapuiv_reply_t

xcb_glx_get_pixel_mapuiv_request_t

xcb_glx_get_pixel_mapusv_cookie_t

xcb_glx_get_pixel_mapusv_reply_t

xcb_glx_get_pixel_mapusv_request_t

xcb_glx_get_polygon_stipple_cookie_t

xcb_glx_get_polygon_stipple_reply_t

xcb_glx_get_polygon_stipple_request_t

xcb_glx_get_query_objectiv_arb_cookie_t

xcb_glx_get_query_objectiv_arb_reply_t

xcb_glx_get_query_objectiv_arb_request_t

xcb_glx_get_query_objectuiv_arb_cookie_t

xcb_glx_get_query_objectuiv_arb_reply_t

xcb_glx_get_query_objectuiv_arb_request_t

xcb_glx_get_queryiv_arb_cookie_t

xcb_glx_get_queryiv_arb_reply_t

xcb_glx_get_queryiv_arb_request_t

xcb_glx_get_separable_filter_cookie_t

xcb_glx_get_separable_filter_reply_t

xcb_glx_get_separable_filter_request_t

xcb_glx_get_string_cookie_t

xcb_glx_get_string_reply_t

xcb_glx_get_string_request_t

xcb_glx_get_tex_envfv_cookie_t

xcb_glx_get_tex_envfv_reply_t

xcb_glx_get_tex_envfv_request_t

xcb_glx_get_tex_enviv_cookie_t

xcb_glx_get_tex_enviv_reply_t

xcb_glx_get_tex_enviv_request_t

xcb_glx_get_tex_gendv_cookie_t

xcb_glx_get_tex_gendv_reply_t

xcb_glx_get_tex_gendv_request_t

xcb_glx_get_tex_genfv_cookie_t

xcb_glx_get_tex_genfv_reply_t

xcb_glx_get_tex_genfv_request_t

xcb_glx_get_tex_geniv_cookie_t

xcb_glx_get_tex_geniv_reply_t

xcb_glx_get_tex_geniv_request_t

xcb_glx_get_tex_image_cookie_t

xcb_glx_get_tex_image_reply_t

xcb_glx_get_tex_image_request_t

xcb_glx_get_tex_level_parameterfv_cookie_t

xcb_glx_get_tex_level_parameterfv_reply_t

xcb_glx_get_tex_level_parameterfv_request_t

xcb_glx_get_tex_level_parameteriv_cookie_t

xcb_glx_get_tex_level_parameteriv_reply_t

xcb_glx_get_tex_level_parameteriv_request_t

xcb_glx_get_tex_parameterfv_cookie_t

xcb_glx_get_tex_parameterfv_reply_t

xcb_glx_get_tex_parameterfv_request_t

xcb_glx_get_tex_parameteriv_cookie_t

xcb_glx_get_tex_parameteriv_reply_t

xcb_glx_get_tex_parameteriv_request_t

xcb_glx_get_visual_configs_cookie_t

xcb_glx_get_visual_configs_reply_t

xcb_glx_get_visual_configs_request_t

xcb_glx_is_direct_cookie_t

xcb_glx_is_direct_reply_t

xcb_glx_is_direct_request_t

xcb_glx_is_list_cookie_t

xcb_glx_is_list_reply_t

xcb_glx_is_list_request_t

xcb_glx_is_query_arb_cookie_t

xcb_glx_is_query_arb_reply_t

xcb_glx_is_query_arb_request_t

xcb_glx_is_texture_cookie_t

xcb_glx_is_texture_reply_t

xcb_glx_is_texture_request_t

xcb_glx_make_context_current_cookie_t

xcb_glx_make_context_current_reply_t

xcb_glx_make_context_current_request_t

xcb_glx_make_current_cookie_t

xcb_glx_make_current_reply_t

xcb_glx_make_current_request_t

xcb_glx_new_list_request_t

xcb_glx_pbuffer_clobber_event_t

xcb_glx_pbuffer_iterator_t

xcb_glx_pixel_storef_request_t

xcb_glx_pixel_storei_request_t

xcb_glx_pixmap_iterator_t

xcb_glx_query_context_cookie_t

xcb_glx_query_context_reply_t

xcb_glx_query_context_request_t

xcb_glx_query_extensions_string_cookie_t

xcb_glx_query_extensions_string_reply_t

xcb_glx_query_extensions_string_request_t

xcb_glx_query_server_string_cookie_t

xcb_glx_query_server_string_reply_t

xcb_glx_query_server_string_request_t

xcb_glx_query_version_cookie_t

xcb_glx_query_version_reply_t

xcb_glx_query_version_request_t

xcb_glx_read_pixels_cookie_t

xcb_glx_read_pixels_reply_t

xcb_glx_read_pixels_request_t

xcb_glx_render_large_request_t

xcb_glx_render_mode_cookie_t

xcb_glx_render_mode_reply_t

xcb_glx_render_mode_request_t

xcb_glx_render_request_t

xcb_glx_select_buffer_request_t

xcb_glx_set_client_info_2arb_request_t

xcb_glx_set_client_info_arb_request_t

xcb_glx_swap_buffers_request_t

xcb_glx_use_x_font_request_t

xcb_glx_vendor_private_request_t

xcb_glx_vendor_private_with_reply_cookie_t

xcb_glx_vendor_private_with_reply_reply_t

xcb_glx_vendor_private_with_reply_request_t

xcb_glx_wait_gl_request_t

xcb_glx_wait_x_request_t

xcb_glx_window_iterator_t

Constants

XCB_GLX_ARE_TEXTURES_RESIDENT

XCB_GLX_BAD_CONTEXT

XCB_GLX_BAD_CONTEXT_STATE

XCB_GLX_BAD_CONTEXT_TAG

XCB_GLX_BAD_CURRENT_DRAWABLE

XCB_GLX_BAD_CURRENT_WINDOW

XCB_GLX_BAD_DRAWABLE

XCB_GLX_BAD_FB_CONFIG

XCB_GLX_BAD_LARGE_REQUEST

XCB_GLX_BAD_PBUFFER

XCB_GLX_BAD_PIXMAP

XCB_GLX_BAD_RENDER_REQUEST

XCB_GLX_BAD_WINDOW

XCB_GLX_BUFFER_SWAP_COMPLETE

XCB_GLX_CHANGE_DRAWABLE_ATTRIBUTES

XCB_GLX_CLIENT_INFO

XCB_GLX_COPY_CONTEXT

XCB_GLX_CREATE_CONTEXT

XCB_GLX_CREATE_CONTEXT_ATTRIBS_ARB

XCB_GLX_CREATE_GLX_PIXMAP

XCB_GLX_CREATE_NEW_CONTEXT

XCB_GLX_CREATE_PBUFFER

XCB_GLX_CREATE_PIXMAP

XCB_GLX_CREATE_WINDOW

XCB_GLX_DELETE_LISTS

XCB_GLX_DELETE_QUERIES_ARB

XCB_GLX_DELETE_TEXTURES

XCB_GLX_DELETE_WINDOW

XCB_GLX_DESTROY_CONTEXT

XCB_GLX_DESTROY_GLX_PIXMAP

XCB_GLX_DESTROY_PBUFFER

XCB_GLX_DESTROY_PIXMAP

XCB_GLX_END_LIST

XCB_GLX_FEEDBACK_BUFFER

XCB_GLX_FINISH

XCB_GLX_FLUSH

XCB_GLX_GC_GL_ACCUM_BUFFER_BIT

XCB_GLX_GC_GL_ALL_ATTRIB_BITS

XCB_GLX_GC_GL_COLOR_BUFFER_BIT

XCB_GLX_GC_GL_CURRENT_BIT

XCB_GLX_GC_GL_DEPTH_BUFFER_BIT

XCB_GLX_GC_GL_ENABLE_BIT

XCB_GLX_GC_GL_EVAL_BIT

XCB_GLX_GC_GL_FOG_BIT

XCB_GLX_GC_GL_HINT_BIT

XCB_GLX_GC_GL_LIGHTING_BIT

XCB_GLX_GC_GL_LINE_BIT

XCB_GLX_GC_GL_LIST_BIT

XCB_GLX_GC_GL_PIXEL_MODE_BIT

XCB_GLX_GC_GL_POINT_BIT

XCB_GLX_GC_GL_POLYGON_BIT

XCB_GLX_GC_GL_POLYGON_STIPPLE_BIT

XCB_GLX_GC_GL_SCISSOR_BIT

XCB_GLX_GC_GL_STENCIL_BUFFER_BIT

XCB_GLX_GC_GL_TEXTURE_BIT

XCB_GLX_GC_GL_TRANSFORM_BIT

XCB_GLX_GC_GL_VIEWPORT_BIT

XCB_GLX_GENERIC

XCB_GLX_GEN_LISTS

XCB_GLX_GEN_QUERIES_ARB

XCB_GLX_GEN_TEXTURES

XCB_GLX_GET_BOOLEANV

XCB_GLX_GET_CLIP_PLANE

XCB_GLX_GET_COLOR_TABLE

XCB_GLX_GET_COLOR_TABLE_PARAMETERFV

XCB_GLX_GET_COLOR_TABLE_PARAMETERIV

XCB_GLX_GET_COMPRESSED_TEX_IMAGE_ARB

XCB_GLX_GET_CONVOLUTION_FILTER

XCB_GLX_GET_CONVOLUTION_PARAMETERFV

XCB_GLX_GET_CONVOLUTION_PARAMETERIV

XCB_GLX_GET_DOUBLEV

XCB_GLX_GET_DRAWABLE_ATTRIBUTES

XCB_GLX_GET_ERROR

XCB_GLX_GET_FB_CONFIGS

XCB_GLX_GET_FLOATV

XCB_GLX_GET_HISTOGRAM

XCB_GLX_GET_HISTOGRAM_PARAMETERFV

XCB_GLX_GET_HISTOGRAM_PARAMETERIV

XCB_GLX_GET_INTEGERV

XCB_GLX_GET_LIGHTFV

XCB_GLX_GET_LIGHTIV

XCB_GLX_GET_MAPDV

XCB_GLX_GET_MAPFV

XCB_GLX_GET_MAPIV

XCB_GLX_GET_MATERIALFV

XCB_GLX_GET_MATERIALIV

XCB_GLX_GET_MINMAX

XCB_GLX_GET_MINMAX_PARAMETERFV

XCB_GLX_GET_MINMAX_PARAMETERIV

XCB_GLX_GET_PIXEL_MAPFV

XCB_GLX_GET_PIXEL_MAPUIV

XCB_GLX_GET_PIXEL_MAPUSV

XCB_GLX_GET_POLYGON_STIPPLE

XCB_GLX_GET_QUERYIV_ARB

XCB_GLX_GET_QUERY_OBJECTIV_ARB

XCB_GLX_GET_QUERY_OBJECTUIV_ARB

XCB_GLX_GET_SEPARABLE_FILTER

XCB_GLX_GET_STRING

XCB_GLX_GET_TEX_ENVFV

XCB_GLX_GET_TEX_ENVIV

XCB_GLX_GET_TEX_GENDV

XCB_GLX_GET_TEX_GENFV

XCB_GLX_GET_TEX_GENIV

XCB_GLX_GET_TEX_IMAGE

XCB_GLX_GET_TEX_LEVEL_PARAMETERFV

XCB_GLX_GET_TEX_LEVEL_PARAMETERIV

XCB_GLX_GET_TEX_PARAMETERFV

XCB_GLX_GET_TEX_PARAMETERIV

XCB_GLX_GET_VISUAL_CONFIGS

XCB_GLX_GLX_BAD_PROFILE_ARB

XCB_GLX_IS_DIRECT

XCB_GLX_IS_LIST

XCB_GLX_IS_QUERY_ARB

XCB_GLX_IS_TEXTURE

XCB_GLX_MAJOR_VERSION

XCB_GLX_MAKE_CONTEXT_CURRENT

XCB_GLX_MAKE_CURRENT

XCB_GLX_MINOR_VERSION

XCB_GLX_NEW_LIST

XCB_GLX_PBCDT_PBUFFER

XCB_GLX_PBCDT_WINDOW

XCB_GLX_PBCET_DAMAGED

XCB_GLX_PBCET_SAVED

XCB_GLX_PBUFFER_CLOBBER

XCB_GLX_PIXEL_STOREF

XCB_GLX_PIXEL_STOREI

XCB_GLX_QUERY_CONTEXT

XCB_GLX_QUERY_EXTENSIONS_STRING

XCB_GLX_QUERY_SERVER_STRING

XCB_GLX_QUERY_VERSION

XCB_GLX_READ_PIXELS

XCB_GLX_RENDER

XCB_GLX_RENDER_LARGE

XCB_GLX_RENDER_MODE

XCB_GLX_RM_GL_FEEDBACK

XCB_GLX_RM_GL_RENDER

XCB_GLX_RM_GL_SELECT

XCB_GLX_SELECT_BUFFER

XCB_GLX_SET_CLIENT_INFO_2ARB

XCB_GLX_SET_CLIENT_INFO_ARB

XCB_GLX_SWAP_BUFFERS

XCB_GLX_UNSUPPORTED_PRIVATE_REQUEST

XCB_GLX_USE_X_FONT

XCB_GLX_VENDOR_PRIVATE

XCB_GLX_VENDOR_PRIVATE_WITH_REPLY

XCB_GLX_WAIT_GL

XCB_GLX_WAIT_X

Statics

xcb_glx_id

Functions

xcb_glx_are_textures_resident

xcb_glx_are_textures_resident_data

xcb_glx_are_textures_resident_data_end

xcb_glx_are_textures_resident_data_length

xcb_glx_are_textures_resident_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_are_textures_resident_unchecked

xcb_glx_bool32_end

xcb_glx_bool32_next

xcb_glx_change_drawable_attributes

xcb_glx_change_drawable_attributes_checked

xcb_glx_client_info

xcb_glx_client_info_checked

xcb_glx_context_end

xcb_glx_context_next

xcb_glx_context_tag_end

xcb_glx_context_tag_next

xcb_glx_copy_context

xcb_glx_copy_context_checked

xcb_glx_create_context

xcb_glx_create_context_attribs_arb

xcb_glx_create_context_attribs_arb_checked

xcb_glx_create_context_checked

xcb_glx_create_glx_pixmap

xcb_glx_create_glx_pixmap_checked

xcb_glx_create_new_context

xcb_glx_create_new_context_checked

xcb_glx_create_pbuffer

xcb_glx_create_pbuffer_checked

xcb_glx_create_pixmap

xcb_glx_create_pixmap_checked

xcb_glx_create_window

xcb_glx_create_window_checked

xcb_glx_delete_lists

xcb_glx_delete_lists_checked

xcb_glx_delete_queries_arb

xcb_glx_delete_queries_arb_checked

xcb_glx_delete_textures

xcb_glx_delete_textures_checked

xcb_glx_delete_window

xcb_glx_delete_window_checked

xcb_glx_destroy_context

xcb_glx_destroy_context_checked

xcb_glx_destroy_glx_pixmap

xcb_glx_destroy_glx_pixmap_checked

xcb_glx_destroy_pbuffer

xcb_glx_destroy_pbuffer_checked

xcb_glx_destroy_pixmap

xcb_glx_destroy_pixmap_checked

xcb_glx_drawable_end

xcb_glx_drawable_next

xcb_glx_end_list

xcb_glx_end_list_checked

xcb_glx_fbconfig_end

xcb_glx_fbconfig_next

xcb_glx_feedback_buffer

xcb_glx_feedback_buffer_checked

xcb_glx_finish

xcb_glx_finish_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_finish_unchecked

xcb_glx_float32_end

xcb_glx_float32_next

xcb_glx_float64_end

xcb_glx_float64_next

xcb_glx_flush

xcb_glx_flush_checked

xcb_glx_gen_lists

xcb_glx_gen_lists_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_gen_lists_unchecked

xcb_glx_gen_queries_arb

xcb_glx_gen_queries_arb_data

xcb_glx_gen_queries_arb_data_end

xcb_glx_gen_queries_arb_data_length

xcb_glx_gen_queries_arb_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_gen_queries_arb_unchecked

xcb_glx_gen_textures

xcb_glx_gen_textures_data

xcb_glx_gen_textures_data_end

xcb_glx_gen_textures_data_length

xcb_glx_gen_textures_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_gen_textures_unchecked

xcb_glx_get_booleanv

xcb_glx_get_booleanv_data

xcb_glx_get_booleanv_data_end

xcb_glx_get_booleanv_data_length

xcb_glx_get_booleanv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_booleanv_unchecked

xcb_glx_get_clip_plane

xcb_glx_get_clip_plane_data

xcb_glx_get_clip_plane_data_end

xcb_glx_get_clip_plane_data_length

xcb_glx_get_clip_plane_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_clip_plane_unchecked

xcb_glx_get_color_table

xcb_glx_get_color_table_data

xcb_glx_get_color_table_data_end

xcb_glx_get_color_table_data_length

xcb_glx_get_color_table_parameterfv

xcb_glx_get_color_table_parameterfv_data

xcb_glx_get_color_table_parameterfv_data_end

xcb_glx_get_color_table_parameterfv_data_length

xcb_glx_get_color_table_parameterfv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_color_table_parameterfv_unchecked

xcb_glx_get_color_table_parameteriv

xcb_glx_get_color_table_parameteriv_data

xcb_glx_get_color_table_parameteriv_data_end

xcb_glx_get_color_table_parameteriv_data_length

xcb_glx_get_color_table_parameteriv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_color_table_parameteriv_unchecked

xcb_glx_get_color_table_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_color_table_unchecked

xcb_glx_get_compressed_tex_image_arb

xcb_glx_get_compressed_tex_image_arb_data

xcb_glx_get_compressed_tex_image_arb_data_end

xcb_glx_get_compressed_tex_image_arb_data_length

xcb_glx_get_compressed_tex_image_arb_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_compressed_tex_image_arb_unchecked

xcb_glx_get_convolution_filter

xcb_glx_get_convolution_filter_data

xcb_glx_get_convolution_filter_data_end

xcb_glx_get_convolution_filter_data_length

xcb_glx_get_convolution_filter_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_convolution_filter_unchecked

xcb_glx_get_convolution_parameterfv

xcb_glx_get_convolution_parameterfv_data

xcb_glx_get_convolution_parameterfv_data_end

xcb_glx_get_convolution_parameterfv_data_length

xcb_glx_get_convolution_parameterfv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_convolution_parameterfv_unchecked

xcb_glx_get_convolution_parameteriv

xcb_glx_get_convolution_parameteriv_data

xcb_glx_get_convolution_parameteriv_data_end

xcb_glx_get_convolution_parameteriv_data_length

xcb_glx_get_convolution_parameteriv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_convolution_parameteriv_unchecked

xcb_glx_get_doublev

xcb_glx_get_doublev_data

xcb_glx_get_doublev_data_end

xcb_glx_get_doublev_data_length

xcb_glx_get_doublev_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_doublev_unchecked

xcb_glx_get_drawable_attributes

xcb_glx_get_drawable_attributes_attribs

xcb_glx_get_drawable_attributes_attribs_end

xcb_glx_get_drawable_attributes_attribs_length

xcb_glx_get_drawable_attributes_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_drawable_attributes_unchecked

xcb_glx_get_error

xcb_glx_get_error_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_error_unchecked

xcb_glx_get_fb_configs

xcb_glx_get_fb_configs_property_list

xcb_glx_get_fb_configs_property_list_end

xcb_glx_get_fb_configs_property_list_length

xcb_glx_get_fb_configs_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_fb_configs_unchecked

xcb_glx_get_floatv

xcb_glx_get_floatv_data

xcb_glx_get_floatv_data_end

xcb_glx_get_floatv_data_length

xcb_glx_get_floatv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_floatv_unchecked

xcb_glx_get_histogram

xcb_glx_get_histogram_data

xcb_glx_get_histogram_data_end

xcb_glx_get_histogram_data_length

xcb_glx_get_histogram_parameterfv

xcb_glx_get_histogram_parameterfv_data

xcb_glx_get_histogram_parameterfv_data_end

xcb_glx_get_histogram_parameterfv_data_length

xcb_glx_get_histogram_parameterfv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_histogram_parameterfv_unchecked

xcb_glx_get_histogram_parameteriv

xcb_glx_get_histogram_parameteriv_data

xcb_glx_get_histogram_parameteriv_data_end

xcb_glx_get_histogram_parameteriv_data_length

xcb_glx_get_histogram_parameteriv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_histogram_parameteriv_unchecked

xcb_glx_get_histogram_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_histogram_unchecked

xcb_glx_get_integerv

xcb_glx_get_integerv_data

xcb_glx_get_integerv_data_end

xcb_glx_get_integerv_data_length

xcb_glx_get_integerv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_integerv_unchecked

xcb_glx_get_lightfv

xcb_glx_get_lightfv_data

xcb_glx_get_lightfv_data_end

xcb_glx_get_lightfv_data_length

xcb_glx_get_lightfv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_lightfv_unchecked

xcb_glx_get_lightiv

xcb_glx_get_lightiv_data

xcb_glx_get_lightiv_data_end

xcb_glx_get_lightiv_data_length

xcb_glx_get_lightiv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_lightiv_unchecked

xcb_glx_get_mapdv

xcb_glx_get_mapdv_data

xcb_glx_get_mapdv_data_end

xcb_glx_get_mapdv_data_length

xcb_glx_get_mapdv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_mapdv_unchecked

xcb_glx_get_mapfv

xcb_glx_get_mapfv_data

xcb_glx_get_mapfv_data_end

xcb_glx_get_mapfv_data_length

xcb_glx_get_mapfv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_mapfv_unchecked

xcb_glx_get_mapiv

xcb_glx_get_mapiv_data

xcb_glx_get_mapiv_data_end

xcb_glx_get_mapiv_data_length

xcb_glx_get_mapiv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_mapiv_unchecked

xcb_glx_get_materialfv

xcb_glx_get_materialfv_data

xcb_glx_get_materialfv_data_end

xcb_glx_get_materialfv_data_length

xcb_glx_get_materialfv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_materialfv_unchecked

xcb_glx_get_materialiv

xcb_glx_get_materialiv_data

xcb_glx_get_materialiv_data_end

xcb_glx_get_materialiv_data_length

xcb_glx_get_materialiv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_materialiv_unchecked

xcb_glx_get_minmax

xcb_glx_get_minmax_data

xcb_glx_get_minmax_data_end

xcb_glx_get_minmax_data_length

xcb_glx_get_minmax_parameterfv

xcb_glx_get_minmax_parameterfv_data

xcb_glx_get_minmax_parameterfv_data_end

xcb_glx_get_minmax_parameterfv_data_length

xcb_glx_get_minmax_parameterfv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_minmax_parameterfv_unchecked

xcb_glx_get_minmax_parameteriv

xcb_glx_get_minmax_parameteriv_data

xcb_glx_get_minmax_parameteriv_data_end

xcb_glx_get_minmax_parameteriv_data_length

xcb_glx_get_minmax_parameteriv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_minmax_parameteriv_unchecked

xcb_glx_get_minmax_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_minmax_unchecked

xcb_glx_get_pixel_mapfv

xcb_glx_get_pixel_mapfv_data

xcb_glx_get_pixel_mapfv_data_end

xcb_glx_get_pixel_mapfv_data_length

xcb_glx_get_pixel_mapfv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_pixel_mapfv_unchecked

xcb_glx_get_pixel_mapuiv

xcb_glx_get_pixel_mapuiv_data

xcb_glx_get_pixel_mapuiv_data_end

xcb_glx_get_pixel_mapuiv_data_length

xcb_glx_get_pixel_mapuiv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_pixel_mapuiv_unchecked

xcb_glx_get_pixel_mapusv

xcb_glx_get_pixel_mapusv_data

xcb_glx_get_pixel_mapusv_data_end

xcb_glx_get_pixel_mapusv_data_length

xcb_glx_get_pixel_mapusv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_pixel_mapusv_unchecked

xcb_glx_get_polygon_stipple

xcb_glx_get_polygon_stipple_data

xcb_glx_get_polygon_stipple_data_end

xcb_glx_get_polygon_stipple_data_length

xcb_glx_get_polygon_stipple_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_polygon_stipple_unchecked

xcb_glx_get_query_objectiv_arb

xcb_glx_get_query_objectiv_arb_data

xcb_glx_get_query_objectiv_arb_data_end

xcb_glx_get_query_objectiv_arb_data_length

xcb_glx_get_query_objectiv_arb_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_query_objectiv_arb_unchecked

xcb_glx_get_query_objectuiv_arb

xcb_glx_get_query_objectuiv_arb_data

xcb_glx_get_query_objectuiv_arb_data_end

xcb_glx_get_query_objectuiv_arb_data_length

xcb_glx_get_query_objectuiv_arb_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_query_objectuiv_arb_unchecked

xcb_glx_get_queryiv_arb

xcb_glx_get_queryiv_arb_data

xcb_glx_get_queryiv_arb_data_end

xcb_glx_get_queryiv_arb_data_length

xcb_glx_get_queryiv_arb_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_queryiv_arb_unchecked

xcb_glx_get_separable_filter

xcb_glx_get_separable_filter_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_separable_filter_rows_and_cols

xcb_glx_get_separable_filter_rows_and_cols_end

xcb_glx_get_separable_filter_rows_and_cols_length

xcb_glx_get_separable_filter_unchecked

xcb_glx_get_string

xcb_glx_get_string_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_string_string

xcb_glx_get_string_string_end

xcb_glx_get_string_string_length

xcb_glx_get_string_unchecked

xcb_glx_get_tex_envfv

xcb_glx_get_tex_envfv_data

xcb_glx_get_tex_envfv_data_end

xcb_glx_get_tex_envfv_data_length

xcb_glx_get_tex_envfv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_tex_envfv_unchecked

xcb_glx_get_tex_enviv

xcb_glx_get_tex_enviv_data

xcb_glx_get_tex_enviv_data_end

xcb_glx_get_tex_enviv_data_length

xcb_glx_get_tex_enviv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_tex_enviv_unchecked

xcb_glx_get_tex_gendv

xcb_glx_get_tex_gendv_data

xcb_glx_get_tex_gendv_data_end

xcb_glx_get_tex_gendv_data_length

xcb_glx_get_tex_gendv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_tex_gendv_unchecked

xcb_glx_get_tex_genfv

xcb_glx_get_tex_genfv_data

xcb_glx_get_tex_genfv_data_end

xcb_glx_get_tex_genfv_data_length

xcb_glx_get_tex_genfv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_tex_genfv_unchecked

xcb_glx_get_tex_geniv

xcb_glx_get_tex_geniv_data

xcb_glx_get_tex_geniv_data_end

xcb_glx_get_tex_geniv_data_length

xcb_glx_get_tex_geniv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_tex_geniv_unchecked

xcb_glx_get_tex_image

xcb_glx_get_tex_image_data

xcb_glx_get_tex_image_data_end

xcb_glx_get_tex_image_data_length

xcb_glx_get_tex_image_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_tex_image_unchecked

xcb_glx_get_tex_level_parameterfv

xcb_glx_get_tex_level_parameterfv_data

xcb_glx_get_tex_level_parameterfv_data_end

xcb_glx_get_tex_level_parameterfv_data_length

xcb_glx_get_tex_level_parameterfv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_tex_level_parameterfv_unchecked

xcb_glx_get_tex_level_parameteriv

xcb_glx_get_tex_level_parameteriv_data

xcb_glx_get_tex_level_parameteriv_data_end

xcb_glx_get_tex_level_parameteriv_data_length

xcb_glx_get_tex_level_parameteriv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_tex_level_parameteriv_unchecked

xcb_glx_get_tex_parameterfv

xcb_glx_get_tex_parameterfv_data

xcb_glx_get_tex_parameterfv_data_end

xcb_glx_get_tex_parameterfv_data_length

xcb_glx_get_tex_parameterfv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_tex_parameterfv_unchecked

xcb_glx_get_tex_parameteriv

xcb_glx_get_tex_parameteriv_data

xcb_glx_get_tex_parameteriv_data_end

xcb_glx_get_tex_parameteriv_data_length

xcb_glx_get_tex_parameteriv_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_tex_parameteriv_unchecked

xcb_glx_get_visual_configs

xcb_glx_get_visual_configs_property_list

xcb_glx_get_visual_configs_property_list_end

xcb_glx_get_visual_configs_property_list_length

xcb_glx_get_visual_configs_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_get_visual_configs_unchecked

xcb_glx_is_direct

xcb_glx_is_direct_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_is_direct_unchecked

xcb_glx_is_list

xcb_glx_is_list_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_is_list_unchecked

xcb_glx_is_query_arb

xcb_glx_is_query_arb_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_is_query_arb_unchecked

xcb_glx_is_texture

xcb_glx_is_texture_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_is_texture_unchecked

xcb_glx_make_context_current

xcb_glx_make_context_current_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_make_context_current_unchecked

xcb_glx_make_current

xcb_glx_make_current_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_make_current_unchecked

xcb_glx_new_list

xcb_glx_new_list_checked

xcb_glx_pbuffer_end

xcb_glx_pbuffer_next

xcb_glx_pixel_storef

xcb_glx_pixel_storef_checked

xcb_glx_pixel_storei

xcb_glx_pixel_storei_checked

xcb_glx_pixmap_end

xcb_glx_pixmap_next

xcb_glx_query_context

xcb_glx_query_context_attribs

xcb_glx_query_context_attribs_end

xcb_glx_query_context_attribs_length

xcb_glx_query_context_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_query_context_unchecked

xcb_glx_query_extensions_string

xcb_glx_query_extensions_string_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_query_extensions_string_unchecked

xcb_glx_query_server_string

xcb_glx_query_server_string_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_query_server_string_string

xcb_glx_query_server_string_string_end

xcb_glx_query_server_string_string_length

xcb_glx_query_server_string_unchecked

xcb_glx_query_version

xcb_glx_query_version_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_query_version_unchecked

xcb_glx_read_pixels

xcb_glx_read_pixels_data

xcb_glx_read_pixels_data_end

xcb_glx_read_pixels_data_length

xcb_glx_read_pixels_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_read_pixels_unchecked

xcb_glx_render

xcb_glx_render_checked

xcb_glx_render_large

xcb_glx_render_large_checked

xcb_glx_render_mode

xcb_glx_render_mode_data

xcb_glx_render_mode_data_end

xcb_glx_render_mode_data_length

xcb_glx_render_mode_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_render_mode_unchecked

xcb_glx_select_buffer

xcb_glx_select_buffer_checked

xcb_glx_set_client_info_2arb

xcb_glx_set_client_info_2arb_checked

xcb_glx_set_client_info_arb

xcb_glx_set_client_info_arb_checked

xcb_glx_swap_buffers

xcb_glx_swap_buffers_checked

xcb_glx_use_x_font

xcb_glx_use_x_font_checked

xcb_glx_vendor_private

xcb_glx_vendor_private_checked

xcb_glx_vendor_private_with_reply

xcb_glx_vendor_private_with_reply_data_2

xcb_glx_vendor_private_with_reply_data_2_end

xcb_glx_vendor_private_with_reply_data_2_length

xcb_glx_vendor_private_with_reply_reply
the returned value must be freed by the caller using libc::free().

xcb_glx_vendor_private_with_reply_unchecked

xcb_glx_wait_gl

xcb_glx_wait_gl_checked

xcb_glx_wait_x

xcb_glx_wait_x_checked

xcb_glx_window_end

xcb_glx_window_next

Type Definitions

xcb_glx_bad_context_error_t

xcb_glx_bad_context_state_error_t

xcb_glx_bad_context_tag_error_t

xcb_glx_bad_current_drawable_error_t

xcb_glx_bad_current_window_error_t

xcb_glx_bad_drawable_error_t

xcb_glx_bad_fb_config_error_t

xcb_glx_bad_large_request_error_t

xcb_glx_bad_pbuffer_error_t

xcb_glx_bad_pixmap_error_t

xcb_glx_bad_render_request_error_t

xcb_glx_bad_window_error_t

xcb_glx_bool32_t

xcb_glx_context_t

xcb_glx_context_tag_t

xcb_glx_drawable_t

xcb_glx_fbconfig_t

xcb_glx_float32_t

xcb_glx_float64_t

xcb_glx_gc_t

xcb_glx_glx_bad_profile_arb_error_t

xcb_glx_pbcdt_t

xcb_glx_pbcet_t

xcb_glx_pbuffer_t

xcb_glx_pixmap_t

xcb_glx_rm_t

xcb_glx_unsupported_private_request_error_t

xcb_glx_window_t