Module xcb::ffi::present

Structs

xcb_present_complete_notify_event_t

xcb_present_configure_notify_event_t

xcb_present_event_iterator_t

xcb_present_generic_event_t

xcb_present_idle_notify_event_t

xcb_present_notify_iterator_t

xcb_present_notify_msc_request_t

xcb_present_notify_t

xcb_present_pixmap_request_t

xcb_present_query_capabilities_cookie_t

xcb_present_query_capabilities_reply_t

xcb_present_query_capabilities_request_t

xcb_present_query_version_cookie_t

xcb_present_query_version_reply_t

xcb_present_query_version_request_t

xcb_present_redirect_notify_event_t

xcb_present_select_input_request_t

Constants

XCB_PRESENT_CAPABILITY_ASYNC

XCB_PRESENT_CAPABILITY_FENCE

XCB_PRESENT_CAPABILITY_NONE

XCB_PRESENT_CAPABILITY_UST

XCB_PRESENT_COMPLETE_KIND_NOTIFY_MSC

XCB_PRESENT_COMPLETE_KIND_PIXMAP

XCB_PRESENT_COMPLETE_MODE_COPY

XCB_PRESENT_COMPLETE_MODE_FLIP

XCB_PRESENT_COMPLETE_MODE_SKIP

XCB_PRESENT_COMPLETE_NOTIFY

XCB_PRESENT_CONFIGURE_NOTIFY

XCB_PRESENT_EVENT_COMPLETE_NOTIFY

XCB_PRESENT_EVENT_CONFIGURE_NOTIFY

XCB_PRESENT_EVENT_IDLE_NOTIFY

XCB_PRESENT_EVENT_MASK_COMPLETE_NOTIFY

XCB_PRESENT_EVENT_MASK_CONFIGURE_NOTIFY

XCB_PRESENT_EVENT_MASK_IDLE_NOTIFY

XCB_PRESENT_EVENT_MASK_NO_EVENT

XCB_PRESENT_EVENT_MASK_REDIRECT_NOTIFY

XCB_PRESENT_EVENT_REDIRECT_NOTIFY

XCB_PRESENT_GENERIC

XCB_PRESENT_IDLE_NOTIFY

XCB_PRESENT_MAJOR_VERSION

XCB_PRESENT_MINOR_VERSION

XCB_PRESENT_NOTIFY_MSC

XCB_PRESENT_OPTION_ASYNC

XCB_PRESENT_OPTION_COPY

XCB_PRESENT_OPTION_NONE

XCB_PRESENT_OPTION_UST

XCB_PRESENT_PIXMAP

XCB_PRESENT_QUERY_CAPABILITIES

XCB_PRESENT_QUERY_VERSION

XCB_PRESENT_REDIRECT_NOTIFY

XCB_PRESENT_SELECT_INPUT

Statics

xcb_present_id

Functions

xcb_present_event_end⚠

xcb_present_event_next⚠

xcb_present_notify_end⚠

xcb_present_notify_msc⚠

xcb_present_notify_msc_checked⚠

xcb_present_notify_next⚠

xcb_present_pixmap⚠

xcb_present_pixmap_checked⚠

xcb_present_query_capabilities⚠

xcb_present_query_capabilities_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_present_query_capabilities_unchecked⚠

xcb_present_query_version⚠

xcb_present_query_version_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_present_query_version_unchecked⚠

xcb_present_redirect_notify_notifies⚠

xcb_present_redirect_notify_notifies_iterator⚠

xcb_present_redirect_notify_notifies_length⚠

xcb_present_select_input⚠

xcb_present_select_input_checked⚠

Type Definitions

xcb_present_capability_t

xcb_present_complete_kind_t

xcb_present_complete_mode_t

xcb_present_event_enum_t

xcb_present_event_mask_t

xcb_present_event_t

xcb_present_option_t