Module xcb::ffi::randr

Structs

xcb_randr_add_output_mode_request_t

xcb_randr_bad_crtc_error_t

xcb_randr_bad_mode_error_t

xcb_randr_bad_output_error_t

xcb_randr_bad_provider_error_t

xcb_randr_change_output_property_request_t

xcb_randr_change_provider_property_request_t

xcb_randr_configure_output_property_request_t

xcb_randr_configure_provider_property_request_t

xcb_randr_create_mode_cookie_t

xcb_randr_create_mode_reply_t

xcb_randr_create_mode_request_t

xcb_randr_crtc_change_iterator_t

xcb_randr_crtc_change_t

xcb_randr_crtc_iterator_t

xcb_randr_delete_output_mode_request_t

xcb_randr_delete_output_property_request_t

xcb_randr_delete_provider_property_request_t

xcb_randr_destroy_mode_request_t

xcb_randr_get_crtc_gamma_cookie_t

xcb_randr_get_crtc_gamma_reply_t

xcb_randr_get_crtc_gamma_request_t

xcb_randr_get_crtc_gamma_size_cookie_t

xcb_randr_get_crtc_gamma_size_reply_t

xcb_randr_get_crtc_gamma_size_request_t

xcb_randr_get_crtc_info_cookie_t

xcb_randr_get_crtc_info_reply_t

xcb_randr_get_crtc_info_request_t

xcb_randr_get_crtc_transform_cookie_t

xcb_randr_get_crtc_transform_reply_t

xcb_randr_get_crtc_transform_request_t

xcb_randr_get_output_info_cookie_t

xcb_randr_get_output_info_reply_t

xcb_randr_get_output_info_request_t

xcb_randr_get_output_primary_cookie_t

xcb_randr_get_output_primary_reply_t

xcb_randr_get_output_primary_request_t

xcb_randr_get_output_property_cookie_t

xcb_randr_get_output_property_reply_t

xcb_randr_get_output_property_request_t

xcb_randr_get_panning_cookie_t

xcb_randr_get_panning_reply_t

xcb_randr_get_panning_request_t

xcb_randr_get_provider_info_cookie_t

xcb_randr_get_provider_info_reply_t

xcb_randr_get_provider_info_request_t

xcb_randr_get_provider_property_cookie_t

xcb_randr_get_provider_property_reply_t

xcb_randr_get_provider_property_request_t

xcb_randr_get_providers_cookie_t

xcb_randr_get_providers_reply_t

xcb_randr_get_providers_request_t

xcb_randr_get_screen_info_cookie_t

xcb_randr_get_screen_info_reply_t

xcb_randr_get_screen_info_request_t

xcb_randr_get_screen_resources_cookie_t

xcb_randr_get_screen_resources_current_cookie_t

xcb_randr_get_screen_resources_current_reply_t

xcb_randr_get_screen_resources_current_request_t

xcb_randr_get_screen_resources_reply_t

xcb_randr_get_screen_resources_request_t

xcb_randr_get_screen_size_range_cookie_t

xcb_randr_get_screen_size_range_reply_t

xcb_randr_get_screen_size_range_request_t

xcb_randr_list_output_properties_cookie_t

xcb_randr_list_output_properties_reply_t

xcb_randr_list_output_properties_request_t

xcb_randr_list_provider_properties_cookie_t

xcb_randr_list_provider_properties_reply_t

xcb_randr_list_provider_properties_request_t

xcb_randr_mode_info_iterator_t

xcb_randr_mode_info_t

xcb_randr_mode_iterator_t

xcb_randr_notify_data_iterator_t

xcb_randr_notify_data_t

xcb_randr_notify_event_t

xcb_randr_output_change_iterator_t

xcb_randr_output_change_t

xcb_randr_output_iterator_t

xcb_randr_output_property_iterator_t

xcb_randr_output_property_t

xcb_randr_provider_change_iterator_t

xcb_randr_provider_change_t

xcb_randr_provider_iterator_t

xcb_randr_provider_property_iterator_t

xcb_randr_provider_property_t

xcb_randr_query_output_property_cookie_t

xcb_randr_query_output_property_reply_t

xcb_randr_query_output_property_request_t

xcb_randr_query_provider_property_cookie_t

xcb_randr_query_provider_property_reply_t

xcb_randr_query_provider_property_request_t

xcb_randr_query_version_cookie_t

xcb_randr_query_version_reply_t

xcb_randr_query_version_request_t

xcb_randr_refresh_rates_iterator_t

xcb_randr_refresh_rates_t

xcb_randr_resource_change_iterator_t

xcb_randr_resource_change_t

xcb_randr_screen_change_notify_event_t

xcb_randr_screen_size_iterator_t

xcb_randr_screen_size_t

xcb_randr_select_input_request_t

xcb_randr_set_crtc_config_cookie_t

xcb_randr_set_crtc_config_reply_t

xcb_randr_set_crtc_config_request_t

xcb_randr_set_crtc_gamma_request_t

xcb_randr_set_crtc_transform_request_t

xcb_randr_set_output_primary_request_t

xcb_randr_set_panning_cookie_t

xcb_randr_set_panning_reply_t

xcb_randr_set_panning_request_t

xcb_randr_set_provider_offload_sink_request_t

xcb_randr_set_provider_output_source_request_t

xcb_randr_set_screen_config_cookie_t

xcb_randr_set_screen_config_reply_t

xcb_randr_set_screen_config_request_t

xcb_randr_set_screen_size_request_t

Constants

XCB_RANDR_ADD_OUTPUT_MODE

XCB_RANDR_BAD_CRTC

XCB_RANDR_BAD_MODE

XCB_RANDR_BAD_OUTPUT

XCB_RANDR_BAD_PROVIDER

XCB_RANDR_CHANGE_OUTPUT_PROPERTY

XCB_RANDR_CHANGE_PROVIDER_PROPERTY

XCB_RANDR_CONFIGURE_OUTPUT_PROPERTY

XCB_RANDR_CONFIGURE_PROVIDER_PROPERTY

XCB_RANDR_CONNECTION_CONNECTED

XCB_RANDR_CONNECTION_DISCONNECTED

XCB_RANDR_CONNECTION_UNKNOWN

XCB_RANDR_CREATE_MODE

XCB_RANDR_DELETE_OUTPUT_MODE

XCB_RANDR_DELETE_OUTPUT_PROPERTY

XCB_RANDR_DELETE_PROVIDER_PROPERTY

XCB_RANDR_DESTROY_MODE

XCB_RANDR_GET_CRTC_GAMMA

XCB_RANDR_GET_CRTC_GAMMA_SIZE

XCB_RANDR_GET_CRTC_INFO

XCB_RANDR_GET_CRTC_TRANSFORM

XCB_RANDR_GET_OUTPUT_INFO

XCB_RANDR_GET_OUTPUT_PRIMARY

XCB_RANDR_GET_OUTPUT_PROPERTY

XCB_RANDR_GET_PANNING

XCB_RANDR_GET_PROVIDERS

XCB_RANDR_GET_PROVIDER_INFO

XCB_RANDR_GET_PROVIDER_PROPERTY

XCB_RANDR_GET_SCREEN_INFO

XCB_RANDR_GET_SCREEN_RESOURCES

XCB_RANDR_GET_SCREEN_RESOURCES_CURRENT

XCB_RANDR_GET_SCREEN_SIZE_RANGE

XCB_RANDR_LIST_OUTPUT_PROPERTIES

XCB_RANDR_LIST_PROVIDER_PROPERTIES

XCB_RANDR_MAJOR_VERSION

XCB_RANDR_MINOR_VERSION

XCB_RANDR_MODE_FLAG_BCAST

XCB_RANDR_MODE_FLAG_CSYNC

XCB_RANDR_MODE_FLAG_CSYNC_NEGATIVE

XCB_RANDR_MODE_FLAG_CSYNC_POSITIVE

XCB_RANDR_MODE_FLAG_DOUBLE_CLOCK

XCB_RANDR_MODE_FLAG_DOUBLE_SCAN

XCB_RANDR_MODE_FLAG_HALVE_CLOCK

XCB_RANDR_MODE_FLAG_HSKEW_PRESENT

XCB_RANDR_MODE_FLAG_HSYNC_NEGATIVE

XCB_RANDR_MODE_FLAG_HSYNC_POSITIVE

XCB_RANDR_MODE_FLAG_INTERLACE

XCB_RANDR_MODE_FLAG_PIXEL_MULTIPLEX

XCB_RANDR_MODE_FLAG_VSYNC_NEGATIVE

XCB_RANDR_MODE_FLAG_VSYNC_POSITIVE

XCB_RANDR_NOTIFY

XCB_RANDR_NOTIFY_CRTC_CHANGE

XCB_RANDR_NOTIFY_MASK_CRTC_CHANGE

XCB_RANDR_NOTIFY_MASK_OUTPUT_CHANGE

XCB_RANDR_NOTIFY_MASK_OUTPUT_PROPERTY

XCB_RANDR_NOTIFY_MASK_PROVIDER_CHANGE

XCB_RANDR_NOTIFY_MASK_PROVIDER_PROPERTY

XCB_RANDR_NOTIFY_MASK_RESOURCE_CHANGE

XCB_RANDR_NOTIFY_MASK_SCREEN_CHANGE

XCB_RANDR_NOTIFY_OUTPUT_CHANGE

XCB_RANDR_NOTIFY_OUTPUT_PROPERTY

XCB_RANDR_NOTIFY_PROVIDER_CHANGE

XCB_RANDR_NOTIFY_PROVIDER_PROPERTY

XCB_RANDR_NOTIFY_RESOURCE_CHANGE

XCB_RANDR_PROVIDER_CAPABILITY_SINK_OFFLOAD

XCB_RANDR_PROVIDER_CAPABILITY_SINK_OUTPUT

XCB_RANDR_PROVIDER_CAPABILITY_SOURCE_OFFLOAD

XCB_RANDR_PROVIDER_CAPABILITY_SOURCE_OUTPUT

XCB_RANDR_QUERY_OUTPUT_PROPERTY

XCB_RANDR_QUERY_PROVIDER_PROPERTY

XCB_RANDR_QUERY_VERSION

XCB_RANDR_ROTATION_REFLECT_X

XCB_RANDR_ROTATION_REFLECT_Y

XCB_RANDR_ROTATION_ROTATE_0

XCB_RANDR_ROTATION_ROTATE_90

XCB_RANDR_ROTATION_ROTATE_180

XCB_RANDR_ROTATION_ROTATE_270

XCB_RANDR_SCREEN_CHANGE_NOTIFY

XCB_RANDR_SELECT_INPUT

XCB_RANDR_SET_CONFIG_FAILED

XCB_RANDR_SET_CONFIG_INVALID_CONFIG_TIME

XCB_RANDR_SET_CONFIG_INVALID_TIME

XCB_RANDR_SET_CONFIG_SUCCESS

XCB_RANDR_SET_CRTC_CONFIG

XCB_RANDR_SET_CRTC_GAMMA

XCB_RANDR_SET_CRTC_TRANSFORM

XCB_RANDR_SET_OUTPUT_PRIMARY

XCB_RANDR_SET_PANNING

XCB_RANDR_SET_PROVIDER_OFFLOAD_SINK

XCB_RANDR_SET_PROVIDER_OUTPUT_SOURCE

XCB_RANDR_SET_SCREEN_CONFIG

XCB_RANDR_SET_SCREEN_SIZE

XCB_RANDR_TRANSFORM_PROJECTIVE

XCB_RANDR_TRANSFORM_SCALE_DOWN

XCB_RANDR_TRANSFORM_SCALE_UP

XCB_RANDR_TRANSFORM_UNIT

Statics

xcb_randr_id

Functions

xcb_randr_add_output_mode⚠

xcb_randr_add_output_mode_checked⚠

xcb_randr_change_output_property⚠

xcb_randr_change_output_property_checked⚠

xcb_randr_change_provider_property⚠

xcb_randr_change_provider_property_checked⚠

xcb_randr_configure_output_property⚠

xcb_randr_configure_output_property_checked⚠

xcb_randr_configure_provider_property⚠

xcb_randr_configure_provider_property_checked⚠

xcb_randr_create_mode⚠

xcb_randr_create_mode_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_create_mode_unchecked⚠

xcb_randr_crtc_change_end⚠

xcb_randr_crtc_change_next⚠

xcb_randr_crtc_end⚠

xcb_randr_crtc_next⚠

xcb_randr_delete_output_mode⚠

xcb_randr_delete_output_mode_checked⚠

xcb_randr_delete_output_property⚠

xcb_randr_delete_output_property_checked⚠

xcb_randr_delete_provider_property⚠

xcb_randr_delete_provider_property_checked⚠

xcb_randr_destroy_mode⚠

xcb_randr_destroy_mode_checked⚠

xcb_randr_get_crtc_gamma⚠

xcb_randr_get_crtc_gamma_blue⚠

xcb_randr_get_crtc_gamma_blue_end⚠

xcb_randr_get_crtc_gamma_blue_length⚠

xcb_randr_get_crtc_gamma_green⚠

xcb_randr_get_crtc_gamma_green_end⚠

xcb_randr_get_crtc_gamma_green_length⚠

xcb_randr_get_crtc_gamma_red⚠

xcb_randr_get_crtc_gamma_red_end⚠

xcb_randr_get_crtc_gamma_red_length⚠

xcb_randr_get_crtc_gamma_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_crtc_gamma_size⚠

xcb_randr_get_crtc_gamma_size_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_crtc_gamma_size_unchecked⚠

xcb_randr_get_crtc_gamma_unchecked⚠

xcb_randr_get_crtc_info⚠

xcb_randr_get_crtc_info_outputs⚠

xcb_randr_get_crtc_info_outputs_end⚠

xcb_randr_get_crtc_info_outputs_length⚠

xcb_randr_get_crtc_info_possible⚠

xcb_randr_get_crtc_info_possible_end⚠

xcb_randr_get_crtc_info_possible_length⚠

xcb_randr_get_crtc_info_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_crtc_info_unchecked⚠

xcb_randr_get_crtc_transform⚠

xcb_randr_get_crtc_transform_current_filter_name⚠

xcb_randr_get_crtc_transform_current_filter_name_end⚠

xcb_randr_get_crtc_transform_current_filter_name_length⚠

xcb_randr_get_crtc_transform_current_params⚠

xcb_randr_get_crtc_transform_current_params_end⚠

xcb_randr_get_crtc_transform_current_params_length⚠

xcb_randr_get_crtc_transform_pending_filter_name⚠

xcb_randr_get_crtc_transform_pending_filter_name_end⚠

xcb_randr_get_crtc_transform_pending_filter_name_length⚠

xcb_randr_get_crtc_transform_pending_params⚠

xcb_randr_get_crtc_transform_pending_params_end⚠

xcb_randr_get_crtc_transform_pending_params_length⚠

xcb_randr_get_crtc_transform_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_crtc_transform_unchecked⚠

xcb_randr_get_output_info⚠

xcb_randr_get_output_info_clones⚠

xcb_randr_get_output_info_clones_end⚠

xcb_randr_get_output_info_clones_length⚠

xcb_randr_get_output_info_crtcs⚠

xcb_randr_get_output_info_crtcs_end⚠

xcb_randr_get_output_info_crtcs_length⚠

xcb_randr_get_output_info_modes⚠

xcb_randr_get_output_info_modes_end⚠

xcb_randr_get_output_info_modes_length⚠

xcb_randr_get_output_info_name⚠

xcb_randr_get_output_info_name_end⚠

xcb_randr_get_output_info_name_length⚠

xcb_randr_get_output_info_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_output_info_unchecked⚠

xcb_randr_get_output_primary⚠

xcb_randr_get_output_primary_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_output_primary_unchecked⚠

xcb_randr_get_output_property⚠

xcb_randr_get_output_property_data⚠

xcb_randr_get_output_property_data_end⚠

xcb_randr_get_output_property_data_length⚠

xcb_randr_get_output_property_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_output_property_unchecked⚠

xcb_randr_get_panning⚠

xcb_randr_get_panning_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_panning_unchecked⚠

xcb_randr_get_provider_info⚠

xcb_randr_get_provider_info_associated_capability⚠

xcb_randr_get_provider_info_associated_capability_end⚠

xcb_randr_get_provider_info_associated_capability_length⚠

xcb_randr_get_provider_info_associated_providers⚠

xcb_randr_get_provider_info_associated_providers_end⚠

xcb_randr_get_provider_info_associated_providers_length⚠

xcb_randr_get_provider_info_crtcs⚠

xcb_randr_get_provider_info_crtcs_end⚠

xcb_randr_get_provider_info_crtcs_length⚠

xcb_randr_get_provider_info_name⚠

xcb_randr_get_provider_info_name_end⚠

xcb_randr_get_provider_info_name_length⚠

xcb_randr_get_provider_info_outputs⚠

xcb_randr_get_provider_info_outputs_end⚠

xcb_randr_get_provider_info_outputs_length⚠

xcb_randr_get_provider_info_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_provider_info_unchecked⚠

xcb_randr_get_provider_property⚠

xcb_randr_get_provider_property_data⚠

xcb_randr_get_provider_property_data_end⚠

xcb_randr_get_provider_property_data_length⚠

xcb_randr_get_provider_property_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_provider_property_unchecked⚠

xcb_randr_get_providers⚠

xcb_randr_get_providers_providers⚠

xcb_randr_get_providers_providers_end⚠

xcb_randr_get_providers_providers_length⚠

xcb_randr_get_providers_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_providers_unchecked⚠

xcb_randr_get_screen_info⚠

xcb_randr_get_screen_info_rates_iterator⚠

xcb_randr_get_screen_info_rates_length⚠

xcb_randr_get_screen_info_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_screen_info_sizes⚠

xcb_randr_get_screen_info_sizes_iterator⚠

xcb_randr_get_screen_info_sizes_length⚠

xcb_randr_get_screen_info_unchecked⚠

xcb_randr_get_screen_resources⚠

xcb_randr_get_screen_resources_crtcs⚠

xcb_randr_get_screen_resources_crtcs_end⚠

xcb_randr_get_screen_resources_crtcs_length⚠

xcb_randr_get_screen_resources_current⚠

xcb_randr_get_screen_resources_current_crtcs⚠

xcb_randr_get_screen_resources_current_crtcs_end⚠

xcb_randr_get_screen_resources_current_crtcs_length⚠

xcb_randr_get_screen_resources_current_modes⚠

xcb_randr_get_screen_resources_current_modes_iterator⚠

xcb_randr_get_screen_resources_current_modes_length⚠

xcb_randr_get_screen_resources_current_names⚠

xcb_randr_get_screen_resources_current_names_end⚠

xcb_randr_get_screen_resources_current_names_length⚠

xcb_randr_get_screen_resources_current_outputs⚠

xcb_randr_get_screen_resources_current_outputs_end⚠

xcb_randr_get_screen_resources_current_outputs_length⚠

xcb_randr_get_screen_resources_current_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_screen_resources_current_unchecked⚠

xcb_randr_get_screen_resources_modes⚠

xcb_randr_get_screen_resources_modes_iterator⚠

xcb_randr_get_screen_resources_modes_length⚠

xcb_randr_get_screen_resources_names⚠

xcb_randr_get_screen_resources_names_end⚠

xcb_randr_get_screen_resources_names_length⚠

xcb_randr_get_screen_resources_outputs⚠

xcb_randr_get_screen_resources_outputs_end⚠

xcb_randr_get_screen_resources_outputs_length⚠

xcb_randr_get_screen_resources_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_screen_resources_unchecked⚠

xcb_randr_get_screen_size_range⚠

xcb_randr_get_screen_size_range_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_get_screen_size_range_unchecked⚠

xcb_randr_list_output_properties⚠

xcb_randr_list_output_properties_atoms⚠

xcb_randr_list_output_properties_atoms_end⚠

xcb_randr_list_output_properties_atoms_length⚠

xcb_randr_list_output_properties_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_list_output_properties_unchecked⚠

xcb_randr_list_provider_properties⚠

xcb_randr_list_provider_properties_atoms⚠

xcb_randr_list_provider_properties_atoms_end⚠

xcb_randr_list_provider_properties_atoms_length⚠

xcb_randr_list_provider_properties_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_list_provider_properties_unchecked⚠

xcb_randr_mode_end⚠

xcb_randr_mode_info_end⚠

xcb_randr_mode_info_next⚠

xcb_randr_mode_next⚠

xcb_randr_notify_data_end⚠

xcb_randr_notify_data_next⚠

xcb_randr_output_change_end⚠

xcb_randr_output_change_next⚠

xcb_randr_output_end⚠

xcb_randr_output_next⚠

xcb_randr_output_property_end⚠

xcb_randr_output_property_next⚠

xcb_randr_provider_change_end⚠

xcb_randr_provider_change_next⚠

xcb_randr_provider_end⚠

xcb_randr_provider_next⚠

xcb_randr_provider_property_end⚠

xcb_randr_provider_property_next⚠

xcb_randr_query_output_property⚠

xcb_randr_query_output_property_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_query_output_property_unchecked⚠

xcb_randr_query_output_property_valid_values⚠

xcb_randr_query_output_property_valid_values_end⚠

xcb_randr_query_output_property_valid_values_length⚠

xcb_randr_query_provider_property⚠

xcb_randr_query_provider_property_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_query_provider_property_unchecked⚠

xcb_randr_query_provider_property_valid_values⚠

xcb_randr_query_provider_property_valid_values_end⚠

xcb_randr_query_provider_property_valid_values_length⚠

xcb_randr_query_version⚠

xcb_randr_query_version_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_query_version_unchecked⚠

xcb_randr_refresh_rates_end⚠

xcb_randr_refresh_rates_next⚠

xcb_randr_refresh_rates_rates⚠

xcb_randr_refresh_rates_rates_end⚠

xcb_randr_refresh_rates_rates_length⚠

xcb_randr_resource_change_end⚠

xcb_randr_resource_change_next⚠

xcb_randr_screen_size_end⚠

xcb_randr_screen_size_next⚠

xcb_randr_select_input⚠

xcb_randr_select_input_checked⚠

xcb_randr_set_crtc_config⚠

xcb_randr_set_crtc_config_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_set_crtc_config_unchecked⚠

xcb_randr_set_crtc_gamma⚠

xcb_randr_set_crtc_gamma_checked⚠

xcb_randr_set_crtc_transform⚠

xcb_randr_set_crtc_transform_checked⚠

xcb_randr_set_output_primary⚠

xcb_randr_set_output_primary_checked⚠

xcb_randr_set_panning⚠

xcb_randr_set_panning_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_set_panning_unchecked⚠

xcb_randr_set_provider_offload_sink⚠

xcb_randr_set_provider_offload_sink_checked⚠

xcb_randr_set_provider_output_source⚠

xcb_randr_set_provider_output_source_checked⚠

xcb_randr_set_screen_config⚠

xcb_randr_set_screen_config_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_randr_set_screen_config_unchecked⚠

xcb_randr_set_screen_size⚠

xcb_randr_set_screen_size_checked⚠

Type Definitions

xcb_randr_connection_t

xcb_randr_crtc_t

xcb_randr_mode_flag_t

xcb_randr_mode_t

xcb_randr_notify_mask_t

xcb_randr_notify_t

xcb_randr_output_t

xcb_randr_provider_capability_t

xcb_randr_provider_t

xcb_randr_rotation_t

xcb_randr_set_config_t

xcb_randr_transform_t