Module xcb::ffi::render

Structs

xcb_render_add_glyphs_request_t

xcb_render_add_traps_request_t

xcb_render_animcursorelt_iterator_t

xcb_render_animcursorelt_t

xcb_render_change_picture_request_t

xcb_render_color_iterator_t

xcb_render_color_t

xcb_render_composite_glyphs_8_request_t

xcb_render_composite_glyphs_16_request_t

xcb_render_composite_glyphs_32_request_t

xcb_render_composite_request_t

xcb_render_create_anim_cursor_request_t

xcb_render_create_conical_gradient_request_t

xcb_render_create_cursor_request_t

xcb_render_create_glyph_set_request_t

xcb_render_create_linear_gradient_request_t

xcb_render_create_picture_request_t

xcb_render_create_radial_gradient_request_t

xcb_render_create_solid_fill_request_t

xcb_render_directformat_iterator_t

xcb_render_directformat_t

xcb_render_fill_rectangles_request_t

xcb_render_fixed_iterator_t

xcb_render_free_glyph_set_request_t

xcb_render_free_glyphs_request_t

xcb_render_free_picture_request_t

xcb_render_glyph_error_t

xcb_render_glyph_iterator_t

xcb_render_glyph_set_error_t

xcb_render_glyphinfo_iterator_t

xcb_render_glyphinfo_t

xcb_render_glyphset_iterator_t

xcb_render_indexvalue_iterator_t

xcb_render_indexvalue_t

xcb_render_linefix_iterator_t

xcb_render_linefix_t

xcb_render_pict_format_error_t

xcb_render_pict_op_error_t

xcb_render_pictdepth_iterator_t

xcb_render_pictdepth_t

xcb_render_pictformat_iterator_t

xcb_render_pictforminfo_iterator_t

xcb_render_pictforminfo_t

xcb_render_pictscreen_iterator_t

xcb_render_pictscreen_t

xcb_render_picture_error_t

xcb_render_picture_iterator_t

xcb_render_pictvisual_iterator_t

xcb_render_pictvisual_t

xcb_render_pointfix_iterator_t

xcb_render_pointfix_t

xcb_render_query_filters_cookie_t

xcb_render_query_filters_reply_t

xcb_render_query_filters_request_t

xcb_render_query_pict_formats_cookie_t

xcb_render_query_pict_formats_reply_t

xcb_render_query_pict_formats_request_t

xcb_render_query_pict_index_values_cookie_t

xcb_render_query_pict_index_values_reply_t

xcb_render_query_pict_index_values_request_t

xcb_render_query_version_cookie_t

xcb_render_query_version_reply_t

xcb_render_query_version_request_t

xcb_render_reference_glyph_set_request_t

xcb_render_set_picture_clip_rectangles_request_t

xcb_render_set_picture_filter_request_t

xcb_render_set_picture_transform_request_t

xcb_render_spanfix_iterator_t

xcb_render_spanfix_t

xcb_render_transform_iterator_t

xcb_render_transform_t

xcb_render_trap_iterator_t

xcb_render_trap_t

xcb_render_trapezoid_iterator_t

xcb_render_trapezoid_t

xcb_render_trapezoids_request_t

xcb_render_tri_fan_request_t

xcb_render_tri_strip_request_t

xcb_render_triangle_iterator_t

xcb_render_triangle_t

xcb_render_triangles_request_t

Constants

XCB_RENDER_ADD_GLYPHS

XCB_RENDER_ADD_TRAPS

XCB_RENDER_CHANGE_PICTURE

XCB_RENDER_COMPOSITE

XCB_RENDER_COMPOSITE_GLYPHS_8

XCB_RENDER_COMPOSITE_GLYPHS_16

XCB_RENDER_COMPOSITE_GLYPHS_32

XCB_RENDER_CP_ALPHA_MAP

XCB_RENDER_CP_ALPHA_X_ORIGIN

XCB_RENDER_CP_ALPHA_Y_ORIGIN

XCB_RENDER_CP_CLIP_MASK

XCB_RENDER_CP_CLIP_X_ORIGIN

XCB_RENDER_CP_CLIP_Y_ORIGIN

XCB_RENDER_CP_COMPONENT_ALPHA

XCB_RENDER_CP_DITHER

XCB_RENDER_CP_GRAPHICS_EXPOSURE

XCB_RENDER_CP_POLY_EDGE

XCB_RENDER_CP_POLY_MODE

XCB_RENDER_CP_REPEAT

XCB_RENDER_CP_SUBWINDOW_MODE

XCB_RENDER_CREATE_ANIM_CURSOR

XCB_RENDER_CREATE_CONICAL_GRADIENT

XCB_RENDER_CREATE_CURSOR

XCB_RENDER_CREATE_GLYPH_SET

XCB_RENDER_CREATE_LINEAR_GRADIENT

XCB_RENDER_CREATE_PICTURE

XCB_RENDER_CREATE_RADIAL_GRADIENT

XCB_RENDER_CREATE_SOLID_FILL

XCB_RENDER_FILL_RECTANGLES

XCB_RENDER_FREE_GLYPHS

XCB_RENDER_FREE_GLYPH_SET

XCB_RENDER_FREE_PICTURE

XCB_RENDER_GLYPH

XCB_RENDER_GLYPH_SET

XCB_RENDER_MAJOR_VERSION

XCB_RENDER_MINOR_VERSION

XCB_RENDER_PICTURE

XCB_RENDER_PICTURE_NONE

XCB_RENDER_PICT_FORMAT

XCB_RENDER_PICT_OP

XCB_RENDER_PICT_OP_ADD

XCB_RENDER_PICT_OP_ATOP

XCB_RENDER_PICT_OP_ATOP_REVERSE

XCB_RENDER_PICT_OP_CLEAR

XCB_RENDER_PICT_OP_COLOR_BURN

XCB_RENDER_PICT_OP_COLOR_DODGE

XCB_RENDER_PICT_OP_CONJOINT_ATOP

XCB_RENDER_PICT_OP_CONJOINT_ATOP_REVERSE

XCB_RENDER_PICT_OP_CONJOINT_CLEAR

XCB_RENDER_PICT_OP_CONJOINT_DST

XCB_RENDER_PICT_OP_CONJOINT_IN

XCB_RENDER_PICT_OP_CONJOINT_IN_REVERSE

XCB_RENDER_PICT_OP_CONJOINT_OUT

XCB_RENDER_PICT_OP_CONJOINT_OUT_REVERSE

XCB_RENDER_PICT_OP_CONJOINT_OVER

XCB_RENDER_PICT_OP_CONJOINT_OVER_REVERSE

XCB_RENDER_PICT_OP_CONJOINT_SRC

XCB_RENDER_PICT_OP_CONJOINT_XOR

XCB_RENDER_PICT_OP_DARKEN

XCB_RENDER_PICT_OP_DIFFERENCE

XCB_RENDER_PICT_OP_DISJOINT_ATOP

XCB_RENDER_PICT_OP_DISJOINT_ATOP_REVERSE

XCB_RENDER_PICT_OP_DISJOINT_CLEAR

XCB_RENDER_PICT_OP_DISJOINT_DST

XCB_RENDER_PICT_OP_DISJOINT_IN

XCB_RENDER_PICT_OP_DISJOINT_IN_REVERSE

XCB_RENDER_PICT_OP_DISJOINT_OUT

XCB_RENDER_PICT_OP_DISJOINT_OUT_REVERSE

XCB_RENDER_PICT_OP_DISJOINT_OVER

XCB_RENDER_PICT_OP_DISJOINT_OVER_REVERSE

XCB_RENDER_PICT_OP_DISJOINT_SRC

XCB_RENDER_PICT_OP_DISJOINT_XOR

XCB_RENDER_PICT_OP_DST

XCB_RENDER_PICT_OP_EXCLUSION

XCB_RENDER_PICT_OP_HARD_LIGHT

XCB_RENDER_PICT_OP_HSL_COLOR

XCB_RENDER_PICT_OP_HSL_HUE

XCB_RENDER_PICT_OP_HSL_LUMINOSITY

XCB_RENDER_PICT_OP_HSL_SATURATION

XCB_RENDER_PICT_OP_IN

XCB_RENDER_PICT_OP_IN_REVERSE

XCB_RENDER_PICT_OP_LIGHTEN

XCB_RENDER_PICT_OP_MULTIPLY

XCB_RENDER_PICT_OP_OUT

XCB_RENDER_PICT_OP_OUT_REVERSE

XCB_RENDER_PICT_OP_OVER

XCB_RENDER_PICT_OP_OVERLAY

XCB_RENDER_PICT_OP_OVER_REVERSE

XCB_RENDER_PICT_OP_SATURATE

XCB_RENDER_PICT_OP_SCREEN

XCB_RENDER_PICT_OP_SOFT_LIGHT

XCB_RENDER_PICT_OP_SRC

XCB_RENDER_PICT_OP_XOR

XCB_RENDER_PICT_TYPE_DIRECT

XCB_RENDER_PICT_TYPE_INDEXED

XCB_RENDER_POLY_EDGE_SHARP

XCB_RENDER_POLY_EDGE_SMOOTH

XCB_RENDER_POLY_MODE_IMPRECISE

XCB_RENDER_POLY_MODE_PRECISE

XCB_RENDER_QUERY_FILTERS

XCB_RENDER_QUERY_PICT_FORMATS

XCB_RENDER_QUERY_PICT_INDEX_VALUES

XCB_RENDER_QUERY_VERSION

XCB_RENDER_REFERENCE_GLYPH_SET

XCB_RENDER_REPEAT_NONE

XCB_RENDER_REPEAT_NORMAL

XCB_RENDER_REPEAT_PAD

XCB_RENDER_REPEAT_REFLECT

XCB_RENDER_SET_PICTURE_CLIP_RECTANGLES

XCB_RENDER_SET_PICTURE_FILTER

XCB_RENDER_SET_PICTURE_TRANSFORM

XCB_RENDER_SUB_PIXEL_HORIZONTAL_BGR

XCB_RENDER_SUB_PIXEL_HORIZONTAL_RGB

XCB_RENDER_SUB_PIXEL_NONE

XCB_RENDER_SUB_PIXEL_UNKNOWN

XCB_RENDER_SUB_PIXEL_VERTICAL_BGR

XCB_RENDER_SUB_PIXEL_VERTICAL_RGB

XCB_RENDER_TRAPEZOIDS

XCB_RENDER_TRIANGLES

XCB_RENDER_TRI_FAN

XCB_RENDER_TRI_STRIP

Statics

xcb_render_id

Functions

xcb_render_add_glyphs⚠

xcb_render_add_glyphs_checked⚠

xcb_render_add_traps⚠

xcb_render_add_traps_checked⚠

xcb_render_animcursorelt_end⚠

xcb_render_animcursorelt_next⚠

xcb_render_change_picture⚠

xcb_render_change_picture_checked⚠

xcb_render_color_end⚠

xcb_render_color_next⚠

xcb_render_composite⚠

xcb_render_composite_checked⚠

xcb_render_composite_glyphs_8⚠

xcb_render_composite_glyphs_8_checked⚠

xcb_render_composite_glyphs_16⚠

xcb_render_composite_glyphs_16_checked⚠

xcb_render_composite_glyphs_32⚠

xcb_render_composite_glyphs_32_checked⚠

xcb_render_create_anim_cursor⚠

xcb_render_create_anim_cursor_checked⚠

xcb_render_create_conical_gradient⚠

xcb_render_create_conical_gradient_checked⚠

xcb_render_create_cursor⚠

xcb_render_create_cursor_checked⚠

xcb_render_create_glyph_set⚠

xcb_render_create_glyph_set_checked⚠

xcb_render_create_linear_gradient⚠

xcb_render_create_linear_gradient_checked⚠

xcb_render_create_picture⚠

xcb_render_create_picture_checked⚠

xcb_render_create_radial_gradient⚠

xcb_render_create_radial_gradient_checked⚠

xcb_render_create_solid_fill⚠

xcb_render_create_solid_fill_checked⚠

xcb_render_directformat_end⚠

xcb_render_directformat_next⚠

xcb_render_fill_rectangles⚠

xcb_render_fill_rectangles_checked⚠

xcb_render_fixed_end⚠

xcb_render_fixed_next⚠

xcb_render_free_glyph_set⚠

xcb_render_free_glyph_set_checked⚠

xcb_render_free_glyphs⚠

xcb_render_free_glyphs_checked⚠

xcb_render_free_picture⚠

xcb_render_free_picture_checked⚠

xcb_render_glyph_end⚠

xcb_render_glyph_next⚠

xcb_render_glyphinfo_end⚠

xcb_render_glyphinfo_next⚠

xcb_render_glyphset_end⚠

xcb_render_glyphset_next⚠

xcb_render_indexvalue_end⚠

xcb_render_indexvalue_next⚠

xcb_render_linefix_end⚠

xcb_render_linefix_next⚠

xcb_render_pictdepth_end⚠

xcb_render_pictdepth_next⚠

xcb_render_pictdepth_visuals⚠

xcb_render_pictdepth_visuals_iterator⚠

xcb_render_pictdepth_visuals_length⚠

xcb_render_pictformat_end⚠

xcb_render_pictformat_next⚠

xcb_render_pictforminfo_end⚠

xcb_render_pictforminfo_next⚠

xcb_render_pictscreen_depths_iterator⚠

xcb_render_pictscreen_depths_length⚠

xcb_render_pictscreen_end⚠

xcb_render_pictscreen_next⚠

xcb_render_picture_end⚠

xcb_render_picture_next⚠

xcb_render_pictvisual_end⚠

xcb_render_pictvisual_next⚠

xcb_render_pointfix_end⚠

xcb_render_pointfix_next⚠

xcb_render_query_filters⚠

xcb_render_query_filters_aliases⚠

xcb_render_query_filters_aliases_end⚠

xcb_render_query_filters_aliases_length⚠

xcb_render_query_filters_filters_iterator⚠

xcb_render_query_filters_filters_length⚠

xcb_render_query_filters_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_render_query_filters_unchecked⚠

xcb_render_query_pict_formats⚠

xcb_render_query_pict_formats_formats⚠

xcb_render_query_pict_formats_formats_iterator⚠

xcb_render_query_pict_formats_formats_length⚠

xcb_render_query_pict_formats_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_render_query_pict_formats_screens_iterator⚠

xcb_render_query_pict_formats_screens_length⚠

xcb_render_query_pict_formats_subpixels⚠

xcb_render_query_pict_formats_subpixels_end⚠

xcb_render_query_pict_formats_subpixels_length⚠

xcb_render_query_pict_formats_unchecked⚠

xcb_render_query_pict_index_values⚠

xcb_render_query_pict_index_values_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_render_query_pict_index_values_unchecked⚠

xcb_render_query_pict_index_values_values⚠

xcb_render_query_pict_index_values_values_iterator⚠

xcb_render_query_pict_index_values_values_length⚠

xcb_render_query_version⚠

xcb_render_query_version_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_render_query_version_unchecked⚠

xcb_render_reference_glyph_set⚠

xcb_render_reference_glyph_set_checked⚠

xcb_render_set_picture_clip_rectangles⚠

xcb_render_set_picture_clip_rectangles_checked⚠

xcb_render_set_picture_filter⚠

xcb_render_set_picture_filter_checked⚠

xcb_render_set_picture_transform⚠

xcb_render_set_picture_transform_checked⚠

xcb_render_spanfix_end⚠

xcb_render_spanfix_next⚠

xcb_render_transform_end⚠

xcb_render_transform_next⚠

xcb_render_trap_end⚠

xcb_render_trap_next⚠

xcb_render_trapezoid_end⚠

xcb_render_trapezoid_next⚠

xcb_render_trapezoids⚠

xcb_render_trapezoids_checked⚠

xcb_render_tri_fan⚠

xcb_render_tri_fan_checked⚠

xcb_render_tri_strip⚠

xcb_render_tri_strip_checked⚠

xcb_render_triangle_end⚠

xcb_render_triangle_next⚠

xcb_render_triangles⚠

xcb_render_triangles_checked⚠

Type Definitions

xcb_render_cp_t

xcb_render_fixed_t

xcb_render_glyph_t

xcb_render_glyphset_t

xcb_render_pict_op_t

xcb_render_pict_type_t

xcb_render_pictformat_t

xcb_render_picture_enum_t

xcb_render_picture_t

xcb_render_poly_edge_t

xcb_render_poly_mode_t

xcb_render_repeat_t

xcb_render_sub_pixel_t