Module xcb::ffi::res

Structs

xcb_res_client_id_spec_iterator_t

xcb_res_client_id_spec_t

xcb_res_client_id_value_iterator_t

xcb_res_client_id_value_t

xcb_res_client_iterator_t

xcb_res_client_t

xcb_res_query_client_ids_cookie_t

xcb_res_query_client_ids_reply_t

xcb_res_query_client_ids_request_t

xcb_res_query_client_pixmap_bytes_cookie_t

xcb_res_query_client_pixmap_bytes_reply_t

xcb_res_query_client_pixmap_bytes_request_t

xcb_res_query_client_resources_cookie_t

xcb_res_query_client_resources_reply_t

xcb_res_query_client_resources_request_t

xcb_res_query_clients_cookie_t

xcb_res_query_clients_reply_t

xcb_res_query_clients_request_t

xcb_res_query_resource_bytes_cookie_t

xcb_res_query_resource_bytes_reply_t

xcb_res_query_resource_bytes_request_t

xcb_res_query_version_cookie_t

xcb_res_query_version_reply_t

xcb_res_query_version_request_t

xcb_res_resource_id_spec_iterator_t

xcb_res_resource_id_spec_t

xcb_res_resource_size_spec_iterator_t

xcb_res_resource_size_spec_t

xcb_res_resource_size_value_iterator_t

xcb_res_resource_size_value_t

xcb_res_type_iterator_t

xcb_res_type_t

Constants

XCB_RES_CLIENT_ID_MASK_CLIENT_XID

XCB_RES_CLIENT_ID_MASK_LOCAL_CLIENT_PID

XCB_RES_MAJOR_VERSION

XCB_RES_MINOR_VERSION

XCB_RES_QUERY_CLIENTS

XCB_RES_QUERY_CLIENT_IDS

XCB_RES_QUERY_CLIENT_PIXMAP_BYTES

XCB_RES_QUERY_CLIENT_RESOURCES

XCB_RES_QUERY_RESOURCE_BYTES

XCB_RES_QUERY_VERSION

Statics

xcb_res_id

Functions

xcb_res_client_end

xcb_res_client_id_spec_end

xcb_res_client_id_spec_next

xcb_res_client_id_value_end

xcb_res_client_id_value_next

xcb_res_client_id_value_value

xcb_res_client_id_value_value_end

xcb_res_client_id_value_value_length

xcb_res_client_next

xcb_res_query_client_ids

xcb_res_query_client_ids_ids_iterator

xcb_res_query_client_ids_ids_length

xcb_res_query_client_ids_reply
the returned value must be freed by the caller using libc::free().

xcb_res_query_client_ids_unchecked

xcb_res_query_client_pixmap_bytes

xcb_res_query_client_pixmap_bytes_reply
the returned value must be freed by the caller using libc::free().

xcb_res_query_client_pixmap_bytes_unchecked

xcb_res_query_client_resources

xcb_res_query_client_resources_reply
the returned value must be freed by the caller using libc::free().

xcb_res_query_client_resources_types

xcb_res_query_client_resources_types_iterator

xcb_res_query_client_resources_types_length

xcb_res_query_client_resources_unchecked

xcb_res_query_clients

xcb_res_query_clients_clients

xcb_res_query_clients_clients_iterator

xcb_res_query_clients_clients_length

xcb_res_query_clients_reply
the returned value must be freed by the caller using libc::free().

xcb_res_query_clients_unchecked

xcb_res_query_resource_bytes

xcb_res_query_resource_bytes_reply
the returned value must be freed by the caller using libc::free().

xcb_res_query_resource_bytes_sizes_iterator

xcb_res_query_resource_bytes_sizes_length

xcb_res_query_resource_bytes_unchecked

xcb_res_query_version

xcb_res_query_version_reply
the returned value must be freed by the caller using libc::free().

xcb_res_query_version_unchecked

xcb_res_resource_id_spec_end

xcb_res_resource_id_spec_next

xcb_res_resource_size_spec_end

xcb_res_resource_size_spec_next

xcb_res_resource_size_value_cross_references

xcb_res_resource_size_value_cross_references_iterator

xcb_res_resource_size_value_cross_references_length

xcb_res_resource_size_value_end

xcb_res_resource_size_value_next

xcb_res_type_end

xcb_res_type_next

Type Definitions

xcb_res_client_id_mask_t