Module xcb::ffi::screensaver

Structs

xcb_screensaver_notify_event_t

xcb_screensaver_query_info_cookie_t

xcb_screensaver_query_info_reply_t

xcb_screensaver_query_info_request_t

xcb_screensaver_query_version_cookie_t

xcb_screensaver_query_version_reply_t

xcb_screensaver_query_version_request_t

xcb_screensaver_select_input_request_t

xcb_screensaver_set_attributes_request_t

xcb_screensaver_suspend_request_t

xcb_screensaver_unset_attributes_request_t

Constants

XCB_SCREENSAVER_EVENT_CYCLE_MASK

XCB_SCREENSAVER_EVENT_NOTIFY_MASK

XCB_SCREENSAVER_KIND_BLANKED

XCB_SCREENSAVER_KIND_EXTERNAL

XCB_SCREENSAVER_KIND_INTERNAL

XCB_SCREENSAVER_MAJOR_VERSION

XCB_SCREENSAVER_MINOR_VERSION

XCB_SCREENSAVER_NOTIFY

XCB_SCREENSAVER_QUERY_INFO

XCB_SCREENSAVER_QUERY_VERSION

XCB_SCREENSAVER_SELECT_INPUT

XCB_SCREENSAVER_SET_ATTRIBUTES

XCB_SCREENSAVER_STATE_CYCLE

XCB_SCREENSAVER_STATE_DISABLED

XCB_SCREENSAVER_STATE_OFF

XCB_SCREENSAVER_STATE_ON

XCB_SCREENSAVER_SUSPEND

XCB_SCREENSAVER_UNSET_ATTRIBUTES

Statics

xcb_screensaver_id

Functions

xcb_screensaver_query_info

xcb_screensaver_query_info_reply
the returned value must be freed by the caller using libc::free().

xcb_screensaver_query_info_unchecked

xcb_screensaver_query_version

xcb_screensaver_query_version_reply
the returned value must be freed by the caller using libc::free().

xcb_screensaver_query_version_unchecked

xcb_screensaver_select_input

xcb_screensaver_select_input_checked

xcb_screensaver_set_attributes

xcb_screensaver_set_attributes_checked

xcb_screensaver_suspend

xcb_screensaver_suspend_checked

xcb_screensaver_unset_attributes

xcb_screensaver_unset_attributes_checked

Type Definitions

xcb_screensaver_event_t

xcb_screensaver_kind_t

xcb_screensaver_state_t