Module xcb::ffi::selinux

Structs

xcb_selinux_get_client_context_cookie_t

xcb_selinux_get_client_context_reply_t

xcb_selinux_get_client_context_request_t

xcb_selinux_get_device_context_cookie_t

xcb_selinux_get_device_context_reply_t

xcb_selinux_get_device_context_request_t

xcb_selinux_get_device_create_context_cookie_t

xcb_selinux_get_device_create_context_reply_t

xcb_selinux_get_device_create_context_request_t

xcb_selinux_get_property_context_cookie_t

xcb_selinux_get_property_context_reply_t

xcb_selinux_get_property_context_request_t

xcb_selinux_get_property_create_context_cookie_t

xcb_selinux_get_property_create_context_reply_t

xcb_selinux_get_property_create_context_request_t

xcb_selinux_get_property_data_context_cookie_t

xcb_selinux_get_property_data_context_reply_t

xcb_selinux_get_property_data_context_request_t

xcb_selinux_get_property_use_context_cookie_t

xcb_selinux_get_property_use_context_reply_t

xcb_selinux_get_property_use_context_request_t

xcb_selinux_get_selection_context_cookie_t

xcb_selinux_get_selection_context_reply_t

xcb_selinux_get_selection_context_request_t

xcb_selinux_get_selection_create_context_cookie_t

xcb_selinux_get_selection_create_context_reply_t

xcb_selinux_get_selection_create_context_request_t

xcb_selinux_get_selection_data_context_cookie_t

xcb_selinux_get_selection_data_context_reply_t

xcb_selinux_get_selection_data_context_request_t

xcb_selinux_get_selection_use_context_cookie_t

xcb_selinux_get_selection_use_context_reply_t

xcb_selinux_get_selection_use_context_request_t

xcb_selinux_get_window_context_cookie_t

xcb_selinux_get_window_context_reply_t

xcb_selinux_get_window_context_request_t

xcb_selinux_get_window_create_context_cookie_t

xcb_selinux_get_window_create_context_reply_t

xcb_selinux_get_window_create_context_request_t

xcb_selinux_list_item_iterator_t

xcb_selinux_list_item_t

xcb_selinux_list_properties_cookie_t

xcb_selinux_list_properties_reply_t

xcb_selinux_list_properties_request_t

xcb_selinux_list_selections_cookie_t

xcb_selinux_list_selections_reply_t

xcb_selinux_list_selections_request_t

xcb_selinux_query_version_cookie_t

xcb_selinux_query_version_reply_t

xcb_selinux_query_version_request_t

xcb_selinux_set_device_context_request_t

xcb_selinux_set_device_create_context_request_t

xcb_selinux_set_property_create_context_request_t

xcb_selinux_set_property_use_context_request_t

xcb_selinux_set_selection_create_context_request_t

xcb_selinux_set_selection_use_context_request_t

xcb_selinux_set_window_create_context_request_t

Constants

XCB_SELINUX_GET_CLIENT_CONTEXT

XCB_SELINUX_GET_DEVICE_CONTEXT

XCB_SELINUX_GET_DEVICE_CREATE_CONTEXT

XCB_SELINUX_GET_PROPERTY_CONTEXT

XCB_SELINUX_GET_PROPERTY_CREATE_CONTEXT

XCB_SELINUX_GET_PROPERTY_DATA_CONTEXT

XCB_SELINUX_GET_PROPERTY_USE_CONTEXT

XCB_SELINUX_GET_SELECTION_CONTEXT

XCB_SELINUX_GET_SELECTION_CREATE_CONTEXT

XCB_SELINUX_GET_SELECTION_DATA_CONTEXT

XCB_SELINUX_GET_SELECTION_USE_CONTEXT

XCB_SELINUX_GET_WINDOW_CONTEXT

XCB_SELINUX_GET_WINDOW_CREATE_CONTEXT

XCB_SELINUX_LIST_PROPERTIES

XCB_SELINUX_LIST_SELECTIONS

XCB_SELINUX_MAJOR_VERSION

XCB_SELINUX_MINOR_VERSION

XCB_SELINUX_QUERY_VERSION

XCB_SELINUX_SET_DEVICE_CONTEXT

XCB_SELINUX_SET_DEVICE_CREATE_CONTEXT

XCB_SELINUX_SET_PROPERTY_CREATE_CONTEXT

XCB_SELINUX_SET_PROPERTY_USE_CONTEXT

XCB_SELINUX_SET_SELECTION_CREATE_CONTEXT

XCB_SELINUX_SET_SELECTION_USE_CONTEXT

XCB_SELINUX_SET_WINDOW_CREATE_CONTEXT

Statics

xcb_selinux_id

Functions

xcb_selinux_get_client_context

xcb_selinux_get_client_context_context

xcb_selinux_get_client_context_context_end

xcb_selinux_get_client_context_context_length

xcb_selinux_get_client_context_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_get_client_context_unchecked

xcb_selinux_get_device_context

xcb_selinux_get_device_context_context

xcb_selinux_get_device_context_context_end

xcb_selinux_get_device_context_context_length

xcb_selinux_get_device_context_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_get_device_context_unchecked

xcb_selinux_get_device_create_context

xcb_selinux_get_device_create_context_context

xcb_selinux_get_device_create_context_context_end

xcb_selinux_get_device_create_context_context_length

xcb_selinux_get_device_create_context_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_get_device_create_context_unchecked

xcb_selinux_get_property_context

xcb_selinux_get_property_context_context

xcb_selinux_get_property_context_context_end

xcb_selinux_get_property_context_context_length

xcb_selinux_get_property_context_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_get_property_context_unchecked

xcb_selinux_get_property_create_context

xcb_selinux_get_property_create_context_context

xcb_selinux_get_property_create_context_context_end

xcb_selinux_get_property_create_context_context_length

xcb_selinux_get_property_create_context_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_get_property_create_context_unchecked

xcb_selinux_get_property_data_context

xcb_selinux_get_property_data_context_context

xcb_selinux_get_property_data_context_context_end

xcb_selinux_get_property_data_context_context_length

xcb_selinux_get_property_data_context_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_get_property_data_context_unchecked

xcb_selinux_get_property_use_context

xcb_selinux_get_property_use_context_context

xcb_selinux_get_property_use_context_context_end

xcb_selinux_get_property_use_context_context_length

xcb_selinux_get_property_use_context_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_get_property_use_context_unchecked

xcb_selinux_get_selection_context

xcb_selinux_get_selection_context_context

xcb_selinux_get_selection_context_context_end

xcb_selinux_get_selection_context_context_length

xcb_selinux_get_selection_context_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_get_selection_context_unchecked

xcb_selinux_get_selection_create_context

xcb_selinux_get_selection_create_context_context

xcb_selinux_get_selection_create_context_context_end

xcb_selinux_get_selection_create_context_context_length

xcb_selinux_get_selection_create_context_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_get_selection_create_context_unchecked

xcb_selinux_get_selection_data_context

xcb_selinux_get_selection_data_context_context

xcb_selinux_get_selection_data_context_context_end

xcb_selinux_get_selection_data_context_context_length

xcb_selinux_get_selection_data_context_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_get_selection_data_context_unchecked

xcb_selinux_get_selection_use_context

xcb_selinux_get_selection_use_context_context

xcb_selinux_get_selection_use_context_context_end

xcb_selinux_get_selection_use_context_context_length

xcb_selinux_get_selection_use_context_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_get_selection_use_context_unchecked

xcb_selinux_get_window_context

xcb_selinux_get_window_context_context

xcb_selinux_get_window_context_context_end

xcb_selinux_get_window_context_context_length

xcb_selinux_get_window_context_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_get_window_context_unchecked

xcb_selinux_get_window_create_context

xcb_selinux_get_window_create_context_context

xcb_selinux_get_window_create_context_context_end

xcb_selinux_get_window_create_context_context_length

xcb_selinux_get_window_create_context_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_get_window_create_context_unchecked

xcb_selinux_list_item_data_context

xcb_selinux_list_item_data_context_end

xcb_selinux_list_item_data_context_length

xcb_selinux_list_item_end

xcb_selinux_list_item_next

xcb_selinux_list_item_object_context

xcb_selinux_list_item_object_context_end

xcb_selinux_list_item_object_context_length

xcb_selinux_list_properties

xcb_selinux_list_properties_properties_iterator

xcb_selinux_list_properties_properties_length

xcb_selinux_list_properties_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_list_properties_unchecked

xcb_selinux_list_selections

xcb_selinux_list_selections_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_list_selections_selections_iterator

xcb_selinux_list_selections_selections_length

xcb_selinux_list_selections_unchecked

xcb_selinux_query_version

xcb_selinux_query_version_reply
the returned value must be freed by the caller using libc::free().

xcb_selinux_query_version_unchecked

xcb_selinux_set_device_context

xcb_selinux_set_device_context_checked

xcb_selinux_set_device_create_context

xcb_selinux_set_device_create_context_checked

xcb_selinux_set_property_create_context

xcb_selinux_set_property_create_context_checked

xcb_selinux_set_property_use_context

xcb_selinux_set_property_use_context_checked

xcb_selinux_set_selection_create_context

xcb_selinux_set_selection_create_context_checked

xcb_selinux_set_selection_use_context

xcb_selinux_set_selection_use_context_checked

xcb_selinux_set_window_create_context

xcb_selinux_set_window_create_context_checked