Module xcb::ffi::shm

Structs

xcb_shm_attach_fd_request_t

xcb_shm_attach_request_t

xcb_shm_completion_event_t

xcb_shm_create_pixmap_request_t

xcb_shm_create_segment_cookie_t

xcb_shm_create_segment_reply_t

xcb_shm_create_segment_request_t

xcb_shm_detach_request_t

xcb_shm_get_image_cookie_t

xcb_shm_get_image_reply_t

xcb_shm_get_image_request_t

xcb_shm_put_image_request_t

xcb_shm_query_version_cookie_t

xcb_shm_query_version_reply_t

xcb_shm_query_version_request_t

xcb_shm_seg_iterator_t

Constants

XCB_SHM_ATTACH

XCB_SHM_ATTACH_FD

XCB_SHM_BAD_SEG

XCB_SHM_COMPLETION

XCB_SHM_CREATE_PIXMAP

XCB_SHM_CREATE_SEGMENT

XCB_SHM_DETACH

XCB_SHM_GET_IMAGE

XCB_SHM_MAJOR_VERSION

XCB_SHM_MINOR_VERSION

XCB_SHM_PUT_IMAGE

XCB_SHM_QUERY_VERSION

Statics

xcb_shm_id

Functions

xcb_shm_attach

xcb_shm_attach_checked

xcb_shm_attach_fd

xcb_shm_attach_fd_checked

xcb_shm_create_pixmap

xcb_shm_create_pixmap_checked

xcb_shm_create_segment

xcb_shm_create_segment_reply
the returned value must be freed by the caller using libc::free().

xcb_shm_create_segment_reply_fds
the returned value must be freed by the caller using libc::free().

xcb_shm_create_segment_unchecked

xcb_shm_detach

xcb_shm_detach_checked

xcb_shm_get_image

xcb_shm_get_image_reply
the returned value must be freed by the caller using libc::free().

xcb_shm_get_image_unchecked

xcb_shm_put_image

xcb_shm_put_image_checked

xcb_shm_query_version

xcb_shm_query_version_reply
the returned value must be freed by the caller using libc::free().

xcb_shm_query_version_unchecked

xcb_shm_seg_end

xcb_shm_seg_next

Type Definitions

xcb_shm_bad_seg_error_t

xcb_shm_seg_t