Module xcb::ffi::xevie

Structs

xcb_xevie_end_cookie_t

xcb_xevie_end_reply_t

xcb_xevie_end_request_t

xcb_xevie_event_iterator_t

xcb_xevie_event_t

xcb_xevie_query_version_cookie_t

xcb_xevie_query_version_reply_t

xcb_xevie_query_version_request_t

xcb_xevie_select_input_cookie_t

xcb_xevie_select_input_reply_t

xcb_xevie_select_input_request_t

xcb_xevie_send_cookie_t

xcb_xevie_send_reply_t

xcb_xevie_send_request_t

xcb_xevie_start_cookie_t

xcb_xevie_start_reply_t

xcb_xevie_start_request_t

Constants

XCB_XEVIE_DATATYPE_MODIFIED

XCB_XEVIE_DATATYPE_UNMODIFIED

XCB_XEVIE_END

XCB_XEVIE_MAJOR_VERSION

XCB_XEVIE_MINOR_VERSION

XCB_XEVIE_QUERY_VERSION

XCB_XEVIE_SELECT_INPUT

XCB_XEVIE_SEND

XCB_XEVIE_START

Statics

xcb_xevie_id

Functions

xcb_xevie_end⚠

xcb_xevie_end_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xevie_end_unchecked⚠

xcb_xevie_event_end⚠

xcb_xevie_event_next⚠

xcb_xevie_query_version⚠

xcb_xevie_query_version_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xevie_query_version_unchecked⚠

xcb_xevie_select_input⚠

xcb_xevie_select_input_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xevie_select_input_unchecked⚠

xcb_xevie_send⚠

xcb_xevie_send_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xevie_send_unchecked⚠

xcb_xevie_start⚠

xcb_xevie_start_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xevie_start_unchecked⚠

Type Definitions

xcb_xevie_datatype_t