Module xcb::ffi::xf86dri

Structs

xcb_xf86dri_auth_connection_cookie_t

xcb_xf86dri_auth_connection_reply_t

xcb_xf86dri_auth_connection_request_t

xcb_xf86dri_close_connection_request_t

xcb_xf86dri_create_context_cookie_t

xcb_xf86dri_create_context_reply_t

xcb_xf86dri_create_context_request_t

xcb_xf86dri_create_drawable_cookie_t

xcb_xf86dri_create_drawable_reply_t

xcb_xf86dri_create_drawable_request_t

xcb_xf86dri_destroy_context_request_t

xcb_xf86dri_destroy_drawable_request_t

xcb_xf86dri_drm_clip_rect_iterator_t

xcb_xf86dri_drm_clip_rect_t

xcb_xf86dri_get_client_driver_name_cookie_t

xcb_xf86dri_get_client_driver_name_reply_t

xcb_xf86dri_get_client_driver_name_request_t

xcb_xf86dri_get_device_info_cookie_t

xcb_xf86dri_get_device_info_reply_t

xcb_xf86dri_get_device_info_request_t

xcb_xf86dri_get_drawable_info_cookie_t

xcb_xf86dri_get_drawable_info_reply_t

xcb_xf86dri_get_drawable_info_request_t

xcb_xf86dri_open_connection_cookie_t

xcb_xf86dri_open_connection_reply_t

xcb_xf86dri_open_connection_request_t

xcb_xf86dri_query_direct_rendering_capable_cookie_t

xcb_xf86dri_query_direct_rendering_capable_reply_t

xcb_xf86dri_query_direct_rendering_capable_request_t

xcb_xf86dri_query_version_cookie_t

xcb_xf86dri_query_version_reply_t

xcb_xf86dri_query_version_request_t

Constants

XCB_XF86DRI_AUTH_CONNECTION

XCB_XF86DRI_CLOSE_CONNECTION

XCB_XF86DRI_CREATE_CONTEXT

XCB_XF86DRI_CREATE_DRAWABLE

XCB_XF86DRI_DESTROY_CONTEXT

XCB_XF86DRI_DESTROY_DRAWABLE

XCB_XF86DRI_GET_CLIENT_DRIVER_NAME

XCB_XF86DRI_GET_DEVICE_INFO

XCB_XF86DRI_GET_DRAWABLE_INFO

XCB_XF86DRI_MAJOR_VERSION

XCB_XF86DRI_MINOR_VERSION

XCB_XF86DRI_OPEN_CONNECTION

XCB_XF86DRI_QUERY_DIRECT_RENDERING_CAPABLE

XCB_XF86DRI_QUERY_VERSION

Statics

xcb_xf86dri_id

Functions

xcb_xf86dri_auth_connection

xcb_xf86dri_auth_connection_reply
the returned value must be freed by the caller using libc::free().

xcb_xf86dri_auth_connection_unchecked

xcb_xf86dri_close_connection

xcb_xf86dri_close_connection_checked

xcb_xf86dri_create_context

xcb_xf86dri_create_context_reply
the returned value must be freed by the caller using libc::free().

xcb_xf86dri_create_context_unchecked

xcb_xf86dri_create_drawable

xcb_xf86dri_create_drawable_reply
the returned value must be freed by the caller using libc::free().

xcb_xf86dri_create_drawable_unchecked

xcb_xf86dri_destroy_context

xcb_xf86dri_destroy_context_checked

xcb_xf86dri_destroy_drawable

xcb_xf86dri_destroy_drawable_checked

xcb_xf86dri_drm_clip_rect_end

xcb_xf86dri_drm_clip_rect_next

xcb_xf86dri_get_client_driver_name

xcb_xf86dri_get_client_driver_name_client_driver_name

xcb_xf86dri_get_client_driver_name_client_driver_name_end

xcb_xf86dri_get_client_driver_name_client_driver_name_length

xcb_xf86dri_get_client_driver_name_reply
the returned value must be freed by the caller using libc::free().

xcb_xf86dri_get_client_driver_name_unchecked

xcb_xf86dri_get_device_info

xcb_xf86dri_get_device_info_device_private

xcb_xf86dri_get_device_info_device_private_end

xcb_xf86dri_get_device_info_device_private_length

xcb_xf86dri_get_device_info_reply
the returned value must be freed by the caller using libc::free().

xcb_xf86dri_get_device_info_unchecked

xcb_xf86dri_get_drawable_info

xcb_xf86dri_get_drawable_info_back_clip_rects

xcb_xf86dri_get_drawable_info_back_clip_rects_iterator

xcb_xf86dri_get_drawable_info_back_clip_rects_length

xcb_xf86dri_get_drawable_info_clip_rects

xcb_xf86dri_get_drawable_info_clip_rects_iterator

xcb_xf86dri_get_drawable_info_clip_rects_length

xcb_xf86dri_get_drawable_info_reply
the returned value must be freed by the caller using libc::free().

xcb_xf86dri_get_drawable_info_unchecked

xcb_xf86dri_open_connection

xcb_xf86dri_open_connection_bus_id

xcb_xf86dri_open_connection_bus_id_end

xcb_xf86dri_open_connection_bus_id_length

xcb_xf86dri_open_connection_reply
the returned value must be freed by the caller using libc::free().

xcb_xf86dri_open_connection_unchecked

xcb_xf86dri_query_direct_rendering_capable

xcb_xf86dri_query_direct_rendering_capable_reply
the returned value must be freed by the caller using libc::free().

xcb_xf86dri_query_direct_rendering_capable_unchecked

xcb_xf86dri_query_version

xcb_xf86dri_query_version_reply
the returned value must be freed by the caller using libc::free().

xcb_xf86dri_query_version_unchecked