Module xcb::ffi::xf86vidmode

Structs

xcb_xf86vidmode_add_mode_line_request_t

xcb_xf86vidmode_bad_clock_error_t

xcb_xf86vidmode_bad_h_timings_error_t

xcb_xf86vidmode_bad_v_timings_error_t

xcb_xf86vidmode_client_not_local_error_t

xcb_xf86vidmode_delete_mode_line_request_t

xcb_xf86vidmode_dotclock_iterator_t

xcb_xf86vidmode_extension_disabled_error_t

xcb_xf86vidmode_get_all_mode_lines_cookie_t

xcb_xf86vidmode_get_all_mode_lines_reply_t

xcb_xf86vidmode_get_all_mode_lines_request_t

xcb_xf86vidmode_get_dot_clocks_cookie_t

xcb_xf86vidmode_get_dot_clocks_reply_t

xcb_xf86vidmode_get_dot_clocks_request_t

xcb_xf86vidmode_get_gamma_cookie_t

xcb_xf86vidmode_get_gamma_ramp_cookie_t

xcb_xf86vidmode_get_gamma_ramp_reply_t

xcb_xf86vidmode_get_gamma_ramp_request_t

xcb_xf86vidmode_get_gamma_ramp_size_cookie_t

xcb_xf86vidmode_get_gamma_ramp_size_reply_t

xcb_xf86vidmode_get_gamma_ramp_size_request_t

xcb_xf86vidmode_get_gamma_reply_t

xcb_xf86vidmode_get_gamma_request_t

xcb_xf86vidmode_get_mode_line_cookie_t

xcb_xf86vidmode_get_mode_line_reply_t

xcb_xf86vidmode_get_mode_line_request_t

xcb_xf86vidmode_get_monitor_cookie_t

xcb_xf86vidmode_get_monitor_reply_t

xcb_xf86vidmode_get_monitor_request_t

xcb_xf86vidmode_get_permissions_cookie_t

xcb_xf86vidmode_get_permissions_reply_t

xcb_xf86vidmode_get_permissions_request_t

xcb_xf86vidmode_get_view_port_cookie_t

xcb_xf86vidmode_get_view_port_reply_t

xcb_xf86vidmode_get_view_port_request_t

xcb_xf86vidmode_lock_mode_switch_request_t

xcb_xf86vidmode_mod_mode_line_request_t

xcb_xf86vidmode_mode_info_iterator_t

xcb_xf86vidmode_mode_info_t

xcb_xf86vidmode_mode_unsuitable_error_t

xcb_xf86vidmode_query_version_cookie_t

xcb_xf86vidmode_query_version_reply_t

xcb_xf86vidmode_query_version_request_t

xcb_xf86vidmode_set_client_version_request_t

xcb_xf86vidmode_set_gamma_ramp_request_t

xcb_xf86vidmode_set_gamma_request_t

xcb_xf86vidmode_set_view_port_request_t

xcb_xf86vidmode_switch_mode_request_t

xcb_xf86vidmode_switch_to_mode_request_t

xcb_xf86vidmode_syncrange_iterator_t

xcb_xf86vidmode_validate_mode_line_cookie_t

xcb_xf86vidmode_validate_mode_line_reply_t

xcb_xf86vidmode_validate_mode_line_request_t

xcb_xf86vidmode_zoom_locked_error_t

Constants

XCB_XF86VIDMODE_ADD_MODE_LINE

XCB_XF86VIDMODE_BAD_CLOCK

XCB_XF86VIDMODE_BAD_H_TIMINGS

XCB_XF86VIDMODE_BAD_V_TIMINGS

XCB_XF86VIDMODE_CLIENT_NOT_LOCAL

XCB_XF86VIDMODE_CLOCK_FLAG_PROGRAMABLE

XCB_XF86VIDMODE_DELETE_MODE_LINE

XCB_XF86VIDMODE_EXTENSION_DISABLED

XCB_XF86VIDMODE_GET_ALL_MODE_LINES

XCB_XF86VIDMODE_GET_DOT_CLOCKS

XCB_XF86VIDMODE_GET_GAMMA

XCB_XF86VIDMODE_GET_GAMMA_RAMP

XCB_XF86VIDMODE_GET_GAMMA_RAMP_SIZE

XCB_XF86VIDMODE_GET_MODE_LINE

XCB_XF86VIDMODE_GET_MONITOR

XCB_XF86VIDMODE_GET_PERMISSIONS

XCB_XF86VIDMODE_GET_VIEW_PORT

XCB_XF86VIDMODE_LOCK_MODE_SWITCH

XCB_XF86VIDMODE_MAJOR_VERSION

XCB_XF86VIDMODE_MINOR_VERSION

XCB_XF86VIDMODE_MODE_FLAG_BROADCAST

XCB_XF86VIDMODE_MODE_FLAG_COMPOSITE_SYNC

XCB_XF86VIDMODE_MODE_FLAG_DOUBLE_CLOCK

XCB_XF86VIDMODE_MODE_FLAG_HALF_CLOCK

XCB_XF86VIDMODE_MODE_FLAG_H_SKEW

XCB_XF86VIDMODE_MODE_FLAG_INTERLACE

XCB_XF86VIDMODE_MODE_FLAG_NEGATIVE_C_SYNC

XCB_XF86VIDMODE_MODE_FLAG_NEGATIVE_H_SYNC

XCB_XF86VIDMODE_MODE_FLAG_NEGATIVE_V_SYNC

XCB_XF86VIDMODE_MODE_FLAG_PIXMUX

XCB_XF86VIDMODE_MODE_FLAG_POSITIVE_C_SYNC

XCB_XF86VIDMODE_MODE_FLAG_POSITIVE_H_SYNC

XCB_XF86VIDMODE_MODE_FLAG_POSITIVE_V_SYNC

XCB_XF86VIDMODE_MODE_UNSUITABLE

XCB_XF86VIDMODE_MOD_MODE_LINE

XCB_XF86VIDMODE_PERMISSION_READ

XCB_XF86VIDMODE_PERMISSION_WRITE

XCB_XF86VIDMODE_QUERY_VERSION

XCB_XF86VIDMODE_SET_CLIENT_VERSION

XCB_XF86VIDMODE_SET_GAMMA

XCB_XF86VIDMODE_SET_GAMMA_RAMP

XCB_XF86VIDMODE_SET_VIEW_PORT

XCB_XF86VIDMODE_SWITCH_MODE

XCB_XF86VIDMODE_SWITCH_TO_MODE

XCB_XF86VIDMODE_VALIDATE_MODE_LINE

XCB_XF86VIDMODE_ZOOM_LOCKED

Statics

xcb_xf86vidmode_id

Functions

xcb_xf86vidmode_add_mode_line⚠

xcb_xf86vidmode_add_mode_line_checked⚠

xcb_xf86vidmode_delete_mode_line⚠

xcb_xf86vidmode_delete_mode_line_checked⚠

xcb_xf86vidmode_dotclock_end⚠

xcb_xf86vidmode_dotclock_next⚠

xcb_xf86vidmode_get_all_mode_lines⚠

xcb_xf86vidmode_get_all_mode_lines_modeinfo⚠

xcb_xf86vidmode_get_all_mode_lines_modeinfo_iterator⚠

xcb_xf86vidmode_get_all_mode_lines_modeinfo_length⚠

xcb_xf86vidmode_get_all_mode_lines_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xf86vidmode_get_all_mode_lines_unchecked⚠

xcb_xf86vidmode_get_dot_clocks⚠

xcb_xf86vidmode_get_dot_clocks_clock⚠

xcb_xf86vidmode_get_dot_clocks_clock_end⚠

xcb_xf86vidmode_get_dot_clocks_clock_length⚠

xcb_xf86vidmode_get_dot_clocks_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xf86vidmode_get_dot_clocks_unchecked⚠

xcb_xf86vidmode_get_gamma⚠

xcb_xf86vidmode_get_gamma_ramp⚠

xcb_xf86vidmode_get_gamma_ramp_blue⚠

xcb_xf86vidmode_get_gamma_ramp_blue_end⚠

xcb_xf86vidmode_get_gamma_ramp_blue_length⚠

xcb_xf86vidmode_get_gamma_ramp_green⚠

xcb_xf86vidmode_get_gamma_ramp_green_end⚠

xcb_xf86vidmode_get_gamma_ramp_green_length⚠

xcb_xf86vidmode_get_gamma_ramp_red⚠

xcb_xf86vidmode_get_gamma_ramp_red_end⚠

xcb_xf86vidmode_get_gamma_ramp_red_length⚠

xcb_xf86vidmode_get_gamma_ramp_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xf86vidmode_get_gamma_ramp_size⚠

xcb_xf86vidmode_get_gamma_ramp_size_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xf86vidmode_get_gamma_ramp_size_unchecked⚠

xcb_xf86vidmode_get_gamma_ramp_unchecked⚠

xcb_xf86vidmode_get_gamma_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xf86vidmode_get_gamma_unchecked⚠

xcb_xf86vidmode_get_mode_line⚠

xcb_xf86vidmode_get_mode_line_private⚠

xcb_xf86vidmode_get_mode_line_private_end⚠

xcb_xf86vidmode_get_mode_line_private_length⚠

xcb_xf86vidmode_get_mode_line_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xf86vidmode_get_mode_line_unchecked⚠

xcb_xf86vidmode_get_monitor⚠

xcb_xf86vidmode_get_monitor_alignment_pad⚠

xcb_xf86vidmode_get_monitor_alignment_pad_end⚠

xcb_xf86vidmode_get_monitor_alignment_pad_length⚠

xcb_xf86vidmode_get_monitor_hsync⚠

xcb_xf86vidmode_get_monitor_hsync_end⚠

xcb_xf86vidmode_get_monitor_hsync_length⚠

xcb_xf86vidmode_get_monitor_model⚠

xcb_xf86vidmode_get_monitor_model_end⚠

xcb_xf86vidmode_get_monitor_model_length⚠

xcb_xf86vidmode_get_monitor_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xf86vidmode_get_monitor_unchecked⚠

xcb_xf86vidmode_get_monitor_vendor⚠

xcb_xf86vidmode_get_monitor_vendor_end⚠

xcb_xf86vidmode_get_monitor_vendor_length⚠

xcb_xf86vidmode_get_monitor_vsync⚠

xcb_xf86vidmode_get_monitor_vsync_end⚠

xcb_xf86vidmode_get_monitor_vsync_length⚠

xcb_xf86vidmode_get_permissions⚠

xcb_xf86vidmode_get_permissions_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xf86vidmode_get_permissions_unchecked⚠

xcb_xf86vidmode_get_view_port⚠

xcb_xf86vidmode_get_view_port_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xf86vidmode_get_view_port_unchecked⚠

xcb_xf86vidmode_lock_mode_switch⚠

xcb_xf86vidmode_lock_mode_switch_checked⚠

xcb_xf86vidmode_mod_mode_line⚠

xcb_xf86vidmode_mod_mode_line_checked⚠

xcb_xf86vidmode_mode_info_end⚠

xcb_xf86vidmode_mode_info_next⚠

xcb_xf86vidmode_query_version⚠

xcb_xf86vidmode_query_version_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xf86vidmode_query_version_unchecked⚠

xcb_xf86vidmode_set_client_version⚠

xcb_xf86vidmode_set_client_version_checked⚠

xcb_xf86vidmode_set_gamma⚠

xcb_xf86vidmode_set_gamma_checked⚠

xcb_xf86vidmode_set_gamma_ramp⚠

xcb_xf86vidmode_set_gamma_ramp_checked⚠

xcb_xf86vidmode_set_view_port⚠

xcb_xf86vidmode_set_view_port_checked⚠

xcb_xf86vidmode_switch_mode⚠

xcb_xf86vidmode_switch_mode_checked⚠

xcb_xf86vidmode_switch_to_mode⚠

xcb_xf86vidmode_switch_to_mode_checked⚠

xcb_xf86vidmode_syncrange_end⚠

xcb_xf86vidmode_syncrange_next⚠

xcb_xf86vidmode_validate_mode_line⚠

xcb_xf86vidmode_validate_mode_line_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xf86vidmode_validate_mode_line_unchecked⚠

Type Definitions

xcb_xf86vidmode_clock_flag_t

xcb_xf86vidmode_dotclock_t

xcb_xf86vidmode_mode_flag_t

xcb_xf86vidmode_permission_t

xcb_xf86vidmode_syncrange_t