Module xcb::ffi::xinerama

Structs

xcb_xinerama_get_screen_count_cookie_t

xcb_xinerama_get_screen_count_reply_t

xcb_xinerama_get_screen_count_request_t

xcb_xinerama_get_screen_size_cookie_t

xcb_xinerama_get_screen_size_reply_t

xcb_xinerama_get_screen_size_request_t

xcb_xinerama_get_state_cookie_t

xcb_xinerama_get_state_reply_t

xcb_xinerama_get_state_request_t

xcb_xinerama_is_active_cookie_t

xcb_xinerama_is_active_reply_t

xcb_xinerama_is_active_request_t

xcb_xinerama_query_screens_cookie_t

xcb_xinerama_query_screens_reply_t

xcb_xinerama_query_screens_request_t

xcb_xinerama_query_version_cookie_t

xcb_xinerama_query_version_reply_t

xcb_xinerama_query_version_request_t

xcb_xinerama_screen_info_iterator_t

xcb_xinerama_screen_info_t

Constants

XCB_XINERAMA_GET_SCREEN_COUNT

XCB_XINERAMA_GET_SCREEN_SIZE

XCB_XINERAMA_GET_STATE

XCB_XINERAMA_IS_ACTIVE

XCB_XINERAMA_MAJOR_VERSION

XCB_XINERAMA_MINOR_VERSION

XCB_XINERAMA_QUERY_SCREENS

XCB_XINERAMA_QUERY_VERSION

Statics

xcb_xinerama_id

Functions

xcb_xinerama_get_screen_count

xcb_xinerama_get_screen_count_reply
the returned value must be freed by the caller using libc::free().

xcb_xinerama_get_screen_count_unchecked

xcb_xinerama_get_screen_size

xcb_xinerama_get_screen_size_reply
the returned value must be freed by the caller using libc::free().

xcb_xinerama_get_screen_size_unchecked

xcb_xinerama_get_state

xcb_xinerama_get_state_reply
the returned value must be freed by the caller using libc::free().

xcb_xinerama_get_state_unchecked

xcb_xinerama_is_active

xcb_xinerama_is_active_reply
the returned value must be freed by the caller using libc::free().

xcb_xinerama_is_active_unchecked

xcb_xinerama_query_screens

xcb_xinerama_query_screens_reply
the returned value must be freed by the caller using libc::free().

xcb_xinerama_query_screens_screen_info

xcb_xinerama_query_screens_screen_info_iterator

xcb_xinerama_query_screens_screen_info_length

xcb_xinerama_query_screens_unchecked

xcb_xinerama_query_version

xcb_xinerama_query_version_reply
the returned value must be freed by the caller using libc::free().

xcb_xinerama_query_version_unchecked

xcb_xinerama_screen_info_end

xcb_xinerama_screen_info_next