Module xcb::ffi::xkb

Structs

_xcb_xkb_get_kbd_by_name_replies__compat_map

_xcb_xkb_get_kbd_by_name_replies__geometry

_xcb_xkb_get_kbd_by_name_replies__indicator_maps

_xcb_xkb_get_kbd_by_name_replies__key_names

_xcb_xkb_get_kbd_by_name_replies__types

xcb_xkb_access_x_notify_event_t

xcb_xkb_action_iterator_t

xcb_xkb_action_message_event_t

xcb_xkb_action_t

xcb_xkb_behavior_iterator_t

xcb_xkb_behavior_t

xcb_xkb_bell_class_spec_iterator_t

xcb_xkb_bell_notify_event_t

xcb_xkb_bell_request_t

xcb_xkb_common_behavior_iterator_t

xcb_xkb_common_behavior_t

xcb_xkb_compat_map_notify_event_t

xcb_xkb_controls_notify_event_t

xcb_xkb_counted_string_16_iterator_t

xcb_xkb_counted_string_16_t

xcb_xkb_default_behavior_iterator_t

xcb_xkb_default_behavior_t

xcb_xkb_device_led_info_iterator_t

xcb_xkb_device_led_info_t

xcb_xkb_device_spec_iterator_t

xcb_xkb_extension_device_notify_event_t

xcb_xkb_get_compat_map_cookie_t

xcb_xkb_get_compat_map_reply_t

xcb_xkb_get_compat_map_request_t

xcb_xkb_get_controls_cookie_t

xcb_xkb_get_controls_reply_t

xcb_xkb_get_controls_request_t

xcb_xkb_get_device_info_cookie_t

xcb_xkb_get_device_info_reply_t

xcb_xkb_get_device_info_request_t

xcb_xkb_get_indicator_map_cookie_t

xcb_xkb_get_indicator_map_reply_t

xcb_xkb_get_indicator_map_request_t

xcb_xkb_get_indicator_state_cookie_t

xcb_xkb_get_indicator_state_reply_t

xcb_xkb_get_indicator_state_request_t

xcb_xkb_get_kbd_by_name_cookie_t

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_t

xcb_xkb_get_kbd_by_name_replies_t

xcb_xkb_get_kbd_by_name_replies_types_map_t

xcb_xkb_get_kbd_by_name_reply_t

xcb_xkb_get_kbd_by_name_request_t

xcb_xkb_get_map_cookie_t

xcb_xkb_get_map_map_t

xcb_xkb_get_map_reply_t

xcb_xkb_get_map_request_t

xcb_xkb_get_named_indicator_cookie_t

xcb_xkb_get_named_indicator_reply_t

xcb_xkb_get_named_indicator_request_t

xcb_xkb_get_names_cookie_t

xcb_xkb_get_names_reply_t

xcb_xkb_get_names_request_t

xcb_xkb_get_names_value_list_t

xcb_xkb_get_state_cookie_t

xcb_xkb_get_state_reply_t

xcb_xkb_get_state_request_t

xcb_xkb_id_spec_iterator_t

xcb_xkb_indicator_map_iterator_t

xcb_xkb_indicator_map_notify_event_t

xcb_xkb_indicator_map_t

xcb_xkb_indicator_state_notify_event_t

xcb_xkb_key_alias_iterator_t

xcb_xkb_key_alias_t

xcb_xkb_key_iterator_t

xcb_xkb_key_mod_map_iterator_t

xcb_xkb_key_mod_map_t

xcb_xkb_key_name_iterator_t

xcb_xkb_key_name_t

xcb_xkb_key_sym_map_iterator_t

xcb_xkb_key_sym_map_t

xcb_xkb_key_t

xcb_xkb_key_type_iterator_t

xcb_xkb_key_type_t

xcb_xkb_key_v_mod_map_iterator_t

xcb_xkb_key_v_mod_map_t

xcb_xkb_keyboard_error_t

xcb_xkb_kt_map_entry_iterator_t

xcb_xkb_kt_map_entry_t

xcb_xkb_kt_set_map_entry_iterator_t

xcb_xkb_kt_set_map_entry_t

xcb_xkb_latch_lock_state_request_t

xcb_xkb_led_class_spec_iterator_t

xcb_xkb_list_components_cookie_t

xcb_xkb_list_components_reply_t

xcb_xkb_list_components_request_t

xcb_xkb_listing_iterator_t

xcb_xkb_listing_t

xcb_xkb_lock_behavior_iterator_t

xcb_xkb_lock_behavior_t

xcb_xkb_map_notify_event_t

xcb_xkb_mod_def_iterator_t

xcb_xkb_mod_def_t

xcb_xkb_names_notify_event_t

xcb_xkb_new_keyboard_notify_event_t

xcb_xkb_outline_iterator_t

xcb_xkb_outline_t

xcb_xkb_overlay_behavior_iterator_t

xcb_xkb_overlay_behavior_t

xcb_xkb_overlay_iterator_t

xcb_xkb_overlay_key_iterator_t

xcb_xkb_overlay_key_t

xcb_xkb_overlay_row_iterator_t

xcb_xkb_overlay_row_t

xcb_xkb_overlay_t

xcb_xkb_per_client_flags_cookie_t

xcb_xkb_per_client_flags_reply_t

xcb_xkb_per_client_flags_request_t

xcb_xkb_permament_lock_behavior_iterator_t

xcb_xkb_permament_lock_behavior_t

xcb_xkb_permament_overlay_behavior_iterator_t

xcb_xkb_permament_overlay_behavior_t

xcb_xkb_permament_radio_group_behavior_iterator_t

xcb_xkb_permament_radio_group_behavior_t

xcb_xkb_radio_group_behavior_iterator_t

xcb_xkb_radio_group_behavior_t

xcb_xkb_row_iterator_t

xcb_xkb_row_t

xcb_xkb_sa_action_message_iterator_t

xcb_xkb_sa_action_message_t

xcb_xkb_sa_device_btn_iterator_t

xcb_xkb_sa_device_btn_t

xcb_xkb_sa_device_valuator_iterator_t

xcb_xkb_sa_device_valuator_t

xcb_xkb_sa_iso_lock_iterator_t

xcb_xkb_sa_iso_lock_t

xcb_xkb_sa_latch_group_iterator_t

xcb_xkb_sa_latch_group_t

xcb_xkb_sa_latch_mods_iterator_t

xcb_xkb_sa_latch_mods_t

xcb_xkb_sa_lock_controls_iterator_t

xcb_xkb_sa_lock_controls_t

xcb_xkb_sa_lock_device_btn_iterator_t

xcb_xkb_sa_lock_device_btn_t

xcb_xkb_sa_lock_group_iterator_t

xcb_xkb_sa_lock_group_t

xcb_xkb_sa_lock_mods_iterator_t

xcb_xkb_sa_lock_mods_t

xcb_xkb_sa_lock_ptr_btn_iterator_t

xcb_xkb_sa_lock_ptr_btn_t

xcb_xkb_sa_move_ptr_iterator_t

xcb_xkb_sa_move_ptr_t

xcb_xkb_sa_no_action_iterator_t

xcb_xkb_sa_no_action_t

xcb_xkb_sa_ptr_btn_iterator_t

xcb_xkb_sa_ptr_btn_t

xcb_xkb_sa_redirect_key_iterator_t

xcb_xkb_sa_redirect_key_t

xcb_xkb_sa_set_controls_iterator_t

xcb_xkb_sa_set_controls_t

xcb_xkb_sa_set_group_iterator_t

xcb_xkb_sa_set_group_t

xcb_xkb_sa_set_mods_iterator_t

xcb_xkb_sa_set_mods_t

xcb_xkb_sa_set_ptr_dflt_iterator_t

xcb_xkb_sa_set_ptr_dflt_t

xcb_xkb_sa_switch_screen_iterator_t

xcb_xkb_sa_switch_screen_t

xcb_xkb_sa_terminate_iterator_t

xcb_xkb_sa_terminate_t

xcb_xkb_select_events_details_t

xcb_xkb_select_events_request_t

xcb_xkb_set_behavior_iterator_t

xcb_xkb_set_behavior_t

xcb_xkb_set_compat_map_request_t

xcb_xkb_set_controls_request_t

xcb_xkb_set_debugging_flags_cookie_t

xcb_xkb_set_debugging_flags_reply_t

xcb_xkb_set_debugging_flags_request_t

xcb_xkb_set_device_info_request_t

xcb_xkb_set_explicit_iterator_t

xcb_xkb_set_explicit_t

xcb_xkb_set_indicator_map_request_t

xcb_xkb_set_key_type_iterator_t

xcb_xkb_set_key_type_t

xcb_xkb_set_map_request_t

xcb_xkb_set_map_values_t

xcb_xkb_set_named_indicator_request_t

xcb_xkb_set_names_request_t

xcb_xkb_set_names_values_t

xcb_xkb_shape_iterator_t

xcb_xkb_shape_t

xcb_xkb_si_action_iterator_t

xcb_xkb_si_action_t

xcb_xkb_state_notify_event_t

xcb_xkb_string8_iterator_t

xcb_xkb_sym_interpret_iterator_t

xcb_xkb_sym_interpret_t

xcb_xkb_use_extension_cookie_t

xcb_xkb_use_extension_reply_t

xcb_xkb_use_extension_request_t

Constants

XCB_XKB_ACCESS_X_NOTIFY

XCB_XKB_ACTION_MESSAGE

XCB_XKB_ACTION_MESSAGE_FLAG_GEN_KEY_EVENT

XCB_XKB_ACTION_MESSAGE_FLAG_ON_PRESS

XCB_XKB_ACTION_MESSAGE_FLAG_ON_RELEASE

XCB_XKB_AXN_DETAIL_AXK_WARNING

XCB_XKB_AXN_DETAIL_BK_ACCEPT

XCB_XKB_AXN_DETAIL_BK_REJECT

XCB_XKB_AXN_DETAIL_SK_ACCEPT

XCB_XKB_AXN_DETAIL_SK_PRESS

XCB_XKB_AXN_DETAIL_SK_REJECT

XCB_XKB_AXN_DETAIL_SK_RELEASE

XCB_XKB_AX_OPTION_BK_REJECT_FB

XCB_XKB_AX_OPTION_DUMB_BELL

XCB_XKB_AX_OPTION_FEATURE_FB

XCB_XKB_AX_OPTION_INDICATOR_FB

XCB_XKB_AX_OPTION_LATCH_TO_LOCK

XCB_XKB_AX_OPTION_SK_ACCEPT_FB

XCB_XKB_AX_OPTION_SK_PRESS_FB

XCB_XKB_AX_OPTION_SK_REJECT_FB

XCB_XKB_AX_OPTION_SK_RELEASE_FB

XCB_XKB_AX_OPTION_SLOW_WARN_FB

XCB_XKB_AX_OPTION_STICKY_KEYS_FB

XCB_XKB_AX_OPTION_TWO_KEYS

XCB_XKB_BEHAVIOR_TYPE_DEFAULT

XCB_XKB_BEHAVIOR_TYPE_LOCK

XCB_XKB_BEHAVIOR_TYPE_OVERLAY_1

XCB_XKB_BEHAVIOR_TYPE_OVERLAY_2

XCB_XKB_BEHAVIOR_TYPE_PERMAMENT_LOCK

XCB_XKB_BEHAVIOR_TYPE_PERMAMENT_OVERLAY_1

XCB_XKB_BEHAVIOR_TYPE_PERMAMENT_OVERLAY_2

XCB_XKB_BEHAVIOR_TYPE_PERMAMENT_RADIO_GROUP

XCB_XKB_BEHAVIOR_TYPE_RADIO_GROUP

XCB_XKB_BELL

XCB_XKB_BELL_CLASS_BELL_FEEDBACK_CLASS

XCB_XKB_BELL_CLASS_DFLT_XI_CLASS

XCB_XKB_BELL_CLASS_KBD_FEEDBACK_CLASS

XCB_XKB_BELL_CLASS_RESULT_BELL_FEEDBACK_CLASS

XCB_XKB_BELL_CLASS_RESULT_KBD_FEEDBACK_CLASS

XCB_XKB_BELL_NOTIFY

XCB_XKB_BOOL_CTRLS_HIGH_ACCESS_X_FEEDBACK

XCB_XKB_BOOL_CTRLS_HIGH_AUDIBLE_BELL

XCB_XKB_BOOL_CTRLS_HIGH_IGNORE_GROUP_LOCK

XCB_XKB_BOOL_CTRLS_HIGH_OVERLAY_1

XCB_XKB_BOOL_CTRLS_HIGH_OVERLAY_2

XCB_XKB_BOOL_CTRLS_LOW_ACCESS_X_KEYS

XCB_XKB_BOOL_CTRLS_LOW_ACCESS_X_TIMEOUT

XCB_XKB_BOOL_CTRLS_LOW_BOUNCE_KEYS

XCB_XKB_BOOL_CTRLS_LOW_MOUSE_KEYS

XCB_XKB_BOOL_CTRLS_LOW_MOUSE_KEYS_ACCEL

XCB_XKB_BOOL_CTRLS_LOW_REPEAT_KEYS

XCB_XKB_BOOL_CTRLS_LOW_SLOW_KEYS

XCB_XKB_BOOL_CTRLS_LOW_STICKY_KEYS

XCB_XKB_BOOL_CTRL_ACCESS_X_FEEDBACK_MASK

XCB_XKB_BOOL_CTRL_ACCESS_X_KEYS

XCB_XKB_BOOL_CTRL_ACCESS_X_TIMEOUT_MASK

XCB_XKB_BOOL_CTRL_AUDIBLE_BELL_MASK

XCB_XKB_BOOL_CTRL_BOUNCE_KEYS

XCB_XKB_BOOL_CTRL_IGNORE_GROUP_LOCK_MASK

XCB_XKB_BOOL_CTRL_MOUSE_KEYS

XCB_XKB_BOOL_CTRL_MOUSE_KEYS_ACCEL

XCB_XKB_BOOL_CTRL_OVERLAY_1_MASK

XCB_XKB_BOOL_CTRL_OVERLAY_2_MASK

XCB_XKB_BOOL_CTRL_REPEAT_KEYS

XCB_XKB_BOOL_CTRL_SLOW_KEYS

XCB_XKB_BOOL_CTRL_STICKY_KEYS

XCB_XKB_CM_DETAIL_GROUP_COMPAT

XCB_XKB_CM_DETAIL_SYM_INTERP

XCB_XKB_COMPAT_MAP_NOTIFY

XCB_XKB_CONST_KEY_NAME_LENGTH

XCB_XKB_CONST_MAX_LEGAL_KEY_CODE

XCB_XKB_CONST_PER_KEY_BIT_ARRAY_SIZE

XCB_XKB_CONTROLS_NOTIFY

XCB_XKB_CONTROL_CONTROLS_ENABLED

XCB_XKB_CONTROL_GROUPS_WRAP

XCB_XKB_CONTROL_IGNORE_LOCK_MODS

XCB_XKB_CONTROL_INTERNAL_MODS

XCB_XKB_CONTROL_PER_KEY_REPEAT

XCB_XKB_DOODAD_TYPE_INDICATOR

XCB_XKB_DOODAD_TYPE_LOGO

XCB_XKB_DOODAD_TYPE_OUTLINE

XCB_XKB_DOODAD_TYPE_SOLID

XCB_XKB_DOODAD_TYPE_TEXT

XCB_XKB_ERROR_BAD_CLASS

XCB_XKB_ERROR_BAD_DEVICE

XCB_XKB_ERROR_BAD_ID

XCB_XKB_EVENT_TYPE_ACCESS_X_NOTIFY

XCB_XKB_EVENT_TYPE_ACTION_MESSAGE

XCB_XKB_EVENT_TYPE_BELL_NOTIFY

XCB_XKB_EVENT_TYPE_COMPAT_MAP_NOTIFY

XCB_XKB_EVENT_TYPE_CONTROLS_NOTIFY

XCB_XKB_EVENT_TYPE_EXTENSION_DEVICE_NOTIFY

XCB_XKB_EVENT_TYPE_INDICATOR_MAP_NOTIFY

XCB_XKB_EVENT_TYPE_INDICATOR_STATE_NOTIFY

XCB_XKB_EVENT_TYPE_MAP_NOTIFY

XCB_XKB_EVENT_TYPE_NAMES_NOTIFY

XCB_XKB_EVENT_TYPE_NEW_KEYBOARD_NOTIFY

XCB_XKB_EVENT_TYPE_STATE_NOTIFY

XCB_XKB_EXPLICIT_AUTO_REPEAT

XCB_XKB_EXPLICIT_BEHAVIOR

XCB_XKB_EXPLICIT_INTERPRET

XCB_XKB_EXPLICIT_KEY_TYPE_1

XCB_XKB_EXPLICIT_KEY_TYPE_2

XCB_XKB_EXPLICIT_KEY_TYPE_3

XCB_XKB_EXPLICIT_KEY_TYPE_4

XCB_XKB_EXPLICIT_V_MOD_MAP

XCB_XKB_EXTENSION_DEVICE_NOTIFY

XCB_XKB_GBN_DETAIL_CLIENT_SYMBOLS

XCB_XKB_GBN_DETAIL_COMPAT_MAP

XCB_XKB_GBN_DETAIL_GEOMETRY

XCB_XKB_GBN_DETAIL_INDICATOR_MAPS

XCB_XKB_GBN_DETAIL_KEY_NAMES

XCB_XKB_GBN_DETAIL_OTHER_NAMES

XCB_XKB_GBN_DETAIL_SERVER_SYMBOLS

XCB_XKB_GBN_DETAIL_TYPES

XCB_XKB_GET_COMPAT_MAP

XCB_XKB_GET_CONTROLS

XCB_XKB_GET_DEVICE_INFO

XCB_XKB_GET_INDICATOR_MAP

XCB_XKB_GET_INDICATOR_STATE

XCB_XKB_GET_KBD_BY_NAME

XCB_XKB_GET_MAP

XCB_XKB_GET_NAMED_INDICATOR

XCB_XKB_GET_NAMES

XCB_XKB_GET_STATE

XCB_XKB_GROUPS_ALL

XCB_XKB_GROUPS_ANY

XCB_XKB_GROUPS_WRAP_CLAMP_INTO_RANGE

XCB_XKB_GROUPS_WRAP_REDIRECT_INTO_RANGE

XCB_XKB_GROUPS_WRAP_WRAP_INTO_RANGE

XCB_XKB_GROUP_1

XCB_XKB_GROUP_2

XCB_XKB_GROUP_3

XCB_XKB_GROUP_4

XCB_XKB_ID_ALL_XI_CLASS

XCB_XKB_ID_ALL_XI_ID

XCB_XKB_ID_DFLT_XI_CLASS

XCB_XKB_ID_DFLT_XI_ID

XCB_XKB_ID_USE_CORE_KBD

XCB_XKB_ID_USE_CORE_PTR

XCB_XKB_ID_XI_NONE

XCB_XKB_IM_FLAG_LED_DRIVES_KB

XCB_XKB_IM_FLAG_NO_AUTOMATIC

XCB_XKB_IM_FLAG_NO_EXPLICIT

XCB_XKB_IM_GROUPS_WHICH_USE_BASE

XCB_XKB_IM_GROUPS_WHICH_USE_COMPAT

XCB_XKB_IM_GROUPS_WHICH_USE_EFFECTIVE

XCB_XKB_IM_GROUPS_WHICH_USE_LATCHED

XCB_XKB_IM_GROUPS_WHICH_USE_LOCKED

XCB_XKB_IM_MODS_WHICH_USE_BASE

XCB_XKB_IM_MODS_WHICH_USE_COMPAT

XCB_XKB_IM_MODS_WHICH_USE_EFFECTIVE

XCB_XKB_IM_MODS_WHICH_USE_LATCHED

XCB_XKB_IM_MODS_WHICH_USE_LOCKED

XCB_XKB_INDICATOR_MAP_NOTIFY

XCB_XKB_INDICATOR_STATE_NOTIFY

XCB_XKB_KEYBOARD

XCB_XKB_LATCH_LOCK_STATE

XCB_XKB_LED_CLASS_ALL_XI_CLASSES

XCB_XKB_LED_CLASS_DFLT_XI_CLASS

XCB_XKB_LED_CLASS_KBD_FEEDBACK_CLASS

XCB_XKB_LED_CLASS_LED_FEEDBACK_CLASS

XCB_XKB_LED_CLASS_RESULT_KBD_FEEDBACK_CLASS

XCB_XKB_LED_CLASS_RESULT_LED_FEEDBACK_CLASS

XCB_XKB_LIST_COMPONENTS

XCB_XKB_LOCK_DEVICE_FLAGS_NO_LOCK

XCB_XKB_LOCK_DEVICE_FLAGS_NO_UNLOCK

XCB_XKB_MAJOR_VERSION

XCB_XKB_MAP_NOTIFY

XCB_XKB_MAP_PART_EXPLICIT_COMPONENTS

XCB_XKB_MAP_PART_KEY_ACTIONS

XCB_XKB_MAP_PART_KEY_BEHAVIORS

XCB_XKB_MAP_PART_KEY_SYMS

XCB_XKB_MAP_PART_KEY_TYPES

XCB_XKB_MAP_PART_MODIFIER_MAP

XCB_XKB_MAP_PART_VIRTUAL_MODS

XCB_XKB_MAP_PART_VIRTUAL_MOD_MAP

XCB_XKB_MINOR_VERSION

XCB_XKB_NAMES_NOTIFY

XCB_XKB_NAME_DETAIL_COMPAT

XCB_XKB_NAME_DETAIL_GEOMETRY

XCB_XKB_NAME_DETAIL_GROUP_NAMES

XCB_XKB_NAME_DETAIL_INDICATOR_NAMES

XCB_XKB_NAME_DETAIL_KEYCODES

XCB_XKB_NAME_DETAIL_KEY_ALIASES

XCB_XKB_NAME_DETAIL_KEY_NAMES

XCB_XKB_NAME_DETAIL_KEY_TYPE_NAMES

XCB_XKB_NAME_DETAIL_KT_LEVEL_NAMES

XCB_XKB_NAME_DETAIL_PHYS_SYMBOLS

XCB_XKB_NAME_DETAIL_RG_NAMES

XCB_XKB_NAME_DETAIL_SYMBOLS

XCB_XKB_NAME_DETAIL_TYPES

XCB_XKB_NAME_DETAIL_VIRTUAL_MOD_NAMES

XCB_XKB_NEW_KEYBOARD_NOTIFY

XCB_XKB_NKN_DETAIL_DEVICE_ID

XCB_XKB_NKN_DETAIL_GEOMETRY

XCB_XKB_NKN_DETAIL_KEYCODES

XCB_XKB_PER_CLIENT_FLAGS

XCB_XKB_PER_CLIENT_FLAG_AUTO_RESET_CONTROLS

XCB_XKB_PER_CLIENT_FLAG_DETECTABLE_AUTO_REPEAT

XCB_XKB_PER_CLIENT_FLAG_GRABS_USE_XKB_STATE

XCB_XKB_PER_CLIENT_FLAG_LOOKUP_STATE_WHEN_GRABBED

XCB_XKB_PER_CLIENT_FLAG_SEND_EVENT_USES_XKB_STATE

XCB_XKB_SA_CLEAR_LOCKS

XCB_XKB_SA_GROUP_ABSOLUTE

XCB_XKB_SA_ISO_LOCK_FLAG_GROUP_ABSOLUTE

XCB_XKB_SA_ISO_LOCK_FLAG_ISO_DFLT_IS_GROUP

XCB_XKB_SA_ISO_LOCK_FLAG_NO_LOCK

XCB_XKB_SA_ISO_LOCK_FLAG_NO_UNLOCK

XCB_XKB_SA_ISO_LOCK_FLAG_USE_MOD_MAP_MODS

XCB_XKB_SA_ISO_LOCK_NO_AFFECT_CTRLS

XCB_XKB_SA_ISO_LOCK_NO_AFFECT_GROUP

XCB_XKB_SA_ISO_LOCK_NO_AFFECT_MODS

XCB_XKB_SA_ISO_LOCK_NO_AFFECT_PTR

XCB_XKB_SA_LATCH_TO_LOCK

XCB_XKB_SA_MOVE_PTR_FLAG_MOVE_ABSOLUTE_X

XCB_XKB_SA_MOVE_PTR_FLAG_MOVE_ABSOLUTE_Y

XCB_XKB_SA_MOVE_PTR_FLAG_NO_ACCELERATION

XCB_XKB_SA_SET_PTR_DFLT_FLAG_AFFECT_DFLT_BUTTON

XCB_XKB_SA_SET_PTR_DFLT_FLAG_DFLT_BTN_ABSOLUTE

XCB_XKB_SA_TYPE_ACTION_MESSAGE

XCB_XKB_SA_TYPE_DEVICE_BTN

XCB_XKB_SA_TYPE_DEVICE_VALUATOR

XCB_XKB_SA_TYPE_ISO_LOCK

XCB_XKB_SA_TYPE_LATCH_GROUP

XCB_XKB_SA_TYPE_LATCH_MODS

XCB_XKB_SA_TYPE_LOCK_CONTROLS

XCB_XKB_SA_TYPE_LOCK_DEVICE_BTN

XCB_XKB_SA_TYPE_LOCK_GROUP

XCB_XKB_SA_TYPE_LOCK_MODS

XCB_XKB_SA_TYPE_LOCK_PTR_BTN

XCB_XKB_SA_TYPE_MOVE_PTR

XCB_XKB_SA_TYPE_NO_ACTION

XCB_XKB_SA_TYPE_PTR_BTN

XCB_XKB_SA_TYPE_REDIRECT_KEY

XCB_XKB_SA_TYPE_SET_CONTROLS

XCB_XKB_SA_TYPE_SET_GROUP

XCB_XKB_SA_TYPE_SET_MODS

XCB_XKB_SA_TYPE_SET_PTR_DFLT

XCB_XKB_SA_TYPE_SWITCH_SCREEN

XCB_XKB_SA_TYPE_TERMINATE

XCB_XKB_SA_USE_MOD_MAP_MODS

XCB_XKB_SA_VAL_WHAT_IGNORE_VAL

XCB_XKB_SA_VAL_WHAT_SET_VAL_ABSOLUTE

XCB_XKB_SA_VAL_WHAT_SET_VAL_CENTER

XCB_XKB_SA_VAL_WHAT_SET_VAL_MAX

XCB_XKB_SA_VAL_WHAT_SET_VAL_MIN

XCB_XKB_SA_VAL_WHAT_SET_VAL_RELATIVE

XCB_XKB_SELECT_EVENTS

XCB_XKB_SET_COMPAT_MAP

XCB_XKB_SET_CONTROLS

XCB_XKB_SET_DEBUGGING_FLAGS

XCB_XKB_SET_DEVICE_INFO

XCB_XKB_SET_INDICATOR_MAP

XCB_XKB_SET_MAP

XCB_XKB_SET_MAP_FLAGS_RECOMPUTE_ACTIONS

XCB_XKB_SET_MAP_FLAGS_RESIZE_TYPES

XCB_XKB_SET_NAMED_INDICATOR

XCB_XKB_SET_NAMES

XCB_XKB_SET_OF_GROUPS_ANY

XCB_XKB_SET_OF_GROUP_GROUP_1

XCB_XKB_SET_OF_GROUP_GROUP_2

XCB_XKB_SET_OF_GROUP_GROUP_3

XCB_XKB_SET_OF_GROUP_GROUP_4

XCB_XKB_STATE_NOTIFY

XCB_XKB_STATE_PART_COMPAT_GRAB_MODS

XCB_XKB_STATE_PART_COMPAT_LOOKUP_MODS

XCB_XKB_STATE_PART_COMPAT_STATE

XCB_XKB_STATE_PART_GRAB_MODS

XCB_XKB_STATE_PART_GROUP_BASE

XCB_XKB_STATE_PART_GROUP_LATCH

XCB_XKB_STATE_PART_GROUP_LOCK

XCB_XKB_STATE_PART_GROUP_STATE

XCB_XKB_STATE_PART_LOOKUP_MODS

XCB_XKB_STATE_PART_MODIFIER_BASE

XCB_XKB_STATE_PART_MODIFIER_LATCH

XCB_XKB_STATE_PART_MODIFIER_LOCK

XCB_XKB_STATE_PART_MODIFIER_STATE

XCB_XKB_STATE_PART_POINTER_BUTTONS

XCB_XKB_SWITCH_SCREEN_FLAG_ABSOLUTE

XCB_XKB_SWITCH_SCREEN_FLAG_APPLICATION

XCB_XKB_SYM_INTERPRET_MATCH_ALL_OF

XCB_XKB_SYM_INTERPRET_MATCH_ANY_OF

XCB_XKB_SYM_INTERPRET_MATCH_ANY_OF_OR_NONE

XCB_XKB_SYM_INTERPRET_MATCH_EXACTLY

XCB_XKB_SYM_INTERPRET_MATCH_NONE_OF

XCB_XKB_SYM_INTERP_MATCH_LEVEL_ONE_ONLY

XCB_XKB_SYM_INTERP_MATCH_OP_MASK

XCB_XKB_USE_EXTENSION

XCB_XKB_V_MODS_HIGH_8

XCB_XKB_V_MODS_HIGH_9

XCB_XKB_V_MODS_HIGH_10

XCB_XKB_V_MODS_HIGH_11

XCB_XKB_V_MODS_HIGH_12

XCB_XKB_V_MODS_HIGH_13

XCB_XKB_V_MODS_HIGH_14

XCB_XKB_V_MODS_HIGH_15

XCB_XKB_V_MODS_LOW_0

XCB_XKB_V_MODS_LOW_1

XCB_XKB_V_MODS_LOW_2

XCB_XKB_V_MODS_LOW_3

XCB_XKB_V_MODS_LOW_4

XCB_XKB_V_MODS_LOW_5

XCB_XKB_V_MODS_LOW_6

XCB_XKB_V_MODS_LOW_7

XCB_XKB_V_MOD_0

XCB_XKB_V_MOD_1

XCB_XKB_V_MOD_2

XCB_XKB_V_MOD_3

XCB_XKB_V_MOD_4

XCB_XKB_V_MOD_5

XCB_XKB_V_MOD_6

XCB_XKB_V_MOD_7

XCB_XKB_V_MOD_8

XCB_XKB_V_MOD_9

XCB_XKB_V_MOD_10

XCB_XKB_V_MOD_11

XCB_XKB_V_MOD_12

XCB_XKB_V_MOD_13

XCB_XKB_V_MOD_14

XCB_XKB_V_MOD_15

XCB_XKB_XI_FEATURE_BUTTON_ACTIONS

XCB_XKB_XI_FEATURE_INDICATOR_MAPS

XCB_XKB_XI_FEATURE_INDICATOR_NAMES

XCB_XKB_XI_FEATURE_INDICATOR_STATE

XCB_XKB_XI_FEATURE_KEYBOARDS

Statics

xcb_xkb_id

Functions

xcb_xkb_action_end⚠

xcb_xkb_action_next⚠

xcb_xkb_behavior_end⚠

xcb_xkb_behavior_next⚠

xcb_xkb_bell⚠

xcb_xkb_bell_checked⚠

xcb_xkb_bell_class_spec_end⚠

xcb_xkb_bell_class_spec_next⚠

xcb_xkb_common_behavior_end⚠

xcb_xkb_common_behavior_next⚠

xcb_xkb_counted_string_16_alignment_pad⚠

xcb_xkb_counted_string_16_alignment_pad_end⚠

xcb_xkb_counted_string_16_alignment_pad_length⚠

xcb_xkb_counted_string_16_end⚠

xcb_xkb_counted_string_16_next⚠

xcb_xkb_counted_string_16_string⚠

xcb_xkb_counted_string_16_string_end⚠

xcb_xkb_counted_string_16_string_length⚠

xcb_xkb_default_behavior_end⚠

xcb_xkb_default_behavior_next⚠

xcb_xkb_device_led_info_end⚠

xcb_xkb_device_led_info_maps⚠

xcb_xkb_device_led_info_maps_iterator⚠

xcb_xkb_device_led_info_maps_length⚠

xcb_xkb_device_led_info_names⚠

xcb_xkb_device_led_info_names_end⚠

xcb_xkb_device_led_info_names_length⚠

xcb_xkb_device_led_info_next⚠

xcb_xkb_device_spec_end⚠

xcb_xkb_device_spec_next⚠

xcb_xkb_get_compat_map⚠

xcb_xkb_get_compat_map_group_rtrn⚠

xcb_xkb_get_compat_map_group_rtrn_iterator⚠

xcb_xkb_get_compat_map_group_rtrn_length⚠

xcb_xkb_get_compat_map_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_get_compat_map_si_rtrn⚠

xcb_xkb_get_compat_map_si_rtrn_iterator⚠

xcb_xkb_get_compat_map_si_rtrn_length⚠

xcb_xkb_get_compat_map_unchecked⚠

xcb_xkb_get_controls⚠

xcb_xkb_get_controls_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_get_controls_unchecked⚠

xcb_xkb_get_device_info⚠

xcb_xkb_get_device_info_btn_actions⚠

xcb_xkb_get_device_info_btn_actions_iterator⚠

xcb_xkb_get_device_info_btn_actions_length⚠

xcb_xkb_get_device_info_leds_iterator⚠

xcb_xkb_get_device_info_leds_length⚠

xcb_xkb_get_device_info_name⚠

xcb_xkb_get_device_info_name_end⚠

xcb_xkb_get_device_info_name_length⚠

xcb_xkb_get_device_info_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_get_device_info_unchecked⚠

xcb_xkb_get_indicator_map⚠

xcb_xkb_get_indicator_map_maps⚠

xcb_xkb_get_indicator_map_maps_iterator⚠

xcb_xkb_get_indicator_map_maps_length⚠

xcb_xkb_get_indicator_map_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_get_indicator_map_unchecked⚠

xcb_xkb_get_indicator_state⚠

xcb_xkb_get_indicator_state_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_get_indicator_state_unchecked⚠

xcb_xkb_get_kbd_by_name⚠

xcb_xkb_get_kbd_by_name_replies⚠

xcb_xkb_get_kbd_by_name_replies_compat_map_group_rtrn⚠

xcb_xkb_get_kbd_by_name_replies_compat_map_group_rtrn_iterator⚠

xcb_xkb_get_kbd_by_name_replies_compat_map_group_rtrn_length⚠

xcb_xkb_get_kbd_by_name_replies_compat_map_si_rtrn⚠

xcb_xkb_get_kbd_by_name_replies_compat_map_si_rtrn_iterator⚠

xcb_xkb_get_kbd_by_name_replies_compat_map_si_rtrn_length⚠

xcb_xkb_get_kbd_by_name_replies_geometry_label_font⚠

xcb_xkb_get_kbd_by_name_replies_indicator_maps_maps⚠

xcb_xkb_get_kbd_by_name_replies_indicator_maps_maps_iterator⚠

xcb_xkb_get_kbd_by_name_replies_indicator_maps_maps_length⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_groups⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_groups_end⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_groups_length⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_indicator_names⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_indicator_names_end⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_indicator_names_length⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_key_aliases⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_key_aliases_iterator⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_key_aliases_length⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_key_names⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_key_names_iterator⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_key_names_length⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_kt_level_names⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_kt_level_names_end⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_kt_level_names_length⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_n_levels_per_type⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_n_levels_per_type_end⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_n_levels_per_type_length⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_radio_group_names⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_radio_group_names_end⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_radio_group_names_length⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_type_names⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_type_names_end⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_type_names_length⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_virtual_mod_names⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_virtual_mod_names_end⚠

xcb_xkb_get_kbd_by_name_replies_key_names_value_list_virtual_mod_names_length⚠

xcb_xkb_get_kbd_by_name_replies_types_map⚠

xcb_xkb_get_kbd_by_name_replies_types_map_acts_rtrn_acts⚠

xcb_xkb_get_kbd_by_name_replies_types_map_acts_rtrn_acts_iterator⚠

xcb_xkb_get_kbd_by_name_replies_types_map_acts_rtrn_acts_length⚠

xcb_xkb_get_kbd_by_name_replies_types_map_acts_rtrn_count⚠

xcb_xkb_get_kbd_by_name_replies_types_map_acts_rtrn_count_end⚠

xcb_xkb_get_kbd_by_name_replies_types_map_acts_rtrn_count_length⚠

xcb_xkb_get_kbd_by_name_replies_types_map_behaviors_rtrn⚠

xcb_xkb_get_kbd_by_name_replies_types_map_behaviors_rtrn_iterator⚠

xcb_xkb_get_kbd_by_name_replies_types_map_behaviors_rtrn_length⚠

xcb_xkb_get_kbd_by_name_replies_types_map_explicit_rtrn⚠

xcb_xkb_get_kbd_by_name_replies_types_map_explicit_rtrn_iterator⚠

xcb_xkb_get_kbd_by_name_replies_types_map_explicit_rtrn_length⚠

xcb_xkb_get_kbd_by_name_replies_types_map_modmap_rtrn⚠

xcb_xkb_get_kbd_by_name_replies_types_map_modmap_rtrn_iterator⚠

xcb_xkb_get_kbd_by_name_replies_types_map_modmap_rtrn_length⚠

xcb_xkb_get_kbd_by_name_replies_types_map_syms_rtrn_iterator⚠

xcb_xkb_get_kbd_by_name_replies_types_map_syms_rtrn_length⚠

xcb_xkb_get_kbd_by_name_replies_types_map_types_rtrn_iterator⚠

xcb_xkb_get_kbd_by_name_replies_types_map_types_rtrn_length⚠

xcb_xkb_get_kbd_by_name_replies_types_map_vmodmap_rtrn⚠

xcb_xkb_get_kbd_by_name_replies_types_map_vmodmap_rtrn_iterator⚠

xcb_xkb_get_kbd_by_name_replies_types_map_vmodmap_rtrn_length⚠

xcb_xkb_get_kbd_by_name_replies_types_map_vmods_rtrn⚠

xcb_xkb_get_kbd_by_name_replies_types_map_vmods_rtrn_end⚠

xcb_xkb_get_kbd_by_name_replies_types_map_vmods_rtrn_length⚠

xcb_xkb_get_kbd_by_name_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_get_kbd_by_name_unchecked⚠

xcb_xkb_get_map⚠

xcb_xkb_get_map_map⚠

xcb_xkb_get_map_map_acts_rtrn_acts⚠

xcb_xkb_get_map_map_acts_rtrn_acts_iterator⚠

xcb_xkb_get_map_map_acts_rtrn_acts_length⚠

xcb_xkb_get_map_map_acts_rtrn_count⚠

xcb_xkb_get_map_map_acts_rtrn_count_end⚠

xcb_xkb_get_map_map_acts_rtrn_count_length⚠

xcb_xkb_get_map_map_behaviors_rtrn⚠

xcb_xkb_get_map_map_behaviors_rtrn_iterator⚠

xcb_xkb_get_map_map_behaviors_rtrn_length⚠

xcb_xkb_get_map_map_explicit_rtrn⚠

xcb_xkb_get_map_map_explicit_rtrn_iterator⚠

xcb_xkb_get_map_map_explicit_rtrn_length⚠

xcb_xkb_get_map_map_modmap_rtrn⚠

xcb_xkb_get_map_map_modmap_rtrn_iterator⚠

xcb_xkb_get_map_map_modmap_rtrn_length⚠

xcb_xkb_get_map_map_syms_rtrn_iterator⚠

xcb_xkb_get_map_map_syms_rtrn_length⚠

xcb_xkb_get_map_map_types_rtrn_iterator⚠

xcb_xkb_get_map_map_types_rtrn_length⚠

xcb_xkb_get_map_map_vmodmap_rtrn⚠

xcb_xkb_get_map_map_vmodmap_rtrn_iterator⚠

xcb_xkb_get_map_map_vmodmap_rtrn_length⚠

xcb_xkb_get_map_map_vmods_rtrn⚠

xcb_xkb_get_map_map_vmods_rtrn_end⚠

xcb_xkb_get_map_map_vmods_rtrn_length⚠

xcb_xkb_get_map_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_get_map_unchecked⚠

xcb_xkb_get_named_indicator⚠

xcb_xkb_get_named_indicator_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_get_named_indicator_unchecked⚠

xcb_xkb_get_names⚠

xcb_xkb_get_names_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_get_names_unchecked⚠

xcb_xkb_get_names_value_list⚠

xcb_xkb_get_names_value_list_alignment_pad⚠

xcb_xkb_get_names_value_list_alignment_pad_end⚠

xcb_xkb_get_names_value_list_alignment_pad_length⚠

xcb_xkb_get_names_value_list_groups⚠

xcb_xkb_get_names_value_list_groups_end⚠

xcb_xkb_get_names_value_list_groups_length⚠

xcb_xkb_get_names_value_list_indicator_names⚠

xcb_xkb_get_names_value_list_indicator_names_end⚠

xcb_xkb_get_names_value_list_indicator_names_length⚠

xcb_xkb_get_names_value_list_key_aliases⚠

xcb_xkb_get_names_value_list_key_aliases_iterator⚠

xcb_xkb_get_names_value_list_key_aliases_length⚠

xcb_xkb_get_names_value_list_key_names⚠

xcb_xkb_get_names_value_list_key_names_iterator⚠

xcb_xkb_get_names_value_list_key_names_length⚠

xcb_xkb_get_names_value_list_kt_level_names⚠

xcb_xkb_get_names_value_list_kt_level_names_end⚠

xcb_xkb_get_names_value_list_kt_level_names_length⚠

xcb_xkb_get_names_value_list_n_levels_per_type⚠

xcb_xkb_get_names_value_list_n_levels_per_type_end⚠

xcb_xkb_get_names_value_list_n_levels_per_type_length⚠

xcb_xkb_get_names_value_list_radio_group_names⚠

xcb_xkb_get_names_value_list_radio_group_names_end⚠

xcb_xkb_get_names_value_list_radio_group_names_length⚠

xcb_xkb_get_names_value_list_type_names⚠

xcb_xkb_get_names_value_list_type_names_end⚠

xcb_xkb_get_names_value_list_type_names_length⚠

xcb_xkb_get_names_value_list_virtual_mod_names⚠

xcb_xkb_get_names_value_list_virtual_mod_names_end⚠

xcb_xkb_get_names_value_list_virtual_mod_names_length⚠

xcb_xkb_get_state⚠

xcb_xkb_get_state_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_get_state_unchecked⚠

xcb_xkb_id_spec_end⚠

xcb_xkb_id_spec_next⚠

xcb_xkb_indicator_map_end⚠

xcb_xkb_indicator_map_next⚠

xcb_xkb_key_alias_end⚠

xcb_xkb_key_alias_next⚠

xcb_xkb_key_end⚠

xcb_xkb_key_mod_map_end⚠

xcb_xkb_key_mod_map_next⚠

xcb_xkb_key_name_end⚠

xcb_xkb_key_name_next⚠

xcb_xkb_key_next⚠

xcb_xkb_key_sym_map_end⚠

xcb_xkb_key_sym_map_next⚠

xcb_xkb_key_sym_map_syms⚠

xcb_xkb_key_sym_map_syms_end⚠

xcb_xkb_key_sym_map_syms_length⚠

xcb_xkb_key_type_end⚠

xcb_xkb_key_type_map⚠

xcb_xkb_key_type_map_iterator⚠

xcb_xkb_key_type_map_length⚠

xcb_xkb_key_type_next⚠

xcb_xkb_key_type_preserve⚠

xcb_xkb_key_type_preserve_iterator⚠

xcb_xkb_key_type_preserve_length⚠

xcb_xkb_key_v_mod_map_end⚠

xcb_xkb_key_v_mod_map_next⚠

xcb_xkb_kt_map_entry_end⚠

xcb_xkb_kt_map_entry_next⚠

xcb_xkb_kt_set_map_entry_end⚠

xcb_xkb_kt_set_map_entry_next⚠

xcb_xkb_latch_lock_state⚠

xcb_xkb_latch_lock_state_checked⚠

xcb_xkb_led_class_spec_end⚠

xcb_xkb_led_class_spec_next⚠

xcb_xkb_list_components⚠

xcb_xkb_list_components_compat_maps_iterator⚠

xcb_xkb_list_components_compat_maps_length⚠

xcb_xkb_list_components_geometries_iterator⚠

xcb_xkb_list_components_geometries_length⚠

xcb_xkb_list_components_keycodes_iterator⚠

xcb_xkb_list_components_keycodes_length⚠

xcb_xkb_list_components_keymaps_iterator⚠

xcb_xkb_list_components_keymaps_length⚠

xcb_xkb_list_components_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_list_components_symbols_iterator⚠

xcb_xkb_list_components_symbols_length⚠

xcb_xkb_list_components_types_iterator⚠

xcb_xkb_list_components_types_length⚠

xcb_xkb_list_components_unchecked⚠

xcb_xkb_listing_end⚠

xcb_xkb_listing_next⚠

xcb_xkb_listing_string⚠

xcb_xkb_listing_string_end⚠

xcb_xkb_listing_string_length⚠

xcb_xkb_lock_behavior_end⚠

xcb_xkb_lock_behavior_next⚠

xcb_xkb_mod_def_end⚠

xcb_xkb_mod_def_next⚠

xcb_xkb_outline_end⚠

xcb_xkb_outline_next⚠

xcb_xkb_outline_points⚠

xcb_xkb_outline_points_iterator⚠

xcb_xkb_outline_points_length⚠

xcb_xkb_overlay_behavior_end⚠

xcb_xkb_overlay_behavior_next⚠

xcb_xkb_overlay_end⚠

xcb_xkb_overlay_key_end⚠

xcb_xkb_overlay_key_next⚠

xcb_xkb_overlay_next⚠

xcb_xkb_overlay_row_end⚠

xcb_xkb_overlay_row_keys⚠

xcb_xkb_overlay_row_keys_iterator⚠

xcb_xkb_overlay_row_keys_length⚠

xcb_xkb_overlay_row_next⚠

xcb_xkb_overlay_rows_iterator⚠

xcb_xkb_overlay_rows_length⚠

xcb_xkb_per_client_flags⚠

xcb_xkb_per_client_flags_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_per_client_flags_unchecked⚠

xcb_xkb_permament_lock_behavior_end⚠

xcb_xkb_permament_lock_behavior_next⚠

xcb_xkb_permament_overlay_behavior_end⚠

xcb_xkb_permament_overlay_behavior_next⚠

xcb_xkb_permament_radio_group_behavior_end⚠

xcb_xkb_permament_radio_group_behavior_next⚠

xcb_xkb_radio_group_behavior_end⚠

xcb_xkb_radio_group_behavior_next⚠

xcb_xkb_row_end⚠

xcb_xkb_row_keys⚠

xcb_xkb_row_keys_iterator⚠

xcb_xkb_row_keys_length⚠

xcb_xkb_row_next⚠

xcb_xkb_sa_action_message_end⚠

xcb_xkb_sa_action_message_next⚠

xcb_xkb_sa_device_btn_end⚠

xcb_xkb_sa_device_btn_next⚠

xcb_xkb_sa_device_valuator_end⚠

xcb_xkb_sa_device_valuator_next⚠

xcb_xkb_sa_iso_lock_end⚠

xcb_xkb_sa_iso_lock_next⚠

xcb_xkb_sa_latch_group_end⚠

xcb_xkb_sa_latch_group_next⚠

xcb_xkb_sa_latch_mods_end⚠

xcb_xkb_sa_latch_mods_next⚠

xcb_xkb_sa_lock_controls_end⚠

xcb_xkb_sa_lock_controls_next⚠

xcb_xkb_sa_lock_device_btn_end⚠

xcb_xkb_sa_lock_device_btn_next⚠

xcb_xkb_sa_lock_group_end⚠

xcb_xkb_sa_lock_group_next⚠

xcb_xkb_sa_lock_mods_end⚠

xcb_xkb_sa_lock_mods_next⚠

xcb_xkb_sa_lock_ptr_btn_end⚠

xcb_xkb_sa_lock_ptr_btn_next⚠

xcb_xkb_sa_move_ptr_end⚠

xcb_xkb_sa_move_ptr_next⚠

xcb_xkb_sa_no_action_end⚠

xcb_xkb_sa_no_action_next⚠

xcb_xkb_sa_ptr_btn_end⚠

xcb_xkb_sa_ptr_btn_next⚠

xcb_xkb_sa_redirect_key_end⚠

xcb_xkb_sa_redirect_key_next⚠

xcb_xkb_sa_set_controls_end⚠

xcb_xkb_sa_set_controls_next⚠

xcb_xkb_sa_set_group_end⚠

xcb_xkb_sa_set_group_next⚠

xcb_xkb_sa_set_mods_end⚠

xcb_xkb_sa_set_mods_next⚠

xcb_xkb_sa_set_ptr_dflt_end⚠

xcb_xkb_sa_set_ptr_dflt_next⚠

xcb_xkb_sa_switch_screen_end⚠

xcb_xkb_sa_switch_screen_next⚠

xcb_xkb_sa_terminate_end⚠

xcb_xkb_sa_terminate_next⚠

xcb_xkb_select_events⚠

xcb_xkb_select_events_checked⚠

xcb_xkb_set_behavior_end⚠

xcb_xkb_set_behavior_next⚠

xcb_xkb_set_compat_map⚠

xcb_xkb_set_compat_map_checked⚠

xcb_xkb_set_controls⚠

xcb_xkb_set_controls_checked⚠

xcb_xkb_set_debugging_flags⚠

xcb_xkb_set_debugging_flags_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_set_debugging_flags_unchecked⚠

xcb_xkb_set_device_info⚠

xcb_xkb_set_device_info_checked⚠

xcb_xkb_set_explicit_end⚠

xcb_xkb_set_explicit_next⚠

xcb_xkb_set_indicator_map⚠

xcb_xkb_set_indicator_map_checked⚠

xcb_xkb_set_key_type_end⚠

xcb_xkb_set_key_type_entries⚠

xcb_xkb_set_key_type_entries_iterator⚠

xcb_xkb_set_key_type_entries_length⚠

xcb_xkb_set_key_type_next⚠

xcb_xkb_set_key_type_preserve_entries⚠

xcb_xkb_set_key_type_preserve_entries_iterator⚠

xcb_xkb_set_key_type_preserve_entries_length⚠

xcb_xkb_set_map⚠

xcb_xkb_set_map_checked⚠

xcb_xkb_set_map_values_actions⚠

xcb_xkb_set_map_values_actions_count⚠

xcb_xkb_set_map_values_actions_count_end⚠

xcb_xkb_set_map_values_actions_count_length⚠

xcb_xkb_set_map_values_actions_iterator⚠

xcb_xkb_set_map_values_actions_length⚠

xcb_xkb_set_map_values_behaviors⚠

xcb_xkb_set_map_values_behaviors_iterator⚠

xcb_xkb_set_map_values_behaviors_length⚠

xcb_xkb_set_map_values_explicit⚠

xcb_xkb_set_map_values_explicit_iterator⚠

xcb_xkb_set_map_values_explicit_length⚠

xcb_xkb_set_map_values_modmap⚠

xcb_xkb_set_map_values_modmap_iterator⚠

xcb_xkb_set_map_values_modmap_length⚠

xcb_xkb_set_map_values_syms_iterator⚠

xcb_xkb_set_map_values_syms_length⚠

xcb_xkb_set_map_values_types_iterator⚠

xcb_xkb_set_map_values_types_length⚠

xcb_xkb_set_map_values_vmodmap⚠

xcb_xkb_set_map_values_vmodmap_iterator⚠

xcb_xkb_set_map_values_vmodmap_length⚠

xcb_xkb_set_map_values_vmods⚠

xcb_xkb_set_map_values_vmods_end⚠

xcb_xkb_set_map_values_vmods_length⚠

xcb_xkb_set_named_indicator⚠

xcb_xkb_set_named_indicator_checked⚠

xcb_xkb_set_names⚠

xcb_xkb_set_names_checked⚠

xcb_xkb_set_names_values_groups⚠

xcb_xkb_set_names_values_groups_end⚠

xcb_xkb_set_names_values_groups_length⚠

xcb_xkb_set_names_values_indicator_names⚠

xcb_xkb_set_names_values_indicator_names_end⚠

xcb_xkb_set_names_values_indicator_names_length⚠

xcb_xkb_set_names_values_key_aliases⚠

xcb_xkb_set_names_values_key_aliases_iterator⚠

xcb_xkb_set_names_values_key_aliases_length⚠

xcb_xkb_set_names_values_key_names⚠

xcb_xkb_set_names_values_key_names_iterator⚠

xcb_xkb_set_names_values_key_names_length⚠

xcb_xkb_set_names_values_kt_level_names⚠

xcb_xkb_set_names_values_kt_level_names_end⚠

xcb_xkb_set_names_values_kt_level_names_length⚠

xcb_xkb_set_names_values_n_levels_per_type⚠

xcb_xkb_set_names_values_n_levels_per_type_end⚠

xcb_xkb_set_names_values_n_levels_per_type_length⚠

xcb_xkb_set_names_values_radio_group_names⚠

xcb_xkb_set_names_values_radio_group_names_end⚠

xcb_xkb_set_names_values_radio_group_names_length⚠

xcb_xkb_set_names_values_type_names⚠

xcb_xkb_set_names_values_type_names_end⚠

xcb_xkb_set_names_values_type_names_length⚠

xcb_xkb_set_names_values_virtual_mod_names⚠

xcb_xkb_set_names_values_virtual_mod_names_end⚠

xcb_xkb_set_names_values_virtual_mod_names_length⚠

xcb_xkb_shape_end⚠

xcb_xkb_shape_next⚠

xcb_xkb_shape_outlines_iterator⚠

xcb_xkb_shape_outlines_length⚠

xcb_xkb_si_action_end⚠

xcb_xkb_si_action_next⚠

xcb_xkb_string8_end⚠

xcb_xkb_string8_next⚠

xcb_xkb_sym_interpret_end⚠

xcb_xkb_sym_interpret_next⚠

xcb_xkb_use_extension⚠

xcb_xkb_use_extension_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xkb_use_extension_unchecked⚠

Type Definitions

xcb_xkb_action_message_flag_t

xcb_xkb_ax_option_t

xcb_xkb_axn_detail_t

xcb_xkb_behavior_type_t

xcb_xkb_bell_class_result_t

xcb_xkb_bell_class_spec_t

xcb_xkb_bell_class_t

xcb_xkb_bool_ctrl_t

xcb_xkb_bool_ctrls_high_t

xcb_xkb_bool_ctrls_low_t

xcb_xkb_cm_detail_t

xcb_xkb_const_t

xcb_xkb_control_t

xcb_xkb_device_spec_t

xcb_xkb_doodad_type_t

xcb_xkb_error_t

xcb_xkb_event_type_t

xcb_xkb_explicit_t

xcb_xkb_gbn_detail_t

xcb_xkb_group_t

xcb_xkb_groups_t

xcb_xkb_groups_wrap_t

xcb_xkb_id_spec_t

xcb_xkb_id_t

xcb_xkb_im_flag_t

xcb_xkb_im_groups_which_t

xcb_xkb_im_mods_which_t

xcb_xkb_led_class_result_t

xcb_xkb_led_class_spec_t

xcb_xkb_led_class_t

xcb_xkb_lock_device_flags_t

xcb_xkb_map_part_t

xcb_xkb_name_detail_t

xcb_xkb_nkn_detail_t

xcb_xkb_per_client_flag_t

xcb_xkb_sa_iso_lock_flag_t

xcb_xkb_sa_iso_lock_no_affect_t

xcb_xkb_sa_move_ptr_flag_t

xcb_xkb_sa_set_ptr_dflt_flag_t

xcb_xkb_sa_t

xcb_xkb_sa_type_t

xcb_xkb_sa_val_what_t

xcb_xkb_set_map_flags_t

xcb_xkb_set_of_group_t

xcb_xkb_set_of_groups_t

xcb_xkb_state_part_t

xcb_xkb_string8_t

xcb_xkb_switch_screen_flag_t

xcb_xkb_sym_interp_match_t

xcb_xkb_sym_interpret_match_t

xcb_xkb_v_mod_t

xcb_xkb_v_mods_high_t

xcb_xkb_v_mods_low_t

xcb_xkb_xi_feature_t