Function xcb::ffi::xproto::xcb_circulate_window_checked

pub unsafe extern "C" fn xcb_circulate_window_checked(
    c: *mut xcb_connection_t, 
    direction: u8, 
    window: xcb_window_t
) -> xcb_void_cookie_t

Change window stacking order

If direction is XCB_CIRCULATE_RAISE_LOWEST, the lowest mapped child (if any) will be raised to the top of the stack.

If direction is XCB_CIRCULATE_LOWER_HIGHEST, the highest mapped child will be lowered to the bottom of the stack.