Function xcb::ffi::xproto::xcb_create_glyph_cursor

pub unsafe extern "C" fn xcb_create_glyph_cursor(
    c: *mut xcb_connection_t, 
    cid: xcb_cursor_t, 
    source_font: xcb_font_t, 
    mask_font: xcb_font_t, 
    source_char: u16, 
    mask_char: u16, 
    fore_red: u16, 
    fore_green: u16, 
    fore_blue: u16, 
    back_red: u16, 
    back_green: u16, 
    back_blue: u16
) -> xcb_void_cookie_t

create cursor

Creates a cursor from a font glyph. X provides a set of standard cursor shapes in a special font named cursor. Applications are encouraged to use this interface for their cursors because the font can be customized for the individual display type.

All pixels which are set to 1 in the source will use the foreground color (as specified by fore_red, fore_green and fore_blue). All pixels set to 0 will use the background color (as specified by back_red, back_green and back_blue).