Function xcb::ffi::xproto::xcb_create_window

pub unsafe extern "C" fn xcb_create_window(
    c: *mut xcb_connection_t, 
    depth: u8, 
    wid: xcb_window_t, 
    parent: xcb_window_t, 
    x: i16, 
    y: i16, 
    width: u16, 
    height: u16, 
    border_width: u16, 
    class: u16, 
    visual: xcb_visualid_t, 
    value_mask: u32, 
    value_list: *const u32
) -> xcb_void_cookie_t

Creates a window

Creates an unmapped window as child of the specified parent window. A CreateNotify event will be generated. The new window is placed on top in the stacking order with respect to siblings.

The coordinate system has the X axis horizontal and the Y axis vertical with the origin [0, 0] at the upper-left corner. Coordinates are integral, in terms of pixels, and coincide with pixel centers. Each window and pixmap has its own coordinate system. For a window, the origin is inside the border at the inside, upper-left corner.

The created window is not yet displayed (mapped), call xcb_map_window to display it.

The created window will initially use the same cursor as its parent.