Function xcb::ffi::xproto::xcb_warp_pointer

pub unsafe extern "C" fn xcb_warp_pointer(
    c: *mut xcb_connection_t, 
    src_window: xcb_window_t, 
    dst_window: xcb_window_t, 
    src_x: i16, 
    src_y: i16, 
    src_width: u16, 
    src_height: u16, 
    dst_x: i16, 
    dst_y: i16
) -> xcb_void_cookie_t

move mouse pointer

Moves the mouse pointer to the specified position.

If src_window is not XCB_NONE (TODO), the move will only take place if the pointer is inside src_window and within the rectangle specified by (src_x, src_y, src_width, src_height). The rectangle coordinates are relative to src_window.

If dst_window is not XCB_NONE (TODO), the pointer will be moved to the offsets (dst_x, dst_y) relative to dst_window. If dst_window is XCB_NONE (TODO), the pointer will be moved by the offsets (dst_x, dst_y) relative to the current position of the pointer.