Struct xcb::ffi::xproto::xcb_configure_notify_event_t

#[repr(C)]

pub struct xcb_configure_notify_event_t {
    pub response_type: u8,
    pub pad0: u8,
    pub sequence: u16,
    pub event: xcb_window_t,
    pub window: xcb_window_t,
    pub above_sibling: xcb_window_t,
    pub x: i16,
    pub y: i16,
    pub width: u16,
    pub height: u16,
    pub border_width: u16,
    pub override_redirect: u8,
    pub pad1: u8,
}

NOT YET DOCUMENTED

Fields
response_type: u8 pad0: u8 sequence: u16 event: xcb_window_t

The reconfigured window or its parent, depending on whether StructureNotify or SubstructureNotify was selected.

window: xcb_window_t

The window whose size, position, border, and/or stacking order was changed.

above_sibling: xcb_window_t

If XCB_NONE, the window is on the bottom of the stack with respect to sibling windows. However, if set to a sibling window, the window is placed on top of this sibling window.

x: i16

The X coordinate of the upper-left outside corner of window, relative to the parent window's origin.

y: i16

The Y coordinate of the upper-left outside corner of window, relative to the parent window's origin.

width: u16

The inside width of window, not including the border.

height: u16

The inside height of window, not including the border.

border_width: u16

The border width of window.

override_redirect: u8

Window managers should ignore this window if override_redirect is 1.

pad1: u8