Struct xcb::ffi::xproto::xcb_create_window_request_t

#[repr(C)]

pub struct xcb_create_window_request_t {
    pub major_opcode: u8,
    pub depth: u8,
    pub length: u16,
    pub wid: xcb_window_t,
    pub parent: xcb_window_t,
    pub x: i16,
    pub y: i16,
    pub width: u16,
    pub height: u16,
    pub border_width: u16,
    pub class: u16,
    pub visual: xcb_visualid_t,
    pub value_mask: u32,
}

Creates a window

Creates an unmapped window as child of the specified parent window. A CreateNotify event will be generated. The new window is placed on top in the stacking order with respect to siblings.

The coordinate system has the X axis horizontal and the Y axis vertical with the origin [0, 0] at the upper-left corner. Coordinates are integral, in terms of pixels, and coincide with pixel centers. Each window and pixmap has its own coordinate system. For a window, the origin is inside the border at the inside, upper-left corner.

The created window is not yet displayed (mapped), call xcb_map_window to display it.

The created window will initially use the same cursor as its parent.

Fields
major_opcode: u8 depth: u8

Specifies the new window's depth (TODO: what unit?).

The special value XCB_COPY_FROM_PARENT means the depth is taken from the parent window.

length: u16 wid: xcb_window_t

The ID with which you will refer to the new window, created by xcb_generate_id.

parent: xcb_window_t

The parent window of the new window.

x: i16

The X coordinate of the new window.

y: i16

The Y coordinate of the new window.

width: u16

The width of the new window.

height: u16

The height of the new window.

border_width: u16

TODO:

Must be zero if the class is InputOnly or a xcb_match_error_t occurs.

class: u16

visual: xcb_visualid_t

Specifies the id for the new window's visual.

The special value XCB_COPY_FROM_PARENT means the visual is taken from the parent window.

value_mask: u32