Struct xcb::ffi::xproto::xcb_get_property_request_t

pub struct xcb_get_property_request_t {
    pub major_opcode: u8,
    pub delete: u8,
    pub length: u16,
    pub window: xcb_window_t,
    pub property: xcb_atom_t,
    pub type_: xcb_atom_t,
    pub long_offset: u32,
    pub long_length: u32,
}

Gets a window property

Gets the specified property from the specified window. Properties are for example the window title (WM_NAME) or its minimum size (WM_NORMAL_HINTS). Protocols such as EWMH also use properties - for example EWMH defines the window title, encoded as UTF-8 string, in the _NET_WM_NAME property.

TODO: talk about type

TODO: talk about delete

TODO: talk about the offset/length thing. what's a valid use case?

Fields

major_opcode

delete

Whether the property should actually be deleted. For deleting a property, the specified type has to match the actual property type.

length

window

The window whose property you want to get.

property

The property you want to get (an atom).

type_

The type of the property you want to get (an atom).

long_offset

Specifies the offset (in 32-bit multiples) in the specified property where the data is to be retrieved.

long_length

Specifies how many 32-bit multiples of data should be retrieved (e.g. if you set long_length to 4, you will receive 16 bytes of data).