Struct xcb::ffi::xproto::xcb_image_text_16_request_t

#[repr(C)]

pub struct xcb_image_text_16_request_t {
    pub major_opcode: u8,
    pub string_len: u8,
    pub length: u16,
    pub drawable: xcb_drawable_t,
    pub gc: xcb_gcontext_t,
    pub x: i16,
    pub y: i16,
}

Draws text

Fills the destination rectangle with the background pixel from gc, then paints the text with the foreground pixel from gc. The upper-left corner of the filled rectangle is at [x, y - font-ascent]. The width is overall-width, the height is font-ascent + font-descent. The overall-width, font-ascent and font-descent are as returned by xcb_query_text_extents (TODO).

Note that using X core fonts is deprecated (but still supported) in favor of client-side rendering using Xft.

Fields
major_opcode: u8 string_len: u8

The length of the string in characters. Note that this parameter limited by 255 due to using 8 bits!

length: u16 drawable: xcb_drawable_t

The drawable (Window or Pixmap) to draw text on.

gc: xcb_gcontext_t

The graphics context to use.

The following graphics context components are used: plane-mask, foreground, background, font, subwindow-mode, clip-x-origin, clip-y-origin, and clip-mask.

x: i16

The x coordinate of the first character, relative to the origin of drawable.

y: i16

The y coordinate of the first character, relative to the origin of drawable.