Struct xcb::ffi::xproto::xcb_list_fonts_with_info_reply_t

pub struct xcb_list_fonts_with_info_reply_t {
    pub response_type: u8,
    pub name_len: u8,
    pub sequence: u16,
    pub length: u32,
    pub min_bounds: xcb_charinfo_t,
    pub pad0: [u8; 4],
    pub max_bounds: xcb_charinfo_t,
    pub pad1: [u8; 4],
    pub min_char_or_byte2: u16,
    pub max_char_or_byte2: u16,
    pub default_char: u16,
    pub properties_len: u16,
    pub draw_direction: u8,
    pub min_byte1: u8,
    pub max_byte1: u8,
    pub all_chars_exist: u8,
    pub font_ascent: i16,
    pub font_descent: i16,
    pub replies_hint: u32,
}

Fields
response_type: u8 name_len: u8

The number of matched font names.

sequence: u16 length: u32 min_bounds: xcb_charinfo_t

minimum bounds over all existing char

pad0: [u8; 4] max_bounds: xcb_charinfo_t

maximum bounds over all existing char

pad1: [u8; 4] min_char_or_byte2: u16

first character

max_char_or_byte2: u16

last character

default_char: u16

char to print for undefined character

properties_len: u16

how many properties there are

draw_direction: u8

min_byte1: u8 max_byte1: u8 all_chars_exist: u8

flag if all characters have nonzero size

font_ascent: i16

baseline to top edge of raster

font_descent: i16

baseline to bottom edge of raster

replies_hint: u32

An indication of how many more fonts will be returned. This is only a hint and may be larger or smaller than the number of fonts actually returned. A zero value does not guarantee that no more fonts will be returned.