Struct xcb::ffi::xproto::xcb_map_window_request_t

pub struct xcb_map_window_request_t {
    pub major_opcode: u8,
    pub pad0: u8,
    pub length: u16,
    pub window: xcb_window_t,
}

Makes a window visible

Maps the specified window. This means making the window visible (as long as its parent is visible).

This MapWindow request will be translated to a MapRequest request if a window manager is running. The window manager then decides to either map the window or not. Set the override-redirect window attribute to true if you want to bypass this mechanism.

If the window manager decides to map the window (or if no window manager is running), a MapNotify event is generated.

If the window becomes viewable and no earlier contents for it are remembered, the X server tiles the window with its background. If the window's background is undefined, the existing screen contents are not altered, and the X server generates zero or more Expose events.

If the window type is InputOutput, an Expose event will be generated when the window becomes visible. The normal response to an Expose event should be to repaint the window.

Fields

major_opcode

pad0

length

window

The window to make visible.