Struct xcb::ffi::xproto::xcb_poly_fill_rectangle_request_t

#[repr(C)]

pub struct xcb_poly_fill_rectangle_request_t {
    pub major_opcode: u8,
    pub pad0: u8,
    pub length: u16,
    pub drawable: xcb_drawable_t,
    pub gc: xcb_gcontext_t,
}

Fills rectangles

Fills the specified rectangle(s) in the order listed in the array. For any given rectangle, each pixel is not drawn more than once. If rectangles intersect, the intersecting pixels are drawn multiple times.

Fields
major_opcode: u8 pad0: u8 length: u16 drawable: xcb_drawable_t

The drawable (Window or Pixmap) to draw on.

gc: xcb_gcontext_t

The graphics context to use.

The following graphics context components are used: function, plane-mask, fill-style, subwindow-mode, clip-x-origin, clip-y-origin, and clip-mask.

The following graphics context mode-dependent components are used: foreground, background, tile, stipple, tile-stipple-x-origin, and tile-stipple-y-origin.