Struct xcb::ffi::xproto::xcb_query_pointer_reply_t

pub struct xcb_query_pointer_reply_t {
    pub response_type: u8,
    pub same_screen: u8,
    pub sequence: u16,
    pub length: u32,
    pub root: xcb_window_t,
    pub child: xcb_window_t,
    pub root_x: i16,
    pub root_y: i16,
    pub win_x: i16,
    pub win_y: i16,
    pub mask: u16,
    pub pad0: [u8; 2],
}

Fields
response_type: u8 same_screen: u8

If same_screen is False, then the pointer is not on the same screen as the argument window, child is None, and win_x and win_y are zero. If same_screen is True, then win_x and win_y are the pointer coordinates relative to the argument window's origin, and child is the child containing the pointer, if any.

sequence: u16 length: u32 root: xcb_window_t

The root window the pointer is logically on.

child: xcb_window_t

The child window containing the pointer, if any, if same_screen is true. If same_screen is false, XCB_NONE is returned.

root_x: i16

The pointer X position, relative to root.

root_y: i16

The pointer Y position, relative to root.

win_x: i16

The pointer X coordinate, relative to child, if same_screen is true. Zero otherwise.

win_y: i16

The pointer Y coordinate, relative to child, if same_screen is true. Zero otherwise.

mask: u16

The current logical state of the modifier keys and the buttons. Note that the logical state of a device (as seen by means of the protocol) may lag the physical state if device event processing is frozen.

pad0: [u8; 2]