Struct xcb::ffi::xproto::xcb_set_selection_owner_request_t

#[repr(C)]

pub struct xcb_set_selection_owner_request_t {
    pub major_opcode: u8,
    pub pad0: u8,
    pub length: u16,
    pub owner: xcb_window_t,
    pub selection: xcb_atom_t,
    pub time: xcb_timestamp_t,
}

Sets the owner of a selection

Makes window the owner of the selection selection and updates the last-change time of the specified selection.

TODO: briefly explain what a selection is.

Fields
major_opcode: u8 pad0: u8 length: u16 owner: xcb_window_t

The new owner of the selection.

The special value XCB_NONE means that the selection will have no owner.

selection: xcb_atom_t

The selection.

time: xcb_timestamp_t

Timestamp to avoid race conditions when running X over the network.

The selection will not be changed if time is earlier than the current last-change time of the selection or is later than the current X server time. Otherwise, the last-change time is set to the specified time.

The special value XCB_CURRENT_TIME will be replaced with the current server time.