xcb::ffi::xproto

Type Alias xcb_button_release_event_t

pub type xcb_button_release_event_t = xcb_button_press_event_t;

Aliased Type

struct xcb_button_release_event_t {

    pub response_type: u8,
    pub detail: u8,
    pub sequence: u16,
    pub time: u32,
    pub root: u32,
    pub event: u32,
    pub child: u32,
    pub root_x: i16,
    pub root_y: i16,
    pub event_x: i16,
    pub event_y: i16,
    pub state: u16,
    pub same_screen: u8,
    pub pad0: u8,

}

Fields
response_type: u8detail: u8

The keycode (a number representing a physical key on the keyboard) of the key which was pressed.

sequence: u16time: u32

Time when the event was generated (in milliseconds).

root: u32

The root window of child.

event: u32child: u32root_x: i16

The X coordinate of the pointer relative to the root window at the time of the event.

root_y: i16

The Y coordinate of the pointer relative to the root window at the time of the event.

event_x: i16

If same_screen is true, this is the X coordinate relative to the event window’s origin. Otherwise, event_x will be set to zero.

event_y: i16

If same_screen is true, this is the Y coordinate relative to the event window’s origin. Otherwise, event_y will be set to zero.

state: u16

The logical state of the pointer buttons and modifier keys just prior to the event.

same_screen: u8

Whether the event window is on the same screen as the root window.

pad0: u8