Module xcb::ffi::xtest

Structs

xcb_test_compare_cursor_cookie_t

xcb_test_compare_cursor_reply_t

xcb_test_compare_cursor_request_t

xcb_test_fake_input_request_t

xcb_test_get_version_cookie_t

xcb_test_get_version_reply_t

xcb_test_get_version_request_t

xcb_test_grab_control_request_t

Constants

XCB_TEST_COMPARE_CURSOR

XCB_TEST_CURSOR_CURRENT

XCB_TEST_CURSOR_NONE

XCB_TEST_FAKE_INPUT

XCB_TEST_GET_VERSION

XCB_TEST_GRAB_CONTROL

XCB_TEST_MAJOR_VERSION

XCB_TEST_MINOR_VERSION

Statics

xcb_test_id

Functions

xcb_test_compare_cursor

xcb_test_compare_cursor_reply
the returned value must be freed by the caller using libc::free().

xcb_test_compare_cursor_unchecked

xcb_test_fake_input

xcb_test_fake_input_checked

xcb_test_get_version

xcb_test_get_version_reply
the returned value must be freed by the caller using libc::free().

xcb_test_get_version_unchecked

xcb_test_grab_control

xcb_test_grab_control_checked

Type Definitions

xcb_test_cursor_t