Module xcb::ffi::xv

Structs

xcb_xv_adaptor_info_iterator_t

xcb_xv_adaptor_info_t

xcb_xv_attribute_info_iterator_t

xcb_xv_attribute_info_t

xcb_xv_bad_control_error_t

xcb_xv_bad_encoding_error_t

xcb_xv_bad_port_error_t

xcb_xv_encoding_info_iterator_t

xcb_xv_encoding_info_t

xcb_xv_encoding_iterator_t

xcb_xv_format_iterator_t

xcb_xv_format_t

xcb_xv_get_port_attribute_cookie_t

xcb_xv_get_port_attribute_reply_t

xcb_xv_get_port_attribute_request_t

xcb_xv_get_still_request_t

xcb_xv_get_video_request_t

xcb_xv_grab_port_cookie_t

xcb_xv_grab_port_reply_t

xcb_xv_grab_port_request_t

xcb_xv_image_format_info_iterator_t

xcb_xv_image_format_info_t

xcb_xv_image_iterator_t

xcb_xv_image_t

xcb_xv_list_image_formats_cookie_t

xcb_xv_list_image_formats_reply_t

xcb_xv_list_image_formats_request_t

xcb_xv_port_iterator_t

xcb_xv_port_notify_event_t

xcb_xv_put_image_request_t

xcb_xv_put_still_request_t

xcb_xv_put_video_request_t

xcb_xv_query_adaptors_cookie_t

xcb_xv_query_adaptors_reply_t

xcb_xv_query_adaptors_request_t

xcb_xv_query_best_size_cookie_t

xcb_xv_query_best_size_reply_t

xcb_xv_query_best_size_request_t

xcb_xv_query_encodings_cookie_t

xcb_xv_query_encodings_reply_t

xcb_xv_query_encodings_request_t

xcb_xv_query_extension_cookie_t

xcb_xv_query_extension_reply_t

xcb_xv_query_extension_request_t

xcb_xv_query_image_attributes_cookie_t

xcb_xv_query_image_attributes_reply_t

xcb_xv_query_image_attributes_request_t

xcb_xv_query_port_attributes_cookie_t

xcb_xv_query_port_attributes_reply_t

xcb_xv_query_port_attributes_request_t

xcb_xv_rational_iterator_t

xcb_xv_rational_t

xcb_xv_select_port_notify_request_t

xcb_xv_select_video_notify_request_t

xcb_xv_set_port_attribute_request_t

xcb_xv_shm_put_image_request_t

xcb_xv_stop_video_request_t

xcb_xv_ungrab_port_request_t

xcb_xv_video_notify_event_t

Constants

XCB_XV_ATTRIBUTE_FLAG_GETTABLE

XCB_XV_ATTRIBUTE_FLAG_SETTABLE

XCB_XV_BAD_CONTROL

XCB_XV_BAD_ENCODING

XCB_XV_BAD_PORT

XCB_XV_GET_PORT_ATTRIBUTE

XCB_XV_GET_STILL

XCB_XV_GET_VIDEO

XCB_XV_GRAB_PORT

XCB_XV_GRAB_PORT_STATUS_ALREADY_GRABBED

XCB_XV_GRAB_PORT_STATUS_BAD_ALLOC

XCB_XV_GRAB_PORT_STATUS_BAD_EXTENSION

XCB_XV_GRAB_PORT_STATUS_BAD_REPLY

XCB_XV_GRAB_PORT_STATUS_INVALID_TIME

XCB_XV_GRAB_PORT_STATUS_SUCCESS

XCB_XV_IMAGE_FORMAT_INFO_FORMAT_PACKED

XCB_XV_IMAGE_FORMAT_INFO_FORMAT_PLANAR

XCB_XV_IMAGE_FORMAT_INFO_TYPE_RGB

XCB_XV_IMAGE_FORMAT_INFO_TYPE_YUV

XCB_XV_LIST_IMAGE_FORMATS

XCB_XV_MAJOR_VERSION

XCB_XV_MINOR_VERSION

XCB_XV_PORT_NOTIFY

XCB_XV_PUT_IMAGE

XCB_XV_PUT_STILL

XCB_XV_PUT_VIDEO

XCB_XV_QUERY_ADAPTORS

XCB_XV_QUERY_BEST_SIZE

XCB_XV_QUERY_ENCODINGS

XCB_XV_QUERY_EXTENSION

XCB_XV_QUERY_IMAGE_ATTRIBUTES

XCB_XV_QUERY_PORT_ATTRIBUTES

XCB_XV_SCANLINE_ORDER_BOTTOM_TO_TOP

XCB_XV_SCANLINE_ORDER_TOP_TO_BOTTOM

XCB_XV_SELECT_PORT_NOTIFY

XCB_XV_SELECT_VIDEO_NOTIFY

XCB_XV_SET_PORT_ATTRIBUTE

XCB_XV_SHM_PUT_IMAGE

XCB_XV_STOP_VIDEO

XCB_XV_TYPE_IMAGE_MASK

XCB_XV_TYPE_INPUT_MASK

XCB_XV_TYPE_OUTPUT_MASK

XCB_XV_TYPE_STILL_MASK

XCB_XV_TYPE_VIDEO_MASK

XCB_XV_UNGRAB_PORT

XCB_XV_VIDEO_NOTIFY

XCB_XV_VIDEO_NOTIFY_REASON_BUSY

XCB_XV_VIDEO_NOTIFY_REASON_HARD_ERROR

XCB_XV_VIDEO_NOTIFY_REASON_PREEMPTED

XCB_XV_VIDEO_NOTIFY_REASON_STARTED

XCB_XV_VIDEO_NOTIFY_REASON_STOPPED

Statics

xcb_xv_id

Functions

xcb_xv_adaptor_info_end⚠

xcb_xv_adaptor_info_formats⚠

xcb_xv_adaptor_info_formats_iterator⚠

xcb_xv_adaptor_info_formats_length⚠

xcb_xv_adaptor_info_name⚠

xcb_xv_adaptor_info_name_end⚠

xcb_xv_adaptor_info_name_length⚠

xcb_xv_adaptor_info_next⚠

xcb_xv_attribute_info_end⚠

xcb_xv_attribute_info_name⚠

xcb_xv_attribute_info_name_end⚠

xcb_xv_attribute_info_name_length⚠

xcb_xv_attribute_info_next⚠

xcb_xv_encoding_end⚠

xcb_xv_encoding_info_end⚠

xcb_xv_encoding_info_name⚠

xcb_xv_encoding_info_name_end⚠

xcb_xv_encoding_info_name_length⚠

xcb_xv_encoding_info_next⚠

xcb_xv_encoding_next⚠

xcb_xv_format_end⚠

xcb_xv_format_next⚠

xcb_xv_get_port_attribute⚠

xcb_xv_get_port_attribute_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xv_get_port_attribute_unchecked⚠

xcb_xv_get_still⚠

xcb_xv_get_still_checked⚠

xcb_xv_get_video⚠

xcb_xv_get_video_checked⚠

xcb_xv_grab_port⚠

xcb_xv_grab_port_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xv_grab_port_unchecked⚠

xcb_xv_image_data⚠

xcb_xv_image_data_end⚠

xcb_xv_image_data_length⚠

xcb_xv_image_end⚠

xcb_xv_image_format_info_end⚠

xcb_xv_image_format_info_next⚠

xcb_xv_image_next⚠

xcb_xv_image_offsets⚠

xcb_xv_image_offsets_end⚠

xcb_xv_image_offsets_length⚠

xcb_xv_image_pitches⚠

xcb_xv_image_pitches_end⚠

xcb_xv_image_pitches_length⚠

xcb_xv_list_image_formats⚠

xcb_xv_list_image_formats_format⚠

xcb_xv_list_image_formats_format_iterator⚠

xcb_xv_list_image_formats_format_length⚠

xcb_xv_list_image_formats_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xv_list_image_formats_unchecked⚠

xcb_xv_port_end⚠

xcb_xv_port_next⚠

xcb_xv_put_image⚠

xcb_xv_put_image_checked⚠

xcb_xv_put_still⚠

xcb_xv_put_still_checked⚠

xcb_xv_put_video⚠

xcb_xv_put_video_checked⚠

xcb_xv_query_adaptors⚠

xcb_xv_query_adaptors_info_iterator⚠

xcb_xv_query_adaptors_info_length⚠

xcb_xv_query_adaptors_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xv_query_adaptors_unchecked⚠

xcb_xv_query_best_size⚠

xcb_xv_query_best_size_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xv_query_best_size_unchecked⚠

xcb_xv_query_encodings⚠

xcb_xv_query_encodings_info_iterator⚠

xcb_xv_query_encodings_info_length⚠

xcb_xv_query_encodings_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xv_query_encodings_unchecked⚠

xcb_xv_query_extension⚠

xcb_xv_query_extension_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xv_query_extension_unchecked⚠

xcb_xv_query_image_attributes⚠

xcb_xv_query_image_attributes_offsets⚠

xcb_xv_query_image_attributes_offsets_end⚠

xcb_xv_query_image_attributes_offsets_length⚠

xcb_xv_query_image_attributes_pitches⚠

xcb_xv_query_image_attributes_pitches_end⚠

xcb_xv_query_image_attributes_pitches_length⚠

xcb_xv_query_image_attributes_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xv_query_image_attributes_unchecked⚠

xcb_xv_query_port_attributes⚠

xcb_xv_query_port_attributes_attributes_iterator⚠

xcb_xv_query_port_attributes_attributes_length⚠

xcb_xv_query_port_attributes_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xv_query_port_attributes_unchecked⚠

xcb_xv_rational_end⚠

xcb_xv_rational_next⚠

xcb_xv_select_port_notify⚠

xcb_xv_select_port_notify_checked⚠

xcb_xv_select_video_notify⚠

xcb_xv_select_video_notify_checked⚠

xcb_xv_set_port_attribute⚠

xcb_xv_set_port_attribute_checked⚠

xcb_xv_shm_put_image⚠

xcb_xv_shm_put_image_checked⚠

xcb_xv_stop_video⚠

xcb_xv_stop_video_checked⚠

xcb_xv_ungrab_port⚠

xcb_xv_ungrab_port_checked⚠

Type Definitions

xcb_xv_attribute_flag_t

xcb_xv_encoding_t

xcb_xv_grab_port_status_t

xcb_xv_image_format_info_format_t

xcb_xv_image_format_info_type_t

xcb_xv_port_t

xcb_xv_scanline_order_t

xcb_xv_type_t

xcb_xv_video_notify_reason_t