Module xcb::ffi::xvmc

Structs

xcb_xvmc_context_iterator_t

xcb_xvmc_create_context_cookie_t

xcb_xvmc_create_context_reply_t

xcb_xvmc_create_context_request_t

xcb_xvmc_create_subpicture_cookie_t

xcb_xvmc_create_subpicture_reply_t

xcb_xvmc_create_subpicture_request_t

xcb_xvmc_create_surface_cookie_t

xcb_xvmc_create_surface_reply_t

xcb_xvmc_create_surface_request_t

xcb_xvmc_destroy_context_request_t

xcb_xvmc_destroy_subpicture_request_t

xcb_xvmc_destroy_surface_request_t

xcb_xvmc_list_subpicture_types_cookie_t

xcb_xvmc_list_subpicture_types_reply_t

xcb_xvmc_list_subpicture_types_request_t

xcb_xvmc_list_surface_types_cookie_t

xcb_xvmc_list_surface_types_reply_t

xcb_xvmc_list_surface_types_request_t

xcb_xvmc_query_version_cookie_t

xcb_xvmc_query_version_reply_t

xcb_xvmc_query_version_request_t

xcb_xvmc_subpicture_iterator_t

xcb_xvmc_surface_info_iterator_t

xcb_xvmc_surface_info_t

xcb_xvmc_surface_iterator_t

Constants

XCB_XVMC_CREATE_CONTEXT

XCB_XVMC_CREATE_SUBPICTURE

XCB_XVMC_CREATE_SURFACE

XCB_XVMC_DESTROY_CONTEXT

XCB_XVMC_DESTROY_SUBPICTURE

XCB_XVMC_DESTROY_SURFACE

XCB_XVMC_LIST_SUBPICTURE_TYPES

XCB_XVMC_LIST_SURFACE_TYPES

XCB_XVMC_MAJOR_VERSION

XCB_XVMC_MINOR_VERSION

XCB_XVMC_QUERY_VERSION

Statics

xcb_xvmc_id

Functions

xcb_xvmc_context_end⚠

xcb_xvmc_context_next⚠

xcb_xvmc_create_context⚠

xcb_xvmc_create_context_priv_data⚠

xcb_xvmc_create_context_priv_data_end⚠

xcb_xvmc_create_context_priv_data_length⚠

xcb_xvmc_create_context_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xvmc_create_context_unchecked⚠

xcb_xvmc_create_subpicture⚠

xcb_xvmc_create_subpicture_priv_data⚠

xcb_xvmc_create_subpicture_priv_data_end⚠

xcb_xvmc_create_subpicture_priv_data_length⚠

xcb_xvmc_create_subpicture_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xvmc_create_subpicture_unchecked⚠

xcb_xvmc_create_surface⚠

xcb_xvmc_create_surface_priv_data⚠

xcb_xvmc_create_surface_priv_data_end⚠

xcb_xvmc_create_surface_priv_data_length⚠

xcb_xvmc_create_surface_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xvmc_create_surface_unchecked⚠

xcb_xvmc_destroy_context⚠

xcb_xvmc_destroy_context_checked⚠

xcb_xvmc_destroy_subpicture⚠

xcb_xvmc_destroy_subpicture_checked⚠

xcb_xvmc_destroy_surface⚠

xcb_xvmc_destroy_surface_checked⚠

xcb_xvmc_list_subpicture_types⚠

xcb_xvmc_list_subpicture_types_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xvmc_list_subpicture_types_types⚠

xcb_xvmc_list_subpicture_types_types_iterator⚠

xcb_xvmc_list_subpicture_types_types_length⚠

xcb_xvmc_list_subpicture_types_unchecked⚠

xcb_xvmc_list_surface_types⚠

xcb_xvmc_list_surface_types_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xvmc_list_surface_types_surfaces⚠

xcb_xvmc_list_surface_types_surfaces_iterator⚠

xcb_xvmc_list_surface_types_surfaces_length⚠

xcb_xvmc_list_surface_types_unchecked⚠

xcb_xvmc_query_version⚠

xcb_xvmc_query_version_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xvmc_query_version_unchecked⚠

xcb_xvmc_subpicture_end⚠

xcb_xvmc_subpicture_next⚠

xcb_xvmc_surface_end⚠

xcb_xvmc_surface_info_end⚠

xcb_xvmc_surface_info_next⚠

xcb_xvmc_surface_next⚠

Type Definitions

xcb_xvmc_context_t

xcb_xvmc_subpicture_t

xcb_xvmc_surface_t