xcb

Module glx

The GLX X extension.

Accessible with the glx cargo feature.

Structs

AreTexturesResident

The AreTexturesResident request.

AreTexturesResidentCookie

Cookie type for AreTexturesResident.

AreTexturesResidentCookieUnchecked

Unchecked cookie type for AreTexturesResident.

AreTexturesResidentReply

Reply type for AreTexturesResident.

BufferSwapCompleteEvent

The BufferSwapCompleteEvent event.

ChangeDrawableAttributes

The ChangeDrawableAttributes request.

ClientInfo

The ClientInfo request.

Context

CopyContext

The CopyContext request.

CreateContext

The CreateContext request.

CreateContextAttribsArb

The CreateContextAttribsArb request.

CreateGlxPixmap

The CreateGlxPixmap request.

CreateNewContext

The CreateNewContext request.

CreatePbuffer

The CreatePbuffer request.

CreatePixmap

The CreatePixmap request.

CreateWindow

The CreateWindow request.

DeleteLists

The DeleteLists request.

DeleteQueriesArb

The DeleteQueriesArb request.

DeleteTextures

The DeleteTextures request.

DeleteWindow

The DeleteWindow request.

DestroyContext

The DestroyContext request.

DestroyGlxPixmap

The DestroyGlxPixmap request.

DestroyPbuffer

The DestroyPbuffer request.

DestroyPixmap

The DestroyPixmap request.

EndList

The EndList request.

Fbconfig

FeedbackBuffer

The FeedbackBuffer request.

Finish

The Finish request.

FinishCookie

Cookie type for Finish.

FinishCookieUnchecked

Unchecked cookie type for Finish.

FinishReply

Reply type for Finish.

Flush

The Flush request.

Gc

GenLists

The GenLists request.

GenListsCookie

Cookie type for GenLists.

GenListsCookieUnchecked

Unchecked cookie type for GenLists.

GenListsReply

Reply type for GenLists.

GenQueriesArb

The GenQueriesArb request.

GenQueriesArbCookie

Cookie type for GenQueriesArb.

GenQueriesArbCookieUnchecked

Unchecked cookie type for GenQueriesArb.

GenQueriesArbReply

Reply type for GenQueriesArb.

GenTextures

The GenTextures request.

GenTexturesCookie

Cookie type for GenTextures.

GenTexturesCookieUnchecked

Unchecked cookie type for GenTextures.

GenTexturesReply

Reply type for GenTextures.

GenericError

The GenericError error.

GetBooleanv

The GetBooleanv request.

GetBooleanvCookie

Cookie type for GetBooleanv.

GetBooleanvCookieUnchecked

Unchecked cookie type for GetBooleanv.

GetBooleanvReply

Reply type for GetBooleanv.

GetClipPlane

The GetClipPlane request.

GetClipPlaneCookie

Cookie type for GetClipPlane.

GetClipPlaneCookieUnchecked

Unchecked cookie type for GetClipPlane.

GetClipPlaneReply

Reply type for GetClipPlane.

GetColorTable

The GetColorTable request.

GetColorTableCookie

Cookie type for GetColorTable.

GetColorTableCookieUnchecked

Unchecked cookie type for GetColorTable.

GetColorTableParameterfv

The GetColorTableParameterfv request.

GetColorTableParameterfvCookie

Cookie type for GetColorTableParameterfv.

GetColorTableParameterfvCookieUnchecked

Unchecked cookie type for GetColorTableParameterfv.

GetColorTableParameterfvReply

Reply type for GetColorTableParameterfv.

GetColorTableParameteriv

The GetColorTableParameteriv request.

GetColorTableParameterivCookie

Cookie type for GetColorTableParameteriv.

GetColorTableParameterivCookieUnchecked

Unchecked cookie type for GetColorTableParameteriv.

GetColorTableParameterivReply

Reply type for GetColorTableParameteriv.

GetColorTableReply

Reply type for GetColorTable.

GetCompressedTexImageArb

The GetCompressedTexImageArb request.

GetCompressedTexImageArbCookie

Cookie type for GetCompressedTexImageArb.

GetCompressedTexImageArbCookieUnchecked

Unchecked cookie type for GetCompressedTexImageArb.

GetCompressedTexImageArbReply

Reply type for GetCompressedTexImageArb.

GetConvolutionFilter

The GetConvolutionFilter request.

GetConvolutionFilterCookie

Cookie type for GetConvolutionFilter.

GetConvolutionFilterCookieUnchecked

Unchecked cookie type for GetConvolutionFilter.

GetConvolutionFilterReply

Reply type for GetConvolutionFilter.

GetConvolutionParameterfv

The GetConvolutionParameterfv request.

GetConvolutionParameterfvCookie

Cookie type for GetConvolutionParameterfv.

GetConvolutionParameterfvCookieUnchecked

Unchecked cookie type for GetConvolutionParameterfv.

GetConvolutionParameterfvReply

Reply type for GetConvolutionParameterfv.

GetConvolutionParameteriv

The GetConvolutionParameteriv request.

GetConvolutionParameterivCookie

Cookie type for GetConvolutionParameteriv.

GetConvolutionParameterivCookieUnchecked

Unchecked cookie type for GetConvolutionParameteriv.

GetConvolutionParameterivReply

Reply type for GetConvolutionParameteriv.

GetDoublev

The GetDoublev request.

GetDoublevCookie

Cookie type for GetDoublev.

GetDoublevCookieUnchecked

Unchecked cookie type for GetDoublev.

GetDoublevReply

Reply type for GetDoublev.

GetDrawableAttributes

The GetDrawableAttributes request.

GetDrawableAttributesCookie

Cookie type for GetDrawableAttributes.

GetDrawableAttributesCookieUnchecked

Unchecked cookie type for GetDrawableAttributes.

GetDrawableAttributesReply

Reply type for GetDrawableAttributes.

GetError

The GetError request.

GetErrorCookie

Cookie type for GetError.

GetErrorCookieUnchecked

Unchecked cookie type for GetError.

GetErrorReply

Reply type for GetError.

GetFbConfigs

The GetFbConfigs request.

GetFbConfigsCookie

Cookie type for GetFbConfigs.

GetFbConfigsCookieUnchecked

Unchecked cookie type for GetFbConfigs.

GetFbConfigsReply

Reply type for GetFbConfigs.

GetFloatv

The GetFloatv request.

GetFloatvCookie

Cookie type for GetFloatv.

GetFloatvCookieUnchecked

Unchecked cookie type for GetFloatv.

GetFloatvReply

Reply type for GetFloatv.

GetHistogram

The GetHistogram request.

GetHistogramCookie

Cookie type for GetHistogram.

GetHistogramCookieUnchecked

Unchecked cookie type for GetHistogram.

GetHistogramParameterfv

The GetHistogramParameterfv request.

GetHistogramParameterfvCookie

Cookie type for GetHistogramParameterfv.

GetHistogramParameterfvCookieUnchecked

Unchecked cookie type for GetHistogramParameterfv.

GetHistogramParameterfvReply

Reply type for GetHistogramParameterfv.

GetHistogramParameteriv

The GetHistogramParameteriv request.

GetHistogramParameterivCookie

Cookie type for GetHistogramParameteriv.

GetHistogramParameterivCookieUnchecked

Unchecked cookie type for GetHistogramParameteriv.

GetHistogramParameterivReply

Reply type for GetHistogramParameteriv.

GetHistogramReply

Reply type for GetHistogram.

GetIntegerv

The GetIntegerv request.

GetIntegervCookie

Cookie type for GetIntegerv.

GetIntegervCookieUnchecked

Unchecked cookie type for GetIntegerv.

GetIntegervReply

Reply type for GetIntegerv.

GetLightfv

The GetLightfv request.

GetLightfvCookie

Cookie type for GetLightfv.

GetLightfvCookieUnchecked

Unchecked cookie type for GetLightfv.

GetLightfvReply

Reply type for GetLightfv.

GetLightiv

The GetLightiv request.

GetLightivCookie

Cookie type for GetLightiv.

GetLightivCookieUnchecked

Unchecked cookie type for GetLightiv.

GetLightivReply

Reply type for GetLightiv.

GetMapdv

The GetMapdv request.

GetMapdvCookie

Cookie type for GetMapdv.

GetMapdvCookieUnchecked

Unchecked cookie type for GetMapdv.

GetMapdvReply

Reply type for GetMapdv.

GetMapfv

The GetMapfv request.

GetMapfvCookie

Cookie type for GetMapfv.

GetMapfvCookieUnchecked

Unchecked cookie type for GetMapfv.

GetMapfvReply

Reply type for GetMapfv.

GetMapiv

The GetMapiv request.

GetMapivCookie

Cookie type for GetMapiv.

GetMapivCookieUnchecked

Unchecked cookie type for GetMapiv.

GetMapivReply

Reply type for GetMapiv.

GetMaterialfv

The GetMaterialfv request.

GetMaterialfvCookie

Cookie type for GetMaterialfv.

GetMaterialfvCookieUnchecked

Unchecked cookie type for GetMaterialfv.

GetMaterialfvReply

Reply type for GetMaterialfv.

GetMaterialiv

The GetMaterialiv request.

GetMaterialivCookie

Cookie type for GetMaterialiv.

GetMaterialivCookieUnchecked

Unchecked cookie type for GetMaterialiv.

GetMaterialivReply

Reply type for GetMaterialiv.

GetMinmax

The GetMinmax request.

GetMinmaxCookie

Cookie type for GetMinmax.

GetMinmaxCookieUnchecked

Unchecked cookie type for GetMinmax.

GetMinmaxParameterfv

The GetMinmaxParameterfv request.

GetMinmaxParameterfvCookie

Cookie type for GetMinmaxParameterfv.

GetMinmaxParameterfvCookieUnchecked

Unchecked cookie type for GetMinmaxParameterfv.

GetMinmaxParameterfvReply

Reply type for GetMinmaxParameterfv.

GetMinmaxParameteriv

The GetMinmaxParameteriv request.

GetMinmaxParameterivCookie

Cookie type for GetMinmaxParameteriv.

GetMinmaxParameterivCookieUnchecked

Unchecked cookie type for GetMinmaxParameteriv.

GetMinmaxParameterivReply

Reply type for GetMinmaxParameteriv.

GetMinmaxReply

Reply type for GetMinmax.

GetPixelMapfv

The GetPixelMapfv request.

GetPixelMapfvCookie

Cookie type for GetPixelMapfv.

GetPixelMapfvCookieUnchecked

Unchecked cookie type for GetPixelMapfv.

GetPixelMapfvReply

Reply type for GetPixelMapfv.

GetPixelMapuiv

The GetPixelMapuiv request.

GetPixelMapuivCookie

Cookie type for GetPixelMapuiv.

GetPixelMapuivCookieUnchecked

Unchecked cookie type for GetPixelMapuiv.

GetPixelMapuivReply

Reply type for GetPixelMapuiv.

GetPixelMapusv

The GetPixelMapusv request.

GetPixelMapusvCookie

Cookie type for GetPixelMapusv.

GetPixelMapusvCookieUnchecked

Unchecked cookie type for GetPixelMapusv.

GetPixelMapusvReply

Reply type for GetPixelMapusv.

GetPolygonStipple

The GetPolygonStipple request.

GetPolygonStippleCookie

Cookie type for GetPolygonStipple.

GetPolygonStippleCookieUnchecked

Unchecked cookie type for GetPolygonStipple.

GetPolygonStippleReply

Reply type for GetPolygonStipple.

GetQueryObjectivArb

The GetQueryObjectivArb request.

GetQueryObjectivArbCookie

Cookie type for GetQueryObjectivArb.

GetQueryObjectivArbCookieUnchecked

Unchecked cookie type for GetQueryObjectivArb.

GetQueryObjectivArbReply

Reply type for GetQueryObjectivArb.

GetQueryObjectuivArb

The GetQueryObjectuivArb request.

GetQueryObjectuivArbCookie

Cookie type for GetQueryObjectuivArb.

GetQueryObjectuivArbCookieUnchecked

Unchecked cookie type for GetQueryObjectuivArb.

GetQueryObjectuivArbReply

Reply type for GetQueryObjectuivArb.

GetQueryivArb

The GetQueryivArb request.

GetQueryivArbCookie

Cookie type for GetQueryivArb.

GetQueryivArbCookieUnchecked

Unchecked cookie type for GetQueryivArb.

GetQueryivArbReply

Reply type for GetQueryivArb.

GetSeparableFilter

The GetSeparableFilter request.

GetSeparableFilterCookie

Cookie type for GetSeparableFilter.

GetSeparableFilterCookieUnchecked

Unchecked cookie type for GetSeparableFilter.

GetSeparableFilterReply

Reply type for GetSeparableFilter.

GetString

The GetString request.

GetStringCookie

Cookie type for GetString.

GetStringCookieUnchecked

Unchecked cookie type for GetString.

GetStringReply

Reply type for GetString.

GetTexEnvfv

The GetTexEnvfv request.

GetTexEnvfvCookie

Cookie type for GetTexEnvfv.

GetTexEnvfvCookieUnchecked

Unchecked cookie type for GetTexEnvfv.

GetTexEnvfvReply

Reply type for GetTexEnvfv.

GetTexEnviv

The GetTexEnviv request.

GetTexEnvivCookie

Cookie type for GetTexEnviv.

GetTexEnvivCookieUnchecked

Unchecked cookie type for GetTexEnviv.

GetTexEnvivReply

Reply type for GetTexEnviv.

GetTexGendv

The GetTexGendv request.

GetTexGendvCookie

Cookie type for GetTexGendv.

GetTexGendvCookieUnchecked

Unchecked cookie type for GetTexGendv.

GetTexGendvReply

Reply type for GetTexGendv.

GetTexGenfv

The GetTexGenfv request.

GetTexGenfvCookie

Cookie type for GetTexGenfv.

GetTexGenfvCookieUnchecked

Unchecked cookie type for GetTexGenfv.

GetTexGenfvReply

Reply type for GetTexGenfv.

GetTexGeniv

The GetTexGeniv request.

GetTexGenivCookie

Cookie type for GetTexGeniv.

GetTexGenivCookieUnchecked

Unchecked cookie type for GetTexGeniv.

GetTexGenivReply

Reply type for GetTexGeniv.

GetTexImage

The GetTexImage request.

GetTexImageCookie

Cookie type for GetTexImage.

GetTexImageCookieUnchecked

Unchecked cookie type for GetTexImage.

GetTexImageReply

Reply type for GetTexImage.

GetTexLevelParameterfv

The GetTexLevelParameterfv request.

GetTexLevelParameterfvCookie

Cookie type for GetTexLevelParameterfv.

GetTexLevelParameterfvCookieUnchecked

Unchecked cookie type for GetTexLevelParameterfv.

GetTexLevelParameterfvReply

Reply type for GetTexLevelParameterfv.

GetTexLevelParameteriv

The GetTexLevelParameteriv request.

GetTexLevelParameterivCookie

Cookie type for GetTexLevelParameteriv.

GetTexLevelParameterivCookieUnchecked

Unchecked cookie type for GetTexLevelParameteriv.

GetTexLevelParameterivReply

Reply type for GetTexLevelParameteriv.

GetTexParameterfv

The GetTexParameterfv request.

GetTexParameterfvCookie

Cookie type for GetTexParameterfv.

GetTexParameterfvCookieUnchecked

Unchecked cookie type for GetTexParameterfv.

GetTexParameterfvReply

Reply type for GetTexParameterfv.

GetTexParameteriv

The GetTexParameteriv request.

GetTexParameterivCookie

Cookie type for GetTexParameteriv.

GetTexParameterivCookieUnchecked

Unchecked cookie type for GetTexParameteriv.

GetTexParameterivReply

Reply type for GetTexParameteriv.

GetVisualConfigs

The GetVisualConfigs request.

GetVisualConfigsCookie

Cookie type for GetVisualConfigs.

GetVisualConfigsCookieUnchecked

Unchecked cookie type for GetVisualConfigs.

GetVisualConfigsReply

Reply type for GetVisualConfigs.

IsDirect

The IsDirect request.

IsDirectCookie

Cookie type for IsDirect.

IsDirectCookieUnchecked

Unchecked cookie type for IsDirect.

IsDirectReply

Reply type for IsDirect.

IsEnabled

The IsEnabled request.

IsEnabledCookie

Cookie type for IsEnabled.

IsEnabledCookieUnchecked

Unchecked cookie type for IsEnabled.

IsEnabledReply

Reply type for IsEnabled.

IsList

The IsList request.

IsListCookie

Cookie type for IsList.

IsListCookieUnchecked

Unchecked cookie type for IsList.

IsListReply

Reply type for IsList.

IsQueryArb

The IsQueryArb request.

IsQueryArbCookie

Cookie type for IsQueryArb.

IsQueryArbCookieUnchecked

Unchecked cookie type for IsQueryArb.

IsQueryArbReply

Reply type for IsQueryArb.

IsTexture

The IsTexture request.

IsTextureCookie

Cookie type for IsTexture.

IsTextureCookieUnchecked

Unchecked cookie type for IsTexture.

IsTextureReply

Reply type for IsTexture.

MakeContextCurrent

The MakeContextCurrent request.

MakeContextCurrentCookie

Cookie type for MakeContextCurrent.

MakeContextCurrentCookieUnchecked

Unchecked cookie type for MakeContextCurrent.

MakeContextCurrentReply

Reply type for MakeContextCurrent.

MakeCurrent

The MakeCurrent request.

MakeCurrentCookie

Cookie type for MakeCurrent.

MakeCurrentCookieUnchecked

Unchecked cookie type for MakeCurrent.

MakeCurrentReply

Reply type for MakeCurrent.

NewList

The NewList request.

Pbuffer

PbufferClobberEvent

The PbufferClobberEvent event.

PixelStoref

The PixelStoref request.

PixelStorei

The PixelStorei request.

Pixmap

QueryContext

The QueryContext request.

QueryContextCookie

Cookie type for QueryContext.

QueryContextCookieUnchecked

Unchecked cookie type for QueryContext.

QueryContextReply

Reply type for QueryContext.

QueryExtensionsString

The QueryExtensionsString request.

QueryExtensionsStringCookie

Cookie type for QueryExtensionsString.

QueryExtensionsStringCookieUnchecked

Unchecked cookie type for QueryExtensionsString.

QueryExtensionsStringReply

Reply type for QueryExtensionsString.

QueryServerString

The QueryServerString request.

QueryServerStringCookie

Cookie type for QueryServerString.

QueryServerStringCookieUnchecked

Unchecked cookie type for QueryServerString.

QueryServerStringReply

Reply type for QueryServerString.

QueryVersion

The QueryVersion request.

QueryVersionCookie

Cookie type for QueryVersion.

QueryVersionCookieUnchecked

Unchecked cookie type for QueryVersion.

QueryVersionReply

Reply type for QueryVersion.

ReadPixels

The ReadPixels request.

ReadPixelsCookie

Cookie type for ReadPixels.

ReadPixelsCookieUnchecked

Unchecked cookie type for ReadPixels.

ReadPixelsReply

Reply type for ReadPixels.

Render

The Render request.

RenderLarge

The RenderLarge request.

RenderMode

The RenderMode request.

RenderModeCookie

Cookie type for RenderMode.

RenderModeCookieUnchecked

Unchecked cookie type for RenderMode.

RenderModeReply

Reply type for RenderMode.

SelectBuffer

The SelectBuffer request.

SetClientInfo2arb

The SetClientInfo2arb request.

SetClientInfoArb

The SetClientInfoArb request.

SwapBuffers

The SwapBuffers request.

UseXFont

The UseXFont request.

VendorPrivate

The VendorPrivate request.

VendorPrivateWithReply

The VendorPrivateWithReply request.

VendorPrivateWithReplyCookie

Cookie type for VendorPrivateWithReply.

VendorPrivateWithReplyCookieUnchecked

Unchecked cookie type for VendorPrivateWithReply.

VendorPrivateWithReplyReply

Reply type for VendorPrivateWithReply.

WaitGl

The WaitGl request.

WaitX

The WaitX request.

Window

Enums

Drawable

Error

Unified error type for the Glx extension

Event

Unified event type for the Glx extension

Pbcdt

Pbcet

Rm

Constants

MAJOR_VERSION

The major version of the Glx extension.

MINOR_VERSION

The minor version of the Glx extension.

VERSION_STRING

The version string of the Glx extension.

XNAME

The official identifier for the Glx extension.

Functions

get_extension_data

Fetch server runtime info data of the Glx extension.

prefetch_extension_data

Prefetch server runtime info data of the Glx extension.

Type Aliases

BadContextError

The BadContextError error.

BadContextStateError

The BadContextStateError error.

BadContextTagError

The BadContextTagError error.

BadCurrentDrawableError

The BadCurrentDrawableError error.

BadCurrentWindowError

The BadCurrentWindowError error.

BadDrawableError

The BadDrawableError error.

BadFbConfigError

The BadFbConfigError error.

BadLargeRequestError

The BadLargeRequestError error.

BadPbufferError

The BadPbufferError error.

BadPixmapError

The BadPixmapError error.

BadRenderRequestError

The BadRenderRequestError error.

BadWindowError

The BadWindowError error.

ContextTag

GlxBadProfileArbError

The GlxBadProfileArbError error.

UnsupportedPrivateRequestError

The UnsupportedPrivateRequestError error.