xcb
Module present

The Present X extension.

Accessible with the present cargo feature.

Structs

Capability

CompleteNotifyEvent

The CompleteNotifyEvent event.

ConfigureNotifyEvent

The ConfigureNotifyEvent event.

EventMask

EventXid

GenericEvent

The GenericEvent event.

IdleNotifyEvent

The IdleNotifyEvent event.

Notify

NotifyMsc

The NotifyMsc request.

Option

Pixmap

The Pixmap request.

QueryCapabilities

The QueryCapabilities request.

QueryCapabilitiesCookie

Cookie type for QueryCapabilities.

QueryCapabilitiesCookieUnchecked

Unchecked cookie type for QueryCapabilities.

QueryCapabilitiesReply

Reply type for QueryCapabilities.

QueryVersion

The QueryVersion request.

QueryVersionCookie

Cookie type for QueryVersion.

QueryVersionCookieUnchecked

Unchecked cookie type for QueryVersion.

QueryVersionReply

Reply type for QueryVersion.

RedirectNotifyEvent

The RedirectNotifyEvent event.

SelectInput

The SelectInput request.

Enums

CompleteKind

CompleteMode

Event

Unified event type for the Present extension

EventEnum

Constants

MAJOR_VERSION

The major version of the Present extension.

MINOR_VERSION

The minor version of the Present extension.

VERSION_STRING

The version string of the Present extension.

XNAME

The official identifier for the Present extension.

Functions

get_extension_data

Fetch server runtime info data of the Present extension.

prefetch_extension_data

Prefetch server runtime info data of the Present extension.