Module xcb::shm

The MIT-SHM X extension.

Accessible with the shm cargo feature.

Structs

Attach

The Attach request.

AttachFd

The AttachFd request.

CompletionEvent

The CompletionEvent event.

CreatePixmap

The CreatePixmap request.

CreateSegment

The CreateSegment request.

CreateSegmentCookie

Cookie type for CreateSegment.

CreateSegmentCookieUnchecked

Unchecked cookie type for CreateSegment.

CreateSegmentReply

Reply type for CreateSegment.

Detach

The Detach request.

GetImage

The GetImage request.

GetImageCookie

Cookie type for GetImage.

GetImageCookieUnchecked

Unchecked cookie type for GetImage.

GetImageReply

Reply type for GetImage.

PutImage

The PutImage request.

QueryVersion

The QueryVersion request.

QueryVersionCookie

Cookie type for QueryVersion.

QueryVersionCookieUnchecked

Unchecked cookie type for QueryVersion.

QueryVersionReply

Reply type for QueryVersion.

Seg

Enums

Error

Unified error type for the Shm extension

Event

Unified event type for the Shm extension

Constants

MAJOR_VERSION

The major version of the Shm extension.

MINOR_VERSION

The minor version of the Shm extension.

VERSION_STRING

The version string of the Shm extension.

XNAME

The official identifier for the Shm extension.

Functions

get_extension_data

Fetch server runtime info data of the Shm extension.

prefetch_extension_data

Prefetch server runtime info data of the Shm extension.

Type Aliases

BadSegError

The BadSegError error.