Trait xcb::BaseError

pub trait BaseError: Raw<xcb_generic_error_t> {
    const EXTENSION: Option<Extension>;
    const NUMBER: u32;
}

A trait to designate base protocol errors.

The base errors follow the usual resolution idiom of error_code offset.

This should be completely transparent to the user, as ProtocolError is resolving all types of errors together.

Required Associated Constants

const EXTENSION: Option<Extension>

The extension associated to this error, or None for the main protocol

const NUMBER: u32

The number associated to this error

Object Safety

This trait is not object safe.