Trait xcb::RequestWithReply

pub trait RequestWithReply: Request {
    type Reply: Reply;
    type Cookie: CookieWithReplyChecked<Reply = Self::Reply>;
    type CookieUnchecked: CookieWithReplyUnchecked<Reply = Self::Reply>;
}

Trait for requests that return a reply.

Required Associated Types

type Reply: Reply

Reply associated with the request

type Cookie: CookieWithReplyChecked<Reply = Self::Reply>

Default cookie type for the request, as returned by Connection::send_request.

type CookieUnchecked: CookieWithReplyUnchecked<Reply = Self::Reply>

Unchecked cookie type for the request, as returned by Connection::send_request_unchecked.

Object Safety

This trait is not object safe.