Trait xcb::ResolveWireEvent

pub trait ResolveWireEvent: Sized {
    unsafe fn resolve_wire_event(
        first_event: u8, 
        event: *mut xcb_generic_event_t
    ) -> Self;
}

Trait for the resolution of raw wire event to a unified event enum.

Self is normally an enum of several event subtypes. See crate::x::Event and crate::Event

Required methods

unsafe fn resolve_wire_event(
first_event: u8,
event: *mut xcb_generic_event_t
) -> Self

Resolve a pointer to xcb_generic_event_t to Self, inferring the correct subtype using response_type field and first_event

Panics

Panics if the event subtype cannot be resolved to Self. That is, response_type field must be checked beforehand to be in range with first_event.

Safety

event must be a valid, non-null event returned by xcb_wait_for_event or similar function