Struct xcb::x::FocusInEvent

pub struct FocusInEvent { /* private fields */ }

NOT YET DOCUMENTED

Implementations

impl FocusInEvent

pub fn new( detail: NotifyDetail, event: Window, mode: NotifyMode ) -> FocusInEvent

pub fn response_type(&self) -> u8

pub fn detail(&self) -> NotifyDetail

pub fn sequence(&self) -> u16

pub fn event(&self) -> Window

The window on which the focus event was generated. This is the window used by the X server to report the event.

pub fn mode(&self) -> NotifyMode