xcb::x

Struct QueryPointerReply

pub struct QueryPointerReply { /* private fields */ }

Implementations

impl QueryPointerReply

pub fn response_type(&self) -> u8

pub fn same_screen(&self) -> bool

If same_screen is False, then the pointer is not on the same screen as the argument window, child is None, and win_x and win_y are zero. If same_screen is True, then win_x and win_y are the pointer coordinates relative to the argument window’s origin, and child is the child containing the pointer, if any.

pub fn sequence(&self) -> u16

pub fn length(&self) -> u32

pub fn root(&self) -> Window

The root window the pointer is logically on.

pub fn child(&self) -> Window

The child window containing the pointer, if any, if same_screen is true. If same_screen is false, XCB_NONE is returned.

pub fn root_x(&self) -> i16

The pointer X position, relative to root.

pub fn root_y(&self) -> i16

The pointer Y position, relative to root.

pub fn win_x(&self) -> i16

The pointer X coordinate, relative to child, if same_screen is true. Zero otherwise.

pub fn win_y(&self) -> i16

The pointer Y coordinate, relative to child, if same_screen is true. Zero otherwise.

pub fn mask(&self) -> KeyButMask

The current logical state of the modifier keys and the buttons. Note that the logical state of a device (as seen by means of the protocol) may lag the physical state if device event processing is frozen.