Module xcb::xinerama

The XINERAMA X extension.

Accessible with the xinerama cargo feature.

Structs

GetScreenCount

The GetScreenCount request.

GetScreenCountCookie

Cookie type for GetScreenCount.

GetScreenCountCookieUnchecked

Unchecked cookie type for GetScreenCount.

GetScreenCountReply

Reply type for GetScreenCount.

GetScreenSize

The GetScreenSize request.

GetScreenSizeCookie

Cookie type for GetScreenSize.

GetScreenSizeCookieUnchecked

Unchecked cookie type for GetScreenSize.

GetScreenSizeReply

Reply type for GetScreenSize.

GetState

The GetState request.

GetStateCookie

Cookie type for GetState.

GetStateCookieUnchecked

Unchecked cookie type for GetState.

GetStateReply

Reply type for GetState.

IsActive

The IsActive request.

IsActiveCookie

Cookie type for IsActive.

IsActiveCookieUnchecked

Unchecked cookie type for IsActive.

IsActiveReply

Reply type for IsActive.

QueryScreens

The QueryScreens request.

QueryScreensCookie

Cookie type for QueryScreens.

QueryScreensCookieUnchecked

Unchecked cookie type for QueryScreens.

QueryScreensReply

Reply type for QueryScreens.

QueryVersion

The QueryVersion request.

QueryVersionCookie

Cookie type for QueryVersion.

QueryVersionCookieUnchecked

Unchecked cookie type for QueryVersion.

QueryVersionReply

Reply type for QueryVersion.

ScreenInfo

Constants

MAJOR_VERSION

The major version of the Xinerama extension.

MINOR_VERSION

The minor version of the Xinerama extension.

VERSION_STRING

The version string of the Xinerama extension.

XNAME

The official identifier for the Xinerama extension.

Functions

get_extension_data

Fetch server runtime info data of the Xinerama extension.

prefetch_extension_data

Prefetch server runtime info data of the Xinerama extension.