Function xcb::xproto::allow_events_checked

pub fn allow_events_checked<'a>(
    c: &'a Connection, 
    mode: u8, 
    time: Timestamp
) -> VoidCookie<'a>

release queued events

Releases queued events if the client has caused a device (pointer/keyboard) to freeze due to grabbing it actively. This request has no effect if time is earlier than the last-grab time of the most recent active grab for this client or if time is later than the current X server time.

parameters:

c: The connection object to the server

mode:

time: Timestamp to avoid race conditions when running X over the network.

The special value XCB_CURRENT_TIME will be replaced with the current server time.