Function xcb::xproto::change_gc

pub fn change_gc<'a>(
    c: &'a Connection, 
    gc: Gcontext, 
    value_list: &[(u32, u32)]
) -> VoidCookie<'a>

change graphics context components

Changes the components specified by value_mask for the specified graphics context.

parameters:

c: The connection object to the server

gc: The graphics context to change.

value_list: Values for each of the components specified in the bitmask value_mask. The order has to correspond to the order of possible value_mask bits. See the example.