Function xcb::xproto::circulate_window

pub fn circulate_window<'a>(
    c: &'a Connection, 
    direction: u8, 
    window: Window
) -> VoidCookie<'a>

Change window stacking order

If direction is XCB_CIRCULATE_RAISE_LOWEST, the lowest mapped child (if any) will be raised to the top of the stack.

If direction is XCB_CIRCULATE_LOWER_HIGHEST, the highest mapped child will be lowered to the bottom of the stack.

parameters:

c: The connection object to the server

direction:

window: The window to raise/lower (depending on direction).