Function xcb::xproto::create_pixmap

pub fn create_pixmap<'a>(
    c: &'a Connection, 
    depth: u8, 
    pid: Pixmap, 
    drawable: Drawable, 
    width: u16, 
    height: u16
) -> VoidCookie<'a>

Creates a pixmap

Creates a pixmap. The pixmap can only be used on the same screen as drawable is on and only with drawables of the same depth.

parameters:

c: The connection object to the server

depth: TODO

pid: The ID with which you will refer to the new pixmap, created by xcb_generate_id.

drawable: Drawable to get the screen from.

width: The width of the new pixmap.

height: The height of the new pixmap.