Function xcb::xproto::grab_pointer_unchecked

pub fn grab_pointer_unchecked<'a>(
    c: &'a Connection,
    owner_events: bool,
    grab_window: Window,
    event_mask: u16,
    pointer_mode: u8,
    keyboard_mode: u8,
    confine_to: Window,
    cursor: Cursor,
    time: Timestamp
) -> GrabPointerCookie<'a>

Grab the pointer

Actively grabs control of the pointer. Further pointer events are reported only to the grabbing client. Overrides any active pointer grab by this client.

parameters:

c: The connection object to the server

owner_events: If 1, the grab_window will still get the pointer events. If 0, events are not reported to the grab_window.

grab_window: Specifies the window on which the pointer should be grabbed.

event_mask: Specifies which pointer events are reported to the client.

TODO: which values?

pointer_mode:

keyboard_mode:

confine_to: Specifies the window to confine the pointer in (the user will not be able to move the pointer out of that window).

The special value XCB_NONE means don’t confine the pointer.

cursor: Specifies the cursor that should be displayed or XCB_NONE to not change the cursor.

time: The time argument allows you to avoid certain circumstances that come up if applications take a long time to respond or if there are long network delays. Consider a situation where you have two applications, both of which normally grab the pointer when clicked on. If both applications specify the timestamp from the event, the second application may wake up faster and successfully grab the pointer before the first application. The first application then will get an indication that the other application grabbed the pointer before its request was processed.

The special value XCB_CURRENT_TIME will be replaced with the current server time.