xcb::xproto

Function intern_atom

pub fn intern_atom<'a>(
    c: &'a Connection,
    only_if_exists: bool,
    name: &str,
) -> InternAtomCookie<'a>

Get atom identifier by name

Retrieves the identifier (xcb_atom_t TODO) for the atom with the specified name. Atoms are used in protocols like EWMH, for example to store window titles (_NET_WM_NAME atom) as property of a window.

If only_if_exists is 0, the atom will be created if it does not already exist. If only_if_exists is 1, XCB_ATOM_NONE will be returned if the atom does not yet exist.

parameters:

c: The connection object to the server

only_if_exists: Return a valid atom id only if the atom already exists.

name: The name of the atom.