Function xcb::xproto::poly_fill_rectangle

pub fn poly_fill_rectangle(c: &Connection, drawable: Drawable, gc: Gcontext, rectangles: &[Rectangle]) -> VoidCookie

Fills rectangles

Fills the specified rectangle(s) in the order listed in the array. For any given rectangle, each pixel is not drawn more than once. If rectangles intersect, the intersecting pixels are drawn multiple times.

parameters:

c: The connection object to the server

drawable: The drawable (Window or Pixmap) to draw on.

gc: The graphics context to use.

The following graphics context components are used: function, plane-mask, fill-style, subwindow-mode, clip-x-origin, clip-y-origin, and clip-mask.

The following graphics context mode-dependent components are used: foreground, background, tile, stipple, tile-stipple-x-origin, and tile-stipple-y-origin.

rectangles: The rectangles to fill.