Function xcb::xproto::poly_line

pub fn poly_line(c: &Connection, coordinate_mode: u8, drawable: Drawable, gc: Gcontext, points: &[Point]) -> VoidCookie

draw lines

Draws points_len-1 lines between each pair of points (point[i], point[i+1]) in the points array. The lines are drawn in the order listed in the array. They join correctly at all intermediate points, and if the first and last points coincide, the first and last lines also join correctly. For any given line, a pixel is not drawn more than once. If thin (zero line-width) lines intersect, the intersecting pixels are drawn multiple times. If wide lines intersect, the intersecting pixels are drawn only once, as though the entire request were a single, filled shape.

parameters:

c: The connection object to the server

coordinate_mode:

drawable: The drawable to draw the line(s) on.

gc: The graphics context to use.

points: An array of points.