Function xcb::xproto::query_pointer

pub fn query_pointer<'a>(
    c: &'a Connection, 
    window: Window
) -> QueryPointerCookie<'a>

get pointer coordinates

Gets the root window the pointer is logically on and the pointer coordinates relative to the root window's origin.

parameters:

c: The connection object to the server

window: A window to check if the pointer is on the same screen as window (see the same_screen field in the reply).