Function xcb::xproto::ungrab_key_checked

pub fn ungrab_key_checked(c: &Connection, key: Keycode, grab_window: Window, modifiers: u16) -> VoidCookie

release a key combination

Releases the key combination on grab_window if you grabbed it using xcb_grab_key before.

parameters:

c: The connection object to the server

key: The keycode of the specified key combination.

Using the special value XCB_GRAB_ANY means releasing all possible key codes.

grab_window: The window on which the grabbed key combination will be released.

modifiers: The modifiers of the specified key combination.

Using the special value XCB_MOD_MASK_ANY means releasing the key combination with every possible modifier combination.