Function xcb::xproto::ungrab_pointer_checked

pub fn ungrab_pointer_checked<'a>(
    c: &'a Connection, 
    time: Timestamp
) -> VoidCookie<'a>

release the pointer

Releases the pointer and any queued events if you actively grabbed the pointer before using xcb_grab_pointer, xcb_grab_button or within a normal button press.

EnterNotify and LeaveNotify events are generated.

parameters:

c: The connection object to the server

time: Timestamp to avoid race conditions when running X over the network.

The pointer will not be released if time is earlier than the last-pointer-grab time or later than the current X server time.