Type Definition xcb::xproto::ClientMessageEvent

type ClientMessageEvent = Event<xcb_client_message_event_t>;

NOT YET DOCUMENTED

This event represents a ClientMessage, sent by another X11 client. An example is a client sending the _NET_WM_STATE ClientMessage to the root window to indicate the fullscreen window state, effectively requesting that the window manager puts it into fullscreen mode.

Methods

impl ClientMessageEvent

pub fn format(&self) -> u8

Specifies how to interpret data. Can be either 8, 16 or 32.

pub fn window(&self) -> Window

pub fn type_(&self) -> Atom

An atom which indicates how the data should be interpreted by the receiving client.

pub fn data<'a>(&'a self) -> &'a ClientMessageData

The data itself (20 bytes max).

pub fn new(
format: u8,
window: Window,
type_: Atom,
data: ClientMessageData
) -> ClientMessageEvent

Constructs a new ClientMessageEvent response_type will be set automatically to CLIENT_MESSAGE