Type Definition xcb::xproto::ExposeEvent

type ExposeEvent = Event<xcb_expose_event_t>;

NOT YET DOCUMENTED

Methods

impl ExposeEvent

pub fn window(&self) -> Window

The exposed (damaged) window.

pub fn x(&self) -> u16

The X coordinate of the left-upper corner of the exposed rectangle, relative to the window's origin.

pub fn y(&self) -> u16

The Y coordinate of the left-upper corner of the exposed rectangle, relative to the window's origin.

pub fn width(&self) -> u16

The width of the exposed rectangle.

pub fn height(&self) -> u16

The height of the exposed rectangle.

pub fn count(&self) -> u16

The amount of Expose events following this one. Simple applications that do not want to optimize redisplay by distinguishing between subareas of its window can just ignore all Expose events with nonzero counts and perform full redisplays on events with zero counts.

pub fn new(
window: Window,
x: u16,
y: u16,
width: u16,
height: u16,
count: u16
) -> ExposeEvent

Constructs a new ExposeEvent response_type will be set automatically to EXPOSE