Type Definition xcb::xproto::FocusInEvent

type FocusInEvent = Event<xcb_focus_in_event_t>;

NOT YET DOCUMENTED

Methods

impl FocusInEvent

pub fn detail(&self) -> u8

pub fn event(&self) -> Window

The window on which the focus event was generated. This is the window used by the X server to report the event.

pub fn mode(&self) -> u8

pub fn new(
response_type: u8,
detail: u8,
event: Window,
mode: u8
) -> FocusInEvent

Constructs a new FocusInEvent response_type must be set to one of: - FOCUS_IN - FOCUS_OUT