Type Definition xcb::xproto::GetGeometryReply

pub type GetGeometryReply = Reply<xcb_get_geometry_reply_t>;

Implementations

impl GetGeometryReply

pub fn depth(&self) -> u8

The depth of the drawable (bits per pixel for the object).

pub fn root(&self) -> Window

Root window of the screen containing drawable.

pub fn x(&self) -> i16

The X coordinate of drawable. If drawable is a window, the coordinate specifies the upper-left outer corner relative to its parent’s origin. If drawable is a pixmap, the X coordinate is always 0.

pub fn y(&self) -> i16

The Y coordinate of drawable. If drawable is a window, the coordinate specifies the upper-left outer corner relative to its parent’s origin. If drawable is a pixmap, the Y coordinate is always 0.

pub fn width(&self) -> u16

The width of drawable.

pub fn height(&self) -> u16

The height of drawable.

pub fn border_width(&self) -> u16

The border width (in pixels).