Type Definition xcb::xproto::ListFontsWithInfoReply

pub type ListFontsWithInfoReply = Reply<xcb_list_fonts_with_info_reply_t>;

Implementations

impl ListFontsWithInfoReply

pub fn name_len(&self) -> u8

The number of matched font names.

pub fn min_bounds(&self) -> Charinfo

minimum bounds over all existing char

pub fn max_bounds(&self) -> Charinfo

maximum bounds over all existing char

pub fn min_char_or_byte2(&self) -> u16

first character

pub fn max_char_or_byte2(&self) -> u16

last character

pub fn default_char(&self) -> u16

char to print for undefined character

pub fn properties_len(&self) -> u16

how many properties there are

pub fn draw_direction(&self) -> u8

pub fn min_byte1(&self) -> u8

pub fn max_byte1(&self) -> u8

pub fn all_chars_exist(&self) -> bool

flag if all characters have nonzero size

pub fn font_ascent(&self) -> i16

baseline to top edge of raster

pub fn font_descent(&self) -> i16

baseline to bottom edge of raster

pub fn replies_hint(&self) -> u32

An indication of how many more fonts will be returned. This is only a hint and may be larger or smaller than the number of fonts actually returned. A zero value does not guarantee that no more fonts will be returned.

pub fn properties(&self) -> FontpropIterator

pub fn name(&self) -> &str