Module xcb::base

Structs

Connection

xcb::Connection handles communication with the X server. It wraps an xcb_connection_t object and will call xcb_disconnect when the Connection goes out of scope

RequestFlags

VoidCookie

Enums

ConnError

Error type that is returned by Connection::has_error

EventQueueOwner

Traits

Cookie

CookieGetReply

RawRequest

Reply

Empty marker trait implemented by all replies types (except ())

Request

RequestWithReply

VoidRequest

Wired

trait for types that can serialize themselves over the X wire.

Xid

A X resource trait

XidNew

Type Definitions

ConnResult