Module xcb::base

Structs

AuthInfo

Container for authentication information to connect to the X server

CheckedVoidCookie

Connection

xcb::Connection handles communication with the X server. It wraps an xcb_connection_t object and will call xcb_disconnect when the Connection goes out of scope

DisplayInfo

SpecialEventId

A struct that serve as an identifier for internal special queue in XCB

VoidCookie

Enums

ConnError

Error type that is returned by Connection::has_error

Error

EventQueueOwner

Constants

COPY_FROM_PARENT

COPY_FROM_PARENT can be used for many CreateWindow fields

CURRENT_TIME

CURRENT_TIME can be used in most requests that take a Timestamp

NO_SYMBOL

NO_SYMBOL fills in unused entries in Keysym tables

Traits

BaseError

marker trait to designate base events (aka. non GE_GENERIC events)

BaseEvent

Trait for base events (aka. non GE_GENERIC events)

CheckedCookie

CheckedCookieWithReply

Cookie

FromWireError

trait for converting wire data to an error Self is normally an enum of several errors subtypes

FromWireEvent

trait for converting wire data to an event enum Self is normally an enum of several event subtypes

FromWireGeEvent

trait for converting wire data to a generic event Self is normally an enum of several event subtypes

GeEvent

marker trait to designate GE_GENERIC events

RawRequest

Reply

Request

RequestWithReply

RequestWithoutReply

UncheckedCookieWithReply

Wired

trait for types that can serialize themselves over the X wire.

Xid

A X resource trait

XidNew

Trait for X resources that can be created directly from Connection::generate_id

Functions

parse_display

@brief Parses a display string name in the form documented by X(7x). @param name The name of the display. @param host A pointer to a malloc’d copy of the hostname. @param display A pointer to the display number. @param screen A pointer to the screen number. @return 0 on failure, non 0 otherwise.

Type Definitions

ConnResult

ProtocolResult

Result