Struct xcb::base::Connection

pub struct Connection {
    // some fields omitted
}

xcb::Connection handles communication with the X server. It wraps an xcb_connection_t object and will call xcb_disconnect when the Connection goes out of scope

Methods

impl Connection

fn flush(&self) -> bool

Forces any buffered output to be written to the server. Blocks until the write is complete.

Return true on success, false otherwise.

fn get_maximum_request_length(&self) -> u32

Returns the maximum request length that this server accepts.

In the absence of the BIG-REQUESTS extension, returns the maximum request length field from the connection setup data, which may be as much as 65535. If the server supports BIG-REQUESTS, then the maximum request length field from the reply to the BigRequestsEnable request will be returned instead.

Note that this length is measured in four-byte units, making the theoretical maximum lengths roughly 256kB without BIG-REQUESTS and 16GB with.

fn wait_for_event(&self) -> Option<GenericEvent>

Returns the next event or error from the server.

Returns the next event or error from the server, or returns None in the event of an I/O error. Blocks until either an event or error arrive, or an I/O error occurs.

fn poll_for_event(&self) -> Option<GenericEvent>

Returns the next event or error from the server.

Returns the next event or error from the server, if one is available, or returns None otherwise. If no event is available, that might be because an I/O error like connection close occurred while attempting to read the next event, in which case the connection is shut down when this function returns.

fn poll_for_queued_event(&self) -> Option<GenericEvent>

Returns the next event without reading from the connection.

This is a version of poll_for_event that only examines the event queue for new events. The function doesn't try to read new events from the connection if no queued events are found.

This function is useful for callers that know in advance that all interesting events have already been read from the connection. For example, callers might use wait_for_reply and be interested only of events that preceded a specific reply.

fn get_setup(&self) -> Setup

Access the data returned by the server.

Accessor for the data returned by the server when the Connection was initialized. This data includes - the server's required format for images, - a list of available visuals, - a list of available screens, - the server's maximum request length (in the absence of the BIG-REQUESTS extension), - and other assorted information.

See the X protocol specification for more details.

fn has_error(&self) -> ConnResult<()>

Test whether the connection has shut down due to a fatal error.

Some errors that occur in the context of a Connection are unrecoverable. When such an error occurs, the connection is shut down and further operations on the Connection have no effect, but memory will not be freed until the Connection is dropped.

fn generate_id(&self) -> u32

Allocates an XID for a new object.

Allocates an XID for a new object. Typically used just prior to various object creation functions, such as xcb::create_window.

fn get_raw_conn(&self) -> *mut xcb_connection_t

Returns the inner ffi xcb_connection_t pointer

fn get_raw_dpy(&self) -> *mut Display

Returns the inner ffi xlib::Display pointer.

fn prefetch_extension_data(&self, ext: &mut Extension)

Prefetch of extension data into the extension cache

This function allows a "prefetch" of extension data into the extension cache. Invoking the function may cause a call to xcb_query_extension, but will not block waiting for the reply. xcb_get_extension_data will return the prefetched data after possibly blocking while it is retrieved.

fn get_extension_data(&self, ext: &mut Extension) -> Option<QueryExtensionData>

Caches reply information from QueryExtension requests.

This function is the primary interface to the "extension cache", which caches reply information from QueryExtension requests. Invoking this function may cause a call to xcb_query_extension to retrieve extension information from the server, and may block until extension data is received from the server.

fn set_event_queue_owner(&self, owner: EventQueueOwner)

Sets the owner of the event queue in the case if the connection is opened with the XLib interface. the default owner is XLib.

fn connect_with_xlib_display() -> ConnResult<(Connection, i32)>

Open a new connection with XLib. The event queue owner defaults to XLib One would need to open an XCB connection with Xlib in order to use OpenGL.

unsafe fn new_from_xlib_display(dpy: *mut Display) -> Connection

wraps a xlib::Display and get an XCB connection from an exisiting object xlib::XCloseDisplay will be called when the returned object is dropped