Struct xcb::base::Connection

pub struct Connection { /* fields omitted */ }

xcb::Connection handles communication with the X server. It wraps an xcb_connection_t object and will call xcb_disconnect when the Connection goes out of scope

Implementations

impl Connection

pub fn flush(&self)

@brief Forces any buffered output to be written to the server. @param c The connection to the X server. @return > @c 0 on success, <= @c 0 otherwise.

Forces any buffered output to be written to the server. Blocks until the write is complete.

pub fn get_maximum_request_length(&self) -> u32

@brief Returns the maximum request length that this server accepts. @param c The connection to the X server. @return The maximum request length field.

In the absence of the BIG-REQUESTS extension, returns the maximum request length field from the connection setup data, which may be as much as 65535. If the server supports BIG-REQUESTS, then the maximum request length field from the reply to the BigRequestsEnable request will be returned instead.

Note that this length is measured in four-byte units, making the theoretical maximum lengths roughly 256kB without BIG-REQUESTS and 16GB with.

pub fn prefetch_maximum_request_length(&self)

@brief Prefetch the maximum request length without blocking. @param c The connection to the X server.

Without blocking, does as much work as possible toward computing the maximum request length accepted by the X server.

Invoking this function may cause a call to xcb_big_requests_enable, but will not block waiting for the reply. xcb_get_maximum_request_length will return the prefetched data after possibly blocking while the reply is retrieved.

Note that in order for this function to be fully non-blocking, the application must previously have called xcb_prefetch_extension_data(c, &xcb_big_requests_id) and the reply must have already arrived.

pub fn get_setup(&self) -> &Setup

Access the data returned by the server.

Accessor for the data returned by the server when the xcb_connection_t was initialized. This data includes

the server’s required format for images,

a list of available visuals,

a list of available screens,

the server’s maximum request length (in the absence of the BIG-REQUESTS extension),

and other assorted information.

See the X protocol specification for more details.

The result must not be freed.

pub fn has_error(&self) -> ConnResult<()>

@brief Test whether the connection has shut down due to a fatal error. @param c The connection. @return > 0 if the connection is in an error state; 0 otherwise.

Some errors that occur in the context of an xcb_connection_t are unrecoverable. When such an error occurs, the connection is shut down and further operations on the xcb_connection_t have no effect, but memory will not be freed until xcb_disconnect() is called on the xcb_connection_t.

pub fn connect(displayname: Option<&str>) -> ConnResult<(Connection, i32)>

Connects to the X server. displayname: The name of the display.

Connects to the X server specified by displayname. If displayname is None, uses the value of the DISPLAY environment variable.

Returns Ok(connection object, preferred screen) in case of success, or Err(ConnError) in case of error. If no screen is preferred, the second member of the tuple is set to 0.

pub fn generate_id<T: XidNew>(&self) -> T

Allocates an XID for a new object.

Allocates an XID for a new object. Typically used just prior to various object creation functions, such as xcb_create_window.

pub fn total_read(&self) -> usize

Obtain number of bytes read from the connection.

Returns cumulative number of bytes received from the connection.

This retrieves the total number of bytes read from this connection, to be used for diagnostic/monitoring/informative purposes.

pub fn total_written(&self) -> usize

Obtain number of bytes written to the connection.

Returns cumulative number of bytes sent to the connection.

This retrieves the total number of bytes written to this connection, to be used for diagnostic/monitoring/informative purposes.

pub fn get_raw_conn(&self) -> *mut xcb_connection_t

Returns the inner ffi xcb_connection_t pointer

pub fn into_raw_conn(self) -> *mut xcb_connection_t

Consumes this object, returning the inner ffi xcb_connection_t pointer

pub fn get_raw_dpy(&self) -> *mut Display

Returns the inner ffi xlib::Display pointer.

pub unsafe fn from_raw_conn(conn: *mut xcb_connection_t) -> Connection

builds a new Connection object from an available connection

Safety

The conn pointer must point to a valid xcb_connection_t