Struct xcb::base::Connection

pub struct Connection { /* fields omitted */ }

xcb::Connection handles communication with the X server. It wraps an xcb_connection_t object and will call xcb_disconnect when the Connection goes out of scope

Implementations

impl Connection

pub fn flush(&self)

@brief Forces any buffered output to be written to the server. @param c The connection to the X server. @return > @c 0 on success, <= @c 0 otherwise.

Forces any buffered output to be written to the server. Blocks until the write is complete.

pub fn get_maximum_request_length(&self) -> u32

@brief Returns the maximum request length that this server accepts. @param c The connection to the X server. @return The maximum request length field.

In the absence of the BIG-REQUESTS extension, returns the maximum request length field from the connection setup data, which may be as much as 65535. If the server supports BIG-REQUESTS, then the maximum request length field from the reply to the BigRequestsEnable request will be returned instead.

Note that this length is measured in four-byte units, making the theoretical maximum lengths roughly 256kB without BIG-REQUESTS and 16GB with.

pub fn prefetch_maximum_request_length(&self)

@brief Prefetch the maximum request length without blocking. @param c The connection to the X server.

Without blocking, does as much work as possible toward computing the maximum request length accepted by the X server.

Invoking this function may cause a call to xcb_big_requests_enable, but will not block waiting for the reply. xcb_get_maximum_request_length will return the prefetched data after possibly blocking while the reply is retrieved.

Note that in order for this function to be fully non-blocking, the application must previously have called xcb__extension_data(c, &xcb_big_requests_id) and the reply must have already arrived.

pub fn wait_for_event(&self) -> Result<ProtocolEvent>

@brief Returns the next event or error from the server. @param c The connection to the X server. @return The next event from the server.

Returns the next event or error from the server, or returns null in the event of an I/O error. Blocks until either an event or error arrive, or an I/O error occurs.

pub fn poll_for_event(
&self
) -> ConnResult<Option<ProtocolResult<ProtocolEvent>>>

@brief Returns the next event or error from the server. @param c The connection to the X server. @return The next event from the server.

Returns the next event or error from the server, if one is available, or returns @c NULL otherwise. If no event is available, that might be because an I/O error like connection close occurred while attempting to read the next event, in which case the connection is shut down when this function returns.

pub fn poll_for_queued_event(&self) -> Option<ProtocolResult<ProtocolEvent>>

@brief Returns the next event without reading from the connection. @param c The connection to the X server. @return The next already queued event from the server.

This is a version of xcb_poll_for_event that only examines the event queue for new events. The function doesn’t try to read new events from the connection if no queued events are found.

This function is useful for callers that know in advance that all interesting events have already been read from the connection. For example, callers might use xcb_wait_for_reply and be interested only of events that preceded a specific reply.

pub fn register_for_special_xge<XGE: GeEvent>(&self) -> SpecialEventId

Start listening for a special event. Effectively creates an internal special queue for this event XGE events are only defined in the xinput and present extensions

pub fn unregister_for_special_xge(&self, se: SpecialEventId)

Stop listening to a special event

pub fn wait_for_special_event(
&self,
se: SpecialEventId
) -> Result<ProtocolEvent>

Returns the next event from a special queue, blocking until one arrives

pub fn poll_for_special_event(
&self,
se: SpecialEventId
) -> ConnResult<Option<ProtocolResult<ProtocolEvent>>>

Returns the next event from a special queue

pub fn get_setup(&self) -> &Setup

Access the data returned by the server.

Accessor for the data returned by the server when the xcb_connection_t was initialized. This data includes

the server’s required format for images,

a list of available visuals,

a list of available screens,

the server’s maximum request length (in the absence of the BIG-REQUESTS extension),

and other assorted information.

See the X protocol specification for more details.

The result must not be freed.

pub fn has_error(&self) -> ConnResult<()>

@brief Test whether the connection has shut down due to a fatal error. @param c The connection. @return > 0 if the connection is in an error state; 0 otherwise.

Some errors that occur in the context of an xcb_connection_t are unrecoverable. When such an error occurs, the connection is shut down and further operations on the xcb_connection_t have no effect, but memory will not be freed until xcb_disconnect() is called on the xcb_connection_t.

pub fn send_request<R>(&self, req: &R) -> R::Cookie where
R: Request,

Send a request to the X server.

If the request expect a reply, the cookie can be used to retrieve the reply

pub fn send_request_checked<R>(&self, req: &R) -> CheckedVoidCookie where
R: RequestWithoutReply,

Send a checked request to the X server.

Checked requests do not expect a reply, but the returned cookie can be used to check for errors using Connection::check_request.

pub fn send_request_unchecked<R>(&self, req: &R) -> R::UncheckedCookie where
R: RequestWithReply,

Send an unchecked request to the X server.

Unchecked requests expect a reply that is to be retrieved by Connection::get_reply_unchecked. Unchecked means that the error is not checked when the reply is fetched. Instead, the error will be sent to the event loop

pub fn check_request(&self, cookie: CheckedVoidCookie) -> ProtocolResult<()>

Check a checked request for errors.

@brief Return the error for a request, or NULL if none can ever arrive. @param c The connection to the X server. @param cookie The request cookie. @return The error for the request, or NULL if none can ever arrive.

The xcb_void_cookie_t cookie supplied to this function must have resulted from a call to xcb_[request_name]_checked(). This function will block until one of two conditions happens. If an error is received, it will be returned. If a reply to a subsequent request has already arrived, no error can arrive for this request, so this function will return NULL.

Note that this function will perform a sync if needed to ensure that the sequence number will advance beyond that provided in cookie; this is a convenience to avoid races in determining whether the sync is needed.

pub fn get_reply<C>(&self, cookie: C) -> Result<C::Reply> where
C: CheckedCookieWithReply,

Get the reply of a previous request

pub fn get_reply_unchecked<C>(&self, cookie: C) -> ConnResult<Option<C::Reply>> where
C: UncheckedCookieWithReply,

Get the reply of a previous unchecked request. If an error occured, None is returned and the error will be delivered to the event loop.

pub fn connect(display_name: Option<&str>) -> ConnResult<(Connection, i32)>

Connects to the X server. display_name: The name of the display.

Connects to the X server specified by display_name. If display_name is None, uses the value of the DISPLAY environment variable.

Returns Ok(connection object, preferred screen) in case of success, or Err(ConnError) in case of error. If no screen is preferred, the second member of the tuple is set to 0.

pub fn connect_to_fd(fd: RawFd, auth_info: AuthInfo<'_>) -> ConnResult<Self>

@brief Connects to the X server. @param fd The file descriptor. @param auth_info Authentication data. @return A newly allocated xcb_connection_t structure.

Connects to an X server, given the open socket @p fd and the xcb_auth_info_t @p auth_info. The file descriptor @p fd is bidirectionally connected to an X server. If the connection should be unauthenticated, @p auth_info must be @c NULL.

Always returns a non-NULL pointer to a xcb_connection_t, even on failure. Callers need to use xcb_connection_has_error() to check for failure. When finished, use xcb_disconnect() to close the connection and free the structure.

pub fn connect_to_display_with_auth_info(
display_name: Option<&str>,
auth_info: AuthInfo<'_>
) -> ConnResult<(Connection, i32)>

@brief Connects to the X server, using an authorization information. @param display The name of the display. @param auth The authorization information. @param screen A pointer to a preferred screen number. @return A newly allocated xcb_connection_t structure.

Connects to the X server specified by @p displayname, using the authorization @p auth. If a particular screen on that server is preferred, the int pointed to by @p screenp (if not @c NULL) will be set to that screen; otherwise @p screenp will be set to 0.

Always returns a non-NULL pointer to a xcb_connection_t, even on failure. Callers need to use xcb_connection_has_error() to check for failure. When finished, use xcb_disconnect() to close the connection and free the structure.

pub fn generate_id<T: XidNew>(&self) -> T

Allocates an XID for a new object.

Returned value is typically used in requests such as CreateWindow.

pub fn total_read(&self) -> usize

Obtain number of bytes read from the connection.

Returns cumulative number of bytes received from the connection.

This retrieves the total number of bytes read from this connection, to be used for diagnostic/monitoring/informative purposes.

pub fn total_written(&self) -> usize

Obtain number of bytes written to the connection.

Returns cumulative number of bytes sent to the connection.

This retrieves the total number of bytes written to this connection, to be used for diagnostic/monitoring/informative purposes.

pub fn get_raw_conn(&self) -> *mut xcb_connection_t

Returns the inner ffi xcb_connection_t pointer

pub fn into_raw_conn(self) -> *mut xcb_connection_t

Consumes this object, returning the inner ffi xcb_connection_t pointer

pub fn get_raw_dpy(&self) -> *mut Display

Returns the inner ffi xlib::Display pointer.

pub unsafe fn from_raw_conn(conn: *mut xcb_connection_t) -> Connection

builds a new Connection object from an available connection

Safety

The conn pointer must point to a valid xcb_connection_t