Struct xcb::base::Event

pub struct Event<T> {
    pub ptr: *mut T,
}

Event wraps a pointer to xcb_*_event_t this pointer will be freed when the Event goes out of scope

Fields
ptr: *mut T
Implementations

impl<T> Event<T>

pub fn response_type(&self) -> u8

impl Event<xcb_key_press_event_t>

pub fn detail(&self) -> Keycode

The keycode (a number representing a physical key on the keyboard) of the key which was pressed.

pub fn time(&self) -> Timestamp

Time when the event was generated (in milliseconds).

pub fn root(&self) -> Window

The root window of child.

pub fn event(&self) -> Window

pub fn child(&self) -> Window

pub fn root_x(&self) -> i16

The X coordinate of the pointer relative to the root window at the time of the event.

pub fn root_y(&self) -> i16

The Y coordinate of the pointer relative to the root window at the time of the event.

pub fn event_x(&self) -> i16

If same_screen is true, this is the X coordinate relative to the event window’s origin. Otherwise, event_x will be set to zero.

pub fn event_y(&self) -> i16

If same_screen is true, this is the Y coordinate relative to the event window’s origin. Otherwise, event_y will be set to zero.

pub fn state(&self) -> u16

The logical state of the pointer buttons and modifier keys just prior to the event.

pub fn same_screen(&self) -> bool

Whether the event window is on the same screen as the root window.

pub fn new( response_type: u8, detail: Keycode, time: Timestamp, root: Window, event: Window, child: Window, root_x: i16, root_y: i16, event_x: i16, event_y: i16, state: u16, same_screen: bool ) -> KeyPressEvent

Constructs a new KeyPressEvent response_type must be set to one of: - KEY_PRESS - KEY_RELEASE

impl Event<xcb_button_press_event_t>

pub fn detail(&self) -> Button

The keycode (a number representing a physical key on the keyboard) of the key which was pressed.

pub fn time(&self) -> Timestamp

Time when the event was generated (in milliseconds).

pub fn root(&self) -> Window

The root window of child.

pub fn event(&self) -> Window

pub fn child(&self) -> Window

pub fn root_x(&self) -> i16

The X coordinate of the pointer relative to the root window at the time of the event.

pub fn root_y(&self) -> i16

The Y coordinate of the pointer relative to the root window at the time of the event.

pub fn event_x(&self) -> i16

If same_screen is true, this is the X coordinate relative to the event window’s origin. Otherwise, event_x will be set to zero.

pub fn event_y(&self) -> i16

If same_screen is true, this is the Y coordinate relative to the event window’s origin. Otherwise, event_y will be set to zero.

pub fn state(&self) -> u16

The logical state of the pointer buttons and modifier keys just prior to the event.

pub fn same_screen(&self) -> bool

Whether the event window is on the same screen as the root window.

pub fn new( response_type: u8, detail: Button, time: Timestamp, root: Window, event: Window, child: Window, root_x: i16, root_y: i16, event_x: i16, event_y: i16, state: u16, same_screen: bool ) -> ButtonPressEvent

Constructs a new ButtonPressEvent response_type must be set to one of: - BUTTON_PRESS - BUTTON_RELEASE

impl Event<xcb_motion_notify_event_t>

pub fn detail(&self) -> u8

The keycode (a number representing a physical key on the keyboard) of the key which was pressed.

pub fn time(&self) -> Timestamp

Time when the event was generated (in milliseconds).

pub fn root(&self) -> Window

The root window of child.

pub fn event(&self) -> Window

pub fn child(&self) -> Window

pub fn root_x(&self) -> i16

The X coordinate of the pointer relative to the root window at the time of the event.

pub fn root_y(&self) -> i16

The Y coordinate of the pointer relative to the root window at the time of the event.

pub fn event_x(&self) -> i16

If same_screen is true, this is the X coordinate relative to the event window’s origin. Otherwise, event_x will be set to zero.

pub fn event_y(&self) -> i16

If same_screen is true, this is the Y coordinate relative to the event window’s origin. Otherwise, event_y will be set to zero.

pub fn state(&self) -> u16

The logical state of the pointer buttons and modifier keys just prior to the event.

pub fn same_screen(&self) -> bool

Whether the event window is on the same screen as the root window.

pub fn new( detail: u8, time: Timestamp, root: Window, event: Window, child: Window, root_x: i16, root_y: i16, event_x: i16, event_y: i16, state: u16, same_screen: bool ) -> MotionNotifyEvent

Constructs a new MotionNotifyEvent response_type will be set automatically to MOTION_NOTIFY

impl Event<xcb_enter_notify_event_t>

pub fn detail(&self) -> u8

pub fn time(&self) -> Timestamp

pub fn root(&self) -> Window

The root window for the final cursor position.

pub fn event(&self) -> Window

The window on which the event was generated.

pub fn child(&self) -> Window

If the event window has subwindows and the final pointer position is in one of them, then child is set to that subwindow, XCB_WINDOW_NONE otherwise.

pub fn root_x(&self) -> i16

The pointer X coordinate relative to root’s origin at the time of the event.

pub fn root_y(&self) -> i16

The pointer Y coordinate relative to root’s origin at the time of the event.

pub fn event_x(&self) -> i16

If event is on the same screen as root, this is the pointer X coordinate relative to the event window’s origin.

pub fn event_y(&self) -> i16

If event is on the same screen as root, this is the pointer Y coordinate relative to the event window’s origin.

pub fn state(&self) -> u16

pub fn mode(&self) -> u8

pub fn same_screen_focus(&self) -> u8

pub fn new( response_type: u8, detail: u8, time: Timestamp, root: Window, event: Window, child: Window, root_x: i16, root_y: i16, event_x: i16, event_y: i16, state: u16, mode: u8, same_screen_focus: u8 ) -> EnterNotifyEvent

Constructs a new EnterNotifyEvent response_type must be set to one of: - ENTER_NOTIFY - LEAVE_NOTIFY

impl Event<xcb_focus_in_event_t>

pub fn detail(&self) -> u8

pub fn event(&self) -> Window

The window on which the focus event was generated. This is the window used by the X server to report the event.

pub fn mode(&self) -> u8

pub fn new( response_type: u8, detail: u8, event: Window, mode: u8 ) -> FocusInEvent

Constructs a new FocusInEvent response_type must be set to one of: - FOCUS_IN - FOCUS_OUT

impl Event<xcb_keymap_notify_event_t>

pub fn keys(&self) -> &[u8]

pub fn new(keys: [u8; 31]) -> KeymapNotifyEvent

Constructs a new KeymapNotifyEvent response_type will be set automatically to KEYMAP_NOTIFY

impl Event<xcb_expose_event_t>

pub fn window(&self) -> Window

The exposed (damaged) window.

pub fn x(&self) -> u16

The X coordinate of the left-upper corner of the exposed rectangle, relative to the window’s origin.

pub fn y(&self) -> u16

The Y coordinate of the left-upper corner of the exposed rectangle, relative to the window’s origin.

pub fn width(&self) -> u16

The width of the exposed rectangle.

pub fn height(&self) -> u16

The height of the exposed rectangle.

pub fn count(&self) -> u16

The amount of Expose events following this one. Simple applications that do not want to optimize redisplay by distinguishing between subareas of its window can just ignore all Expose events with nonzero counts and perform full redisplays on events with zero counts.

pub fn new( window: Window, x: u16, y: u16, width: u16, height: u16, count: u16 ) -> ExposeEvent

Constructs a new ExposeEvent response_type will be set automatically to EXPOSE

impl Event<xcb_graphics_exposure_event_t>

pub fn drawable(&self) -> Drawable

pub fn x(&self) -> u16

pub fn y(&self) -> u16

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn minor_opcode(&self) -> u16

pub fn count(&self) -> u16

pub fn major_opcode(&self) -> u8

pub fn new( drawable: Drawable, x: u16, y: u16, width: u16, height: u16, minor_opcode: u16, count: u16, major_opcode: u8 ) -> GraphicsExposureEvent

Constructs a new GraphicsExposureEvent response_type will be set automatically to GRAPHICS_EXPOSURE

impl Event<xcb_no_exposure_event_t>

pub fn drawable(&self) -> Drawable

pub fn minor_opcode(&self) -> u16

pub fn major_opcode(&self) -> u8

pub fn new( drawable: Drawable, minor_opcode: u16, major_opcode: u8 ) -> NoExposureEvent

Constructs a new NoExposureEvent response_type will be set automatically to NO_EXPOSURE

impl Event<xcb_visibility_notify_event_t>

pub fn window(&self) -> Window

pub fn state(&self) -> u8

pub fn new(window: Window, state: u8) -> VisibilityNotifyEvent

Constructs a new VisibilityNotifyEvent response_type will be set automatically to VISIBILITY_NOTIFY

impl Event<xcb_create_notify_event_t>

pub fn parent(&self) -> Window

pub fn window(&self) -> Window

pub fn x(&self) -> i16

pub fn y(&self) -> i16

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn border_width(&self) -> u16

pub fn override_redirect(&self) -> bool

pub fn new( parent: Window, window: Window, x: i16, y: i16, width: u16, height: u16, border_width: u16, override_redirect: bool ) -> CreateNotifyEvent

Constructs a new CreateNotifyEvent response_type will be set automatically to CREATE_NOTIFY

impl Event<xcb_destroy_notify_event_t>

pub fn event(&self) -> Window

The reconfigured window or its parent, depending on whether StructureNotify or SubstructureNotify was selected.

pub fn window(&self) -> Window

The window that is destroyed.

pub fn new(event: Window, window: Window) -> DestroyNotifyEvent

Constructs a new DestroyNotifyEvent response_type will be set automatically to DESTROY_NOTIFY

impl Event<xcb_unmap_notify_event_t>

pub fn event(&self) -> Window

The reconfigured window or its parent, depending on whether StructureNotify or SubstructureNotify was selected.

pub fn window(&self) -> Window

The window that was unmapped.

pub fn from_configure(&self) -> bool

Set to 1 if the event was generated as a result of a resizing of the window’s parent when window had a win_gravity of UnmapGravity.

pub fn new( event: Window, window: Window, from_configure: bool ) -> UnmapNotifyEvent

Constructs a new UnmapNotifyEvent response_type will be set automatically to UNMAP_NOTIFY

impl Event<xcb_map_notify_event_t>

pub fn event(&self) -> Window

The window which was mapped or its parent, depending on whether StructureNotify or SubstructureNotify was selected.

pub fn window(&self) -> Window

The window that was mapped.

pub fn override_redirect(&self) -> bool

Window managers should ignore this window if override_redirect is 1.

pub fn new( event: Window, window: Window, override_redirect: bool ) -> MapNotifyEvent

Constructs a new MapNotifyEvent response_type will be set automatically to MAP_NOTIFY

impl Event<xcb_map_request_event_t>

pub fn parent(&self) -> Window

The parent of window.

pub fn window(&self) -> Window

The window to be mapped.

pub fn new(parent: Window, window: Window) -> MapRequestEvent

Constructs a new MapRequestEvent response_type will be set automatically to MAP_REQUEST

impl Event<xcb_reparent_notify_event_t>

pub fn event(&self) -> Window

pub fn window(&self) -> Window

pub fn parent(&self) -> Window

pub fn x(&self) -> i16

pub fn y(&self) -> i16

pub fn override_redirect(&self) -> bool

pub fn new( event: Window, window: Window, parent: Window, x: i16, y: i16, override_redirect: bool ) -> ReparentNotifyEvent

Constructs a new ReparentNotifyEvent response_type will be set automatically to REPARENT_NOTIFY

impl Event<xcb_configure_notify_event_t>

pub fn event(&self) -> Window

The reconfigured window or its parent, depending on whether StructureNotify or SubstructureNotify was selected.

pub fn window(&self) -> Window

The window whose size, position, border, and/or stacking order was changed.

pub fn above_sibling(&self) -> Window

If XCB_NONE, the window is on the bottom of the stack with respect to sibling windows. However, if set to a sibling window, the window is placed on top of this sibling window.

pub fn x(&self) -> i16

The X coordinate of the upper-left outside corner of window, relative to the parent window’s origin.

pub fn y(&self) -> i16

The Y coordinate of the upper-left outside corner of window, relative to the parent window’s origin.

pub fn width(&self) -> u16

The inside width of window, not including the border.

pub fn height(&self) -> u16

The inside height of window, not including the border.

pub fn border_width(&self) -> u16

The border width of window.

pub fn override_redirect(&self) -> bool

Window managers should ignore this window if override_redirect is 1.

pub fn new( event: Window, window: Window, above_sibling: Window, x: i16, y: i16, width: u16, height: u16, border_width: u16, override_redirect: bool ) -> ConfigureNotifyEvent

Constructs a new ConfigureNotifyEvent response_type will be set automatically to CONFIGURE_NOTIFY

impl Event<xcb_configure_request_event_t>

pub fn stack_mode(&self) -> u8

pub fn parent(&self) -> Window

pub fn window(&self) -> Window

pub fn sibling(&self) -> Window

pub fn x(&self) -> i16

pub fn y(&self) -> i16

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn border_width(&self) -> u16

pub fn value_mask(&self) -> u16

pub fn new( stack_mode: u8, parent: Window, window: Window, sibling: Window, x: i16, y: i16, width: u16, height: u16, border_width: u16, value_mask: u16 ) -> ConfigureRequestEvent

Constructs a new ConfigureRequestEvent response_type will be set automatically to CONFIGURE_REQUEST

impl Event<xcb_gravity_notify_event_t>

pub fn event(&self) -> Window

pub fn window(&self) -> Window

pub fn x(&self) -> i16

pub fn y(&self) -> i16

pub fn new(event: Window, window: Window, x: i16, y: i16) -> GravityNotifyEvent

Constructs a new GravityNotifyEvent response_type will be set automatically to GRAVITY_NOTIFY

impl Event<xcb_resize_request_event_t>

pub fn window(&self) -> Window

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn new(window: Window, width: u16, height: u16) -> ResizeRequestEvent

Constructs a new ResizeRequestEvent response_type will be set automatically to RESIZE_REQUEST

impl Event<xcb_circulate_notify_event_t>

pub fn event(&self) -> Window

Either the restacked window or its parent, depending on whether StructureNotify or SubstructureNotify was selected.

pub fn window(&self) -> Window

The restacked window.

pub fn place(&self) -> u8

pub fn new( response_type: u8, event: Window, window: Window, place: u8 ) -> CirculateNotifyEvent

Constructs a new CirculateNotifyEvent response_type must be set to one of: - CIRCULATE_NOTIFY - CIRCULATE_REQUEST

impl Event<xcb_property_notify_event_t>

pub fn window(&self) -> Window

The window whose associated property was changed.

pub fn atom(&self) -> Atom

The property’s atom, to indicate which property was changed.

pub fn time(&self) -> Timestamp

A timestamp of the server time when the property was changed.

pub fn state(&self) -> u8

pub fn new( window: Window, atom: Atom, time: Timestamp, state: u8 ) -> PropertyNotifyEvent

Constructs a new PropertyNotifyEvent response_type will be set automatically to PROPERTY_NOTIFY

impl Event<xcb_selection_clear_event_t>

pub fn time(&self) -> Timestamp

pub fn owner(&self) -> Window

pub fn selection(&self) -> Atom

pub fn new( time: Timestamp, owner: Window, selection: Atom ) -> SelectionClearEvent

Constructs a new SelectionClearEvent response_type will be set automatically to SELECTION_CLEAR

impl Event<xcb_selection_request_event_t>

pub fn time(&self) -> Timestamp

pub fn owner(&self) -> Window

pub fn requestor(&self) -> Window

pub fn selection(&self) -> Atom

pub fn target(&self) -> Atom

pub fn property(&self) -> Atom

pub fn new( time: Timestamp, owner: Window, requestor: Window, selection: Atom, target: Atom, property: Atom ) -> SelectionRequestEvent

Constructs a new SelectionRequestEvent response_type will be set automatically to SELECTION_REQUEST

impl Event<xcb_selection_notify_event_t>

pub fn time(&self) -> Timestamp

pub fn requestor(&self) -> Window

pub fn selection(&self) -> Atom

pub fn target(&self) -> Atom

pub fn property(&self) -> Atom

pub fn new( time: Timestamp, requestor: Window, selection: Atom, target: Atom, property: Atom ) -> SelectionNotifyEvent

Constructs a new SelectionNotifyEvent response_type will be set automatically to SELECTION_NOTIFY

impl Event<xcb_colormap_notify_event_t>

pub fn window(&self) -> Window

The window whose associated colormap is changed, installed or uninstalled.

pub fn colormap(&self) -> Colormap

The colormap which is changed, installed or uninstalled. This is XCB_NONE when the colormap is changed by a call to FreeColormap.

pub fn new_(&self) -> bool

pub fn state(&self) -> u8

pub fn new( window: Window, colormap: Colormap, new_: bool, state: u8 ) -> ColormapNotifyEvent

Constructs a new ColormapNotifyEvent response_type will be set automatically to COLORMAP_NOTIFY

impl Event<xcb_client_message_event_t>

pub fn format(&self) -> u8

Specifies how to interpret data. Can be either 8, 16 or 32.

pub fn window(&self) -> Window

pub fn type_(&self) -> Atom

An atom which indicates how the data should be interpreted by the receiving client.

pub fn data<'a>(&'a self) -> &'a ClientMessageData

The data itself (20 bytes max).

pub fn new( format: u8, window: Window, type_: Atom, data: ClientMessageData ) -> ClientMessageEvent

Constructs a new ClientMessageEvent response_type will be set automatically to CLIENT_MESSAGE

impl Event<xcb_mapping_notify_event_t>

pub fn request(&self) -> u8

pub fn first_keycode(&self) -> Keycode

The first number in the range of the altered mapping.

pub fn count(&self) -> u8

The number of keycodes altered.

pub fn new(request: u8, first_keycode: Keycode, count: u8) -> MappingNotifyEvent

Constructs a new MappingNotifyEvent response_type will be set automatically to MAPPING_NOTIFY

impl Event<xcb_ge_generic_event_t>

pub fn new() -> GeGenericEvent

Constructs a new GeGenericEvent response_type will be set automatically to GE_GENERIC

impl Event<xcb_shape_notify_event_t>

pub fn shape_kind(&self) -> Kind

pub fn affected_window(&self) -> Window

pub fn extents_x(&self) -> i16

pub fn extents_y(&self) -> i16

pub fn extents_width(&self) -> u16

pub fn extents_height(&self) -> u16

pub fn server_time(&self) -> Timestamp

pub fn shaped(&self) -> bool

pub fn new( shape_kind: Kind, affected_window: Window, extents_x: i16, extents_y: i16, extents_width: u16, extents_height: u16, server_time: Timestamp, shaped: bool ) -> NotifyEvent

Constructs a new NotifyEvent response_type will be set automatically to NOTIFY

impl Event<xcb_xfixes_selection_notify_event_t>

pub fn subtype(&self) -> u8

pub fn window(&self) -> Window

pub fn owner(&self) -> Window

pub fn selection(&self) -> Atom

pub fn timestamp(&self) -> Timestamp

pub fn selection_timestamp(&self) -> Timestamp

pub fn new( subtype: u8, window: Window, owner: Window, selection: Atom, timestamp: Timestamp, selection_timestamp: Timestamp ) -> SelectionNotifyEvent

Constructs a new SelectionNotifyEvent response_type will be set automatically to SELECTION_NOTIFY

impl Event<xcb_xfixes_cursor_notify_event_t>

pub fn subtype(&self) -> u8

pub fn window(&self) -> Window

pub fn cursor_serial(&self) -> u32

pub fn timestamp(&self) -> Timestamp

pub fn name(&self) -> Atom

pub fn new( subtype: u8, window: Window, cursor_serial: u32, timestamp: Timestamp, name: Atom ) -> CursorNotifyEvent

Constructs a new CursorNotifyEvent response_type will be set automatically to CURSOR_NOTIFY