Struct xcb::base::Reply

pub struct Reply<T> {
    pub ptr: *mut T,
}

Wraps a pointer to a xcb_*_reply_t the pointer is freed when the Reply goes out of scope

Fields
ptr: *mut T
Implementations

impl Reply<xcb_get_window_attributes_reply_t>

pub fn backing_store(&self) -> u8

pub fn visual(&self) -> Visualid

The associated visual structure of window.

pub fn class(&self) -> u16

pub fn bit_gravity(&self) -> u8

pub fn win_gravity(&self) -> u8

pub fn backing_planes(&self) -> u32

Planes to be preserved if possible.

pub fn backing_pixel(&self) -> u32

Value to be used when restoring planes.

pub fn save_under(&self) -> bool

Boolean, should bits under be saved?

pub fn map_is_installed(&self) -> bool

pub fn map_state(&self) -> u8

pub fn override_redirect(&self) -> bool

Window managers should ignore this window if override_redirect is 1.

pub fn colormap(&self) -> Colormap

Color map to be associated with window.

pub fn all_event_masks(&self) -> u32

Set of events all people have interest in.

pub fn your_event_mask(&self) -> u32

My event mask.

pub fn do_not_propagate_mask(&self) -> u16

Set of events that should not propagate.

impl Reply<xcb_get_geometry_reply_t>

pub fn depth(&self) -> u8

The depth of the drawable (bits per pixel for the object).

pub fn root(&self) -> Window

Root window of the screen containing drawable.

pub fn x(&self) -> i16

The X coordinate of drawable. If drawable is a window, the coordinate specifies the upper-left outer corner relative to its parent’s origin. If drawable is a pixmap, the X coordinate is always 0.

pub fn y(&self) -> i16

The Y coordinate of drawable. If drawable is a window, the coordinate specifies the upper-left outer corner relative to its parent’s origin. If drawable is a pixmap, the Y coordinate is always 0.

pub fn width(&self) -> u16

The width of drawable.

pub fn height(&self) -> u16

The height of drawable.

pub fn border_width(&self) -> u16

The border width (in pixels).

impl Reply<xcb_query_tree_reply_t>

pub fn root(&self) -> Window

The root window of window.

pub fn parent(&self) -> Window

The parent window of window.

pub fn children_len(&self) -> u16

The number of child windows.

pub fn children(&self) -> &[Window]

impl Reply<xcb_intern_atom_reply_t>

pub fn atom(&self) -> Atom

impl Reply<xcb_get_atom_name_reply_t>

pub fn name_len(&self) -> u16

pub fn name(&self) -> &str

impl Reply<xcb_get_property_reply_t>

pub fn format(&self) -> u8

Specifies whether the data should be viewed as a list of 8-bit, 16-bit, or 32-bit quantities. Possible values are 8, 16, and 32. This information allows the X server to correctly perform byte-swap operations as necessary.

pub fn type_(&self) -> Atom

The actual type of the property (an atom).

pub fn bytes_after(&self) -> u32

The number of bytes remaining to be read in the property if a partial read was performed.

pub fn value_len(&self) -> u32

The length of value. You should use the corresponding accessor instead of this field.

pub fn value<T>(&self) -> &[T]

impl Reply<xcb_list_properties_reply_t>

pub fn atoms_len(&self) -> u16

pub fn atoms(&self) -> &[Atom]

impl Reply<xcb_get_selection_owner_reply_t>

pub fn owner(&self) -> Window

The current selection owner window.

impl Reply<xcb_grab_pointer_reply_t>

pub fn status(&self) -> u8

impl Reply<xcb_grab_keyboard_reply_t>

pub fn status(&self) -> u8

impl Reply<xcb_query_pointer_reply_t>

pub fn same_screen(&self) -> bool

If same_screen is False, then the pointer is not on the same screen as the argument window, child is None, and win_x and win_y are zero. If same_screen is True, then win_x and win_y are the pointer coordinates relative to the argument window’s origin, and child is the child containing the pointer, if any.

pub fn root(&self) -> Window

The root window the pointer is logically on.

pub fn child(&self) -> Window

The child window containing the pointer, if any, if same_screen is true. If same_screen is false, XCB_NONE is returned.

pub fn root_x(&self) -> i16

The pointer X position, relative to root.

pub fn root_y(&self) -> i16

The pointer Y position, relative to root.

pub fn win_x(&self) -> i16

The pointer X coordinate, relative to child, if same_screen is true. Zero otherwise.

pub fn win_y(&self) -> i16

The pointer Y coordinate, relative to child, if same_screen is true. Zero otherwise.

pub fn mask(&self) -> u16

The current logical state of the modifier keys and the buttons. Note that the logical state of a device (as seen by means of the protocol) may lag the physical state if device event processing is frozen.

impl Reply<xcb_get_motion_events_reply_t>

pub fn events_len(&self) -> u32

pub fn events(&self) -> TimecoordIterator

impl Reply<xcb_translate_coordinates_reply_t>

pub fn same_screen(&self) -> bool

pub fn child(&self) -> Window

pub fn dst_x(&self) -> i16

pub fn dst_y(&self) -> i16

impl Reply<xcb_get_input_focus_reply_t>

pub fn revert_to(&self) -> u8

pub fn focus(&self) -> Window

impl Reply<xcb_query_keymap_reply_t>

pub fn keys(&self) -> &[u8]

impl Reply<xcb_query_font_reply_t>

pub fn min_bounds(&self) -> Charinfo

minimum bounds over all existing char

pub fn max_bounds(&self) -> Charinfo

maximum bounds over all existing char

pub fn min_char_or_byte2(&self) -> u16

first character

pub fn max_char_or_byte2(&self) -> u16

last character

pub fn default_char(&self) -> u16

char to print for undefined character

pub fn properties_len(&self) -> u16

how many properties there are

pub fn draw_direction(&self) -> u8

pub fn min_byte1(&self) -> u8

pub fn max_byte1(&self) -> u8

pub fn all_chars_exist(&self) -> bool

flag if all characters have nonzero size

pub fn font_ascent(&self) -> i16

baseline to top edge of raster

pub fn font_descent(&self) -> i16

baseline to bottom edge of raster

pub fn char_infos_len(&self) -> u32

pub fn properties(&self) -> FontpropIterator

pub fn char_infos(&self) -> CharinfoIterator

impl Reply<xcb_query_text_extents_reply_t>

pub fn draw_direction(&self) -> u8

pub fn font_ascent(&self) -> i16

pub fn font_descent(&self) -> i16

pub fn overall_ascent(&self) -> i16

pub fn overall_descent(&self) -> i16

pub fn overall_width(&self) -> i32

pub fn overall_left(&self) -> i32

pub fn overall_right(&self) -> i32

impl Reply<xcb_list_fonts_reply_t>

pub fn names_len(&self) -> u16

The number of font names.

pub fn names(&self) -> StrIterator<'_>

impl Reply<xcb_list_fonts_with_info_reply_t>

pub fn name_len(&self) -> u8

The number of matched font names.

pub fn min_bounds(&self) -> Charinfo

minimum bounds over all existing char

pub fn max_bounds(&self) -> Charinfo

maximum bounds over all existing char

pub fn min_char_or_byte2(&self) -> u16

first character

pub fn max_char_or_byte2(&self) -> u16

last character

pub fn default_char(&self) -> u16

char to print for undefined character

pub fn properties_len(&self) -> u16

how many properties there are

pub fn draw_direction(&self) -> u8

pub fn min_byte1(&self) -> u8

pub fn max_byte1(&self) -> u8

pub fn all_chars_exist(&self) -> bool

flag if all characters have nonzero size

pub fn font_ascent(&self) -> i16

baseline to top edge of raster

pub fn font_descent(&self) -> i16

baseline to bottom edge of raster

pub fn replies_hint(&self) -> u32

An indication of how many more fonts will be returned. This is only a hint and may be larger or smaller than the number of fonts actually returned. A zero value does not guarantee that no more fonts will be returned.

pub fn properties(&self) -> FontpropIterator

pub fn name(&self) -> &str

impl Reply<xcb_get_font_path_reply_t>

pub fn path_len(&self) -> u16

pub fn path(&self) -> StrIterator<'_>

impl Reply<xcb_get_image_reply_t>

pub fn depth(&self) -> u8

pub fn visual(&self) -> Visualid

pub fn data(&self) -> &[u8]

impl Reply<xcb_list_installed_colormaps_reply_t>

pub fn cmaps_len(&self) -> u16

pub fn cmaps(&self) -> &[Colormap]

impl Reply<xcb_alloc_color_reply_t>

pub fn red(&self) -> u16

pub fn green(&self) -> u16

pub fn blue(&self) -> u16

pub fn pixel(&self) -> u32

impl Reply<xcb_alloc_named_color_reply_t>

pub fn pixel(&self) -> u32

pub fn exact_red(&self) -> u16

pub fn exact_green(&self) -> u16

pub fn exact_blue(&self) -> u16

pub fn visual_red(&self) -> u16

pub fn visual_green(&self) -> u16

pub fn visual_blue(&self) -> u16

impl Reply<xcb_alloc_color_cells_reply_t>

pub fn pixels_len(&self) -> u16

pub fn masks_len(&self) -> u16

pub fn pixels(&self) -> &[u32]

pub fn masks(&self) -> &[u32]

impl Reply<xcb_alloc_color_planes_reply_t>

pub fn pixels_len(&self) -> u16

pub fn red_mask(&self) -> u32

pub fn green_mask(&self) -> u32

pub fn blue_mask(&self) -> u32

pub fn pixels(&self) -> &[u32]

impl Reply<xcb_query_colors_reply_t>

pub fn colors_len(&self) -> u16

pub fn colors(&self) -> RgbIterator

impl Reply<xcb_lookup_color_reply_t>

pub fn exact_red(&self) -> u16

pub fn exact_green(&self) -> u16

pub fn exact_blue(&self) -> u16

pub fn visual_red(&self) -> u16

pub fn visual_green(&self) -> u16

pub fn visual_blue(&self) -> u16

impl Reply<xcb_query_best_size_reply_t>

pub fn width(&self) -> u16

pub fn height(&self) -> u16

impl Reply<xcb_query_extension_reply_t>

pub fn present(&self) -> bool

Whether the extension is present on this X11 server.

pub fn major_opcode(&self) -> u8

The major opcode for requests.

pub fn first_event(&self) -> u8

The first event code, if any.

pub fn first_error(&self) -> u8

The first error code, if any.

impl Reply<xcb_list_extensions_reply_t>

pub fn names_len(&self) -> u8

pub fn names(&self) -> StrIterator<'_>

impl Reply<xcb_get_keyboard_mapping_reply_t>

pub fn keysyms_per_keycode(&self) -> u8

pub fn keysyms(&self) -> &[Keysym]

impl Reply<xcb_get_keyboard_control_reply_t>

pub fn global_auto_repeat(&self) -> u8

pub fn led_mask(&self) -> u32

pub fn key_click_percent(&self) -> u8

pub fn bell_percent(&self) -> u8

pub fn bell_pitch(&self) -> u16

pub fn bell_duration(&self) -> u16

pub fn auto_repeats(&self) -> &[u8]

impl Reply<xcb_get_pointer_control_reply_t>

pub fn acceleration_numerator(&self) -> u16

pub fn acceleration_denominator(&self) -> u16

pub fn threshold(&self) -> u16

impl Reply<xcb_get_screen_saver_reply_t>

pub fn timeout(&self) -> u16

pub fn interval(&self) -> u16

pub fn prefer_blanking(&self) -> u8

pub fn allow_exposures(&self) -> u8

impl Reply<xcb_list_hosts_reply_t>

pub fn mode(&self) -> u8

pub fn hosts_len(&self) -> u16

pub fn hosts(&self) -> HostIterator<'_>

impl Reply<xcb_set_pointer_mapping_reply_t>

pub fn status(&self) -> u8

impl Reply<xcb_get_pointer_mapping_reply_t>

pub fn map_len(&self) -> u8

pub fn map(&self) -> &[u8]

impl Reply<xcb_set_modifier_mapping_reply_t>

pub fn status(&self) -> u8

impl Reply<xcb_get_modifier_mapping_reply_t>

pub fn keycodes_per_modifier(&self) -> u8

pub fn keycodes(&self) -> &[Keycode]

impl Reply<xcb_big_requests_enable_reply_t>

pub fn maximum_request_length(&self) -> u32

impl Reply<xcb_xc_misc_get_version_reply_t>

pub fn server_major_version(&self) -> u16

pub fn server_minor_version(&self) -> u16

impl Reply<xcb_xc_misc_get_xid_range_reply_t>

pub fn start_id(&self) -> u32

pub fn count(&self) -> u32

impl Reply<xcb_xc_misc_get_xid_list_reply_t>

pub fn ids_len(&self) -> u32

pub fn ids(&self) -> &[u32]

impl Reply<xcb_render_query_version_reply_t>

pub fn major_version(&self) -> u32

pub fn minor_version(&self) -> u32

impl Reply<xcb_render_query_pict_formats_reply_t>

pub fn num_formats(&self) -> u32

pub fn num_screens(&self) -> u32

pub fn num_depths(&self) -> u32

pub fn num_visuals(&self) -> u32

pub fn num_subpixel(&self) -> u32

pub fn formats(&self) -> PictforminfoIterator

pub fn screens(&self) -> PictscreenIterator<'_>

pub fn subpixels(&self) -> &[u32]

impl Reply<xcb_render_query_pict_index_values_reply_t>

pub fn num_values(&self) -> u32

pub fn values(&self) -> IndexvalueIterator

impl Reply<xcb_render_query_filters_reply_t>

pub fn num_aliases(&self) -> u32

pub fn num_filters(&self) -> u32

pub fn aliases(&self) -> &[u16]

pub fn filters(&self) -> StrIterator<'_>

impl Reply<xcb_shape_query_version_reply_t>

pub fn major_version(&self) -> u16

pub fn minor_version(&self) -> u16

impl Reply<xcb_shape_query_extents_reply_t>

pub fn bounding_shaped(&self) -> bool

pub fn clip_shaped(&self) -> bool

pub fn bounding_shape_extents_x(&self) -> i16

pub fn bounding_shape_extents_y(&self) -> i16

pub fn bounding_shape_extents_width(&self) -> u16

pub fn bounding_shape_extents_height(&self) -> u16

pub fn clip_shape_extents_x(&self) -> i16

pub fn clip_shape_extents_y(&self) -> i16

pub fn clip_shape_extents_width(&self) -> u16

pub fn clip_shape_extents_height(&self) -> u16

impl Reply<xcb_shape_input_selected_reply_t>

pub fn enabled(&self) -> bool

impl Reply<xcb_shape_get_rectangles_reply_t>

pub fn ordering(&self) -> u8

pub fn rectangles_len(&self) -> u32

pub fn rectangles(&self) -> RectangleIterator

impl Reply<xcb_xfixes_query_version_reply_t>

pub fn major_version(&self) -> u32

pub fn minor_version(&self) -> u32

impl Reply<xcb_xfixes_get_cursor_image_reply_t>

pub fn x(&self) -> i16

pub fn y(&self) -> i16

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn xhot(&self) -> u16

pub fn yhot(&self) -> u16

pub fn cursor_serial(&self) -> u32

pub fn cursor_image(&self) -> &[u32]

impl Reply<xcb_xfixes_fetch_region_reply_t>

pub fn extents(&self) -> Rectangle

pub fn rectangles(&self) -> RectangleIterator

impl Reply<xcb_xfixes_get_cursor_name_reply_t>

pub fn atom(&self) -> Atom

pub fn nbytes(&self) -> u16

pub fn name(&self) -> &str

impl Reply<xcb_xfixes_get_cursor_image_and_name_reply_t>

pub fn x(&self) -> i16

pub fn y(&self) -> i16

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn xhot(&self) -> u16

pub fn yhot(&self) -> u16

pub fn cursor_serial(&self) -> u32

pub fn cursor_atom(&self) -> Atom

pub fn nbytes(&self) -> u16

pub fn name(&self) -> &str

pub fn cursor_image(&self) -> &[u32]