Struct xcb::base::Reply

pub struct Reply<T> {
    pub ptr: *mut T,
}

Wraps a pointer to a xcb_*_reply_t the pointer is freed when the Reply goes out of scope

Fields
ptr: *mut T
Implementations

impl Reply<xcb_get_window_attributes_reply_t>

pub fn backing_store(&self) -> u8

pub fn visual(&self) -> Visualid

The associated visual structure of window.

pub fn class(&self) -> u16

pub fn bit_gravity(&self) -> u8

pub fn win_gravity(&self) -> u8

pub fn backing_planes(&self) -> u32

Planes to be preserved if possible.

pub fn backing_pixel(&self) -> u32

Value to be used when restoring planes.

pub fn save_under(&self) -> bool

Boolean, should bits under be saved?

pub fn map_is_installed(&self) -> bool

pub fn map_state(&self) -> u8

pub fn override_redirect(&self) -> bool

Window managers should ignore this window if override_redirect is 1.

pub fn colormap(&self) -> Colormap

Color map to be associated with window.

pub fn all_event_masks(&self) -> u32

Set of events all people have interest in.

pub fn your_event_mask(&self) -> u32

My event mask.

pub fn do_not_propagate_mask(&self) -> u16

Set of events that should not propagate.

impl Reply<xcb_get_geometry_reply_t>

pub fn depth(&self) -> u8

The depth of the drawable (bits per pixel for the object).

pub fn root(&self) -> Window

Root window of the screen containing drawable.

pub fn x(&self) -> i16

The X coordinate of drawable. If drawable is a window, the coordinate specifies the upper-left outer corner relative to its parent’s origin. If drawable is a pixmap, the X coordinate is always 0.

pub fn y(&self) -> i16

The Y coordinate of drawable. If drawable is a window, the coordinate specifies the upper-left outer corner relative to its parent’s origin. If drawable is a pixmap, the Y coordinate is always 0.

pub fn width(&self) -> u16

The width of drawable.

pub fn height(&self) -> u16

The height of drawable.

pub fn border_width(&self) -> u16

The border width (in pixels).

impl Reply<xcb_query_tree_reply_t>

pub fn root(&self) -> Window

The root window of window.

pub fn parent(&self) -> Window

The parent window of window.

pub fn children_len(&self) -> u16

The number of child windows.

pub fn children(&self) -> &[Window]

impl Reply<xcb_intern_atom_reply_t>

pub fn atom(&self) -> Atom

impl Reply<xcb_get_atom_name_reply_t>

pub fn name_len(&self) -> u16

pub fn name(&self) -> &str

impl Reply<xcb_get_property_reply_t>

pub fn format(&self) -> u8

Specifies whether the data should be viewed as a list of 8-bit, 16-bit, or 32-bit quantities. Possible values are 8, 16, and 32. This information allows the X server to correctly perform byte-swap operations as necessary.

pub fn type_(&self) -> Atom

The actual type of the property (an atom).

pub fn bytes_after(&self) -> u32

The number of bytes remaining to be read in the property if a partial read was performed.

pub fn value_len(&self) -> u32

The length of value. You should use the corresponding accessor instead of this field.

pub fn value<T>(&self) -> &[T]

impl Reply<xcb_list_properties_reply_t>

pub fn atoms_len(&self) -> u16

pub fn atoms(&self) -> &[Atom]

impl Reply<xcb_get_selection_owner_reply_t>

pub fn owner(&self) -> Window

The current selection owner window.

impl Reply<xcb_grab_pointer_reply_t>

pub fn status(&self) -> u8

impl Reply<xcb_grab_keyboard_reply_t>

pub fn status(&self) -> u8

impl Reply<xcb_query_pointer_reply_t>

pub fn same_screen(&self) -> bool

If same_screen is False, then the pointer is not on the same screen as the argument window, child is None, and win_x and win_y are zero. If same_screen is True, then win_x and win_y are the pointer coordinates relative to the argument window’s origin, and child is the child containing the pointer, if any.

pub fn root(&self) -> Window

The root window the pointer is logically on.

pub fn child(&self) -> Window

The child window containing the pointer, if any, if same_screen is true. If same_screen is false, XCB_NONE is returned.

pub fn root_x(&self) -> i16

The pointer X position, relative to root.

pub fn root_y(&self) -> i16

The pointer Y position, relative to root.

pub fn win_x(&self) -> i16

The pointer X coordinate, relative to child, if same_screen is true. Zero otherwise.

pub fn win_y(&self) -> i16

The pointer Y coordinate, relative to child, if same_screen is true. Zero otherwise.

pub fn mask(&self) -> u16

The current logical state of the modifier keys and the buttons. Note that the logical state of a device (as seen by means of the protocol) may lag the physical state if device event processing is frozen.

impl Reply<xcb_get_motion_events_reply_t>

pub fn events_len(&self) -> u32

pub fn events(&self) -> TimecoordIterator

impl Reply<xcb_translate_coordinates_reply_t>

pub fn same_screen(&self) -> bool

pub fn child(&self) -> Window

pub fn dst_x(&self) -> i16

pub fn dst_y(&self) -> i16

impl Reply<xcb_get_input_focus_reply_t>

pub fn revert_to(&self) -> u8

pub fn focus(&self) -> Window

impl Reply<xcb_query_keymap_reply_t>

pub fn keys(&self) -> &[u8]

impl Reply<xcb_query_font_reply_t>

pub fn min_bounds(&self) -> Charinfo

minimum bounds over all existing char

pub fn max_bounds(&self) -> Charinfo

maximum bounds over all existing char

pub fn min_char_or_byte2(&self) -> u16

first character

pub fn max_char_or_byte2(&self) -> u16

last character

pub fn default_char(&self) -> u16

char to print for undefined character

pub fn properties_len(&self) -> u16

how many properties there are

pub fn draw_direction(&self) -> u8

pub fn min_byte1(&self) -> u8

pub fn max_byte1(&self) -> u8

pub fn all_chars_exist(&self) -> bool

flag if all characters have nonzero size

pub fn font_ascent(&self) -> i16

baseline to top edge of raster

pub fn font_descent(&self) -> i16

baseline to bottom edge of raster

pub fn char_infos_len(&self) -> u32

pub fn properties(&self) -> FontpropIterator

pub fn char_infos(&self) -> CharinfoIterator

impl Reply<xcb_query_text_extents_reply_t>

pub fn draw_direction(&self) -> u8

pub fn font_ascent(&self) -> i16

pub fn font_descent(&self) -> i16

pub fn overall_ascent(&self) -> i16

pub fn overall_descent(&self) -> i16

pub fn overall_width(&self) -> i32

pub fn overall_left(&self) -> i32

pub fn overall_right(&self) -> i32

impl Reply<xcb_list_fonts_reply_t>

pub fn names_len(&self) -> u16

The number of font names.

pub fn names(&self) -> StrIterator<'_>

impl Reply<xcb_list_fonts_with_info_reply_t>

pub fn name_len(&self) -> u8

The number of matched font names.

pub fn min_bounds(&self) -> Charinfo

minimum bounds over all existing char

pub fn max_bounds(&self) -> Charinfo

maximum bounds over all existing char

pub fn min_char_or_byte2(&self) -> u16

first character

pub fn max_char_or_byte2(&self) -> u16

last character

pub fn default_char(&self) -> u16

char to print for undefined character

pub fn properties_len(&self) -> u16

how many properties there are

pub fn draw_direction(&self) -> u8

pub fn min_byte1(&self) -> u8

pub fn max_byte1(&self) -> u8

pub fn all_chars_exist(&self) -> bool

flag if all characters have nonzero size

pub fn font_ascent(&self) -> i16

baseline to top edge of raster

pub fn font_descent(&self) -> i16

baseline to bottom edge of raster

pub fn replies_hint(&self) -> u32

An indication of how many more fonts will be returned. This is only a hint and may be larger or smaller than the number of fonts actually returned. A zero value does not guarantee that no more fonts will be returned.

pub fn properties(&self) -> FontpropIterator

pub fn name(&self) -> &str

impl Reply<xcb_get_font_path_reply_t>

pub fn path_len(&self) -> u16

pub fn path(&self) -> StrIterator<'_>

impl Reply<xcb_get_image_reply_t>

pub fn depth(&self) -> u8

pub fn visual(&self) -> Visualid

pub fn data(&self) -> &[u8]

impl Reply<xcb_list_installed_colormaps_reply_t>

pub fn cmaps_len(&self) -> u16

pub fn cmaps(&self) -> &[Colormap]

impl Reply<xcb_alloc_color_reply_t>

pub fn red(&self) -> u16

pub fn green(&self) -> u16

pub fn blue(&self) -> u16

pub fn pixel(&self) -> u32

impl Reply<xcb_alloc_named_color_reply_t>

pub fn pixel(&self) -> u32

pub fn exact_red(&self) -> u16

pub fn exact_green(&self) -> u16

pub fn exact_blue(&self) -> u16

pub fn visual_red(&self) -> u16

pub fn visual_green(&self) -> u16

pub fn visual_blue(&self) -> u16

impl Reply<xcb_alloc_color_cells_reply_t>

pub fn pixels_len(&self) -> u16

pub fn masks_len(&self) -> u16

pub fn pixels(&self) -> &[u32]

pub fn masks(&self) -> &[u32]

impl Reply<xcb_alloc_color_planes_reply_t>

pub fn pixels_len(&self) -> u16

pub fn red_mask(&self) -> u32

pub fn green_mask(&self) -> u32

pub fn blue_mask(&self) -> u32

pub fn pixels(&self) -> &[u32]

impl Reply<xcb_query_colors_reply_t>

pub fn colors_len(&self) -> u16

pub fn colors(&self) -> RgbIterator

impl Reply<xcb_lookup_color_reply_t>

pub fn exact_red(&self) -> u16

pub fn exact_green(&self) -> u16

pub fn exact_blue(&self) -> u16

pub fn visual_red(&self) -> u16

pub fn visual_green(&self) -> u16

pub fn visual_blue(&self) -> u16

impl Reply<xcb_query_best_size_reply_t>

pub fn width(&self) -> u16

pub fn height(&self) -> u16

impl Reply<xcb_query_extension_reply_t>

pub fn present(&self) -> bool

Whether the extension is present on this X11 server.

pub fn major_opcode(&self) -> u8

The major opcode for requests.

pub fn first_event(&self) -> u8

The first event code, if any.

pub fn first_error(&self) -> u8

The first error code, if any.

impl Reply<xcb_list_extensions_reply_t>

pub fn names_len(&self) -> u8

pub fn names(&self) -> StrIterator<'_>

impl Reply<xcb_get_keyboard_mapping_reply_t>

pub fn keysyms_per_keycode(&self) -> u8

pub fn keysyms(&self) -> &[Keysym]

impl Reply<xcb_get_keyboard_control_reply_t>

pub fn global_auto_repeat(&self) -> u8

pub fn led_mask(&self) -> u32

pub fn key_click_percent(&self) -> u8

pub fn bell_percent(&self) -> u8

pub fn bell_pitch(&self) -> u16

pub fn bell_duration(&self) -> u16

pub fn auto_repeats(&self) -> &[u8]

impl Reply<xcb_get_pointer_control_reply_t>

pub fn acceleration_numerator(&self) -> u16

pub fn acceleration_denominator(&self) -> u16

pub fn threshold(&self) -> u16

impl Reply<xcb_get_screen_saver_reply_t>

pub fn timeout(&self) -> u16

pub fn interval(&self) -> u16

pub fn prefer_blanking(&self) -> u8

pub fn allow_exposures(&self) -> u8

impl Reply<xcb_list_hosts_reply_t>

pub fn mode(&self) -> u8

pub fn hosts_len(&self) -> u16

pub fn hosts(&self) -> HostIterator<'_>

impl Reply<xcb_set_pointer_mapping_reply_t>

pub fn status(&self) -> u8

impl Reply<xcb_get_pointer_mapping_reply_t>

pub fn map_len(&self) -> u8

pub fn map(&self) -> &[u8]

impl Reply<xcb_set_modifier_mapping_reply_t>

pub fn status(&self) -> u8

impl Reply<xcb_get_modifier_mapping_reply_t>

pub fn keycodes_per_modifier(&self) -> u8

pub fn keycodes(&self) -> &[Keycode]

impl Reply<xcb_big_requests_enable_reply_t>

pub fn maximum_request_length(&self) -> u32

impl Reply<xcb_xc_misc_get_version_reply_t>

pub fn server_major_version(&self) -> u16

pub fn server_minor_version(&self) -> u16

impl Reply<xcb_xc_misc_get_xid_range_reply_t>

pub fn start_id(&self) -> u32

pub fn count(&self) -> u32

impl Reply<xcb_xc_misc_get_xid_list_reply_t>

pub fn ids_len(&self) -> u32

pub fn ids(&self) -> &[u32]

impl Reply<xcb_composite_query_version_reply_t>

pub fn major_version(&self) -> u32

pub fn minor_version(&self) -> u32

impl Reply<xcb_composite_get_overlay_window_reply_t>

pub fn overlay_win(&self) -> Window

impl Reply<xcb_damage_query_version_reply_t>

pub fn major_version(&self) -> u32

pub fn minor_version(&self) -> u32

impl Reply<xcb_dpms_get_version_reply_t>

pub fn server_major_version(&self) -> u16

pub fn server_minor_version(&self) -> u16

impl Reply<xcb_dpms_capable_reply_t>

pub fn capable(&self) -> bool

impl Reply<xcb_dpms_get_timeouts_reply_t>

pub fn standby_timeout(&self) -> u16

pub fn suspend_timeout(&self) -> u16

pub fn off_timeout(&self) -> u16

impl Reply<xcb_dpms_info_reply_t>

pub fn power_level(&self) -> u16

pub fn state(&self) -> bool

impl Reply<xcb_dri2_query_version_reply_t>

pub fn major_version(&self) -> u32

pub fn minor_version(&self) -> u32

impl Reply<xcb_dri2_connect_reply_t>

pub fn driver_name_length(&self) -> u32

pub fn device_name_length(&self) -> u32

pub fn driver_name(&self) -> &str

pub fn alignment_pad<T>(&self) -> &[T]

pub fn device_name(&self) -> &str

impl Reply<xcb_dri2_authenticate_reply_t>

pub fn authenticated(&self) -> u32

impl Reply<xcb_dri2_get_buffers_reply_t>

pub fn width(&self) -> u32

pub fn height(&self) -> u32

pub fn count(&self) -> u32

pub fn buffers(&self) -> Dri2BufferIterator

impl Reply<xcb_dri2_get_buffers_with_format_reply_t>

pub fn width(&self) -> u32

pub fn height(&self) -> u32

pub fn count(&self) -> u32

pub fn buffers(&self) -> Dri2BufferIterator

impl Reply<xcb_dri2_swap_buffers_reply_t>

pub fn swap_hi(&self) -> u32

pub fn swap_lo(&self) -> u32

impl Reply<xcb_dri2_get_msc_reply_t>

pub fn ust_hi(&self) -> u32

pub fn ust_lo(&self) -> u32

pub fn msc_hi(&self) -> u32

pub fn msc_lo(&self) -> u32

pub fn sbc_hi(&self) -> u32

pub fn sbc_lo(&self) -> u32

impl Reply<xcb_dri2_wait_msc_reply_t>

pub fn ust_hi(&self) -> u32

pub fn ust_lo(&self) -> u32

pub fn msc_hi(&self) -> u32

pub fn msc_lo(&self) -> u32

pub fn sbc_hi(&self) -> u32

pub fn sbc_lo(&self) -> u32

impl Reply<xcb_dri2_wait_sbc_reply_t>

pub fn ust_hi(&self) -> u32

pub fn ust_lo(&self) -> u32

pub fn msc_hi(&self) -> u32

pub fn msc_lo(&self) -> u32

pub fn sbc_hi(&self) -> u32

pub fn sbc_lo(&self) -> u32

impl Reply<xcb_dri2_get_param_reply_t>

pub fn is_param_recognized(&self) -> bool

pub fn value_hi(&self) -> u32

pub fn value_lo(&self) -> u32

impl Reply<xcb_dri3_query_version_reply_t>

pub fn major_version(&self) -> u32

pub fn minor_version(&self) -> u32

impl Reply<xcb_dri3_open_reply_t>

pub fn device_fds(&self, c: &Connection) -> &[i32]

impl Reply<xcb_dri3_buffer_from_pixmap_reply_t>

pub fn size(&self) -> u32

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn stride(&self) -> u16

pub fn depth(&self) -> u8

pub fn bpp(&self) -> u8

pub fn pixmap_fds(&self, c: &Connection) -> &[i32]

impl Reply<xcb_dri3_fd_from_fence_reply_t>

pub fn fence_fds(&self, c: &Connection) -> &[i32]

impl Reply<xcb_genericevent_query_version_reply_t>

pub fn major_version(&self) -> u16

pub fn minor_version(&self) -> u16

impl Reply<xcb_glx_make_current_reply_t>

pub fn context_tag(&self) -> ContextTag

impl Reply<xcb_glx_is_direct_reply_t>

pub fn is_direct(&self) -> bool

impl Reply<xcb_glx_query_version_reply_t>

pub fn major_version(&self) -> u32

pub fn minor_version(&self) -> u32

impl Reply<xcb_glx_get_visual_configs_reply_t>

pub fn num_visuals(&self) -> u32

pub fn num_properties(&self) -> u32

pub fn property_list(&self) -> &[u32]

impl Reply<xcb_glx_vendor_private_with_reply_reply_t>

pub fn retval(&self) -> u32

pub fn data1(&self) -> &[u8]

pub fn data2(&self) -> &[u8]

impl Reply<xcb_glx_query_extensions_string_reply_t>

pub fn n(&self) -> u32

impl Reply<xcb_glx_query_server_string_reply_t>

pub fn str_len(&self) -> u32

pub fn string(&self) -> &str

impl Reply<xcb_glx_get_fb_configs_reply_t>

pub fn num_fb_configs(&self) -> u32

pub fn num_properties(&self) -> u32

pub fn property_list(&self) -> &[u32]

impl Reply<xcb_glx_query_context_reply_t>

pub fn num_attribs(&self) -> u32

pub fn attribs(&self) -> &[u32]

impl Reply<xcb_glx_make_context_current_reply_t>

pub fn context_tag(&self) -> ContextTag

impl Reply<xcb_glx_get_drawable_attributes_reply_t>

pub fn num_attribs(&self) -> u32

pub fn attribs(&self) -> &[u32]

impl Reply<xcb_glx_gen_lists_reply_t>

pub fn ret_val(&self) -> u32

impl Reply<xcb_glx_render_mode_reply_t>

pub fn ret_val(&self) -> u32

pub fn n(&self) -> u32

pub fn new_mode(&self) -> u32

pub fn data(&self) -> &[u32]

impl Reply<xcb_glx_read_pixels_reply_t>

pub fn data(&self) -> &[u8]

impl Reply<xcb_glx_get_booleanv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> bool

pub fn data(&self) -> Vec<bool>

impl Reply<xcb_glx_get_clip_plane_reply_t>

pub fn data(&self) -> &[Float64]

impl Reply<xcb_glx_get_doublev_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float64

pub fn data(&self) -> &[Float64]

impl Reply<xcb_glx_get_error_reply_t>

pub fn error(&self) -> i32

impl Reply<xcb_glx_get_floatv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float32

pub fn data(&self) -> &[Float32]

impl Reply<xcb_glx_get_integerv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_get_lightfv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float32

pub fn data(&self) -> &[Float32]

impl Reply<xcb_glx_get_lightiv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_get_mapdv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float64

pub fn data(&self) -> &[Float64]

impl Reply<xcb_glx_get_mapfv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float32

pub fn data(&self) -> &[Float32]

impl Reply<xcb_glx_get_mapiv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_get_materialfv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float32

pub fn data(&self) -> &[Float32]

impl Reply<xcb_glx_get_materialiv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_get_pixel_mapfv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float32

pub fn data(&self) -> &[Float32]

impl Reply<xcb_glx_get_pixel_mapuiv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> u32

pub fn data(&self) -> &[u32]

impl Reply<xcb_glx_get_pixel_mapusv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> u16

pub fn data(&self) -> &[u16]

impl Reply<xcb_glx_get_polygon_stipple_reply_t>

pub fn data(&self) -> &[u8]

impl Reply<xcb_glx_get_string_reply_t>

pub fn n(&self) -> u32

pub fn string(&self) -> &str

impl Reply<xcb_glx_get_tex_envfv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float32

pub fn data(&self) -> &[Float32]

impl Reply<xcb_glx_get_tex_enviv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_get_tex_gendv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float64

pub fn data(&self) -> &[Float64]

impl Reply<xcb_glx_get_tex_genfv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float32

pub fn data(&self) -> &[Float32]

impl Reply<xcb_glx_get_tex_geniv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_get_tex_image_reply_t>

pub fn width(&self) -> i32

pub fn height(&self) -> i32

pub fn depth(&self) -> i32

pub fn data(&self) -> &[u8]

impl Reply<xcb_glx_get_tex_parameterfv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float32

pub fn data(&self) -> &[Float32]

impl Reply<xcb_glx_get_tex_parameteriv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_get_tex_level_parameterfv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float32

pub fn data(&self) -> &[Float32]

impl Reply<xcb_glx_get_tex_level_parameteriv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_is_list_reply_t>

pub fn ret_val(&self) -> Bool32

impl Reply<xcb_glx_are_textures_resident_reply_t>

pub fn ret_val(&self) -> Bool32

pub fn data(&self) -> Vec<bool>

impl Reply<xcb_glx_gen_textures_reply_t>

pub fn data(&self) -> &[u32]

impl Reply<xcb_glx_is_texture_reply_t>

pub fn ret_val(&self) -> Bool32

impl Reply<xcb_glx_get_color_table_reply_t>

pub fn width(&self) -> i32

pub fn data(&self) -> &[u8]

impl Reply<xcb_glx_get_color_table_parameterfv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float32

pub fn data(&self) -> &[Float32]

impl Reply<xcb_glx_get_color_table_parameteriv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_get_convolution_filter_reply_t>

pub fn width(&self) -> i32

pub fn height(&self) -> i32

pub fn data(&self) -> &[u8]

impl Reply<xcb_glx_get_convolution_parameterfv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float32

pub fn data(&self) -> &[Float32]

impl Reply<xcb_glx_get_convolution_parameteriv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_get_separable_filter_reply_t>

pub fn row_w(&self) -> i32

pub fn col_h(&self) -> i32

pub fn rows_and_cols(&self) -> &[u8]

impl Reply<xcb_glx_get_histogram_reply_t>

pub fn width(&self) -> i32

pub fn data(&self) -> &[u8]

impl Reply<xcb_glx_get_histogram_parameterfv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float32

pub fn data(&self) -> &[Float32]

impl Reply<xcb_glx_get_histogram_parameteriv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_get_minmax_reply_t>

pub fn data(&self) -> &[u8]

impl Reply<xcb_glx_get_minmax_parameterfv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> Float32

pub fn data(&self) -> &[Float32]

impl Reply<xcb_glx_get_minmax_parameteriv_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_get_compressed_tex_image_arb_reply_t>

pub fn size(&self) -> i32

pub fn data(&self) -> &[u8]

impl Reply<xcb_glx_gen_queries_arb_reply_t>

pub fn data(&self) -> &[u32]

impl Reply<xcb_glx_is_query_arb_reply_t>

pub fn ret_val(&self) -> Bool32

impl Reply<xcb_glx_get_queryiv_arb_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_get_query_objectiv_arb_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> i32

pub fn data(&self) -> &[i32]

impl Reply<xcb_glx_get_query_objectuiv_arb_reply_t>

pub fn n(&self) -> u32

pub fn datum(&self) -> u32

pub fn data(&self) -> &[u32]

impl Reply<xcb_present_query_version_reply_t>

pub fn major_version(&self) -> u32

pub fn minor_version(&self) -> u32

impl Reply<xcb_present_query_capabilities_reply_t>

pub fn capabilities(&self) -> u32

impl Reply<xcb_randr_query_version_reply_t>

pub fn major_version(&self) -> u32

pub fn minor_version(&self) -> u32

impl Reply<xcb_randr_set_screen_config_reply_t>

pub fn status(&self) -> u8

pub fn new_timestamp(&self) -> Timestamp

pub fn config_timestamp(&self) -> Timestamp

pub fn root(&self) -> Window

pub fn subpixel_order(&self) -> u16

impl Reply<xcb_randr_get_screen_info_reply_t>

pub fn rotations(&self) -> u8

pub fn root(&self) -> Window

pub fn timestamp(&self) -> Timestamp

pub fn config_timestamp(&self) -> Timestamp

pub fn n_sizes(&self) -> u16

pub fn size_id(&self) -> u16

pub fn rotation(&self) -> u16

pub fn rate(&self) -> u16

pub fn n_info(&self) -> u16

pub fn sizes(&self) -> ScreenSizeIterator

pub fn rates(&self) -> RefreshRatesIterator<'_>

impl Reply<xcb_randr_get_screen_size_range_reply_t>

pub fn min_width(&self) -> u16

pub fn min_height(&self) -> u16

pub fn max_width(&self) -> u16

pub fn max_height(&self) -> u16

impl Reply<xcb_randr_get_screen_resources_reply_t>

pub fn timestamp(&self) -> Timestamp

pub fn config_timestamp(&self) -> Timestamp

pub fn num_crtcs(&self) -> u16

pub fn num_outputs(&self) -> u16

pub fn num_modes(&self) -> u16

pub fn names_len(&self) -> u16

pub fn crtcs(&self) -> &[Crtc]

pub fn outputs(&self) -> &[Output]

pub fn modes(&self) -> ModeInfoIterator

pub fn names(&self) -> &[u8]

impl Reply<xcb_randr_get_output_info_reply_t>

pub fn status(&self) -> u8

pub fn timestamp(&self) -> Timestamp

pub fn crtc(&self) -> Crtc

pub fn mm_width(&self) -> u32

pub fn mm_height(&self) -> u32

pub fn connection(&self) -> u8

pub fn subpixel_order(&self) -> u8

pub fn num_crtcs(&self) -> u16

pub fn num_modes(&self) -> u16

pub fn num_preferred(&self) -> u16

pub fn num_clones(&self) -> u16

pub fn name_len(&self) -> u16

pub fn crtcs(&self) -> &[Crtc]

pub fn modes(&self) -> &[Mode]

pub fn clones(&self) -> &[Output]

pub fn name(&self) -> &[u8]

impl Reply<xcb_randr_list_output_properties_reply_t>

pub fn num_atoms(&self) -> u16

pub fn atoms(&self) -> &[Atom]

impl Reply<xcb_randr_query_output_property_reply_t>

pub fn pending(&self) -> bool

pub fn range(&self) -> bool

pub fn immutable(&self) -> bool

pub fn valid_values(&self) -> &[i32]

impl Reply<xcb_randr_get_output_property_reply_t>

pub fn format(&self) -> u8

pub fn type_(&self) -> Atom

pub fn bytes_after(&self) -> u32

pub fn num_items(&self) -> u32

pub fn data(&self) -> &[u8]

impl Reply<xcb_randr_create_mode_reply_t>

pub fn mode(&self) -> Mode

impl Reply<xcb_randr_get_crtc_info_reply_t>

pub fn status(&self) -> u8

pub fn timestamp(&self) -> Timestamp

pub fn x(&self) -> i16

pub fn y(&self) -> i16

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn mode(&self) -> Mode

pub fn rotation(&self) -> u16

pub fn rotations(&self) -> u16

pub fn num_outputs(&self) -> u16

pub fn num_possible_outputs(&self) -> u16

pub fn outputs(&self) -> &[Output]

pub fn possible(&self) -> &[Output]

impl Reply<xcb_randr_set_crtc_config_reply_t>

pub fn status(&self) -> u8

pub fn timestamp(&self) -> Timestamp

impl Reply<xcb_randr_get_crtc_gamma_size_reply_t>

pub fn size(&self) -> u16

impl Reply<xcb_randr_get_crtc_gamma_reply_t>

pub fn size(&self) -> u16

pub fn red(&self) -> &[u16]

pub fn green(&self) -> &[u16]

pub fn blue(&self) -> &[u16]

impl Reply<xcb_randr_get_screen_resources_current_reply_t>

pub fn timestamp(&self) -> Timestamp

pub fn config_timestamp(&self) -> Timestamp

pub fn num_crtcs(&self) -> u16

pub fn num_outputs(&self) -> u16

pub fn num_modes(&self) -> u16

pub fn names_len(&self) -> u16

pub fn crtcs(&self) -> &[Crtc]

pub fn outputs(&self) -> &[Output]

pub fn modes(&self) -> ModeInfoIterator

pub fn names(&self) -> &[u8]

impl Reply<xcb_randr_get_crtc_transform_reply_t>

pub fn pending_transform(&self) -> Transform

pub fn has_transforms(&self) -> bool

pub fn current_transform(&self) -> Transform

pub fn pending_len(&self) -> u16

pub fn pending_nparams(&self) -> u16

pub fn current_len(&self) -> u16

pub fn current_nparams(&self) -> u16

pub fn pending_filter_name(&self) -> &str

pub fn pending_params(&self) -> &[Fixed]

pub fn current_filter_name(&self) -> &str

pub fn current_params(&self) -> &[Fixed]

impl Reply<xcb_randr_get_panning_reply_t>

pub fn status(&self) -> u8

pub fn timestamp(&self) -> Timestamp

pub fn left(&self) -> u16

pub fn top(&self) -> u16

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn track_left(&self) -> u16

pub fn track_top(&self) -> u16

pub fn track_width(&self) -> u16

pub fn track_height(&self) -> u16

pub fn border_left(&self) -> i16

pub fn border_top(&self) -> i16

pub fn border_right(&self) -> i16

pub fn border_bottom(&self) -> i16

impl Reply<xcb_randr_set_panning_reply_t>

pub fn status(&self) -> u8

pub fn timestamp(&self) -> Timestamp

impl Reply<xcb_randr_get_output_primary_reply_t>

pub fn output(&self) -> Output

impl Reply<xcb_randr_get_providers_reply_t>

pub fn timestamp(&self) -> Timestamp

pub fn num_providers(&self) -> u16

pub fn providers(&self) -> &[Provider]

impl Reply<xcb_randr_get_provider_info_reply_t>

pub fn status(&self) -> u8

pub fn timestamp(&self) -> Timestamp

pub fn capabilities(&self) -> u32

pub fn num_crtcs(&self) -> u16

pub fn num_outputs(&self) -> u16

pub fn num_associated_providers(&self) -> u16

pub fn name_len(&self) -> u16

pub fn crtcs(&self) -> &[Crtc]

pub fn outputs(&self) -> &[Output]

pub fn associated_providers(&self) -> &[Provider]

pub fn associated_capability(&self) -> &[u32]

pub fn name(&self) -> &str

impl Reply<xcb_randr_list_provider_properties_reply_t>

pub fn num_atoms(&self) -> u16

pub fn atoms(&self) -> &[Atom]

impl Reply<xcb_randr_query_provider_property_reply_t>

pub fn pending(&self) -> bool

pub fn range(&self) -> bool

pub fn immutable(&self) -> bool

pub fn valid_values(&self) -> &[i32]

impl Reply<xcb_randr_get_provider_property_reply_t>

pub fn format(&self) -> u8

pub fn type_(&self) -> Atom

pub fn bytes_after(&self) -> u32

pub fn num_items(&self) -> u32

pub fn data<T>(&self) -> &[T]

impl Reply<xcb_record_query_version_reply_t>

pub fn major_version(&self) -> u16

pub fn minor_version(&self) -> u16

impl Reply<xcb_record_get_context_reply_t>

pub fn enabled(&self) -> bool

pub fn element_header(&self) -> ElementHeader

pub fn num_intercepted_clients(&self) -> u32

pub fn intercepted_clients(&self) -> ClientInfoIterator<'_>

impl Reply<xcb_record_enable_context_reply_t>

pub fn category(&self) -> u8

pub fn element_header(&self) -> ElementHeader

pub fn client_swapped(&self) -> bool

pub fn xid_base(&self) -> u32

pub fn server_time(&self) -> u32

pub fn rec_sequence_num(&self) -> u32

pub fn data(&self) -> &[u8]

impl Reply<xcb_render_query_version_reply_t>

pub fn major_version(&self) -> u32

pub fn minor_version(&self) -> u32

impl Reply<xcb_render_query_pict_formats_reply_t>

pub fn num_formats(&self) -> u32

pub fn num_screens(&self) -> u32

pub fn num_depths(&self) -> u32

pub fn num_visuals(&self) -> u32

pub fn num_subpixel(&self) -> u32

pub fn formats(&self) -> PictforminfoIterator

pub fn screens(&self) -> PictscreenIterator<'_>

pub fn subpixels(&self) -> &[u32]

impl Reply<xcb_render_query_pict_index_values_reply_t>

pub fn num_values(&self) -> u32

pub fn values(&self) -> IndexvalueIterator

impl Reply<xcb_render_query_filters_reply_t>

pub fn num_aliases(&self) -> u32

pub fn num_filters(&self) -> u32

pub fn aliases(&self) -> &[u16]

pub fn filters(&self) -> StrIterator<'_>

impl Reply<xcb_res_query_version_reply_t>

pub fn server_major(&self) -> u16

pub fn server_minor(&self) -> u16

impl Reply<xcb_res_query_clients_reply_t>

pub fn num_clients(&self) -> u32

pub fn clients(&self) -> ClientIterator

impl Reply<xcb_res_query_client_resources_reply_t>

pub fn num_types(&self) -> u32

pub fn types(&self) -> TypeIterator

impl Reply<xcb_res_query_client_pixmap_bytes_reply_t>

pub fn bytes(&self) -> u32

pub fn bytes_overflow(&self) -> u32

impl Reply<xcb_res_query_client_ids_reply_t>

pub fn num_ids(&self) -> u32

pub fn ids(&self) -> ClientIdValueIterator<'_>

impl Reply<xcb_res_query_resource_bytes_reply_t>

pub fn num_sizes(&self) -> u32

pub fn sizes(&self) -> ResourceSizeValueIterator<'_>

impl Reply<xcb_screensaver_query_version_reply_t>

pub fn server_major_version(&self) -> u16

pub fn server_minor_version(&self) -> u16

impl Reply<xcb_screensaver_query_info_reply_t>

pub fn state(&self) -> u8

pub fn saver_window(&self) -> Window

pub fn ms_until_server(&self) -> u32

pub fn ms_since_user_input(&self) -> u32

pub fn event_mask(&self) -> u32

pub fn kind(&self) -> u8

impl Reply<xcb_selinux_query_version_reply_t>

pub fn server_major(&self) -> u16

pub fn server_minor(&self) -> u16

impl Reply<xcb_selinux_get_device_create_context_reply_t>

pub fn context_len(&self) -> u32

pub fn context(&self) -> &str

impl Reply<xcb_selinux_get_device_context_reply_t>

pub fn context_len(&self) -> u32

pub fn context(&self) -> &str

impl Reply<xcb_selinux_get_window_create_context_reply_t>

pub fn context_len(&self) -> u32

pub fn context(&self) -> &str

impl Reply<xcb_selinux_get_window_context_reply_t>

pub fn context_len(&self) -> u32

pub fn context(&self) -> &str

impl Reply<xcb_selinux_get_property_create_context_reply_t>

pub fn context_len(&self) -> u32

pub fn context(&self) -> &str

impl Reply<xcb_selinux_get_property_use_context_reply_t>

pub fn context_len(&self) -> u32

pub fn context(&self) -> &str

impl Reply<xcb_selinux_get_property_context_reply_t>

pub fn context_len(&self) -> u32

pub fn context(&self) -> &str

impl Reply<xcb_selinux_get_property_data_context_reply_t>

pub fn context_len(&self) -> u32

pub fn context(&self) -> &str

impl Reply<xcb_selinux_list_properties_reply_t>

pub fn properties_len(&self) -> u32

pub fn properties(&self) -> ListItemIterator<'_>

impl Reply<xcb_selinux_get_selection_create_context_reply_t>

pub fn context_len(&self) -> u32

pub fn context(&self) -> &str

impl Reply<xcb_selinux_get_selection_use_context_reply_t>

pub fn context_len(&self) -> u32

pub fn context(&self) -> &str

impl Reply<xcb_selinux_get_selection_context_reply_t>

pub fn context_len(&self) -> u32

pub fn context(&self) -> &str

impl Reply<xcb_selinux_get_selection_data_context_reply_t>

pub fn context_len(&self) -> u32

pub fn context(&self) -> &str

impl Reply<xcb_selinux_list_selections_reply_t>

pub fn selections_len(&self) -> u32

pub fn selections(&self) -> ListItemIterator<'_>

impl Reply<xcb_selinux_get_client_context_reply_t>

pub fn context_len(&self) -> u32

pub fn context(&self) -> &str

impl Reply<xcb_shape_query_version_reply_t>

pub fn major_version(&self) -> u16

pub fn minor_version(&self) -> u16

impl Reply<xcb_shape_query_extents_reply_t>

pub fn bounding_shaped(&self) -> bool

pub fn clip_shaped(&self) -> bool

pub fn bounding_shape_extents_x(&self) -> i16

pub fn bounding_shape_extents_y(&self) -> i16

pub fn bounding_shape_extents_width(&self) -> u16

pub fn bounding_shape_extents_height(&self) -> u16

pub fn clip_shape_extents_x(&self) -> i16

pub fn clip_shape_extents_y(&self) -> i16

pub fn clip_shape_extents_width(&self) -> u16

pub fn clip_shape_extents_height(&self) -> u16

impl Reply<xcb_shape_input_selected_reply_t>

pub fn enabled(&self) -> bool

impl Reply<xcb_shape_get_rectangles_reply_t>

pub fn ordering(&self) -> u8

pub fn rectangles_len(&self) -> u32

pub fn rectangles(&self) -> RectangleIterator

impl Reply<xcb_shm_query_version_reply_t>

pub fn shared_pixmaps(&self) -> bool

pub fn major_version(&self) -> u16

pub fn minor_version(&self) -> u16

pub fn uid(&self) -> u16

pub fn gid(&self) -> u16

pub fn pixmap_format(&self) -> u8

impl Reply<xcb_shm_get_image_reply_t>

pub fn depth(&self) -> u8

pub fn visual(&self) -> Visualid

pub fn size(&self) -> u32

impl Reply<xcb_shm_create_segment_reply_t>

pub fn shm_fds(&self, c: &Connection) -> &[i32]

impl Reply<xcb_sync_initialize_reply_t>

pub fn major_version(&self) -> u8

pub fn minor_version(&self) -> u8

impl Reply<xcb_sync_list_system_counters_reply_t>

pub fn counters_len(&self) -> u32

pub fn counters(&self) -> SystemcounterIterator<'_>

impl Reply<xcb_sync_query_counter_reply_t>

pub fn counter_value(&self) -> Int64

impl Reply<xcb_sync_query_alarm_reply_t>

pub fn trigger(&self) -> Trigger

pub fn delta(&self) -> Int64

pub fn events(&self) -> bool

pub fn state(&self) -> u8

impl Reply<xcb_sync_get_priority_reply_t>

pub fn priority(&self) -> i32

impl Reply<xcb_sync_query_fence_reply_t>

pub fn triggered(&self) -> bool

impl Reply<xcb_test_get_version_reply_t>

pub fn major_version(&self) -> u8

pub fn minor_version(&self) -> u16

impl Reply<xcb_test_compare_cursor_reply_t>

pub fn same(&self) -> bool

impl Reply<xcb_x_print_print_query_version_reply_t>

pub fn major_version(&self) -> u16

pub fn minor_version(&self) -> u16

impl Reply<xcb_x_print_print_get_printer_list_reply_t>

pub fn list_count(&self) -> u32

pub fn printers(&self) -> PrinterIterator<'_>

impl Reply<xcb_x_print_print_get_context_reply_t>

pub fn context(&self) -> u32

impl Reply<xcb_x_print_print_get_screen_of_context_reply_t>

pub fn root(&self) -> Window

impl Reply<xcb_x_print_print_get_document_data_reply_t>

pub fn status_code(&self) -> u32

pub fn finished_flag(&self) -> u32

pub fn data_len(&self) -> u32

pub fn data(&self) -> &[u8]

impl Reply<xcb_x_print_print_input_selected_reply_t>

pub fn event_mask(&self) -> u32

pub fn event_list(&self) -> &[u32]

pub fn all_events_mask(&self) -> u32

pub fn all_events_list(&self) -> &[u32]

impl Reply<xcb_x_print_print_get_attributes_reply_t>

pub fn string_len(&self) -> u32

pub fn attributes(&self) -> &[String8]

impl Reply<xcb_x_print_print_get_one_attributes_reply_t>

pub fn value_len(&self) -> u32

pub fn value(&self) -> &[String8]

impl Reply<xcb_x_print_print_get_page_dimensions_reply_t>

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn offset_x(&self) -> u16

pub fn offset_y(&self) -> u16

pub fn reproducible_width(&self) -> u16

pub fn reproducible_height(&self) -> u16

impl Reply<xcb_x_print_print_query_screens_reply_t>

pub fn list_count(&self) -> u32

pub fn roots(&self) -> &[Window]

impl Reply<xcb_x_print_print_set_image_resolution_reply_t>

pub fn status(&self) -> bool

pub fn previous_resolutions(&self) -> u16

impl Reply<xcb_x_print_print_get_image_resolution_reply_t>

pub fn image_resolution(&self) -> u16

impl Reply<xcb_xevie_query_version_reply_t>

pub fn server_major_version(&self) -> u16

pub fn server_minor_version(&self) -> u16

impl Reply<xcb_xf86dri_query_version_reply_t>

pub fn dri_major_version(&self) -> u16

pub fn dri_minor_version(&self) -> u16

pub fn dri_minor_patch(&self) -> u32

impl Reply<xcb_xf86dri_query_direct_rendering_capable_reply_t>

pub fn is_capable(&self) -> bool

impl Reply<xcb_xf86dri_open_connection_reply_t>

pub fn sarea_handle_low(&self) -> u32

pub fn sarea_handle_high(&self) -> u32

pub fn bus_id_len(&self) -> u32

pub fn bus_id(&self) -> &str

impl Reply<xcb_xf86dri_get_client_driver_name_reply_t>

pub fn client_driver_major_version(&self) -> u32

pub fn client_driver_minor_version(&self) -> u32

pub fn client_driver_patch_version(&self) -> u32

pub fn client_driver_name_len(&self) -> u32

pub fn client_driver_name(&self) -> &str

impl Reply<xcb_xf86dri_create_context_reply_t>

pub fn hw_context(&self) -> u32

impl Reply<xcb_xf86dri_create_drawable_reply_t>

pub fn hw_drawable_handle(&self) -> u32

impl Reply<xcb_xf86dri_get_drawable_info_reply_t>

pub fn drawable_table_index(&self) -> u32

pub fn drawable_table_stamp(&self) -> u32

pub fn drawable_origin_x(&self) -> i16

pub fn drawable_origin_y(&self) -> i16

pub fn drawable_size_w(&self) -> i16

pub fn drawable_size_h(&self) -> i16

pub fn num_clip_rects(&self) -> u32

pub fn back_x(&self) -> i16

pub fn back_y(&self) -> i16

pub fn num_back_clip_rects(&self) -> u32

pub fn clip_rects(&self) -> DrmClipRectIterator

pub fn back_clip_rects(&self) -> DrmClipRectIterator

impl Reply<xcb_xf86dri_get_device_info_reply_t>

pub fn framebuffer_handle_low(&self) -> u32

pub fn framebuffer_handle_high(&self) -> u32

pub fn framebuffer_origin_offset(&self) -> u32

pub fn framebuffer_size(&self) -> u32

pub fn framebuffer_stride(&self) -> u32

pub fn device_private_size(&self) -> u32

pub fn device_private(&self) -> &[u32]

impl Reply<xcb_xf86dri_auth_connection_reply_t>

pub fn authenticated(&self) -> u32

impl Reply<xcb_xf86vidmode_query_version_reply_t>

pub fn major_version(&self) -> u16

pub fn minor_version(&self) -> u16

impl Reply<xcb_xf86vidmode_get_mode_line_reply_t>

pub fn dotclock(&self) -> Dotclock

pub fn hdisplay(&self) -> u16

pub fn hsyncstart(&self) -> u16

pub fn hsyncend(&self) -> u16

pub fn htotal(&self) -> u16

pub fn hskew(&self) -> u16

pub fn vdisplay(&self) -> u16

pub fn vsyncstart(&self) -> u16

pub fn vsyncend(&self) -> u16

pub fn vtotal(&self) -> u16

pub fn flags(&self) -> u32

pub fn privsize(&self) -> u32

pub fn private(&self) -> &[u8]

impl Reply<xcb_xf86vidmode_get_monitor_reply_t>

pub fn vendor_length(&self) -> u8

pub fn model_length(&self) -> u8

pub fn num_hsync(&self) -> u8

pub fn num_vsync(&self) -> u8

pub fn hsync(&self) -> &[Syncrange]

pub fn vsync(&self) -> &[Syncrange]

pub fn vendor(&self) -> &str

pub fn alignment_pad<T>(&self) -> &[T]

pub fn model(&self) -> &str

impl Reply<xcb_xf86vidmode_get_all_mode_lines_reply_t>

pub fn modecount(&self) -> u32

pub fn modeinfo(&self) -> ModeInfoIterator

impl Reply<xcb_xf86vidmode_validate_mode_line_reply_t>

pub fn status(&self) -> u32

impl Reply<xcb_xf86vidmode_get_view_port_reply_t>

pub fn x(&self) -> u32

pub fn y(&self) -> u32

impl Reply<xcb_xf86vidmode_get_dot_clocks_reply_t>

pub fn flags(&self) -> u32

pub fn clocks(&self) -> u32

pub fn maxclocks(&self) -> u32

pub fn clock(&self) -> &[u32]

impl Reply<xcb_xf86vidmode_get_gamma_reply_t>

pub fn red(&self) -> u32

pub fn green(&self) -> u32

pub fn blue(&self) -> u32

impl Reply<xcb_xf86vidmode_get_gamma_ramp_reply_t>

pub fn size(&self) -> u16

pub fn red(&self) -> &[u16]

pub fn green(&self) -> &[u16]

pub fn blue(&self) -> &[u16]

impl Reply<xcb_xf86vidmode_get_gamma_ramp_size_reply_t>

pub fn size(&self) -> u16

impl Reply<xcb_xf86vidmode_get_permissions_reply_t>

pub fn permissions(&self) -> u32

impl Reply<xcb_xfixes_query_version_reply_t>

pub fn major_version(&self) -> u32

pub fn minor_version(&self) -> u32

impl Reply<xcb_xfixes_get_cursor_image_reply_t>

pub fn x(&self) -> i16

pub fn y(&self) -> i16

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn xhot(&self) -> u16

pub fn yhot(&self) -> u16

pub fn cursor_serial(&self) -> u32

pub fn cursor_image(&self) -> &[u32]

impl Reply<xcb_xfixes_fetch_region_reply_t>

pub fn extents(&self) -> Rectangle

pub fn rectangles(&self) -> RectangleIterator

impl Reply<xcb_xfixes_get_cursor_name_reply_t>

pub fn atom(&self) -> Atom

pub fn nbytes(&self) -> u16

pub fn name(&self) -> &str

impl Reply<xcb_xfixes_get_cursor_image_and_name_reply_t>

pub fn x(&self) -> i16

pub fn y(&self) -> i16

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn xhot(&self) -> u16

pub fn yhot(&self) -> u16

pub fn cursor_serial(&self) -> u32

pub fn cursor_atom(&self) -> Atom

pub fn nbytes(&self) -> u16

pub fn name(&self) -> &str

pub fn cursor_image(&self) -> &[u32]

impl Reply<xcb_xinerama_query_version_reply_t>

pub fn major(&self) -> u16

pub fn minor(&self) -> u16

impl Reply<xcb_xinerama_get_state_reply_t>

pub fn state(&self) -> u8

pub fn window(&self) -> Window

impl Reply<xcb_xinerama_get_screen_count_reply_t>

pub fn screen_count(&self) -> u8

pub fn window(&self) -> Window

impl Reply<xcb_xinerama_get_screen_size_reply_t>

pub fn width(&self) -> u32

pub fn height(&self) -> u32

pub fn window(&self) -> Window

pub fn screen(&self) -> u32

impl Reply<xcb_xinerama_is_active_reply_t>

pub fn state(&self) -> u32

impl Reply<xcb_xinerama_query_screens_reply_t>

pub fn number(&self) -> u32

pub fn screen_info(&self) -> ScreenInfoIterator

impl Reply<xcb_xkb_use_extension_reply_t>

pub fn supported(&self) -> bool

pub fn server_major(&self) -> u16

pub fn server_minor(&self) -> u16

impl Reply<xcb_xkb_get_state_reply_t>

pub fn device_id(&self) -> u8

pub fn mods(&self) -> u8

pub fn base_mods(&self) -> u8

pub fn latched_mods(&self) -> u8

pub fn locked_mods(&self) -> u8

pub fn group(&self) -> u8

pub fn locked_group(&self) -> u8

pub fn base_group(&self) -> i16

pub fn latched_group(&self) -> i16

pub fn compat_state(&self) -> u8

pub fn grab_mods(&self) -> u8

pub fn compat_grab_mods(&self) -> u8

pub fn lookup_mods(&self) -> u8

pub fn compat_lookup_mods(&self) -> u8

pub fn ptr_btn_state(&self) -> u16

impl Reply<xcb_xkb_get_controls_reply_t>

pub fn device_id(&self) -> u8

pub fn mouse_keys_dflt_btn(&self) -> u8

pub fn num_groups(&self) -> u8

pub fn groups_wrap(&self) -> u8

pub fn internal_mods_mask(&self) -> u8

pub fn ignore_lock_mods_mask(&self) -> u8

pub fn internal_mods_real_mods(&self) -> u8

pub fn ignore_lock_mods_real_mods(&self) -> u8

pub fn internal_mods_vmods(&self) -> u16

pub fn ignore_lock_mods_vmods(&self) -> u16

pub fn repeat_delay(&self) -> u16

pub fn repeat_interval(&self) -> u16

pub fn slow_keys_delay(&self) -> u16

pub fn debounce_delay(&self) -> u16

pub fn mouse_keys_delay(&self) -> u16

pub fn mouse_keys_interval(&self) -> u16

pub fn mouse_keys_time_to_max(&self) -> u16

pub fn mouse_keys_max_speed(&self) -> u16

pub fn mouse_keys_curve(&self) -> i16

pub fn access_x_option(&self) -> u16

pub fn access_x_timeout(&self) -> u16

pub fn access_x_timeout_options_mask(&self) -> u16

pub fn access_x_timeout_options_values(&self) -> u16

pub fn access_x_timeout_mask(&self) -> u32

pub fn access_x_timeout_values(&self) -> u32

pub fn enabled_controls(&self) -> u32

pub fn per_key_repeat(&self) -> &[u8]

impl Reply<xcb_xkb_get_map_reply_t>

pub fn device_id(&self) -> u8

pub fn min_key_code(&self) -> Keycode

pub fn max_key_code(&self) -> Keycode

pub fn present(&self) -> u16

pub fn first_type(&self) -> u8

pub fn n_types(&self) -> u8

pub fn total_types(&self) -> u8

pub fn first_key_sym(&self) -> Keycode

pub fn total_syms(&self) -> u16

pub fn n_key_syms(&self) -> u8

pub fn first_key_action(&self) -> Keycode

pub fn total_actions(&self) -> u16

pub fn n_key_actions(&self) -> u8

pub fn first_key_behavior(&self) -> Keycode

pub fn n_key_behaviors(&self) -> u8

pub fn total_key_behaviors(&self) -> u8

pub fn first_key_explicit(&self) -> Keycode

pub fn n_key_explicit(&self) -> u8

pub fn total_key_explicit(&self) -> u8

pub fn first_mod_map_key(&self) -> Keycode

pub fn n_mod_map_keys(&self) -> u8

pub fn total_mod_map_keys(&self) -> u8

pub fn first_v_mod_map_key(&self) -> Keycode

pub fn n_v_mod_map_keys(&self) -> u8

pub fn total_v_mod_map_keys(&self) -> u8

pub fn virtual_mods(&self) -> u16

impl Reply<xcb_xkb_get_compat_map_reply_t>

pub fn device_id(&self) -> u8

pub fn groups_rtrn(&self) -> u8

pub fn first_si_rtrn(&self) -> u16

pub fn n_si_rtrn(&self) -> u16

pub fn n_total_si(&self) -> u16

pub fn si_rtrn(&self) -> SymInterpretIterator<'_>

pub fn group_rtrn(&self) -> ModDefIterator

impl Reply<xcb_xkb_get_indicator_state_reply_t>

pub fn device_id(&self) -> u8

pub fn state(&self) -> u32

impl Reply<xcb_xkb_get_indicator_map_reply_t>

pub fn device_id(&self) -> u8

pub fn which(&self) -> u32

pub fn real_indicators(&self) -> u32

pub fn n_indicators(&self) -> u8

pub fn maps(&self) -> IndicatorMapIterator

impl Reply<xcb_xkb_get_named_indicator_reply_t>

pub fn device_id(&self) -> u8

pub fn indicator(&self) -> Atom

pub fn found(&self) -> bool

pub fn on(&self) -> bool

pub fn real_indicator(&self) -> bool

pub fn ndx(&self) -> u8

pub fn map_flags(&self) -> u8

pub fn map_which_groups(&self) -> u8

pub fn map_groups(&self) -> u8

pub fn map_which_mods(&self) -> u8

pub fn map_mods(&self) -> u8

pub fn map_real_mods(&self) -> u8

pub fn map_vmod(&self) -> u16

pub fn map_ctrls(&self) -> u32

pub fn supported(&self) -> bool

impl Reply<xcb_xkb_get_names_reply_t>

pub fn device_id(&self) -> u8

pub fn which(&self) -> u32

pub fn min_key_code(&self) -> Keycode

pub fn max_key_code(&self) -> Keycode

pub fn n_types(&self) -> u8

pub fn group_names(&self) -> u8

pub fn virtual_mods(&self) -> u16

pub fn first_key(&self) -> Keycode

pub fn n_keys(&self) -> u8

pub fn indicators(&self) -> u32

pub fn n_radio_groups(&self) -> u8

pub fn n_key_aliases(&self) -> u8

pub fn n_kt_levels(&self) -> u16

impl Reply<xcb_xkb_per_client_flags_reply_t>

pub fn device_id(&self) -> u8

pub fn supported(&self) -> u32

pub fn value(&self) -> u32

pub fn auto_ctrls(&self) -> u32

pub fn auto_ctrls_values(&self) -> u32

impl Reply<xcb_xkb_list_components_reply_t>

pub fn device_id(&self) -> u8

pub fn n_keymaps(&self) -> u16

pub fn n_keycodes(&self) -> u16

pub fn n_types(&self) -> u16

pub fn n_compat_maps(&self) -> u16

pub fn n_symbols(&self) -> u16

pub fn n_geometries(&self) -> u16

pub fn extra(&self) -> u16

pub fn keymaps(&self) -> ListingIterator<'_>

pub fn keycodes(&self) -> ListingIterator<'_>

pub fn types(&self) -> ListingIterator<'_>

pub fn compat_maps(&self) -> ListingIterator<'_>

pub fn symbols(&self) -> ListingIterator<'_>

pub fn geometries(&self) -> ListingIterator<'_>

impl Reply<xcb_xkb_get_kbd_by_name_reply_t>

pub fn device_id(&self) -> u8

pub fn min_key_code(&self) -> Keycode

pub fn max_key_code(&self) -> Keycode

pub fn loaded(&self) -> bool

pub fn new_keyboard(&self) -> bool

pub fn found(&self) -> u16

pub fn reported(&self) -> u16

impl Reply<xcb_xkb_get_device_info_reply_t>

pub fn device_id(&self) -> u8

pub fn present(&self) -> u16

pub fn supported(&self) -> u16

pub fn unsupported(&self) -> u16

pub fn n_device_led_f_bs(&self) -> u16

pub fn first_btn_wanted(&self) -> u8

pub fn n_btns_wanted(&self) -> u8

pub fn first_btn_rtrn(&self) -> u8

pub fn n_btns_rtrn(&self) -> u8

pub fn total_btns(&self) -> u8

pub fn has_own_state(&self) -> bool

pub fn dflt_kbd_fb(&self) -> u16

pub fn dflt_led_fb(&self) -> u16

pub fn dev_type(&self) -> Atom

pub fn name_len(&self) -> u16

pub fn name(&self) -> &[String8]

pub fn btn_actions(&self) -> ActionIterator

pub fn leds(&self) -> DeviceLedInfoIterator<'_>

impl Reply<xcb_xkb_set_debugging_flags_reply_t>

pub fn current_flags(&self) -> u32

pub fn current_ctrls(&self) -> u32

pub fn supported_flags(&self) -> u32

pub fn supported_ctrls(&self) -> u32

impl Reply<xcb_xvmc_query_version_reply_t>

pub fn major(&self) -> u32

pub fn minor(&self) -> u32

impl Reply<xcb_xvmc_list_surface_types_reply_t>

pub fn num(&self) -> u32

pub fn surfaces(&self) -> SurfaceInfoIterator

impl Reply<xcb_xvmc_create_context_reply_t>

pub fn width_actual(&self) -> u16

pub fn height_actual(&self) -> u16

pub fn flags_return(&self) -> u32

pub fn priv_data(&self) -> &[u32]

impl Reply<xcb_xvmc_create_surface_reply_t>

pub fn priv_data(&self) -> &[u32]

impl Reply<xcb_xvmc_create_subpicture_reply_t>

pub fn width_actual(&self) -> u16

pub fn height_actual(&self) -> u16

pub fn num_palette_entries(&self) -> u16

pub fn entry_bytes(&self) -> u16

pub fn component_order(&self) -> &[u8]

pub fn priv_data(&self) -> &[u32]

impl Reply<xcb_xvmc_list_subpicture_types_reply_t>

pub fn num(&self) -> u32

pub fn types(&self) -> ImageFormatInfoIterator<'_>

impl Reply<xcb_xv_query_extension_reply_t>

pub fn major(&self) -> u16

pub fn minor(&self) -> u16

impl Reply<xcb_xv_query_adaptors_reply_t>

pub fn num_adaptors(&self) -> u16

pub fn info(&self) -> AdaptorInfoIterator<'_>

impl Reply<xcb_xv_query_encodings_reply_t>

pub fn num_encodings(&self) -> u16

pub fn info(&self) -> EncodingInfoIterator<'_>

impl Reply<xcb_xv_grab_port_reply_t>

pub fn result(&self) -> u8

impl Reply<xcb_xv_query_best_size_reply_t>

pub fn actual_width(&self) -> u16

pub fn actual_height(&self) -> u16

impl Reply<xcb_xv_get_port_attribute_reply_t>

pub fn value(&self) -> i32

impl Reply<xcb_xv_query_port_attributes_reply_t>

pub fn num_attributes(&self) -> u32

pub fn text_size(&self) -> u32

pub fn attributes(&self) -> AttributeInfoIterator<'_>

impl Reply<xcb_xv_list_image_formats_reply_t>

pub fn num_formats(&self) -> u32

pub fn format(&self) -> ImageFormatInfoIterator<'_>

impl Reply<xcb_xv_query_image_attributes_reply_t>

pub fn num_planes(&self) -> u32

pub fn data_size(&self) -> u32

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn pitches(&self) -> &[u32]

pub fn offsets(&self) -> &[u32]