xcb::base

Struct StructPtr

pub struct StructPtr<'a, T: 'a> {
    pub ptr: *mut T,
    /* private fields */
}

StructPtr is a wrapper for pointer to struct owned by XCB that must not be freed it is instead bound to the lifetime of its parent that it borrows immutably

Fields
ptr: *mut T
Implementations

impl<'a> StructPtr<'a, xcb_depth_t>

pub fn depth(&self) -> u8

pub fn visuals_len(&self) -> u16

pub fn visuals(&self) -> VisualtypeIterator

impl<'a> StructPtr<'a, xcb_screen_t>

pub fn root(&self) -> Window

pub fn default_colormap(&self) -> Colormap

pub fn white_pixel(&self) -> u32

pub fn black_pixel(&self) -> u32

pub fn current_input_masks(&self) -> u32

pub fn width_in_pixels(&self) -> u16

pub fn height_in_pixels(&self) -> u16

pub fn width_in_millimeters(&self) -> u16

pub fn height_in_millimeters(&self) -> u16

pub fn min_installed_maps(&self) -> u16

pub fn max_installed_maps(&self) -> u16

pub fn root_visual(&self) -> Visualid

pub fn backing_stores(&self) -> u8

pub fn save_unders(&self) -> bool

pub fn root_depth(&self) -> u8

pub fn allowed_depths_len(&self) -> u8

pub fn allowed_depths(&self) -> DepthIterator<'a>

impl<'a> StructPtr<'a, xcb_setup_request_t>

pub fn byte_order(&self) -> u8

pub fn protocol_major_version(&self) -> u16

pub fn protocol_minor_version(&self) -> u16

pub fn authorization_protocol_name_len(&self) -> u16

pub fn authorization_protocol_data_len(&self) -> u16

pub fn authorization_protocol_name(&self) -> &str

pub fn authorization_protocol_data(&self) -> &str

impl<'a> StructPtr<'a, xcb_setup_failed_t>

pub fn status(&self) -> u8

pub fn reason_len(&self) -> u8

pub fn protocol_major_version(&self) -> u16

pub fn protocol_minor_version(&self) -> u16

pub fn length(&self) -> u16

pub fn reason(&self) -> &str

impl<'a> StructPtr<'a, xcb_setup_authenticate_t>

pub fn status(&self) -> u8

pub fn length(&self) -> u16

pub fn reason(&self) -> &str

impl<'a> StructPtr<'a, xcb_setup_t>

pub fn status(&self) -> u8

pub fn protocol_major_version(&self) -> u16

pub fn protocol_minor_version(&self) -> u16

pub fn length(&self) -> u16

pub fn release_number(&self) -> u32

pub fn resource_id_base(&self) -> u32

pub fn resource_id_mask(&self) -> u32

pub fn motion_buffer_size(&self) -> u32

pub fn vendor_len(&self) -> u16

pub fn maximum_request_length(&self) -> u16

pub fn roots_len(&self) -> u8

pub fn pixmap_formats_len(&self) -> u8

pub fn image_byte_order(&self) -> u8

pub fn bitmap_format_bit_order(&self) -> u8

pub fn bitmap_format_scanline_unit(&self) -> u8

pub fn bitmap_format_scanline_pad(&self) -> u8

pub fn min_keycode(&self) -> Keycode

pub fn max_keycode(&self) -> Keycode

pub fn vendor(&self) -> &str

pub fn pixmap_formats(&self) -> FormatIterator

pub fn roots(&self) -> ScreenIterator<'a>

impl<'a> StructPtr<'a, xcb_str_t>

pub fn name_len(&self) -> u8

pub fn name(&self) -> &str

impl<'a> StructPtr<'a, xcb_host_t>

pub fn family(&self) -> u8

pub fn address_len(&self) -> u16

pub fn address(&self) -> &[u8]

impl<'a> StructPtr<'a, xcb_render_pictdepth_t>

pub fn depth(&self) -> u8

pub fn num_visuals(&self) -> u16

pub fn visuals(&self) -> PictvisualIterator

impl<'a> StructPtr<'a, xcb_render_pictscreen_t>

pub fn num_depths(&self) -> u32

pub fn fallback(&self) -> Pictformat

pub fn depths(&self) -> PictdepthIterator<'a>