Struct xcb::base::StructPtr

pub struct StructPtr<'a, T: 'a> {
    pub ptr: *mut T,
    // some fields omitted
}

StructPtr is a wrapper for pointer to struct owned by XCB that must not be freed it is instead bound to the lifetime of its parent that it borrows immutably

Fields
ptr: *mut T
Implementations

impl<'a> StructPtr<'a, xcb_depth_t>

pub fn depth(&self) -> u8

pub fn visuals_len(&self) -> u16

pub fn visuals(&self) -> VisualtypeIterator

impl<'a> StructPtr<'a, xcb_screen_t>

pub fn root(&self) -> Window

pub fn default_colormap(&self) -> Colormap

pub fn white_pixel(&self) -> u32

pub fn black_pixel(&self) -> u32

pub fn current_input_masks(&self) -> u32

pub fn width_in_pixels(&self) -> u16

pub fn height_in_pixels(&self) -> u16

pub fn width_in_millimeters(&self) -> u16

pub fn height_in_millimeters(&self) -> u16

pub fn min_installed_maps(&self) -> u16

pub fn max_installed_maps(&self) -> u16

pub fn root_visual(&self) -> Visualid

pub fn backing_stores(&self) -> u8

pub fn save_unders(&self) -> bool

pub fn root_depth(&self) -> u8

pub fn allowed_depths_len(&self) -> u8

pub fn allowed_depths(&self) -> DepthIterator<'a>

impl<'a> StructPtr<'a, xcb_setup_request_t>

pub fn byte_order(&self) -> u8

pub fn protocol_major_version(&self) -> u16

pub fn protocol_minor_version(&self) -> u16

pub fn authorization_protocol_name_len(&self) -> u16

pub fn authorization_protocol_data_len(&self) -> u16

pub fn authorization_protocol_name(&self) -> &str

pub fn authorization_protocol_data(&self) -> &str

impl<'a> StructPtr<'a, xcb_setup_failed_t>

pub fn status(&self) -> u8

pub fn reason_len(&self) -> u8

pub fn protocol_major_version(&self) -> u16

pub fn protocol_minor_version(&self) -> u16

pub fn length(&self) -> u16

pub fn reason(&self) -> &str

impl<'a> StructPtr<'a, xcb_setup_authenticate_t>

pub fn status(&self) -> u8

pub fn length(&self) -> u16

pub fn reason(&self) -> &str

impl<'a> StructPtr<'a, xcb_setup_t>

pub fn status(&self) -> u8

pub fn protocol_major_version(&self) -> u16

pub fn protocol_minor_version(&self) -> u16

pub fn length(&self) -> u16

pub fn release_number(&self) -> u32

pub fn resource_id_base(&self) -> u32

pub fn resource_id_mask(&self) -> u32

pub fn motion_buffer_size(&self) -> u32

pub fn vendor_len(&self) -> u16

pub fn maximum_request_length(&self) -> u16

pub fn roots_len(&self) -> u8

pub fn pixmap_formats_len(&self) -> u8

pub fn image_byte_order(&self) -> u8

pub fn bitmap_format_bit_order(&self) -> u8

pub fn bitmap_format_scanline_unit(&self) -> u8

pub fn bitmap_format_scanline_pad(&self) -> u8

pub fn min_keycode(&self) -> Keycode

pub fn max_keycode(&self) -> Keycode

pub fn vendor(&self) -> &str

pub fn pixmap_formats(&self) -> FormatIterator

pub fn roots(&self) -> ScreenIterator<'a>

impl<'a> StructPtr<'a, xcb_str_t>

pub fn name_len(&self) -> u8

pub fn name(&self) -> &str

impl<'a> StructPtr<'a, xcb_host_t>

pub fn family(&self) -> u8

pub fn address_len(&self) -> u16

pub fn address(&self) -> &[u8]

impl<'a> StructPtr<'a, xcb_randr_refresh_rates_t>

pub fn n_rates(&self) -> u16

pub fn rates(&self) -> &[u16]

impl<'a> StructPtr<'a, xcb_record_client_info_t>

pub fn client_resource(&self) -> ClientSpec

pub fn num_ranges(&self) -> u32

pub fn ranges(&self) -> RangeIterator

impl<'a> StructPtr<'a, xcb_render_pictdepth_t>

pub fn depth(&self) -> u8

pub fn num_visuals(&self) -> u16

pub fn visuals(&self) -> PictvisualIterator

impl<'a> StructPtr<'a, xcb_render_pictscreen_t>

pub fn num_depths(&self) -> u32

pub fn fallback(&self) -> Pictformat

pub fn depths(&self) -> PictdepthIterator<'a>

impl<'a> StructPtr<'a, xcb_res_client_id_value_t>

pub fn spec(&self) -> ClientIdSpec

pub fn length(&self) -> u32

pub fn value(&self) -> &[u32]

impl<'a> StructPtr<'a, xcb_res_resource_size_value_t>

pub fn size(&self) -> ResourceSizeSpec

pub fn num_cross_references(&self) -> u32

pub fn cross_references(&self) -> ResourceSizeSpecIterator

impl<'a> StructPtr<'a, xcb_selinux_list_item_t>

pub fn name(&self) -> Atom

pub fn object_context_len(&self) -> u32

pub fn data_context_len(&self) -> u32

pub fn object_context(&self) -> &str

pub fn data_context(&self) -> &str

impl<'a> StructPtr<'a, xcb_x_print_printer_t>

pub fn name_len(&self) -> u32

pub fn name(&self) -> &[String8]

pub fn desc_len(&self) -> u32

pub fn description(&self) -> &[String8]

impl<'a> StructPtr<'a, xcb_xkb_key_name_t>

pub fn name(&self) -> &[c_char]

impl<'a> StructPtr<'a, xcb_xkb_key_alias_t>

pub fn real(&self) -> &[c_char]

pub fn alias(&self) -> &[c_char]

impl<'a> StructPtr<'a, xcb_xkb_counted_string_16_t>

pub fn length(&self) -> u16

pub fn string(&self) -> &str

pub fn alignment_pad<T>(&self) -> &[T]

impl<'a> StructPtr<'a, xcb_xkb_key_type_t>

pub fn mods_mask(&self) -> u8

pub fn mods_mods(&self) -> u8

pub fn mods_vmods(&self) -> u16

pub fn num_levels(&self) -> u8

pub fn n_map_entries(&self) -> u8

pub fn has_preserve(&self) -> bool

pub fn map(&self) -> KtMapEntryIterator

pub fn preserve(&self) -> ModDefIterator

impl<'a> StructPtr<'a, xcb_xkb_key_sym_map_t>

pub fn kt_index(&self) -> &[u8]

pub fn group_info(&self) -> u8

pub fn width(&self) -> u8

pub fn n_syms(&self) -> u16

pub fn syms(&self) -> &[Keysym]

impl<'a> StructPtr<'a, xcb_xkb_set_behavior_t>

pub fn keycode(&self) -> Keycode

pub fn behavior(&'a self) -> &'a Behavior

impl<'a> StructPtr<'a, xcb_xkb_set_key_type_t>

pub fn mask(&self) -> u8

pub fn real_mods(&self) -> u8

pub fn virtual_mods(&self) -> u16

pub fn num_levels(&self) -> u8

pub fn n_map_entries(&self) -> u8

pub fn preserve(&self) -> bool

pub fn entries(&self) -> KtSetMapEntryIterator

pub fn preserve_entries(&self) -> KtSetMapEntryIterator

impl<'a> StructPtr<'a, xcb_xkb_outline_t>

pub fn n_points(&self) -> u8

pub fn corner_radius(&self) -> u8

pub fn points(&self) -> PointIterator

impl<'a> StructPtr<'a, xcb_xkb_shape_t>

pub fn name(&self) -> Atom

pub fn n_outlines(&self) -> u8

pub fn primary_ndx(&self) -> u8

pub fn approx_ndx(&self) -> u8

pub fn outlines(&self) -> OutlineIterator<'a>

impl<'a> StructPtr<'a, xcb_xkb_key_t>

pub fn name(&self) -> &[String8]

pub fn gap(&self) -> i16

pub fn shape_ndx(&self) -> u8

pub fn color_ndx(&self) -> u8

impl<'a> StructPtr<'a, xcb_xkb_overlay_key_t>

pub fn over(&self) -> &[String8]

pub fn under(&self) -> &[String8]

impl<'a> StructPtr<'a, xcb_xkb_overlay_row_t>

pub fn row_under(&self) -> u8

pub fn n_keys(&self) -> u8

pub fn keys(&self) -> OverlayKeyIterator<'a>

impl<'a> StructPtr<'a, xcb_xkb_overlay_t>

pub fn name(&self) -> Atom

pub fn n_rows(&self) -> u8

pub fn rows(&self) -> OverlayRowIterator<'a>

impl<'a> StructPtr<'a, xcb_xkb_row_t>

pub fn top(&self) -> i16

pub fn left(&self) -> i16

pub fn n_keys(&self) -> u8

pub fn vertical(&self) -> bool

pub fn keys(&self) -> KeyIterator<'a>

impl<'a> StructPtr<'a, xcb_xkb_listing_t>

pub fn flags(&self) -> u16

pub fn length(&self) -> u16

pub fn string(&self) -> &[String8]

impl<'a> StructPtr<'a, xcb_xkb_device_led_info_t>

pub fn led_class(&self) -> LedClassSpec

pub fn led_id(&self) -> IdSpec

pub fn names_present(&self) -> u32

pub fn maps_present(&self) -> u32

pub fn phys_indicators(&self) -> u32

pub fn state(&self) -> u32

pub fn names(&self) -> &[Atom]

pub fn maps(&self) -> IndicatorMapIterator

impl<'a> StructPtr<'a, xcb_xkb_sa_action_message_t>

pub fn type_(&self) -> u8

pub fn flags(&self) -> u8

pub fn message(&self) -> &[u8]

impl<'a> StructPtr<'a, xcb_xkb_si_action_t>

pub fn type_(&self) -> u8

pub fn data(&self) -> &[u8]

impl<'a> StructPtr<'a, xcb_xkb_sym_interpret_t>

pub fn sym(&self) -> Keysym

pub fn mods(&self) -> u8

pub fn match_(&self) -> u8

pub fn virtual_mod(&self) -> u8

pub fn flags(&self) -> u8

pub fn action(&self) -> SiAction<'_>

impl<'a> StructPtr<'a, xcb_xv_adaptor_info_t>

pub fn base_id(&self) -> Port

pub fn name_size(&self) -> u16

pub fn num_ports(&self) -> u16

pub fn num_formats(&self) -> u16

pub fn type_(&self) -> u8

pub fn name(&self) -> &str

pub fn formats(&self) -> FormatIterator

impl<'a> StructPtr<'a, xcb_xv_encoding_info_t>

pub fn encoding(&self) -> Encoding

pub fn name_size(&self) -> u16

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn rate(&self) -> Rational

pub fn name(&self) -> &str

impl<'a> StructPtr<'a, xcb_xv_image_t>

pub fn id(&self) -> u32

pub fn width(&self) -> u16

pub fn height(&self) -> u16

pub fn data_size(&self) -> u32

pub fn num_planes(&self) -> u32

pub fn pitches(&self) -> &[u32]

pub fn offsets(&self) -> &[u32]

pub fn data(&self) -> &[u8]

impl<'a> StructPtr<'a, xcb_xv_attribute_info_t>

pub fn flags(&self) -> u32

pub fn min(&self) -> i32

pub fn max(&self) -> i32

pub fn size(&self) -> u32

pub fn name(&self) -> &str

impl<'a> StructPtr<'a, xcb_xv_image_format_info_t>

pub fn id(&self) -> u32

pub fn type_(&self) -> u8

pub fn byte_order(&self) -> u8

pub fn guid(&self) -> &[u8]

pub fn bpp(&self) -> u8

pub fn num_planes(&self) -> u8

pub fn depth(&self) -> u8

pub fn red_mask(&self) -> u32

pub fn green_mask(&self) -> u32

pub fn blue_mask(&self) -> u32

pub fn format(&self) -> u8

pub fn y_sample_bits(&self) -> u32

pub fn u_sample_bits(&self) -> u32

pub fn v_sample_bits(&self) -> u32

pub fn vhorz_y_period(&self) -> u32

pub fn vhorz_u_period(&self) -> u32

pub fn vhorz_v_period(&self) -> u32

pub fn vvert_y_period(&self) -> u32

pub fn vvert_u_period(&self) -> u32

pub fn vvert_v_period(&self) -> u32

pub fn vcomp_order(&self) -> &[u8]

pub fn vscanline_order(&self) -> u8