Module xcb::composite

The Composite X extension.

Accessible with the composite cargo feature.

Structs

CreateRegionFromBorderClip

The CreateRegionFromBorderClip request.

GetOverlayWindow

The GetOverlayWindow request.

GetOverlayWindowCookie

Cookie type for GetOverlayWindow.

GetOverlayWindowCookieUnchecked

Unchecked cookie type for GetOverlayWindow.

GetOverlayWindowReply

Reply type for GetOverlayWindow.

NameWindowPixmap

The NameWindowPixmap request.

QueryVersion

The QueryVersion request.

QueryVersionCookie

Cookie type for QueryVersion.

QueryVersionCookieUnchecked

Unchecked cookie type for QueryVersion.

QueryVersionReply

Reply type for QueryVersion.

RedirectSubwindows

The RedirectSubwindows request.

RedirectWindow

The RedirectWindow request.

ReleaseOverlayWindow

The ReleaseOverlayWindow request.

UnredirectSubwindows

The UnredirectSubwindows request.

UnredirectWindow

The UnredirectWindow request.

Enums

Redirect

Constants

MAJOR_VERSION

The major version of the Composite extension.

MINOR_VERSION

The minor version of the Composite extension.

VERSION_STRING

The version string of the Composite extension.

XNAME

The official identifier for the Composite extension.

Functions

get_extension_data

Fetch server runtime info data of the Composite extension.

prefetch_extension_data

Prefetch server runtime info data of the Composite extension.