Struct xcb::ffi::base::xcb_auth_info_t

pub struct xcb_auth_info_t {
    pub namelen: c_int,
    pub name: *mut c_char,
    pub datalen: c_int,
    pub data: *mut c_char,
}

Container for authorization information. A container for authorization information to be sent to the X server

Fields

namelen

length of the string name (as returned by strlen)

name

String containing the authentication protocol name, such as "MIT-MAGIC-COOKIE-1" or "XDM-AUTHORIZATION-1".

datalen

length of the data member

data

data interpreted in a protocol specific manner