Module xcb::ffi::dpms

Structs

xcb_dpms_capable_cookie_t

xcb_dpms_capable_reply_t

xcb_dpms_capable_request_t

xcb_dpms_disable_request_t

xcb_dpms_enable_request_t

xcb_dpms_force_level_request_t

xcb_dpms_get_timeouts_cookie_t

xcb_dpms_get_timeouts_reply_t

xcb_dpms_get_timeouts_request_t

xcb_dpms_get_version_cookie_t

xcb_dpms_get_version_reply_t

xcb_dpms_get_version_request_t

xcb_dpms_info_cookie_t

xcb_dpms_info_reply_t

xcb_dpms_info_request_t

xcb_dpms_set_timeouts_request_t

Constants

XCB_DPMS_CAPABLE

XCB_DPMS_DISABLE

XCB_DPMS_DPMS_MODE_OFF

XCB_DPMS_DPMS_MODE_ON

XCB_DPMS_DPMS_MODE_STANDBY

XCB_DPMS_DPMS_MODE_SUSPEND

XCB_DPMS_ENABLE

XCB_DPMS_FORCE_LEVEL

XCB_DPMS_GET_TIMEOUTS

XCB_DPMS_GET_VERSION

XCB_DPMS_INFO

XCB_DPMS_MAJOR_VERSION

XCB_DPMS_MINOR_VERSION

XCB_DPMS_SET_TIMEOUTS

Statics

xcb_dpms_id

Functions

xcb_dpms_capable⚠

xcb_dpms_capable_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_dpms_capable_unchecked⚠

xcb_dpms_disable⚠

xcb_dpms_disable_checked⚠

xcb_dpms_enable⚠

xcb_dpms_enable_checked⚠

xcb_dpms_force_level⚠

xcb_dpms_force_level_checked⚠

xcb_dpms_get_timeouts⚠

xcb_dpms_get_timeouts_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_dpms_get_timeouts_unchecked⚠

xcb_dpms_get_version⚠

xcb_dpms_get_version_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_dpms_get_version_unchecked⚠

xcb_dpms_info⚠

xcb_dpms_info_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_dpms_info_unchecked⚠

xcb_dpms_set_timeouts⚠

xcb_dpms_set_timeouts_checked⚠

Type Definitions

xcb_dpms_dpms_mode_t