Module xcb::ffi::dri3

Structs

xcb_dri3_buffer_from_pixmap_cookie_t

xcb_dri3_buffer_from_pixmap_reply_t

xcb_dri3_buffer_from_pixmap_request_t

xcb_dri3_fd_from_fence_cookie_t

xcb_dri3_fd_from_fence_reply_t

xcb_dri3_fd_from_fence_request_t

xcb_dri3_fence_from_fd_request_t

xcb_dri3_open_cookie_t

xcb_dri3_open_reply_t

xcb_dri3_open_request_t

xcb_dri3_pixmap_from_buffer_request_t

xcb_dri3_query_version_cookie_t

xcb_dri3_query_version_reply_t

xcb_dri3_query_version_request_t

Constants

XCB_DRI3_BUFFER_FROM_PIXMAP

XCB_DRI3_FD_FROM_FENCE

XCB_DRI3_FENCE_FROM_FD

XCB_DRI3_MAJOR_VERSION

XCB_DRI3_MINOR_VERSION

XCB_DRI3_OPEN

XCB_DRI3_PIXMAP_FROM_BUFFER

XCB_DRI3_QUERY_VERSION

Statics

xcb_dri3_id

Functions

xcb_dri3_buffer_from_pixmap⚠

xcb_dri3_buffer_from_pixmap_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_dri3_buffer_from_pixmap_reply_fds⚠

xcb_dri3_buffer_from_pixmap_unchecked⚠

xcb_dri3_fd_from_fence⚠

xcb_dri3_fd_from_fence_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_dri3_fd_from_fence_reply_fds⚠

xcb_dri3_fd_from_fence_unchecked⚠

xcb_dri3_fence_from_fd⚠

xcb_dri3_fence_from_fd_checked⚠

xcb_dri3_open⚠

xcb_dri3_open_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_dri3_open_reply_fds⚠

xcb_dri3_open_unchecked⚠

xcb_dri3_pixmap_from_buffer⚠

xcb_dri3_pixmap_from_buffer_checked⚠

xcb_dri3_query_version⚠

xcb_dri3_query_version_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_dri3_query_version_unchecked⚠