Module xcb::ffi::record

Structs

xcb_record_bad_context_error_t

xcb_record_client_info_iterator_t

xcb_record_client_info_t

xcb_record_client_spec_iterator_t

xcb_record_context_iterator_t

xcb_record_create_context_request_t

xcb_record_disable_context_request_t

xcb_record_element_header_iterator_t

xcb_record_enable_context_cookie_t

xcb_record_enable_context_reply_t

xcb_record_enable_context_request_t

xcb_record_ext_range_iterator_t

xcb_record_ext_range_t

xcb_record_free_context_request_t

xcb_record_get_context_cookie_t

xcb_record_get_context_reply_t

xcb_record_get_context_request_t

xcb_record_query_version_cookie_t

xcb_record_query_version_reply_t

xcb_record_query_version_request_t

xcb_record_range_8_iterator_t

xcb_record_range_8_t

xcb_record_range_16_iterator_t

xcb_record_range_16_t

xcb_record_range_iterator_t

xcb_record_range_t

xcb_record_register_clients_request_t

xcb_record_unregister_clients_request_t

Constants

XCB_RECORD_BAD_CONTEXT

XCB_RECORD_CREATE_CONTEXT

XCB_RECORD_CS_ALL_CLIENTS

XCB_RECORD_CS_CURRENT_CLIENTS

XCB_RECORD_CS_FUTURE_CLIENTS

XCB_RECORD_DISABLE_CONTEXT

XCB_RECORD_ENABLE_CONTEXT

XCB_RECORD_FREE_CONTEXT

XCB_RECORD_GET_CONTEXT

XCB_RECORD_H_TYPE_FROM_CLIENT_SEQUENCE

XCB_RECORD_H_TYPE_FROM_CLIENT_TIME

XCB_RECORD_H_TYPE_FROM_SERVER_TIME

XCB_RECORD_MAJOR_VERSION

XCB_RECORD_MINOR_VERSION

XCB_RECORD_QUERY_VERSION

XCB_RECORD_REGISTER_CLIENTS

XCB_RECORD_UNREGISTER_CLIENTS

Statics

xcb_record_id

Functions

xcb_record_client_info_end⚠

xcb_record_client_info_next⚠

xcb_record_client_info_ranges⚠

xcb_record_client_info_ranges_iterator⚠

xcb_record_client_info_ranges_length⚠

xcb_record_client_spec_end⚠

xcb_record_client_spec_next⚠

xcb_record_context_end⚠

xcb_record_context_next⚠

xcb_record_create_context⚠

xcb_record_create_context_checked⚠

xcb_record_disable_context⚠

xcb_record_disable_context_checked⚠

xcb_record_element_header_end⚠

xcb_record_element_header_next⚠

xcb_record_enable_context⚠

xcb_record_enable_context_data⚠

xcb_record_enable_context_data_end⚠

xcb_record_enable_context_data_length⚠

xcb_record_enable_context_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_record_enable_context_unchecked⚠

xcb_record_ext_range_end⚠

xcb_record_ext_range_next⚠

xcb_record_free_context⚠

xcb_record_free_context_checked⚠

xcb_record_get_context⚠

xcb_record_get_context_intercepted_clients_iterator⚠

xcb_record_get_context_intercepted_clients_length⚠

xcb_record_get_context_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_record_get_context_unchecked⚠

xcb_record_query_version⚠

xcb_record_query_version_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_record_query_version_unchecked⚠

xcb_record_range_8_end⚠

xcb_record_range_8_next⚠

xcb_record_range_16_end⚠

xcb_record_range_16_next⚠

xcb_record_range_end⚠

xcb_record_range_next⚠

xcb_record_register_clients⚠

xcb_record_register_clients_checked⚠

xcb_record_unregister_clients⚠

xcb_record_unregister_clients_checked⚠

Type Definitions

xcb_record_client_spec_t

xcb_record_context_t

xcb_record_cs_t

xcb_record_element_header_t

xcb_record_h_type_t