xcb::ffi

Module shape

Structs

xcb_shape_combine_request_t

xcb_shape_get_rectangles_cookie_t

xcb_shape_get_rectangles_reply_t

xcb_shape_get_rectangles_request_t

xcb_shape_input_selected_cookie_t

xcb_shape_input_selected_reply_t

xcb_shape_input_selected_request_t

xcb_shape_kind_iterator_t

xcb_shape_mask_request_t

xcb_shape_notify_event_t

xcb_shape_offset_request_t

xcb_shape_op_iterator_t

xcb_shape_query_extents_cookie_t

xcb_shape_query_extents_reply_t

xcb_shape_query_extents_request_t

xcb_shape_query_version_cookie_t

xcb_shape_query_version_reply_t

xcb_shape_query_version_request_t

xcb_shape_rectangles_request_t

xcb_shape_select_input_request_t

Constants

XCB_SHAPE_COMBINE

XCB_SHAPE_GET_RECTANGLES

XCB_SHAPE_INPUT_SELECTED

XCB_SHAPE_MAJOR_VERSION

XCB_SHAPE_MASK

XCB_SHAPE_MINOR_VERSION

XCB_SHAPE_NOTIFY

XCB_SHAPE_OFFSET

XCB_SHAPE_QUERY_EXTENTS

XCB_SHAPE_QUERY_VERSION

XCB_SHAPE_RECTANGLES

XCB_SHAPE_SELECT_INPUT

XCB_SHAPE_SK_BOUNDING

XCB_SHAPE_SK_CLIP

XCB_SHAPE_SK_INPUT

XCB_SHAPE_SO_INTERSECT

XCB_SHAPE_SO_INVERT

XCB_SHAPE_SO_SET

XCB_SHAPE_SO_SUBTRACT

XCB_SHAPE_SO_UNION

Statics

xcb_shape_id⚠

Functions

xcb_shape_combine⚠

xcb_shape_combine_checked⚠

xcb_shape_get_rectangles⚠

xcb_shape_get_rectangles_rectangles⚠

xcb_shape_get_rectangles_rectangles_iterator⚠

xcb_shape_get_rectangles_rectangles_length⚠

xcb_shape_get_rectangles_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_shape_get_rectangles_unchecked⚠

xcb_shape_input_selected⚠

xcb_shape_input_selected_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_shape_input_selected_unchecked⚠

xcb_shape_kind_end⚠

xcb_shape_kind_next⚠

xcb_shape_mask⚠

xcb_shape_mask_checked⚠

xcb_shape_offset⚠

xcb_shape_offset_checked⚠

xcb_shape_op_end⚠

xcb_shape_op_next⚠

xcb_shape_query_extents⚠

xcb_shape_query_extents_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_shape_query_extents_unchecked⚠

xcb_shape_query_version⚠

xcb_shape_query_version_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_shape_query_version_unchecked⚠

xcb_shape_rectangles⚠

xcb_shape_rectangles_checked⚠

xcb_shape_select_input⚠

xcb_shape_select_input_checked⚠

Type Aliases

xcb_shape_kind_t

xcb_shape_op_t

xcb_shape_sk_t

xcb_shape_so_t