Module xcb::ffi::x_print

Structs

xcb_x_print_attribut_notify_event_t

xcb_x_print_bad_context_error_t

xcb_x_print_bad_sequence_error_t

xcb_x_print_create_context_request_t

xcb_x_print_notify_event_t

xcb_x_print_pcontext_iterator_t

xcb_x_print_print_destroy_context_request_t

xcb_x_print_print_end_doc_request_t

xcb_x_print_print_end_job_request_t

xcb_x_print_print_end_page_request_t

xcb_x_print_print_get_attributes_cookie_t

xcb_x_print_print_get_attributes_reply_t

xcb_x_print_print_get_attributes_request_t

xcb_x_print_print_get_context_cookie_t

xcb_x_print_print_get_context_reply_t

xcb_x_print_print_get_context_request_t

xcb_x_print_print_get_document_data_cookie_t

xcb_x_print_print_get_document_data_reply_t

xcb_x_print_print_get_document_data_request_t

xcb_x_print_print_get_image_resolution_cookie_t

xcb_x_print_print_get_image_resolution_reply_t

xcb_x_print_print_get_image_resolution_request_t

xcb_x_print_print_get_one_attributes_cookie_t

xcb_x_print_print_get_one_attributes_reply_t

xcb_x_print_print_get_one_attributes_request_t

xcb_x_print_print_get_page_dimensions_cookie_t

xcb_x_print_print_get_page_dimensions_reply_t

xcb_x_print_print_get_page_dimensions_request_t

xcb_x_print_print_get_printer_list_cookie_t

xcb_x_print_print_get_printer_list_reply_t

xcb_x_print_print_get_printer_list_request_t

xcb_x_print_print_get_screen_of_context_cookie_t

xcb_x_print_print_get_screen_of_context_reply_t

xcb_x_print_print_get_screen_of_context_request_t

xcb_x_print_print_input_selected_cookie_t

xcb_x_print_print_input_selected_reply_t

xcb_x_print_print_input_selected_request_t

xcb_x_print_print_put_document_data_request_t

xcb_x_print_print_query_screens_cookie_t

xcb_x_print_print_query_screens_reply_t

xcb_x_print_print_query_screens_request_t

xcb_x_print_print_query_version_cookie_t

xcb_x_print_print_query_version_reply_t

xcb_x_print_print_query_version_request_t

xcb_x_print_print_rehash_printer_list_request_t

xcb_x_print_print_select_input_request_t

xcb_x_print_print_set_attributes_request_t

xcb_x_print_print_set_context_request_t

xcb_x_print_print_set_image_resolution_cookie_t

xcb_x_print_print_set_image_resolution_reply_t

xcb_x_print_print_set_image_resolution_request_t

xcb_x_print_print_start_doc_request_t

xcb_x_print_print_start_job_request_t

xcb_x_print_print_start_page_request_t

xcb_x_print_printer_iterator_t

xcb_x_print_printer_t

xcb_x_print_string8_iterator_t

Constants

XCB_X_PRINT_ATTRIBUT_NOTIFY

XCB_X_PRINT_ATTR_DOC_ATTR

XCB_X_PRINT_ATTR_JOB_ATTR

XCB_X_PRINT_ATTR_MEDIUM_ATTR

XCB_X_PRINT_ATTR_PAGE_ATTR

XCB_X_PRINT_ATTR_PRINTER_ATTR

XCB_X_PRINT_ATTR_SERVER_ATTR

XCB_X_PRINT_ATTR_SPOOLER_ATTR

XCB_X_PRINT_BAD_CONTEXT

XCB_X_PRINT_BAD_SEQUENCE

XCB_X_PRINT_CREATE_CONTEXT

XCB_X_PRINT_DETAIL_END_DOC_NOTIFY

XCB_X_PRINT_DETAIL_END_JOB_NOTIFY

XCB_X_PRINT_DETAIL_END_PAGE_NOTIFY

XCB_X_PRINT_DETAIL_START_DOC_NOTIFY

XCB_X_PRINT_DETAIL_START_JOB_NOTIFY

XCB_X_PRINT_DETAIL_START_PAGE_NOTIFY

XCB_X_PRINT_EV_MASK_ATTRIBUTE_MASK

XCB_X_PRINT_EV_MASK_NO_EVENT_MASK

XCB_X_PRINT_EV_MASK_PRINT_MASK

XCB_X_PRINT_GET_DOC_FINISHED

XCB_X_PRINT_GET_DOC_SECOND_CONSUMER

XCB_X_PRINT_MAJOR_VERSION

XCB_X_PRINT_MINOR_VERSION

XCB_X_PRINT_NOTIFY

XCB_X_PRINT_PRINT_DESTROY_CONTEXT

XCB_X_PRINT_PRINT_END_DOC

XCB_X_PRINT_PRINT_END_JOB

XCB_X_PRINT_PRINT_END_PAGE

XCB_X_PRINT_PRINT_GET_ATTRIBUTES

XCB_X_PRINT_PRINT_GET_CONTEXT

XCB_X_PRINT_PRINT_GET_DOCUMENT_DATA

XCB_X_PRINT_PRINT_GET_IMAGE_RESOLUTION

XCB_X_PRINT_PRINT_GET_ONE_ATTRIBUTES

XCB_X_PRINT_PRINT_GET_PAGE_DIMENSIONS

XCB_X_PRINT_PRINT_GET_PRINTER_LIST

XCB_X_PRINT_PRINT_GET_SCREEN_OF_CONTEXT

XCB_X_PRINT_PRINT_INPUT_SELECTED

XCB_X_PRINT_PRINT_PUT_DOCUMENT_DATA

XCB_X_PRINT_PRINT_QUERY_SCREENS

XCB_X_PRINT_PRINT_QUERY_VERSION

XCB_X_PRINT_PRINT_REHASH_PRINTER_LIST

XCB_X_PRINT_PRINT_SELECT_INPUT

XCB_X_PRINT_PRINT_SET_ATTRIBUTES

XCB_X_PRINT_PRINT_SET_CONTEXT

XCB_X_PRINT_PRINT_SET_IMAGE_RESOLUTION

XCB_X_PRINT_PRINT_START_DOC

XCB_X_PRINT_PRINT_START_JOB

XCB_X_PRINT_PRINT_START_PAGE

Statics

xcb_x_print_id

Functions

xcb_x_print_create_context

xcb_x_print_create_context_checked

xcb_x_print_pcontext_end

xcb_x_print_pcontext_next

xcb_x_print_print_destroy_context

xcb_x_print_print_destroy_context_checked

xcb_x_print_print_end_doc

xcb_x_print_print_end_doc_checked

xcb_x_print_print_end_job

xcb_x_print_print_end_job_checked

xcb_x_print_print_end_page

xcb_x_print_print_end_page_checked

xcb_x_print_print_get_attributes

xcb_x_print_print_get_attributes_attributes

xcb_x_print_print_get_attributes_attributes_end

xcb_x_print_print_get_attributes_attributes_length

xcb_x_print_print_get_attributes_reply
the returned value must be freed by the caller using libc::free().

xcb_x_print_print_get_attributes_unchecked

xcb_x_print_print_get_context

xcb_x_print_print_get_context_reply
the returned value must be freed by the caller using libc::free().

xcb_x_print_print_get_context_unchecked

xcb_x_print_print_get_document_data

xcb_x_print_print_get_document_data_data

xcb_x_print_print_get_document_data_data_end

xcb_x_print_print_get_document_data_data_length

xcb_x_print_print_get_document_data_reply
the returned value must be freed by the caller using libc::free().

xcb_x_print_print_get_document_data_unchecked

xcb_x_print_print_get_image_resolution

xcb_x_print_print_get_image_resolution_reply
the returned value must be freed by the caller using libc::free().

xcb_x_print_print_get_image_resolution_unchecked

xcb_x_print_print_get_one_attributes

xcb_x_print_print_get_one_attributes_reply
the returned value must be freed by the caller using libc::free().

xcb_x_print_print_get_one_attributes_unchecked

xcb_x_print_print_get_one_attributes_value

xcb_x_print_print_get_one_attributes_value_end

xcb_x_print_print_get_one_attributes_value_length

xcb_x_print_print_get_page_dimensions

xcb_x_print_print_get_page_dimensions_reply
the returned value must be freed by the caller using libc::free().

xcb_x_print_print_get_page_dimensions_unchecked

xcb_x_print_print_get_printer_list

xcb_x_print_print_get_printer_list_printers_iterator

xcb_x_print_print_get_printer_list_printers_length

xcb_x_print_print_get_printer_list_reply
the returned value must be freed by the caller using libc::free().

xcb_x_print_print_get_printer_list_unchecked

xcb_x_print_print_get_screen_of_context

xcb_x_print_print_get_screen_of_context_reply
the returned value must be freed by the caller using libc::free().

xcb_x_print_print_get_screen_of_context_unchecked

xcb_x_print_print_input_selected

xcb_x_print_print_input_selected_all_events_list

xcb_x_print_print_input_selected_all_events_list_end

xcb_x_print_print_input_selected_all_events_list_length

xcb_x_print_print_input_selected_event_list

xcb_x_print_print_input_selected_event_list_end

xcb_x_print_print_input_selected_event_list_length

xcb_x_print_print_input_selected_reply
the returned value must be freed by the caller using libc::free().

xcb_x_print_print_input_selected_unchecked

xcb_x_print_print_put_document_data

xcb_x_print_print_put_document_data_checked

xcb_x_print_print_query_screens

xcb_x_print_print_query_screens_reply
the returned value must be freed by the caller using libc::free().

xcb_x_print_print_query_screens_roots

xcb_x_print_print_query_screens_roots_end

xcb_x_print_print_query_screens_roots_length

xcb_x_print_print_query_screens_unchecked

xcb_x_print_print_query_version

xcb_x_print_print_query_version_reply
the returned value must be freed by the caller using libc::free().

xcb_x_print_print_query_version_unchecked

xcb_x_print_print_rehash_printer_list

xcb_x_print_print_rehash_printer_list_checked

xcb_x_print_print_select_input

xcb_x_print_print_select_input_checked

xcb_x_print_print_set_attributes

xcb_x_print_print_set_attributes_checked

xcb_x_print_print_set_context

xcb_x_print_print_set_context_checked

xcb_x_print_print_set_image_resolution

xcb_x_print_print_set_image_resolution_reply
the returned value must be freed by the caller using libc::free().

xcb_x_print_print_set_image_resolution_unchecked

xcb_x_print_print_start_doc

xcb_x_print_print_start_doc_checked

xcb_x_print_print_start_job

xcb_x_print_print_start_job_checked

xcb_x_print_print_start_page

xcb_x_print_print_start_page_checked

xcb_x_print_printer_description

xcb_x_print_printer_description_end

xcb_x_print_printer_description_length

xcb_x_print_printer_end

xcb_x_print_printer_name

xcb_x_print_printer_name_end

xcb_x_print_printer_name_length

xcb_x_print_printer_next

xcb_x_print_string8_end

xcb_x_print_string8_next

Type Definitions

xcb_x_print_attr_t

xcb_x_print_detail_t

xcb_x_print_ev_mask_t

xcb_x_print_get_doc_t

xcb_x_print_pcontext_t

xcb_x_print_string8_t