xcb::ffi

Module xc_misc

Structs

xcb_xc_misc_get_version_cookie_t

xcb_xc_misc_get_version_reply_t

xcb_xc_misc_get_version_request_t

xcb_xc_misc_get_xid_list_cookie_t

xcb_xc_misc_get_xid_list_reply_t

xcb_xc_misc_get_xid_list_request_t

xcb_xc_misc_get_xid_range_cookie_t

xcb_xc_misc_get_xid_range_reply_t

xcb_xc_misc_get_xid_range_request_t

Constants

XCB_XC_MISC_GET_VERSION

XCB_XC_MISC_GET_XID_LIST

XCB_XC_MISC_GET_XID_RANGE

XCB_XC_MISC_MAJOR_VERSION

XCB_XC_MISC_MINOR_VERSION

Statics

xcb_xc_misc_id⚠

Functions

xcb_xc_misc_get_version⚠

xcb_xc_misc_get_version_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xc_misc_get_version_unchecked⚠

xcb_xc_misc_get_xid_list⚠

xcb_xc_misc_get_xid_list_ids⚠

xcb_xc_misc_get_xid_list_ids_end⚠

xcb_xc_misc_get_xid_list_ids_length⚠

xcb_xc_misc_get_xid_list_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xc_misc_get_xid_list_unchecked⚠

xcb_xc_misc_get_xid_range⚠

xcb_xc_misc_get_xid_range_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xc_misc_get_xid_range_unchecked⚠