Module xcb::ffi::xfixes

Structs

xcb_xfixes_bad_region_error_t

xcb_xfixes_barrier_iterator_t

xcb_xfixes_change_cursor_by_name_request_t

xcb_xfixes_change_cursor_request_t

xcb_xfixes_change_save_set_request_t

xcb_xfixes_copy_region_request_t

xcb_xfixes_create_pointer_barrier_request_t

xcb_xfixes_create_region_from_bitmap_request_t

xcb_xfixes_create_region_from_gc_request_t

xcb_xfixes_create_region_from_picture_request_t

xcb_xfixes_create_region_from_window_request_t

xcb_xfixes_create_region_request_t

xcb_xfixes_cursor_notify_event_t

xcb_xfixes_delete_pointer_barrier_request_t

xcb_xfixes_destroy_region_request_t

xcb_xfixes_expand_region_request_t

xcb_xfixes_fetch_region_cookie_t

xcb_xfixes_fetch_region_reply_t

xcb_xfixes_fetch_region_request_t

xcb_xfixes_get_cursor_image_and_name_cookie_t

xcb_xfixes_get_cursor_image_and_name_reply_t

xcb_xfixes_get_cursor_image_and_name_request_t

xcb_xfixes_get_cursor_image_cookie_t

xcb_xfixes_get_cursor_image_reply_t

xcb_xfixes_get_cursor_image_request_t

xcb_xfixes_get_cursor_name_cookie_t

xcb_xfixes_get_cursor_name_reply_t

xcb_xfixes_get_cursor_name_request_t

xcb_xfixes_hide_cursor_request_t

xcb_xfixes_intersect_region_request_t

xcb_xfixes_invert_region_request_t

xcb_xfixes_query_version_cookie_t

xcb_xfixes_query_version_reply_t

xcb_xfixes_query_version_request_t

xcb_xfixes_region_extents_request_t

xcb_xfixes_region_iterator_t

xcb_xfixes_select_cursor_input_request_t

xcb_xfixes_select_selection_input_request_t

xcb_xfixes_selection_notify_event_t

xcb_xfixes_set_cursor_name_request_t

xcb_xfixes_set_gc_clip_region_request_t

xcb_xfixes_set_picture_clip_region_request_t

xcb_xfixes_set_region_request_t

xcb_xfixes_set_window_shape_region_request_t

xcb_xfixes_show_cursor_request_t

xcb_xfixes_subtract_region_request_t

xcb_xfixes_translate_region_request_t

xcb_xfixes_union_region_request_t

Constants

XCB_XFIXES_BAD_REGION

XCB_XFIXES_BARRIER_DIRECTIONS_NEGATIVE_X

XCB_XFIXES_BARRIER_DIRECTIONS_NEGATIVE_Y

XCB_XFIXES_BARRIER_DIRECTIONS_POSITIVE_X

XCB_XFIXES_BARRIER_DIRECTIONS_POSITIVE_Y

XCB_XFIXES_CHANGE_CURSOR

XCB_XFIXES_CHANGE_CURSOR_BY_NAME

XCB_XFIXES_CHANGE_SAVE_SET

XCB_XFIXES_COPY_REGION

XCB_XFIXES_CREATE_POINTER_BARRIER

XCB_XFIXES_CREATE_REGION

XCB_XFIXES_CREATE_REGION_FROM_BITMAP

XCB_XFIXES_CREATE_REGION_FROM_GC

XCB_XFIXES_CREATE_REGION_FROM_PICTURE

XCB_XFIXES_CREATE_REGION_FROM_WINDOW

XCB_XFIXES_CURSOR_NOTIFY

XCB_XFIXES_CURSOR_NOTIFY_DISPLAY_CURSOR

XCB_XFIXES_CURSOR_NOTIFY_MASK_DISPLAY_CURSOR

XCB_XFIXES_DELETE_POINTER_BARRIER

XCB_XFIXES_DESTROY_REGION

XCB_XFIXES_EXPAND_REGION

XCB_XFIXES_FETCH_REGION

XCB_XFIXES_GET_CURSOR_IMAGE

XCB_XFIXES_GET_CURSOR_IMAGE_AND_NAME

XCB_XFIXES_GET_CURSOR_NAME

XCB_XFIXES_HIDE_CURSOR

XCB_XFIXES_INTERSECT_REGION

XCB_XFIXES_INVERT_REGION

XCB_XFIXES_MAJOR_VERSION

XCB_XFIXES_MINOR_VERSION

XCB_XFIXES_QUERY_VERSION

XCB_XFIXES_REGION_EXTENTS

XCB_XFIXES_REGION_NONE

XCB_XFIXES_SAVE_SET_MAPPING_MAP

XCB_XFIXES_SAVE_SET_MAPPING_UNMAP

XCB_XFIXES_SAVE_SET_MODE_DELETE

XCB_XFIXES_SAVE_SET_MODE_INSERT

XCB_XFIXES_SAVE_SET_TARGET_NEAREST

XCB_XFIXES_SAVE_SET_TARGET_ROOT

XCB_XFIXES_SELECTION_EVENT_MASK_SELECTION_CLIENT_CLOSE

XCB_XFIXES_SELECTION_EVENT_MASK_SELECTION_WINDOW_DESTROY

XCB_XFIXES_SELECTION_EVENT_MASK_SET_SELECTION_OWNER

XCB_XFIXES_SELECTION_EVENT_SELECTION_CLIENT_CLOSE

XCB_XFIXES_SELECTION_EVENT_SELECTION_WINDOW_DESTROY

XCB_XFIXES_SELECTION_EVENT_SET_SELECTION_OWNER

XCB_XFIXES_SELECTION_NOTIFY

XCB_XFIXES_SELECT_CURSOR_INPUT

XCB_XFIXES_SELECT_SELECTION_INPUT

XCB_XFIXES_SET_CURSOR_NAME

XCB_XFIXES_SET_GC_CLIP_REGION

XCB_XFIXES_SET_PICTURE_CLIP_REGION

XCB_XFIXES_SET_REGION

XCB_XFIXES_SET_WINDOW_SHAPE_REGION

XCB_XFIXES_SHOW_CURSOR

XCB_XFIXES_SUBTRACT_REGION

XCB_XFIXES_TRANSLATE_REGION

XCB_XFIXES_UNION_REGION

Statics

xcb_xfixes_id

Functions

xcb_xfixes_barrier_end⚠

xcb_xfixes_barrier_next⚠

xcb_xfixes_change_cursor⚠

xcb_xfixes_change_cursor_by_name⚠

xcb_xfixes_change_cursor_by_name_checked⚠

xcb_xfixes_change_cursor_checked⚠

xcb_xfixes_change_save_set⚠

xcb_xfixes_change_save_set_checked⚠

xcb_xfixes_copy_region⚠

xcb_xfixes_copy_region_checked⚠

xcb_xfixes_create_pointer_barrier⚠

xcb_xfixes_create_pointer_barrier_checked⚠

xcb_xfixes_create_region⚠

xcb_xfixes_create_region_checked⚠

xcb_xfixes_create_region_from_bitmap⚠

xcb_xfixes_create_region_from_bitmap_checked⚠

xcb_xfixes_create_region_from_gc⚠

xcb_xfixes_create_region_from_gc_checked⚠

xcb_xfixes_create_region_from_picture⚠

xcb_xfixes_create_region_from_picture_checked⚠

xcb_xfixes_create_region_from_window⚠

xcb_xfixes_create_region_from_window_checked⚠

xcb_xfixes_delete_pointer_barrier⚠

xcb_xfixes_delete_pointer_barrier_checked⚠

xcb_xfixes_destroy_region⚠

xcb_xfixes_destroy_region_checked⚠

xcb_xfixes_expand_region⚠

xcb_xfixes_expand_region_checked⚠

xcb_xfixes_fetch_region⚠

xcb_xfixes_fetch_region_rectangles⚠

xcb_xfixes_fetch_region_rectangles_iterator⚠

xcb_xfixes_fetch_region_rectangles_length⚠

xcb_xfixes_fetch_region_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xfixes_fetch_region_unchecked⚠

xcb_xfixes_get_cursor_image⚠

xcb_xfixes_get_cursor_image_and_name⚠

xcb_xfixes_get_cursor_image_and_name_cursor_image⚠

xcb_xfixes_get_cursor_image_and_name_cursor_image_end⚠

xcb_xfixes_get_cursor_image_and_name_cursor_image_length⚠

xcb_xfixes_get_cursor_image_and_name_name⚠

xcb_xfixes_get_cursor_image_and_name_name_end⚠

xcb_xfixes_get_cursor_image_and_name_name_length⚠

xcb_xfixes_get_cursor_image_and_name_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xfixes_get_cursor_image_and_name_unchecked⚠

xcb_xfixes_get_cursor_image_cursor_image⚠

xcb_xfixes_get_cursor_image_cursor_image_end⚠

xcb_xfixes_get_cursor_image_cursor_image_length⚠

xcb_xfixes_get_cursor_image_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xfixes_get_cursor_image_unchecked⚠

xcb_xfixes_get_cursor_name⚠

xcb_xfixes_get_cursor_name_name⚠

xcb_xfixes_get_cursor_name_name_end⚠

xcb_xfixes_get_cursor_name_name_length⚠

xcb_xfixes_get_cursor_name_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xfixes_get_cursor_name_unchecked⚠

xcb_xfixes_hide_cursor⚠

xcb_xfixes_hide_cursor_checked⚠

xcb_xfixes_intersect_region⚠

xcb_xfixes_intersect_region_checked⚠

xcb_xfixes_invert_region⚠

xcb_xfixes_invert_region_checked⚠

xcb_xfixes_query_version⚠

xcb_xfixes_query_version_reply⚠

the returned value must be freed by the caller using libc::free().

xcb_xfixes_query_version_unchecked⚠

xcb_xfixes_region_end⚠

xcb_xfixes_region_extents⚠

xcb_xfixes_region_extents_checked⚠

xcb_xfixes_region_next⚠

xcb_xfixes_select_cursor_input⚠

xcb_xfixes_select_cursor_input_checked⚠

xcb_xfixes_select_selection_input⚠

xcb_xfixes_select_selection_input_checked⚠

xcb_xfixes_set_cursor_name⚠

xcb_xfixes_set_cursor_name_checked⚠

xcb_xfixes_set_gc_clip_region⚠

xcb_xfixes_set_gc_clip_region_checked⚠

xcb_xfixes_set_picture_clip_region⚠

xcb_xfixes_set_picture_clip_region_checked⚠

xcb_xfixes_set_region⚠

xcb_xfixes_set_region_checked⚠

xcb_xfixes_set_window_shape_region⚠

xcb_xfixes_set_window_shape_region_checked⚠

xcb_xfixes_show_cursor⚠

xcb_xfixes_show_cursor_checked⚠

xcb_xfixes_subtract_region⚠

xcb_xfixes_subtract_region_checked⚠

xcb_xfixes_translate_region⚠

xcb_xfixes_translate_region_checked⚠

xcb_xfixes_union_region⚠

xcb_xfixes_union_region_checked⚠

Type Definitions

xcb_xfixes_barrier_directions_t

xcb_xfixes_barrier_t

xcb_xfixes_cursor_notify_mask_t

xcb_xfixes_cursor_notify_t

xcb_xfixes_region_enum_t

xcb_xfixes_region_t

xcb_xfixes_save_set_mapping_t

xcb_xfixes_save_set_mode_t

xcb_xfixes_save_set_target_t

xcb_xfixes_selection_event_mask_t

xcb_xfixes_selection_event_t