Struct xcb::ffi::xproto::xcb_create_glyph_cursor_request_t

pub struct xcb_create_glyph_cursor_request_t {
    pub major_opcode: u8,
    pub pad0: u8,
    pub length: u16,
    pub cid: xcb_cursor_t,
    pub source_font: xcb_font_t,
    pub mask_font: xcb_font_t,
    pub source_char: u16,
    pub mask_char: u16,
    pub fore_red: u16,
    pub fore_green: u16,
    pub fore_blue: u16,
    pub back_red: u16,
    pub back_green: u16,
    pub back_blue: u16,
}

create cursor

Creates a cursor from a font glyph. X provides a set of standard cursor shapes in a special font named cursor. Applications are encouraged to use this interface for their cursors because the font can be customized for the individual display type.

All pixels which are set to 1 in the source will use the foreground color (as specified by fore_red, fore_green and fore_blue). All pixels set to 0 will use the background color (as specified by back_red, back_green and back_blue).

Fields

major_opcode

pad0

length

cid

The ID with which you will refer to the cursor, created by xcb_generate_id.

source_font

In which font to look for the cursor glyph.

mask_font

In which font to look for the mask glyph.

source_char

The glyph of source_font to use.

mask_char

The glyph of mask_font to use as a mask: Pixels which are set to 1 define which source pixels are displayed. All pixels which are set to 0 are not displayed.

fore_red

The red value of the foreground color.

fore_green

The green value of the foreground color.

fore_blue

The blue value of the foreground color.

back_red

The red value of the background color.

back_green

The green value of the background color.

back_blue

The blue value of the background color.