Struct xcb::ffi::xproto::xcb_intern_atom_request_t

pub struct xcb_intern_atom_request_t {
    pub major_opcode: u8,
    pub only_if_exists: u8,
    pub length: u16,
    pub name_len: u16,
    pub pad0: [u8; 2],
}

Get atom identifier by name

Retrieves the identifier (xcb_atom_t TODO) for the atom with the specified name. Atoms are used in protocols like EWMH, for example to store window titles (_NET_WM_NAME atom) as property of a window.

If only_if_exists is 0, the atom will be created if it does not already exist. If only_if_exists is 1, XCB_ATOM_NONE will be returned if the atom does not yet exist.

Fields

major_opcode

only_if_exists

Return a valid atom id only if the atom already exists.

length

name_len

The length of the following name.

pad0