Macro xcb::atoms_struct

macro_rules! atoms_struct {
    ($vis : vis struct $Atoms : ident { $($field : ident => $name : expr,) * }) => { ... };
}

An helper macro that generate a struct of atoms.

The struct provide a constructor intern_all that takes a Connection as parameter, interns all the atoms and return xcb::Result<[struct name]>. intern_all takes advantage of XCB asynchronous design by sending all the x::InternAtom requests before starting to wait for the first reply. Fields that refer to atoms not existing in the server are set to x::ATOM_NONE (i.e. only_if_exists is always set to true).

Example

xcb::atoms_struct!(
    struct Atoms {
        wm_protocols    => b"WM_PROTOCOLS",
        wm_del_window   => b"WM_DELETE_WINDOW",
    }
);

fn main() -> xcb::Result<()> {
    // ...
    let atoms = Atoms::intern_all(&conn)?;

    conn.check_request(conn.send_request_checked(&x::ChangeProperty {
        mode: x::PropMode::Replace,
        window,
        property: atoms.wm_protocols,
        r#type: x::ATOM_ATOM,
        data: &[atoms.wm_del_window],
    }))?;
    // ...
}