xcb
Module record

The RECORD X extension.

Accessible with the record cargo feature.

Structs

BadContextError

The BadContextError error.

ClientInfo

ClientInfoBuf

ClientInfoIterator

Context

CreateContext

The CreateContext request.

DisableContext

The DisableContext request.

EnableContext

The EnableContext request.

EnableContextCookie

Cookie type for EnableContext.

EnableContextCookieUnchecked

Unchecked cookie type for EnableContext.

EnableContextReply

Reply type for EnableContext.

ExtRange

FreeContext

The FreeContext request.

GetContext

The GetContext request.

GetContextCookie

Cookie type for GetContext.

GetContextCookieUnchecked

Unchecked cookie type for GetContext.

GetContextReply

Reply type for GetContext.

HType

QueryVersion

The QueryVersion request.

QueryVersionCookie

Cookie type for QueryVersion.

QueryVersionCookieUnchecked

Unchecked cookie type for QueryVersion.

QueryVersionReply

Reply type for QueryVersion.

Range

Range8

Range16

RegisterClients

The RegisterClients request.

UnregisterClients

The UnregisterClients request.

Enums

Cs

Error

Unified error type for the Record extension

Constants

MAJOR_VERSION

The major version of the Record extension.

MINOR_VERSION

The minor version of the Record extension.

VERSION_STRING

The version string of the Record extension.

XNAME

The official identifier for the Record extension.

Functions

get_extension_data

Fetch server runtime info data of the Record extension.

prefetch_extension_data

Prefetch server runtime info data of the Record extension.

Type Aliases

ClientSpec

ElementHeader