Module xcb::render

The RENDER X extension.

Accessible with the render cargo feature.

Structs

AddGlyphs

The AddGlyphs request.

AddTraps

The AddTraps request.

Animcursorelt

ChangePicture

The ChangePicture request.

Color

Composite

The Composite request.

CompositeGlyphs8

The CompositeGlyphs8 request.

CompositeGlyphs16

The CompositeGlyphs16 request.

CompositeGlyphs32

The CompositeGlyphs32 request.

CpMask

CpParams

CreateAnimCursor

The CreateAnimCursor request.

CreateConicalGradient

The CreateConicalGradient request.

CreateCursor

The CreateCursor request.

CreateGlyphSet

The CreateGlyphSet request.

CreateLinearGradient

The CreateLinearGradient request.

CreatePicture

The CreatePicture request.

CreateRadialGradient

The CreateRadialGradient request.

CreateSolidFill

The CreateSolidFill request.

Directformat

FillRectangles

The FillRectangles request.

FreeGlyphSet

The FreeGlyphSet request.

FreeGlyphs

The FreeGlyphs request.

FreePicture

The FreePicture request.

GlyphError

The GlyphError error.

GlyphSetError

The GlyphSetError error.

Glyphinfo

Glyphset

Indexvalue

Linefix

PictFormatError

The PictFormatError error.

PictOpError

The PictOpError error.

Pictdepth

PictdepthBuf

PictdepthIterator

Pictformat

Pictforminfo

Pictscreen

PictscreenBuf

PictscreenIterator

Picture

PictureError

The PictureError error.

Pictvisual

Pointfix

QueryFilters

The QueryFilters request.

QueryFiltersCookie

Cookie type for QueryFilters.

QueryFiltersCookieUnchecked

Unchecked cookie type for QueryFilters.

QueryFiltersReply

Reply type for QueryFilters.

QueryPictFormats

The QueryPictFormats request.

QueryPictFormatsCookie

Cookie type for QueryPictFormats.

QueryPictFormatsCookieUnchecked

Unchecked cookie type for QueryPictFormats.

QueryPictFormatsReply

Reply type for QueryPictFormats.

QueryPictIndexValues

The QueryPictIndexValues request.

QueryPictIndexValuesCookie

Cookie type for QueryPictIndexValues.

QueryPictIndexValuesCookieUnchecked

Unchecked cookie type for QueryPictIndexValues.

QueryPictIndexValuesReply

Reply type for QueryPictIndexValues.

QueryVersion

The QueryVersion request.

QueryVersionCookie

Cookie type for QueryVersion.

QueryVersionCookieUnchecked

Unchecked cookie type for QueryVersion.

QueryVersionReply

Reply type for QueryVersion.

ReferenceGlyphSet

The ReferenceGlyphSet request.

SetPictureClipRectangles

The SetPictureClipRectangles request.

SetPictureFilter

The SetPictureFilter request.

SetPictureTransform

The SetPictureTransform request.

Spanfix

Transform

Trap

Trapezoid

Trapezoids

The Trapezoids request.

TriFan

The TriFan request.

TriStrip

The TriStrip request.

Triangle

Triangles

The Triangles request.

Enums

Cp

Error

Unified error type for the Render extension

PictOp

PictType

PolyEdge

PolyMode

Repeat

SubPixel

Constants

MAJOR_VERSION

The major version of the Render extension.

MINOR_VERSION

The minor version of the Render extension.

PICTURE_NONE

VERSION_STRING

The version string of the Render extension.

XNAME

The official identifier for the Render extension.

Functions

get_extension_data

Fetch server runtime info data of the Render extension.

prefetch_extension_data

Prefetch server runtime info data of the Render extension.

Type Aliases

Fixed

Glyph