xcb
Module res

The X-Resource X extension.

Accessible with the res cargo feature.

Structs

Client

ClientIdMask

ClientIdSpec

ClientIdValue

ClientIdValueBuf

ClientIdValueIterator

QueryClientIds

The QueryClientIds request.

QueryClientIdsCookie

Cookie type for QueryClientIds.

QueryClientIdsCookieUnchecked

Unchecked cookie type for QueryClientIds.

QueryClientIdsReply

Reply type for QueryClientIds.

QueryClientPixmapBytes

The QueryClientPixmapBytes request.

QueryClientPixmapBytesCookie

Cookie type for QueryClientPixmapBytes.

QueryClientPixmapBytesCookieUnchecked

Unchecked cookie type for QueryClientPixmapBytes.

QueryClientPixmapBytesReply

Reply type for QueryClientPixmapBytes.

QueryClientResources

The QueryClientResources request.

QueryClientResourcesCookie

Cookie type for QueryClientResources.

QueryClientResourcesCookieUnchecked

Unchecked cookie type for QueryClientResources.

QueryClientResourcesReply

Reply type for QueryClientResources.

QueryClients

The QueryClients request.

QueryClientsCookie

Cookie type for QueryClients.

QueryClientsCookieUnchecked

Unchecked cookie type for QueryClients.

QueryClientsReply

Reply type for QueryClients.

QueryResourceBytes

The QueryResourceBytes request.

QueryResourceBytesCookie

Cookie type for QueryResourceBytes.

QueryResourceBytesCookieUnchecked

Unchecked cookie type for QueryResourceBytes.

QueryResourceBytesReply

Reply type for QueryResourceBytes.

QueryVersion

The QueryVersion request.

QueryVersionCookie

Cookie type for QueryVersion.

QueryVersionCookieUnchecked

Unchecked cookie type for QueryVersion.

QueryVersionReply

Reply type for QueryVersion.

ResourceIdSpec

ResourceSizeSpec

ResourceSizeValue

ResourceSizeValueBuf

ResourceSizeValueIterator

Type

Constants

MAJOR_VERSION

The major version of the Res extension.

MINOR_VERSION

The minor version of the Res extension.

VERSION_STRING

The version string of the Res extension.

XNAME

The official identifier for the Res extension.

Functions

get_extension_data

Fetch server runtime info data of the Res extension.

prefetch_extension_data

Prefetch server runtime info data of the Res extension.