Module xcb::shape

The SHAPE X extension.

Accessible with the shape cargo feature.

Structs

Combine

The Combine request.

GetRectangles

The GetRectangles request.

GetRectanglesCookie

Cookie type for GetRectangles.

GetRectanglesCookieUnchecked

Unchecked cookie type for GetRectangles.

GetRectanglesReply

Reply type for GetRectangles.

InputSelected

The InputSelected request.

InputSelectedCookie

Cookie type for InputSelected.

InputSelectedCookieUnchecked

Unchecked cookie type for InputSelected.

InputSelectedReply

Reply type for InputSelected.

Mask

The Mask request.

NotifyEvent

The NotifyEvent event.

Offset

The Offset request.

QueryExtents

The QueryExtents request.

QueryExtentsCookie

Cookie type for QueryExtents.

QueryExtentsCookieUnchecked

Unchecked cookie type for QueryExtents.

QueryExtentsReply

Reply type for QueryExtents.

QueryVersion

The QueryVersion request.

QueryVersionCookie

Cookie type for QueryVersion.

QueryVersionCookieUnchecked

Unchecked cookie type for QueryVersion.

QueryVersionReply

Reply type for QueryVersion.

Rectangles

The Rectangles request.

SelectInput

The SelectInput request.

Enums

Event

Unified event type for the Shape extension

Sk

So

Constants

MAJOR_VERSION

The major version of the Shape extension.

MINOR_VERSION

The minor version of the Shape extension.

VERSION_STRING

The version string of the Shape extension.

XNAME

The official identifier for the Shape extension.

Functions

get_extension_data

Fetch server runtime info data of the Shape extension.

prefetch_extension_data

Prefetch server runtime info data of the Shape extension.

Type Aliases

Kind

Op