Module xcb::sync

The SYNC X extension.

Accessible with the sync cargo feature.

Structs

Alarm

AlarmError

The AlarmError error.

AlarmNotifyEvent

The AlarmNotifyEvent event.

Await

The Await request.

AwaitFence

The AwaitFence request.

CaMask

CaParams

ChangeAlarm

The ChangeAlarm request.

ChangeCounter

The ChangeCounter request.

Counter

CounterError

The CounterError error.

CounterNotifyEvent

The CounterNotifyEvent event.

CreateAlarm

The CreateAlarm request.

CreateCounter

The CreateCounter request.

CreateFence

The CreateFence request.

DestroyAlarm

The DestroyAlarm request.

DestroyCounter

The DestroyCounter request.

DestroyFence

The DestroyFence request.

Fence

GetPriority

The GetPriority request.

GetPriorityCookie

Cookie type for GetPriority.

GetPriorityCookieUnchecked

Unchecked cookie type for GetPriority.

GetPriorityReply

Reply type for GetPriority.

Initialize

The Initialize request.

InitializeCookie

Cookie type for Initialize.

InitializeCookieUnchecked

Unchecked cookie type for Initialize.

InitializeReply

Reply type for Initialize.

Int64

ListSystemCounters

The ListSystemCounters request.

ListSystemCountersCookie

Cookie type for ListSystemCounters.

ListSystemCountersCookieUnchecked

Unchecked cookie type for ListSystemCounters.

ListSystemCountersReply

Reply type for ListSystemCounters.

QueryAlarm

The QueryAlarm request.

QueryAlarmCookie

Cookie type for QueryAlarm.

QueryAlarmCookieUnchecked

Unchecked cookie type for QueryAlarm.

QueryAlarmReply

Reply type for QueryAlarm.

QueryCounter

The QueryCounter request.

QueryCounterCookie

Cookie type for QueryCounter.

QueryCounterCookieUnchecked

Unchecked cookie type for QueryCounter.

QueryCounterReply

Reply type for QueryCounter.

QueryFence

The QueryFence request.

QueryFenceCookie

Cookie type for QueryFence.

QueryFenceCookieUnchecked

Unchecked cookie type for QueryFence.

QueryFenceReply

Reply type for QueryFence.

ResetFence

The ResetFence request.

SetCounter

The SetCounter request.

SetPriority

The SetPriority request.

Systemcounter

SystemcounterBuf

SystemcounterIterator

Trigger

TriggerFence

The TriggerFence request.

Waitcondition

Enums

Alarmstate

Ca

Error

Unified error type for the Sync extension

Event

Unified event type for the Sync extension

Testtype

Valuetype

Constants

MAJOR_VERSION

The major version of the Sync extension.

MINOR_VERSION

The minor version of the Sync extension.

VERSION_STRING

The version string of the Sync extension.

XNAME

The official identifier for the Sync extension.

Functions

get_extension_data

Fetch server runtime info data of the Sync extension.

prefetch_extension_data

Prefetch server runtime info data of the Sync extension.