xcb

Trait BaseEvent

pub trait BaseEvent: Raw<xcb_generic_event_t> {
    const EXTENSION: Option<Extension>;
    const NUMBER: u32;

    // Provided method
    fn is_from_send_event(&self) -> bool { ... }
}

Trait for base events (aka. non GE_GENERIC events)

Required Associated Constants

const EXTENSION: Option<Extension>

The extension associated to this event, or None for the main protocol

const NUMBER: u32

The number associated to this event

Provided Methods

fn is_from_send_event(&self) -> bool

Check whether this event was emitted by the SendEvent request See [crate::x::SendEvent].

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.