Trait xcb::CheckedCookie

pub unsafe trait CheckedCookie: Cookie { }

A marker trait for a cookie that allows synchronized error checking.

Safety

Cookies not implementing this trait acknowledge that the error is sent to the event loop

See also Connection::send_request, Connection::send_request_checked and Connection::check_request