xcb

Trait CookieChecked

pub unsafe trait CookieChecked: Cookie { }

A marker trait for a cookie that allows synchronized error checking.

Safety

Cookies not implementing this trait acknowledge that the error is sent to the event loop

See also Connection::send_request, Connection::send_request_checked and Connection::check_request

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.