Trait xcb::CookieWithReplyChecked

pub unsafe trait CookieWithReplyChecked: CookieChecked {
    type Reply: Reply;
}

A trait for checked cookies of requests that send a reply.

Safety

Cookies implementing this trait acknowledge that their error is checked when the reply is fetched from the server. This is the default cookie type for requests with reply.

See also Connection::send_request, Connection::wait_for_reply

Associated Types

type Reply: Reply

The reply type associated with the cookie