Trait xcb::GeEvent

pub trait GeEvent: Raw<xcb_ge_generic_event_t> {
    const EXTENSION: Extension;
    const NUMBER: u32;
}

A trait for GE_GENERIC events

A GE_GENERIC event is an extension event that does not follow the regular response_type offset. This system was introduced because the protocol eventually run out of event numbers.

This should be completely transparent to the user, as Event is resolving all types of events together.

Required Associated Constants

const EXTENSION: Extension

The extension associated to this event

const NUMBER: u32

The number associated to this event