Trait xcb::Raw

pub trait Raw<T>: Sized {
    // Required methods
    unsafe fn from_raw(raw: *mut T) -> Self;
    fn as_raw(&self) -> *mut T;

    // Provided method
    fn into_raw(self) -> *mut T { ... }
}

Trait for types that own a C allocated pointer and are represented by the data pointed to.

Required Methods

unsafe fn from_raw(raw: *mut T) -> Self

Build Self from a raw pointer

Safety

raw must be a valid pointer to the representation of Self, and be allocated with libc::malloc

fn as_raw(&self) -> *mut T

Obtain the raw pointer representation

Provided Methods

fn into_raw(self) -> *mut T

Convert self into a raw pointer

Returned value should be freed with libc::free or sent back to from_raw to avoid memory leak.