Trait xcb::RawRequest

pub unsafe trait RawRequest {
    // Required method
    fn raw_request(&self, conn: &Connection, checked: bool) -> u64;
}

Trait implemented by all requests to send the serialized data over the wire.

Safety

Types implementing this trait acknowledge that the returned value of raw_request correspond to a cookie for Self request and is checked or unchecked depending on the checked flag value.

Required Methods

fn raw_request(&self, conn: &Connection, checked: bool) -> u64

Actual implementation of the request sending

Send the request over the conn wire and return a cookie sequence fitting with the checked flag of Self