Trait xcb::Reply

pub trait Reply {
    // Required methods
    unsafe fn from_raw(raw: *const u8) -> Self;
    unsafe fn into_raw(self) -> *const u8;
    unsafe fn as_raw(&self) -> *const u8;
}

Trait for request replies

Required Methods

unsafe fn from_raw(raw: *const u8) -> Self

Build the reply struct from a raw pointer.

Safety

raw must be a pointer to a valid wire representation of Self, allocated with libc::malloc.

unsafe fn into_raw(self) -> *const u8

Consume the reply struct into a raw pointer.

Safety

The returned pointer must be freed with libc::free to avoid any memory leak, or be used to build another reply.

unsafe fn as_raw(&self) -> *const u8

Get the raw pointer of the XCB reply.

Safety

The returned pointer must NOT be freed. Passing this pointer to libc::free will result in a double free when the reply is dropped.

Object Safety

This trait is not object safe.