Trait xcb::RequestWithReply

pub trait RequestWithReply: Request {
    type Reply: Reply;
    type Cookie: CheckedCookieWithReply<Reply = Self::Reply>;
    type UncheckedCookie: UncheckedCookieWithReply<Reply = Self::Reply>;
}

Trait for requests that return a reply.

Associated Types

type Reply: Reply

Reply associated with the request

type Cookie: CheckedCookieWithReply<Reply = Self::Reply>

Default cookie type for the request, as returned by Connection::send_request.

type UncheckedCookie: UncheckedCookieWithReply<Reply = Self::Reply>

Unchecked cookie type for the request, as returned by Connection::send_request_unchecked.