Trait xcb::UncheckedCookieWithReply

pub unsafe trait UncheckedCookieWithReply: Cookie {
    type Reply: Reply;
}

A trait for unchecked cookies of requests that send a reply.

Safety

Cookies implementing this trait acknowledge that their error is not checked when the reply is fetched from the server but in the event loop.

See also Connection::send_request_unchecked, Connection::wait_for_event

Associated Types

type Reply: Reply

The reply type associated with the cookie