Trait xcb::x::PropEl

pub trait PropEl {
    const FORMAT: u8;
}

Trait for element in a property list

In events (e.g. GetProperty::value), it allows to assert that the format correspond to the type cast and therefore to do the cast safely at runtime.

In request (e.g. ChangeProperty::data), it allows to infer the format value from the type of passed data.

Associated Constants

const FORMAT: u8