Module xcb::xevie

The XEVIE X extension.

Accessible with the xevie cargo feature.

Structs

End

The End request.

EndCookie

Cookie type for End.

EndCookieUnchecked

Unchecked cookie type for End.

EndReply

Reply type for End.

Event

QueryVersion

The QueryVersion request.

QueryVersionCookie

Cookie type for QueryVersion.

QueryVersionCookieUnchecked

Unchecked cookie type for QueryVersion.

QueryVersionReply

Reply type for QueryVersion.

SelectInput

The SelectInput request.

SelectInputCookie

Cookie type for SelectInput.

SelectInputCookieUnchecked

Unchecked cookie type for SelectInput.

SelectInputReply

Reply type for SelectInput.

Send

The Send request.

SendCookie

Cookie type for Send.

SendCookieUnchecked

Unchecked cookie type for Send.

SendReply

Reply type for Send.

Start

The Start request.

StartCookie

Cookie type for Start.

StartCookieUnchecked

Unchecked cookie type for Start.

StartReply

Reply type for Start.

Enums

Datatype

Constants

MAJOR_VERSION

The major version of the Xevie extension.

MINOR_VERSION

The minor version of the Xevie extension.

VERSION_STRING

The version string of the Xevie extension.

XNAME

The official identifier for the Xevie extension.

Functions

get_extension_data

Fetch server runtime info data of the Xevie extension.

prefetch_extension_data

Prefetch server runtime info data of the Xevie extension.