Module xcb::xkb

The XKEYBOARD X extension.

Accessible with the xkb cargo feature.

Structs

AccessXNotifyEvent

The AccessXNotifyEvent event.

ActionMessageEvent

The ActionMessageEvent event.

ActionMessageFlag

AxOption

AxnDetail

Bell

The Bell request.

BellNotifyEvent

The BellNotifyEvent event.

BoolCtrl

BoolCtrlsHigh

BoolCtrlsLow

CmDetail

CompatMapNotifyEvent

The CompatMapNotifyEvent event.

Control

ControlsNotifyEvent

The ControlsNotifyEvent event.

CountedString16

CountedString16Buf

CountedString16Iterator

DeviceLedInfo

DeviceLedInfoBuf

DeviceLedInfoIterator

EventType

Explicit

ExtensionDeviceNotifyEvent

The ExtensionDeviceNotifyEvent event.

GbnDetail

GetCompatMap

The GetCompatMap request.

GetCompatMapCookie

Cookie type for GetCompatMap.

GetCompatMapCookieUnchecked

Unchecked cookie type for GetCompatMap.

GetCompatMapReply

Reply type for GetCompatMap.

GetControls

The GetControls request.

GetControlsCookie

Cookie type for GetControls.

GetControlsCookieUnchecked

Unchecked cookie type for GetControls.

GetControlsReply

Reply type for GetControls.

GetDeviceInfo

The GetDeviceInfo request.

GetDeviceInfoCookie

Cookie type for GetDeviceInfo.

GetDeviceInfoCookieUnchecked

Unchecked cookie type for GetDeviceInfo.

GetDeviceInfoReply

Reply type for GetDeviceInfo.

GetIndicatorMap

The GetIndicatorMap request.

GetIndicatorMapCookie

Cookie type for GetIndicatorMap.

GetIndicatorMapCookieUnchecked

Unchecked cookie type for GetIndicatorMap.

GetIndicatorMapReply

Reply type for GetIndicatorMap.

GetIndicatorState

The GetIndicatorState request.

GetIndicatorStateCookie

Cookie type for GetIndicatorState.

GetIndicatorStateCookieUnchecked

Unchecked cookie type for GetIndicatorState.

GetIndicatorStateReply

Reply type for GetIndicatorState.

GetKbdByName

The GetKbdByName request.

GetKbdByNameCookie

Cookie type for GetKbdByName.

GetKbdByNameCookieUnchecked

Unchecked cookie type for GetKbdByName.

GetKbdByNameReply

Reply type for GetKbdByName.

GetKbdByNameReplyRepliesMapParams

GetKbdByNameReplyRepliesParams

GetKbdByNameReplyRepliesValueListParams

GetMap

The GetMap request.

GetMapCookie

Cookie type for GetMap.

GetMapCookieUnchecked

Unchecked cookie type for GetMap.

GetMapReply

Reply type for GetMap.

GetMapReplyMapParams

GetNamedIndicator

The GetNamedIndicator request.

GetNamedIndicatorCookie

Cookie type for GetNamedIndicator.

GetNamedIndicatorCookieUnchecked

Unchecked cookie type for GetNamedIndicator.

GetNamedIndicatorReply

Reply type for GetNamedIndicator.

GetNames

The GetNames request.

GetNamesCookie

Cookie type for GetNames.

GetNamesCookieUnchecked

Unchecked cookie type for GetNames.

GetNamesReply

Reply type for GetNames.

GetNamesReplyValueListParams

GetState

The GetState request.

GetStateCookie

Cookie type for GetState.

GetStateCookieUnchecked

Unchecked cookie type for GetState.

GetStateReply

Reply type for GetState.

GroupsWrap

ImFlag

ImGroupsWhich

ImModsWhich

IndicatorMap

IndicatorMapNotifyEvent

The IndicatorMapNotifyEvent event.

IndicatorStateNotifyEvent

The IndicatorStateNotifyEvent event.

Key

KeyAlias

KeyModMap

KeyName

KeySymMap

KeySymMapBuf

KeySymMapIterator

KeyType

KeyTypeBuf

KeyTypeIterator

KeyVModMap

KeyboardError

The KeyboardError error.

KtMapEntry

KtSetMapEntry

LatchLockState

The LatchLockState request.

ListComponents

The ListComponents request.

ListComponentsCookie

Cookie type for ListComponents.

ListComponentsCookieUnchecked

Unchecked cookie type for ListComponents.

ListComponentsReply

Reply type for ListComponents.

Listing

ListingBuf

ListingIterator

LockDeviceFlags

MapNotifyEvent

The MapNotifyEvent event.

MapPart

ModDef

NameDetail

NamesNotifyEvent

The NamesNotifyEvent event.

NewKeyboardNotifyEvent

The NewKeyboardNotifyEvent event.

NknDetail

Outline

OutlineBuf

OutlineIterator

Overlay

OverlayBuf

OverlayIterator

OverlayKey

OverlayRow

OverlayRowBuf

OverlayRowIterator

PerClientFlag

PerClientFlags

The PerClientFlags request.

PerClientFlagsCookie

Cookie type for PerClientFlags.

PerClientFlagsCookieUnchecked

Unchecked cookie type for PerClientFlags.

PerClientFlagsReply

Reply type for PerClientFlags.

Row

RowBuf

RowIterator

Sa

SaIsoLockFlag

SaIsoLockNoAffect

SaMovePtrFlag

SaSetPtrDfltFlag

SelectEvents

The SelectEvents request.

SelectEventsDetailsParams

SetBehavior

SetCompatMap

The SetCompatMap request.

SetControls

The SetControls request.

SetDebuggingFlags

The SetDebuggingFlags request.

SetDebuggingFlagsCookie

Cookie type for SetDebuggingFlags.

SetDebuggingFlagsCookieUnchecked

Unchecked cookie type for SetDebuggingFlags.

SetDebuggingFlagsReply

Reply type for SetDebuggingFlags.

SetDeviceInfo

The SetDeviceInfo request.

SetExplicit

SetIndicatorMap

The SetIndicatorMap request.

SetKeyType

SetKeyTypeBuf

SetKeyTypeIterator

SetMap

The SetMap request.

SetMapFlags

SetMapValuesParams

SetNamedIndicator

The SetNamedIndicator request.

SetNames

The SetNames request.

SetNamesValuesParams

SetOfGroup

SetOfGroups

Shape

ShapeBuf

ShapeIterator

SiAction

StateNotifyEvent

The StateNotifyEvent event.

StatePart

SwitchScreenFlag

SymInterpMatch

SymInterpret

UseExtension

The UseExtension request.

UseExtensionCookie

Cookie type for UseExtension.

UseExtensionCookieUnchecked

Unchecked cookie type for UseExtension.

UseExtensionReply

Reply type for UseExtension.

VMod

VModsHigh

VModsLow

XiFeature

Enums

Action

Behavior

BehaviorType

BellClass

BellClassResult

Const

DoodadType

Error

Unified error type for the Xkb extension

ErrorEnum

Event

Unified event type for the Xkb extension

GetKbdByNameReplyReplies

GetKbdByNameReplyRepliesMap

GetKbdByNameReplyRepliesValueList

GetMapReplyMap

GetNamesReplyValueList

Group

Groups

Id

LedClass

LedClassResult

SaType

SaValWhat

SelectEventsDetails

SetMapValues

SetNamesValues

SymInterpretMatch

Constants

MAJOR_VERSION

The major version of the xkb extension.

MINOR_VERSION

The minor version of the xkb extension.

VERSION_STRING

The version string of the xkb extension.

XNAME

The official identifier for the xkb extension.

Functions

get_extension_data

Fetch server runtime info data of the xkb extension.

prefetch_extension_data

Prefetch server runtime info data of the xkb extension.

Type Definitions

BellClassSpec

DeviceSpec

IdSpec

LedClassSpec

String8