Module xcb::xproto

Structs

AccessError

AllocError

Arc

AtomError

Char2b

Charinfo

ColormapError

CursorError

DrawableError

FontError

Fontprop

GContextError

IdChoiceError

ImplementationError

LengthError

MatchError

NameError

PixmapError

Point

Rectangle

RequestError

Segment

Timecoord

ValueError

WindowError

Constants

ACCESS

ACCESS_CONTROL_DISABLE

ACCESS_CONTROL_ENABLE

ALLOC

ALLOC_COLOR

ALLOC_COLOR_CELLS

ALLOC_COLOR_PLANES

ALLOC_NAMED_COLOR

ALLOW_ASYNC_BOTH
For AsyncBoth, if the pointer and the keyboard are frozen by the client, event processing for both devices continues normally. If a device is frozen twice by the client on behalf of two separate grabs, AsyncBoth thaws for both. AsyncBoth has no effect unless both pointer and keyboard are frozen by the client.

ALLOW_ASYNC_KEYBOARD
For AsyncKeyboard, if the keyboard is frozen by the client, keyboard event processing continues normally. If the keyboard is frozen twice by the client on behalf of two separate grabs, AsyncKeyboard thaws for both. AsyncKeyboard has no effect if the keyboard is not frozen by the client, but the keyboard need not be grabbed by the client.

ALLOW_ASYNC_POINTER
For AsyncPointer, if the pointer is frozen by the client, pointer event processing continues normally. If the pointer is frozen twice by the client on behalf of two separate grabs, AsyncPointer thaws for both. AsyncPointer has no effect if the pointer is not frozen by the client, but the pointer need not be grabbed by the client.

ALLOW_EVENTS

ALLOW_REPLAY_KEYBOARD
For ReplayKeyboard, if the keyboard is actively grabbed by the client and is frozen as the result of an event having been sent to the client (either from the activation of a GrabKey or from a previous AllowEvents with mode SyncKeyboard but not from a GrabKeyboard), then the keyboard grab is released and that event is completely reprocessed, this time ignoring any passive grabs at or above (towards the root) the grab-window of the grab just released. The request has no effect if the keyboard is not grabbed by the client or if the keyboard is not frozen as the result of an event.

ALLOW_REPLAY_POINTER
For ReplayPointer, if the pointer is actively grabbed by the client and is frozen as the result of an event having been sent to the client (either from the activation of a GrabButton or from a previous AllowEvents with mode SyncPointer but not from a GrabPointer), then the pointer grab is released and that event is completely reprocessed, this time ignoring any passive grabs at or above (towards the root) the grab-window of the grab just released. The request has no effect if the pointer is not grabbed by the client or if the pointer is not frozen as the result of an event.

ALLOW_SYNC_BOTH
For SyncBoth, if both pointer and keyboard are frozen by the client, event processing (for both devices) continues normally until the next ButtonPress, ButtonRelease, KeyPress, or KeyRelease event is reported to the client for a grabbed device (button event for the pointer, key event for the keyboard), at which time the devices again appear to freeze. However, if the reported event causes the grab to be released, then the devices do not freeze (but if the other device is still grabbed, then a subsequent event for it will still cause both devices to freeze). SyncBoth has no effect unless both pointer and keyboard are frozen by the client. If the pointer or keyboard is frozen twice by the client on behalf of two separate grabs, SyncBoth thaws for both (but a subsequent freeze for SyncBoth will only freeze each device once).

ALLOW_SYNC_KEYBOARD
For SyncKeyboard, if the keyboard is frozen and actively grabbed by the client, keyboard event processing continues normally until the next KeyPress or KeyRelease event is reported to the client, at which time the keyboard again appears to freeze. However, if the reported event causes the keyboard grab to be released, then the keyboard does not freeze. SyncKeyboard has no effect if the keyboard is not frozen by the client or if the keyboard is not grabbed by the client.

ALLOW_SYNC_POINTER
For SyncPointer, if the pointer is frozen and actively grabbed by the client, pointer event processing continues normally until the next ButtonPress or ButtonRelease event is reported to the client, at which time the pointer again appears to freeze. However, if the reported event causes the pointer grab to be released, then the pointer does not freeze. SyncPointer has no effect if the pointer is not frozen by the client or if the pointer is not grabbed by the client.

ARC_MODE_CHORD

ARC_MODE_PIE_SLICE

ATOM

ATOM_ANY

ATOM_ARC

ATOM_ATOM

ATOM_BITMAP

ATOM_CAP_HEIGHT

ATOM_CARDINAL

ATOM_COLORMAP

ATOM_COPYRIGHT

ATOM_CURSOR

ATOM_CUT_BUFFER0

ATOM_CUT_BUFFER1

ATOM_CUT_BUFFER2

ATOM_CUT_BUFFER3

ATOM_CUT_BUFFER4

ATOM_CUT_BUFFER5

ATOM_CUT_BUFFER6

ATOM_CUT_BUFFER7

ATOM_DRAWABLE

ATOM_END_SPACE

ATOM_FAMILY_NAME

ATOM_FONT

ATOM_FONT_NAME

ATOM_FULL_NAME

ATOM_INTEGER

ATOM_ITALIC_ANGLE

ATOM_MAX_SPACE

ATOM_MIN_SPACE

ATOM_NONE

ATOM_NORM_SPACE

ATOM_NOTICE

ATOM_PIXMAP

ATOM_POINT

ATOM_POINT_SIZE

ATOM_PRIMARY

ATOM_QUAD_WIDTH

ATOM_RECTANGLE

ATOM_RESOLUTION

ATOM_RESOURCE_MANAGER

ATOM_RGB_BEST_MAP

ATOM_RGB_BLUE_MAP

ATOM_RGB_COLOR_MAP

ATOM_RGB_DEFAULT_MAP

ATOM_RGB_GRAY_MAP

ATOM_RGB_GREEN_MAP

ATOM_RGB_RED_MAP

ATOM_SECONDARY

ATOM_STRIKEOUT_ASCENT

ATOM_STRIKEOUT_DESCENT

ATOM_STRING

ATOM_SUBSCRIPT_X

ATOM_SUBSCRIPT_Y

ATOM_SUPERSCRIPT_X

ATOM_SUPERSCRIPT_Y

ATOM_UNDERLINE_POSITION

ATOM_UNDERLINE_THICKNESS

ATOM_VISUALID

ATOM_WEIGHT

ATOM_WINDOW

ATOM_WM_CLASS

ATOM_WM_CLIENT_MACHINE

ATOM_WM_COMMAND

ATOM_WM_HINTS

ATOM_WM_ICON_NAME

ATOM_WM_ICON_SIZE

ATOM_WM_NAME

ATOM_WM_NORMAL_HINTS

ATOM_WM_SIZE_HINTS

ATOM_WM_TRANSIENT_FOR

ATOM_WM_ZOOM_HINTS

ATOM_X_HEIGHT

AUTO_REPEAT_MODE_DEFAULT

AUTO_REPEAT_MODE_OFF

AUTO_REPEAT_MODE_ON

BACKING_STORE_ALWAYS

BACKING_STORE_NOT_USEFUL

BACKING_STORE_WHEN_MAPPED

BACK_PIXMAP_NONE

BACK_PIXMAP_PARENT_RELATIVE

BELL

BLANKING_DEFAULT

BLANKING_NOT_PREFERRED

BLANKING_PREFERRED

BUTTON_INDEX_1
The left mouse button.

BUTTON_INDEX_2
The right mouse button.

BUTTON_INDEX_3
The middle mouse button.

BUTTON_INDEX_4
Scroll wheel. TODO: direction?

BUTTON_INDEX_5
Scroll wheel. TODO: direction?

BUTTON_INDEX_ANY
Any of the following (or none):

BUTTON_MASK_1

BUTTON_MASK_2

BUTTON_MASK_3

BUTTON_MASK_4

BUTTON_MASK_5

BUTTON_MASK_ANY

BUTTON_PRESS

BUTTON_RELEASE

CAP_STYLE_BUTT

CAP_STYLE_NOT_LAST

CAP_STYLE_PROJECTING

CAP_STYLE_ROUND

CHANGE_ACTIVE_POINTER_GRAB

CHANGE_GC

CHANGE_HOSTS

CHANGE_KEYBOARD_CONTROL

CHANGE_KEYBOARD_MAPPING

CHANGE_POINTER_CONTROL

CHANGE_PROPERTY

CHANGE_SAVE_SET

CHANGE_WINDOW_ATTRIBUTES

CIRCULATE_LOWER_HIGHEST

CIRCULATE_NOTIFY

CIRCULATE_RAISE_LOWEST

CIRCULATE_REQUEST

CIRCULATE_WINDOW

CLEAR_AREA

CLIENT_MESSAGE

CLIP_ORDERING_UNSORTED

CLIP_ORDERING_YX_BANDED

CLIP_ORDERING_YX_SORTED

CLIP_ORDERING_Y_SORTED

CLOSE_DOWN_DESTROY_ALL

CLOSE_DOWN_RETAIN_PERMANENT

CLOSE_DOWN_RETAIN_TEMPORARY

CLOSE_FONT

COLORMAP

COLORMAP_ALLOC_ALL

COLORMAP_ALLOC_NONE

COLORMAP_NONE

COLORMAP_NOTIFY

COLORMAP_STATE_INSTALLED
The colormap was installed.

COLORMAP_STATE_UNINSTALLED
The colormap was uninstalled.

COLOR_FLAG_BLUE

COLOR_FLAG_GREEN

COLOR_FLAG_RED

CONFIGURE_NOTIFY

CONFIGURE_REQUEST

CONFIGURE_WINDOW

CONFIG_WINDOW_BORDER_WIDTH

CONFIG_WINDOW_HEIGHT

CONFIG_WINDOW_SIBLING

CONFIG_WINDOW_STACK_MODE

CONFIG_WINDOW_WIDTH

CONFIG_WINDOW_X

CONFIG_WINDOW_Y

CONVERT_SELECTION

COORD_MODE_ORIGIN
Treats all coordinates as relative to the origin.

COORD_MODE_PREVIOUS
Treats all coordinates after the first as relative to the previous coordinate.

COPY_AREA

COPY_COLORMAP_AND_FREE

COPY_GC

COPY_PLANE

CREATE_COLORMAP

CREATE_CURSOR

CREATE_GC

CREATE_GLYPH_CURSOR

CREATE_NOTIFY

CREATE_PIXMAP

CREATE_WINDOW

CURSOR

CURSOR_NONE

CW_BACKING_PIXEL
The backing-pixel specifies what value to use in planes not covered by backing-planes. The server is free to save only the specified bit planes in the backing-store or save-under and regenerate the remaining planes with the specified pixel value. Any bits beyond the specified depth of the window in these values are simply ignored.

CW_BACKING_PLANES
The backing-planes indicates (with bits set to 1) which bit planes of the window hold dynamic data that must be preserved in backing-stores and during save-unders.

CW_BACKING_STORE
A backing-store of WhenMapped advises the server that maintaining contents of obscured regions when the window is mapped would be beneficial. A backing-store of Always advises the server that maintaining contents even when the window is unmapped would be beneficial. In this case, the server may generate an exposure event when the window is created. A value of NotUseful advises the server that maintaining contents is unnecessary, although a server may still choose to maintain contents while the window is mapped. Note that if the server maintains contents, then the server should maintain complete contents not just the region within the parent boundaries, even if the window is larger than its parent. While the server maintains contents, exposure events will not normally be generated, but the server may stop maintaining contents at any time.

CW_BACK_PIXEL
Overrides BackPixmap. A pixmap of undefined size filled with the specified background pixel is used for the background. Range-checking is not performed, the background pixel is truncated to the appropriate number of bits.

CW_BACK_PIXMAP
Overrides the default background-pixmap. The background pixmap and window must have the same root and same depth. Any size pixmap can be used, although some sizes may be faster than others.

CW_BIT_GRAVITY
Defines which region of the window should be retained if the window is resized.

CW_BORDER_PIXEL
Overrides BorderPixmap. A pixmap of undefined size filled with the specified border pixel is used for the border. Range checking is not performed on the border-pixel value, it is truncated to the appropriate number of bits.

CW_BORDER_PIXMAP
Overrides the default border-pixmap. The border pixmap and window must have the same root and the same depth. Any size pixmap can be used, although some sizes may be faster than others.

CW_COLORMAP
The colormap specifies the colormap that best reflects the true colors of the window. Servers capable of supporting multiple hardware colormaps may use this information, and window man- agers may use it for InstallColormap requests. The colormap must have the same visual type and root as the window (or a Match error results). If CopyFromParent is specified, the parent's colormap is copied (subsequent changes to the parent's colormap attribute do not affect the child). However, the window must have the same visual type as the parent (or a Match error results), and the parent must not have a colormap of None (or a Match error results). For an explanation of None, see FreeColormap request. The colormap is copied by sharing the colormap object between the child and the parent, not by making a complete copy of the colormap contents.

CW_CURSOR
If a cursor is specified, it will be used whenever the pointer is in the window. If None is speci- fied, the parent's cursor will be used when the pointer is in the window, and any change in the parent's cursor will cause an immediate change in the displayed cursor.

CW_DONT_PROPAGATE
The do-not-propagate-mask defines which events should not be propagated to ancestor windows when no client has the event type selected in this window.

CW_EVENT_MASK
The event-mask defines which events the client is interested in for this window (or for some event types, inferiors of the window).

CW_OVERRIDE_REDIRECT
The override-redirect specifies whether map and configure requests on this window should override a SubstructureRedirect on the parent, typically to inform a window manager not to tamper with the window.

CW_SAVE_UNDER
If 1, the server is advised that when this window is mapped, saving the contents of windows it obscures would be beneficial.

CW_WIN_GRAVITY
Defines how the window should be repositioned if the parent is resized (see ConfigureWindow).

DELETE_PROPERTY

DESTROY_NOTIFY

DESTROY_SUBWINDOWS

DESTROY_WINDOW

DRAWABLE

ENTER_NOTIFY

EVENT_MASK_BUTTON_1_MOTION

EVENT_MASK_BUTTON_2_MOTION

EVENT_MASK_BUTTON_3_MOTION

EVENT_MASK_BUTTON_4_MOTION

EVENT_MASK_BUTTON_5_MOTION

EVENT_MASK_BUTTON_MOTION

EVENT_MASK_BUTTON_PRESS

EVENT_MASK_BUTTON_RELEASE

EVENT_MASK_COLOR_MAP_CHANGE

EVENT_MASK_ENTER_WINDOW

EVENT_MASK_EXPOSURE

EVENT_MASK_FOCUS_CHANGE

EVENT_MASK_KEYMAP_STATE

EVENT_MASK_KEY_PRESS

EVENT_MASK_KEY_RELEASE

EVENT_MASK_LEAVE_WINDOW

EVENT_MASK_NO_EVENT

EVENT_MASK_OWNER_GRAB_BUTTON

EVENT_MASK_POINTER_MOTION

EVENT_MASK_POINTER_MOTION_HINT

EVENT_MASK_PROPERTY_CHANGE

EVENT_MASK_RESIZE_REDIRECT

EVENT_MASK_STRUCTURE_NOTIFY

EVENT_MASK_SUBSTRUCTURE_NOTIFY

EVENT_MASK_SUBSTRUCTURE_REDIRECT

EVENT_MASK_VISIBILITY_CHANGE

EXPOSE

EXPOSURES_ALLOWED

EXPOSURES_DEFAULT

EXPOSURES_NOT_ALLOWED

FAMILY_CHAOS

FAMILY_DE_CNET

FAMILY_INTERNET

FAMILY_INTERNET_6

FAMILY_SERVER_INTERPRETED

FILL_POLY

FILL_RULE_EVEN_ODD

FILL_RULE_WINDING

FILL_STYLE_OPAQUE_STIPPLED

FILL_STYLE_SOLID

FILL_STYLE_STIPPLED

FILL_STYLE_TILED

FOCUS_IN

FOCUS_OUT

FONT

FONT_DRAW_LEFT_TO_RIGHT

FONT_DRAW_RIGHT_TO_LEFT

FONT_NONE

FORCE_SCREEN_SAVER

FREE_COLORMAP

FREE_COLORS

FREE_CURSOR

FREE_GC

FREE_PIXMAP

GC_ARC_MODE
TODO

GC_BACKGROUND
Background colorpixel.

GC_CAP_STYLE
The cap-style defines how the endpoints of a path are drawn: NotLast The result is equivalent to Butt, except that for a line-width of zero the final endpoint is not drawn. Butt The result is square at the endpoint (perpendicular to the slope of the line) with no projection beyond. Round The result is a circular arc with its diameter equal to the line-width, centered on the endpoint; it is equivalent to Butt for line-width zero. Projecting The result is square at the end, but the path continues beyond the endpoint for a distance equal to half the line-width; it is equivalent to Butt for line-width zero.

GC_CLIP_MASK
The clip-mask restricts writes to the destination drawable. Only pixels where the clip-mask has bits set to 1 are drawn. Pixels are not drawn outside the area covered by the clip-mask or where the clip-mask has bits set to 0. The clip-mask affects all graphics requests, but it does not clip sources. The clip-mask origin is interpreted relative to the origin of whatever destination drawable is specified in a graphics request. If a pixmap is specified as the clip-mask, it must have depth 1 and have the same root as the gcontext (or a Match error results). If clip-mask is None, then pixels are always drawn, regardless of the clip origin. The clip-mask can also be set with the SetClipRectangles request.

GC_CLIP_ORIGIN_X
TODO

GC_CLIP_ORIGIN_Y
TODO

GC_DASH_LIST
TODO

GC_DASH_OFFSET
TODO

GC_FILL_RULE

GC_FILL_STYLE
The fill-style defines the contents of the source for line, text, and fill requests. For all text and fill requests (for example, PolyText8, PolyText16, PolyFillRectangle, FillPoly, and PolyFillArc) as well as for line requests with line-style Solid, (for example, PolyLine, PolySegment, PolyRectangle, PolyArc) and for the even dashes for line requests with line-style OnOffDash or DoubleDash: Solid Foreground Tiled Tile OpaqueStippled A tile with the same width and height as stipple but with background everywhere stipple has a zero and with foreground everywhere stipple has a one Stippled Foreground masked by stipple For the odd dashes for line requests with line-style DoubleDash: Solid Background Tiled Same as for even dashes OpaqueStippled Same as for even dashes Stippled Background masked by stipple

GC_FONT
Which font to use for the ImageText8 and ImageText16 requests.

GC_FOREGROUND
Foreground colorpixel.

GC_FUNCTION
TODO: Refer to GX

GC_GRAPHICS_EXPOSURES
Whether ExposureEvents should be generated (1) or not (0).

GC_JOIN_STYLE
The join-style defines how corners are drawn for wide lines: Miter The outer edges of the two lines extend to meet at an angle. However, if the angle is less than 11 degrees, a Bevel join-style is used instead. Round The result is a circular arc with a diameter equal to the line-width, centered on the joinpoint. Bevel The result is Butt endpoint styles, and then the triangular notch is filled.

GC_LINE_STYLE
The line-style defines which sections of a line are drawn: Solid The full path of the line is drawn. DoubleDash The full path of the line is drawn, but the even dashes are filled differently than the odd dashes (see fill-style), with Butt cap-style used where even and odd dashes meet. OnOffDash Only the even dashes are drawn, and cap-style applies to all internal ends of the individual dashes (except NotLast is treated as Butt).

GC_LINE_WIDTH
The line-width is measured in pixels and can be greater than or equal to one, a wide line, or the special value zero, a thin line.

GC_PLANE_MASK
In graphics operations, given a source and destination pixel, the result is computed bitwise on corresponding bits of the pixels; that is, a Boolean operation is performed in each bit plane. The plane-mask restricts the operation to a subset of planes, so the result is:

    ((src FUNC dst) AND plane-mask) OR (dst AND (NOT plane-mask))

GC_STIPPLE
The tile/stipple represents an infinite two-dimensional plane with the tile/stipple replicated in all dimensions. When that plane is superimposed on the drawable for use in a graphics operation, the upper-left corner of some instance of the tile/stipple is at the coordinates within the drawable specified by the tile/stipple origin. The tile/stipple and clip origins are interpreted relative to the origin of whatever destination drawable is specified in a graphics request. The tile pixmap must have the same root and depth as the gcontext (or a Match error results). The stipple pixmap must have depth one and must have the same root as the gcontext (or a Match error results). For fill-style Stippled (but not fill-style OpaqueStippled), the stipple pattern is tiled in a single plane and acts as an additional clip mask to be ANDed with the clip-mask. Any size pixmap can be used for tiling or stippling, although some sizes may be faster to use than others.

GC_SUBWINDOW_MODE
For ClipByChildren, both source and destination windows are additionally clipped by all viewable InputOutput children. For IncludeInferiors, neither source nor destination window is clipped by inferiors. This will result in including subwindow contents in the source and drawing through subwindow boundaries of the destination. The use of IncludeInferiors with a source or destination window of one depth with mapped inferiors of differing depth is not illegal, but the semantics is undefined by the core protocol.

GC_TILE
The tile/stipple represents an infinite two-dimensional plane with the tile/stipple replicated in all dimensions. When that plane is superimposed on the drawable for use in a graphics operation, the upper-left corner of some instance of the tile/stipple is at the coordinates within the drawable specified by the tile/stipple origin. The tile/stipple and clip origins are interpreted relative to the origin of whatever destination drawable is specified in a graphics request. The tile pixmap must have the same root and depth as the gcontext (or a Match error results). The stipple pixmap must have depth one and must have the same root as the gcontext (or a Match error results). For fill-style Stippled (but not fill-style OpaqueStippled), the stipple pattern is tiled in a single plane and acts as an additional clip mask to be ANDed with the clip-mask. Any size pixmap can be used for tiling or stippling, although some sizes may be faster to use than others.

GC_TILE_STIPPLE_ORIGIN_X
TODO

GC_TILE_STIPPLE_ORIGIN_Y
TODO

GET_ATOM_NAME

GET_FONT_PATH

GET_GEOMETRY

GET_IMAGE

GET_INPUT_FOCUS

GET_KEYBOARD_CONTROL

GET_KEYBOARD_MAPPING

GET_MODIFIER_MAPPING

GET_MOTION_EVENTS

GET_POINTER_CONTROL

GET_POINTER_MAPPING

GET_PROPERTY

GET_PROPERTY_TYPE_ANY

GET_SCREEN_SAVER

GET_SELECTION_OWNER

GET_WINDOW_ATTRIBUTES

GE_GENERIC

GRAB_ANY

GRAB_BUTTON

GRAB_KEY

GRAB_KEYBOARD

GRAB_MODE_ASYNC
Keyboard event processing continues normally.

GRAB_MODE_SYNC
The state of the keyboard appears to freeze: No further keyboard events are generated by the server until the grabbing client issues a releasing AllowEvents request or until the keyboard grab is released.

GRAB_POINTER

GRAB_SERVER

GRAB_STATUS_ALREADY_GRABBED

GRAB_STATUS_FROZEN

GRAB_STATUS_INVALID_TIME

GRAB_STATUS_NOT_VIEWABLE

GRAB_STATUS_SUCCESS

GRAPHICS_EXPOSURE

GRAVITY_BIT_FORGET

GRAVITY_CENTER

GRAVITY_EAST

GRAVITY_NORTH

GRAVITY_NORTH_EAST

GRAVITY_NORTH_WEST

GRAVITY_NOTIFY

GRAVITY_SOUTH

GRAVITY_SOUTH_EAST

GRAVITY_SOUTH_WEST

GRAVITY_STATIC

GRAVITY_WEST

GRAVITY_WIN_UNMAP

GX_AND

GX_AND_INVERTED

GX_AND_REVERSE

GX_CLEAR

GX_COPY

GX_COPY_INVERTED

GX_EQUIV

GX_INVERT

GX_NAND

GX_NOOP

GX_NOR

GX_OR

GX_OR_INVERTED

GX_OR_REVERSE

GX_SET

GX_XOR

G_CONTEXT

HOST_MODE_DELETE

HOST_MODE_INSERT

ID_CHOICE

IMAGE_FORMAT_XY_BITMAP

IMAGE_FORMAT_XY_PIXMAP

IMAGE_FORMAT_Z_PIXMAP

IMAGE_ORDER_LSB_FIRST

IMAGE_ORDER_MSB_FIRST

IMAGE_TEXT_16

IMAGE_TEXT_8

IMPLEMENTATION

INPUT_FOCUS_FOLLOW_KEYBOARD
NOT YET DOCUMENTED. Only relevant for the xinput extension.

INPUT_FOCUS_NONE
The focus reverts to XCB_NONE, so no window will have the input focus.

INPUT_FOCUS_PARENT
The focus reverts to the parent (or closest viewable ancestor) and the new revert_to value is XCB_INPUT_FOCUS_NONE.

INPUT_FOCUS_POINTER_ROOT
The focus reverts to XCB_POINTER_ROOT respectively. When the focus reverts, FocusIn and FocusOut events are generated, but the last-focus-change time is not changed.

INSTALL_COLORMAP

INTERN_ATOM

JOIN_STYLE_BEVEL

JOIN_STYLE_MITER

JOIN_STYLE_ROUND

KB_AUTO_REPEAT_MODE

KB_BELL_DURATION

KB_BELL_PERCENT

KB_BELL_PITCH

KB_KEY

KB_KEY_CLICK_PERCENT

KB_LED

KB_LED_MODE

KEYMAP_NOTIFY

KEY_BUT_MASK_BUTTON_1

KEY_BUT_MASK_BUTTON_2

KEY_BUT_MASK_BUTTON_3

KEY_BUT_MASK_BUTTON_4

KEY_BUT_MASK_BUTTON_5

KEY_BUT_MASK_CONTROL

KEY_BUT_MASK_LOCK

KEY_BUT_MASK_MOD_1

KEY_BUT_MASK_MOD_2

KEY_BUT_MASK_MOD_3

KEY_BUT_MASK_MOD_4

KEY_BUT_MASK_MOD_5

KEY_BUT_MASK_SHIFT

KEY_PRESS

KEY_RELEASE

KILL_ALL_TEMPORARY

KILL_CLIENT

LEAVE_NOTIFY

LED_MODE_OFF

LED_MODE_ON

LENGTH

LINE_STYLE_DOUBLE_DASH

LINE_STYLE_ON_OFF_DASH

LINE_STYLE_SOLID

LIST_EXTENSIONS

LIST_FONTS

LIST_FONTS_WITH_INFO

LIST_HOSTS

LIST_INSTALLED_COLORMAPS

LIST_PROPERTIES

LOOKUP_COLOR

MAPPING_KEYBOARD

MAPPING_MODIFIER

MAPPING_NOTIFY

MAPPING_POINTER

MAPPING_STATUS_BUSY

MAPPING_STATUS_FAILURE

MAPPING_STATUS_SUCCESS

MAP_INDEX_1

MAP_INDEX_2

MAP_INDEX_3

MAP_INDEX_4

MAP_INDEX_5

MAP_INDEX_CONTROL

MAP_INDEX_LOCK

MAP_INDEX_SHIFT

MAP_NOTIFY

MAP_REQUEST

MAP_STATE_UNMAPPED

MAP_STATE_UNVIEWABLE

MAP_STATE_VIEWABLE

MAP_SUBWINDOWS

MAP_WINDOW

MATCH

MOD_MASK_1

MOD_MASK_2

MOD_MASK_3

MOD_MASK_4

MOD_MASK_5

MOD_MASK_ANY

MOD_MASK_CONTROL

MOD_MASK_LOCK

MOD_MASK_SHIFT

MOTION_HINT

MOTION_NORMAL

MOTION_NOTIFY

NAME

NOTIFY_DETAIL_ANCESTOR

NOTIFY_DETAIL_INFERIOR

NOTIFY_DETAIL_NONE

NOTIFY_DETAIL_NONLINEAR

NOTIFY_DETAIL_NONLINEAR_VIRTUAL

NOTIFY_DETAIL_POINTER

NOTIFY_DETAIL_POINTER_ROOT

NOTIFY_DETAIL_VIRTUAL

NOTIFY_MODE_GRAB

NOTIFY_MODE_NORMAL

NOTIFY_MODE_UNGRAB

NOTIFY_MODE_WHILE_GRABBED

NO_EXPOSURE

NO_OPERATION

OPEN_FONT

PIXMAP

PIXMAP_NONE

PLACE_ON_BOTTOM
The window is now below all siblings.

PLACE_ON_TOP
The window is now on top of all siblings.

POLY_ARC

POLY_FILL_ARC

POLY_FILL_RECTANGLE

POLY_LINE

POLY_POINT

POLY_RECTANGLE

POLY_SEGMENT

POLY_SHAPE_COMPLEX

POLY_SHAPE_CONVEX

POLY_SHAPE_NONCONVEX

POLY_TEXT_16

POLY_TEXT_8

PROPERTY_DELETE

PROPERTY_NEW_VALUE

PROPERTY_NOTIFY

PROP_MODE_APPEND
Insert the new data after the beginning of existing data. The format must match existing property value. If the property is undefined, it is treated as defined with the correct type and format with zero-length data.

PROP_MODE_PREPEND
Insert the new data before the beginning of existing data. The format must match existing property value. If the property is undefined, it is treated as defined with the correct type and format with zero-length data.

PROP_MODE_REPLACE
Discard the previous property value and store the new data.

PUT_IMAGE

QUERY_BEST_SIZE

QUERY_COLORS

QUERY_EXTENSION

QUERY_FONT

QUERY_KEYMAP

QUERY_POINTER

QUERY_SHAPE_OF_FASTEST_STIPPLE

QUERY_SHAPE_OF_FASTEST_TILE

QUERY_SHAPE_OF_LARGEST_CURSOR

QUERY_TEXT_EXTENTS

QUERY_TREE

RECOLOR_CURSOR

REPARENT_NOTIFY

REPARENT_WINDOW

REQUEST

RESIZE_REQUEST

ROTATE_PROPERTIES

SCREEN_SAVER_ACTIVE

SCREEN_SAVER_RESET

SELECTION_CLEAR

SELECTION_NOTIFY

SELECTION_REQUEST

SEND_EVENT

SEND_EVENT_DEST_ITEM_FOCUS

SEND_EVENT_DEST_POINTER_WINDOW

SET_ACCESS_CONTROL

SET_CLIP_RECTANGLES

SET_CLOSE_DOWN_MODE

SET_DASHES

SET_FONT_PATH

SET_INPUT_FOCUS

SET_MODE_DELETE

SET_MODE_INSERT

SET_MODIFIER_MAPPING

SET_POINTER_MAPPING

SET_SCREEN_SAVER

SET_SELECTION_OWNER

STACK_MODE_ABOVE

STACK_MODE_BELOW

STACK_MODE_BOTTOM_IF

STACK_MODE_OPPOSITE

STACK_MODE_TOP_IF

STORE_COLORS

STORE_NAMED_COLOR

SUBWINDOW_MODE_CLIP_BY_CHILDREN

SUBWINDOW_MODE_INCLUDE_INFERIORS

TIME_CURRENT_TIME

TRANSLATE_COORDINATES

UNGRAB_BUTTON

UNGRAB_KEY

UNGRAB_KEYBOARD

UNGRAB_POINTER

UNGRAB_SERVER

UNINSTALL_COLORMAP

UNMAP_NOTIFY

UNMAP_SUBWINDOWS

UNMAP_WINDOW

VALUE

VISIBILITY_FULLY_OBSCURED

VISIBILITY_NOTIFY

VISIBILITY_PARTIALLY_OBSCURED

VISIBILITY_UNOBSCURED

VISUAL_CLASS_DIRECT_COLOR

VISUAL_CLASS_GRAY_SCALE

VISUAL_CLASS_PSEUDO_COLOR

VISUAL_CLASS_STATIC_COLOR

VISUAL_CLASS_STATIC_GRAY

VISUAL_CLASS_TRUE_COLOR

WARP_POINTER

WINDOW

WINDOW_CLASS_COPY_FROM_PARENT

WINDOW_CLASS_INPUT_ONLY

WINDOW_CLASS_INPUT_OUTPUT

WINDOW_NONE

Functions

alloc_color
Allocate a color

alloc_color_cells

alloc_color_cells_unchecked

alloc_color_planes

alloc_color_planes_unchecked

alloc_color_unchecked
Allocate a color

alloc_named_color

alloc_named_color_unchecked

allow_events
release queued events

allow_events_checked
release queued events

bell

bell_checked

change_active_pointer_grab

change_active_pointer_grab_checked

change_gc
change graphics context components

change_gc_checked
change graphics context components

change_hosts

change_hosts_checked

change_keyboard_control

change_keyboard_control_checked

change_keyboard_mapping

change_keyboard_mapping_checked

change_pointer_control

change_pointer_control_checked

change_property
Changes a window property

change_property_checked
Changes a window property

change_save_set
Changes a client's save set

change_save_set_checked
Changes a client's save set

change_window_attributes
change window attributes

change_window_attributes_checked
change window attributes

circulate_window
Change window stacking order

circulate_window_checked
Change window stacking order

clear_area

clear_area_checked

close_font

close_font_checked

configure_window
Configures window attributes

configure_window_checked
Configures window attributes

convert_selection

convert_selection_checked

copy_area
copy areas

copy_area_checked
copy areas

copy_colormap_and_free

copy_colormap_and_free_checked

copy_gc

copy_gc_checked

copy_plane

copy_plane_checked

create_colormap

create_colormap_checked

create_cursor

create_cursor_checked

create_gc
Creates a graphics context

create_gc_checked
Creates a graphics context

create_glyph_cursor
create cursor

create_glyph_cursor_checked
create cursor

create_pixmap
Creates a pixmap

create_pixmap_checked
Creates a pixmap

create_window
Creates a window

create_window_checked
Creates a window

delete_property

delete_property_checked

destroy_subwindows

destroy_subwindows_checked

destroy_window
Destroys a window

destroy_window_checked
Destroys a window

fill_poly

fill_poly_checked

force_screen_saver

force_screen_saver_checked

free_colormap

free_colormap_checked

free_colors

free_colors_checked

free_cursor
Deletes a cursor

free_cursor_checked
Deletes a cursor

free_gc
Destroys a graphics context

free_gc_checked
Destroys a graphics context

free_pixmap
Destroys a pixmap

free_pixmap_checked
Destroys a pixmap

get_atom_name

get_atom_name_unchecked

get_font_path

get_font_path_unchecked

get_geometry
Get current window geometry

get_geometry_unchecked
Get current window geometry

get_image

get_image_unchecked

get_input_focus

get_input_focus_unchecked

get_keyboard_control

get_keyboard_control_unchecked

get_keyboard_mapping

get_keyboard_mapping_unchecked

get_modifier_mapping

get_modifier_mapping_unchecked

get_motion_events

get_motion_events_unchecked

get_pointer_control

get_pointer_control_unchecked

get_pointer_mapping

get_pointer_mapping_unchecked

get_property
Gets a window property

get_property_unchecked
Gets a window property

get_screen_saver

get_screen_saver_unchecked

get_selection_owner
Gets the owner of a selection

get_selection_owner_unchecked
Gets the owner of a selection

get_window_attributes
Gets window attributes

get_window_attributes_unchecked
Gets window attributes

grab_button
Grab pointer button(s)

grab_button_checked
Grab pointer button(s)

grab_key
Grab keyboard key(s)

grab_key_checked
Grab keyboard key(s)

grab_keyboard
Grab the keyboard

grab_keyboard_unchecked
Grab the keyboard

grab_pointer
Grab the pointer

grab_pointer_unchecked
Grab the pointer

grab_server

grab_server_checked

image_text_16
Draws text

image_text_16_checked
Draws text

image_text_8
Draws text

image_text_8_checked
Draws text

install_colormap

install_colormap_checked

intern_atom
Get atom identifier by name

intern_atom_unchecked
Get atom identifier by name

kill_client
kills a client

kill_client_checked
kills a client

list_extensions

list_extensions_unchecked

list_fonts
get matching font names

list_fonts_unchecked
get matching font names

list_fonts_with_info
get matching font names and information

list_fonts_with_info_unchecked
get matching font names and information

list_hosts

list_hosts_unchecked

list_installed_colormaps

list_installed_colormaps_unchecked

list_properties

list_properties_unchecked

lookup_color

lookup_color_unchecked

map_subwindows

map_subwindows_checked

map_window
Makes a window visible

map_window_checked
Makes a window visible

no_operation

no_operation_checked

open_font
opens a font

open_font_checked
opens a font

poly_arc

poly_arc_checked

poly_fill_arc

poly_fill_arc_checked

poly_fill_rectangle
Fills rectangles

poly_fill_rectangle_checked
Fills rectangles

poly_line
draw lines

poly_line_checked
draw lines

poly_point

poly_point_checked

poly_rectangle

poly_rectangle_checked

poly_segment
draw lines

poly_segment_checked
draw lines

poly_text_16

poly_text_16_checked

poly_text_8

poly_text_8_checked

put_image

put_image_checked

query_best_size

query_best_size_unchecked

query_colors

query_colors_unchecked

query_extension
check if extension is present

query_extension_unchecked
check if extension is present

query_font
query font metrics

query_font_unchecked
query font metrics

query_keymap

query_keymap_unchecked

query_pointer
get pointer coordinates

query_pointer_unchecked
get pointer coordinates

query_text_extents
get text extents

query_text_extents_unchecked
get text extents

query_tree
query the window tree

query_tree_unchecked
query the window tree

recolor_cursor

recolor_cursor_checked

reparent_window
Reparents a window

reparent_window_checked
Reparents a window

rotate_properties

rotate_properties_checked

send_event
send an event

send_event_checked
send an event

set_access_control

set_access_control_checked

set_clip_rectangles

set_clip_rectangles_checked

set_close_down_mode

set_close_down_mode_checked

set_dashes

set_dashes_checked

set_font_path

set_font_path_checked

set_input_focus
Sets input focus

set_input_focus_checked
Sets input focus

set_modifier_mapping

set_modifier_mapping_unchecked

set_pointer_mapping

set_pointer_mapping_unchecked

set_screen_saver

set_screen_saver_checked

set_selection_owner
Sets the owner of a selection

set_selection_owner_checked
Sets the owner of a selection

store_colors

store_colors_checked

store_named_color

store_named_color_checked

translate_coordinates

translate_coordinates_unchecked

ungrab_button

ungrab_button_checked

ungrab_key
release a key combination

ungrab_key_checked
release a key combination

ungrab_keyboard

ungrab_keyboard_checked

ungrab_pointer
release the pointer

ungrab_pointer_checked
release the pointer

ungrab_server

ungrab_server_checked

uninstall_colormap

uninstall_colormap_checked

unmap_subwindows

unmap_subwindows_checked

unmap_window
Makes a window invisible

unmap_window_checked
Makes a window invisible

warp_pointer
move mouse pointer

warp_pointer_checked
move mouse pointer

Type Definitions

AccessControl

AllocColorCellsCookie

AllocColorCellsReply

AllocColorCookie

AllocColorPlanesCookie

AllocColorPlanesReply

AllocColorReply

AllocNamedColorCookie

AllocNamedColorReply

Allow

ArcIterator

ArcMode

Atom

AtomEnum

AutoRepeatMode

BackPixmap

BackingStore

Blanking

Button

ButtonIndex

ButtonMask

ButtonPressEvent
a mouse button was pressed/released

ButtonReleaseEvent
a mouse button was pressed/released

CapStyle

Char2bIterator

CharinfoIterator

Circulate

CirculateNotifyEvent
NOT YET DOCUMENTED

CirculateRequestEvent
NOT YET DOCUMENTED

ClientMessageData

ClientMessageDataIterator

ClientMessageEvent
NOT YET DOCUMENTED

ClipOrdering

CloseDown

ColorFlag

Coloritem

ColoritemIterator

Colormap

ColormapAlloc

ColormapEnum

ColormapNotifyEvent
the colormap for some window changed

ColormapState

ConfigWindow

ConfigureNotifyEvent
NOT YET DOCUMENTED

ConfigureRequestEvent

CoordMode

CreateNotifyEvent

Cursor

CursorEnum

Cw

Depth

DepthIterator

DestroyNotifyEvent
a window is destroyed

Drawable

EnterNotifyEvent
the pointer is in a different window

EventMask

ExposeEvent
NOT YET DOCUMENTED

Exposures

Family

FillRule

FillStyle

FocusInEvent
NOT YET DOCUMENTED

FocusOutEvent
NOT YET DOCUMENTED

Font

FontDraw

FontEnum

Fontable

FontpropIterator

Format

FormatIterator

Gc

Gcontext

GeGenericEvent
generic event (with length)

GetAtomNameCookie

GetAtomNameReply

GetFontPathCookie

GetFontPathReply

GetGeometryCookie

GetGeometryReply

GetImageCookie

GetImageReply

GetInputFocusCookie

GetInputFocusReply

GetKeyboardControlCookie

GetKeyboardControlReply

GetKeyboardMappingCookie

GetKeyboardMappingReply

GetModifierMappingCookie

GetModifierMappingReply

GetMotionEventsCookie

GetMotionEventsReply

GetPointerControlCookie

GetPointerControlReply

GetPointerMappingCookie

GetPointerMappingReply

GetPropertyCookie

GetPropertyReply

GetPropertyType

GetScreenSaverCookie

GetScreenSaverReply

GetSelectionOwnerCookie

GetSelectionOwnerReply

GetWindowAttributesCookie

GetWindowAttributesReply

Grab

GrabKeyboardCookie

GrabKeyboardReply

GrabMode

GrabPointerCookie

GrabPointerReply

GrabStatus

GraphicsExposureEvent

Gravity

GravityNotifyEvent

Gx

Host

HostIterator

HostMode

ImageFormat

ImageOrder

InputFocus

InternAtomCookie

InternAtomReply

JoinStyle

Kb

KeyButMask

KeyPressEvent
a key was pressed/released

KeyReleaseEvent
a key was pressed/released

Keycode

KeymapNotifyEvent

Keysym

Kill

LeaveNotifyEvent
the pointer is in a different window

LedMode

LineStyle

ListExtensionsCookie

ListExtensionsReply

ListFontsCookie

ListFontsReply

ListFontsWithInfoCookie

ListFontsWithInfoReply

ListHostsCookie

ListHostsReply

ListInstalledColormapsCookie

ListInstalledColormapsReply

ListPropertiesCookie

ListPropertiesReply

LookupColorCookie

LookupColorReply

MapIndex

MapNotifyEvent
a window was mapped

MapRequestEvent
window wants to be mapped

MapState

Mapping

MappingNotifyEvent
keyboard mapping changed

MappingStatus

ModMask

Motion

MotionNotifyEvent
a key was pressed

NoExposureEvent

NotifyDetail

NotifyMode

Pixmap

PixmapEnum

Place

PointIterator

PolyShape

PropMode

Property

PropertyNotifyEvent
a window property changed

QueryBestSizeCookie

QueryBestSizeReply

QueryColorsCookie

QueryColorsReply

QueryExtensionCookie

QueryExtensionReply

QueryFontCookie

QueryFontReply

QueryKeymapCookie

QueryKeymapReply

QueryPointerCookie

QueryPointerReply

QueryShapeOf

QueryTextExtentsCookie

QueryTextExtentsReply

QueryTreeCookie

QueryTreeReply

RectangleIterator

ReparentNotifyEvent

ResizeRequestEvent

Rgb

RgbIterator

Screen

ScreenIterator

ScreenSaver

SegmentIterator

SelectionClearEvent

SelectionNotifyEvent

SelectionRequestEvent

SendEventDest

SetMode

SetModifierMappingCookie

SetModifierMappingReply

SetPointerMappingCookie

SetPointerMappingReply

Setup

SetupAuthenticate

SetupAuthenticateIterator

SetupFailed

SetupFailedIterator

SetupIterator

SetupRequest

SetupRequestIterator

StackMode

Str

StrIterator

SubwindowMode

Time

TimecoordIterator

Timestamp

TranslateCoordinatesCookie

TranslateCoordinatesReply

UnmapNotifyEvent
a window is unmapped

Visibility

VisibilityNotifyEvent

VisualClass

Visualid

Visualtype

VisualtypeIterator

Window

WindowClass

WindowEnum