Function xcb::xproto::change_property_checked

pub fn change_property_checked<T>(c: &Connection, mode: u8, window: Window, property: Atom, type_: Atom, format: u8, data: &[T]) -> VoidCookie

Changes a window property

Sets or updates a property on the specified window. Properties are for example the window title (WM_NAME) or its minimum size (WM_NORMAL_HINTS). Protocols such as EWMH also use properties - for example EWMH defines the window title, encoded as UTF-8 string, in the _NET_WM_NAME property.

parameters:

c: The connection object to the server

mode:

window: The window whose property you want to change.

property: The property you want to change (an atom).

type: The type of the property you want to change (an atom).

format: Specifies whether the data should be viewed as a list of 8-bit, 16-bit or 32-bit quantities. Possible values are 8, 16 and 32. This information allows the X server to correctly perform byte-swap operations as necessary.

data: The property data.