Function xcb::xproto::grab_keyboard_unchecked

pub fn grab_keyboard_unchecked<'a>(
    c: &'a Connection, 
    owner_events: bool, 
    grab_window: Window, 
    time: Timestamp, 
    pointer_mode: u8, 
    keyboard_mode: u8
) -> GrabKeyboardCookie<'a>

Grab the keyboard

Actively grabs control of the keyboard and generates FocusIn and FocusOut events. Further key events are reported only to the grabbing client.

Any active keyboard grab by this client is overridden. If the keyboard is actively grabbed by some other client, AlreadyGrabbed is returned. If grab_window is not viewable, GrabNotViewable is returned. If the keyboard is frozen by an active grab of another client, GrabFrozen is returned. If the specified time is earlier than the last-keyboard-grab time or later than the current X server time, GrabInvalidTime is returned. Otherwise, the last-keyboard-grab time is set to the specified time.

parameters:

c: The connection object to the server

owner_events: If 1, the grab_window will still get the pointer events. If 0, events are not reported to the grab_window.

grab_window: Specifies the window on which the pointer should be grabbed.

time: Timestamp to avoid race conditions when running X over the network.

The special value XCB_CURRENT_TIME will be replaced with the current server time.

pointer_mode:

keyboard_mode: