Function xcb::xproto::image_text_16_checked

pub fn image_text_16_checked(c: &Connection, drawable: Drawable, gc: Gcontext, x: i16, y: i16, string: &[Char2b]) -> VoidCookie

Draws text

Fills the destination rectangle with the background pixel from gc, then paints the text with the foreground pixel from gc. The upper-left corner of the filled rectangle is at [x, y - font-ascent]. The width is overall-width, the height is font-ascent + font-descent. The overall-width, font-ascent and font-descent are as returned by xcb_query_text_extents (TODO).

Note that using X core fonts is deprecated (but still supported) in favor of client-side rendering using Xft.

parameters:

c: The connection object to the server

drawable: The drawable (Window or Pixmap) to draw text on.

gc: The graphics context to use.

The following graphics context components are used: plane-mask, foreground, background, font, subwindow-mode, clip-x-origin, clip-y-origin, and clip-mask.

x: The x coordinate of the first character, relative to the origin of drawable.

y: The y coordinate of the first character, relative to the origin of drawable.

string: The string to draw. Only the first 255 characters are relevant due to the data type of string_len. Every character uses 2 bytes (hence the 16 in this request's name).