xcb::xproto

Function poly_segment

pub fn poly_segment<'a>(
    c: &'a Connection,
    drawable: Drawable,
    gc: Gcontext,
    segments: &[Segment],
) -> VoidCookie<'a>

draw lines

Draws multiple, unconnected lines. For each segment, a line is drawn between (x1, y1) and (x2, y2). The lines are drawn in the order listed in the array of xcb_segment_t structures and does not perform joining at coincident endpoints. For any given line, a pixel is not drawn more than once. If lines intersect, the intersecting pixels are drawn multiple times.

TODO: include the xcb_segment_t data structure

TODO: an example

parameters:

c: The connection object to the server

drawable: A drawable (Window or Pixmap) to draw on.

gc: The graphics context to use.

TODO: document which attributes of a gc are used

segments: An array of xcb_segment_t structures.