Function xcb::xproto::set_selection_owner

pub fn set_selection_owner<'a>(
    c: &'a Connection, 
    owner: Window, 
    selection: Atom, 
    time: Timestamp
) -> VoidCookie<'a>

Sets the owner of a selection

Makes window the owner of the selection selection and updates the last-change time of the specified selection.

TODO: briefly explain what a selection is.

parameters:

c: The connection object to the server

owner: The new owner of the selection.

The special value XCB_NONE means that the selection will have no owner.

selection: The selection.

time: Timestamp to avoid race conditions when running X over the network.

The selection will not be changed if time is earlier than the current last-change time of the selection or is later than the current X server time. Otherwise, the last-change time is set to the specified time.

The special value XCB_CURRENT_TIME will be replaced with the current server time.