Type Definition xcb::xproto::ColormapNotifyEvent

type ColormapNotifyEvent = Event<xcb_colormap_notify_event_t>;

the colormap for some window changed

Methods

impl ColormapNotifyEvent

pub fn window(&self) -> Window

The window whose associated colormap is changed, installed or uninstalled.

pub fn colormap(&self) -> Colormap

The colormap which is changed, installed or uninstalled. This is XCB_NONE when the colormap is changed by a call to FreeColormap.

pub fn new_(&self) -> bool

pub fn state(&self) -> u8

pub fn new(
window: Window,
colormap: Colormap,
new_: bool,
state: u8
) -> ColormapNotifyEvent

Constructs a new ColormapNotifyEvent response_type will be set automatically to COLORMAP_NOTIFY