Type Definition xcb::xproto::ConfigureNotifyEvent

type ConfigureNotifyEvent = Event<xcb_configure_notify_event_t>;

NOT YET DOCUMENTED

Implementations

impl ConfigureNotifyEvent

pub fn event(&self) -> Window

The reconfigured window or its parent, depending on whether StructureNotify or SubstructureNotify was selected.

pub fn window(&self) -> Window

The window whose size, position, border, and/or stacking order was changed.

pub fn above_sibling(&self) -> Window

If XCB_NONE, the window is on the bottom of the stack with respect to sibling windows. However, if set to a sibling window, the window is placed on top of this sibling window.

pub fn x(&self) -> i16

The X coordinate of the upper-left outside corner of window, relative to the parent window’s origin.

pub fn y(&self) -> i16

The Y coordinate of the upper-left outside corner of window, relative to the parent window’s origin.

pub fn width(&self) -> u16

The inside width of window, not including the border.

pub fn height(&self) -> u16

The inside height of window, not including the border.

pub fn border_width(&self) -> u16

The border width of window.

pub fn override_redirect(&self) -> bool

Window managers should ignore this window if override_redirect is 1.

pub fn new(
event: Window,
window: Window,
above_sibling: Window,
x: i16,
y: i16,
width: u16,
height: u16,
border_width: u16,
override_redirect: bool
) -> ConfigureNotifyEvent

Constructs a new ConfigureNotifyEvent response_type will be set automatically to CONFIGURE_NOTIFY