Type Definition xcb::xproto::EnterNotifyEvent

type EnterNotifyEvent = Event<xcb_enter_notify_event_t>;

the pointer is in a different window

Implementations

impl EnterNotifyEvent

pub fn detail(&self) -> u8

pub fn time(&self) -> Timestamp

pub fn root(&self) -> Window

The root window for the final cursor position.

pub fn event(&self) -> Window

The window on which the event was generated.

pub fn child(&self) -> Window

If the event window has subwindows and the final pointer position is in one of them, then child is set to that subwindow, XCB_WINDOW_NONE otherwise.

pub fn root_x(&self) -> i16

The pointer X coordinate relative to root’s origin at the time of the event.

pub fn root_y(&self) -> i16

The pointer Y coordinate relative to root’s origin at the time of the event.

pub fn event_x(&self) -> i16

If event is on the same screen as root, this is the pointer X coordinate relative to the event window’s origin.

pub fn event_y(&self) -> i16

If event is on the same screen as root, this is the pointer Y coordinate relative to the event window’s origin.

pub fn state(&self) -> u16

pub fn mode(&self) -> u8

pub fn same_screen_focus(&self) -> u8

pub fn new(
response_type: u8,
detail: u8,
time: Timestamp,
root: Window,
event: Window,
child: Window,
root_x: i16,
root_y: i16,
event_x: i16,
event_y: i16,
state: u16,
mode: u8,
same_screen_focus: u8
) -> EnterNotifyEvent

Constructs a new EnterNotifyEvent response_type must be set to one of: - ENTER_NOTIFY - LEAVE_NOTIFY