Type Definition xcb::xproto::KeyPressEvent

pub type KeyPressEvent = Event<xcb_key_press_event_t>;

a key was pressed/released

Implementations

impl KeyPressEvent

pub fn detail(&self) -> Keycode

The keycode (a number representing a physical key on the keyboard) of the key which was pressed.

pub fn time(&self) -> Timestamp

Time when the event was generated (in milliseconds).

pub fn root(&self) -> Window

The root window of child.

pub fn event(&self) -> Window

pub fn child(&self) -> Window

pub fn root_x(&self) -> i16

The X coordinate of the pointer relative to the root window at the time of the event.

pub fn root_y(&self) -> i16

The Y coordinate of the pointer relative to the root window at the time of the event.

pub fn event_x(&self) -> i16

If same_screen is true, this is the X coordinate relative to the event window’s origin. Otherwise, event_x will be set to zero.

pub fn event_y(&self) -> i16

If same_screen is true, this is the Y coordinate relative to the event window’s origin. Otherwise, event_y will be set to zero.

pub fn state(&self) -> u16

The logical state of the pointer buttons and modifier keys just prior to the event.

pub fn same_screen(&self) -> bool

Whether the event window is on the same screen as the root window.

pub fn new(
response_type: u8,
detail: Keycode,
time: Timestamp,
root: Window,
event: Window,
child: Window,
root_x: i16,
root_y: i16,
event_x: i16,
event_y: i16,
state: u16,
same_screen: bool
) -> KeyPressEvent

Constructs a new KeyPressEvent response_type must be set to one of: - KEY_PRESS - KEY_RELEASE