Type Definition xcb::xproto::UnmapNotifyEvent

pub type UnmapNotifyEvent = Event<xcb_unmap_notify_event_t>;

a window is unmapped

Implementations

impl UnmapNotifyEvent

pub fn event(&self) -> Window

The reconfigured window or its parent, depending on whether StructureNotify or SubstructureNotify was selected.

pub fn window(&self) -> Window

The window that was unmapped.

pub fn from_configure(&self) -> bool

Set to 1 if the event was generated as a result of a resizing of the window’s parent when window had a win_gravity of UnmapGravity.

pub fn new(
event: Window,
window: Window,
from_configure: bool
) -> UnmapNotifyEvent

Constructs a new UnmapNotifyEvent response_type will be set automatically to UNMAP_NOTIFY