Module xcb::xtest

The XTEST X extension.

Accessible with the xtest cargo feature.

Structs

CompareCursor

The CompareCursor request.

CompareCursorCookie

Cookie type for CompareCursor.

CompareCursorCookieUnchecked

Unchecked cookie type for CompareCursor.

CompareCursorReply

Reply type for CompareCursor.

FakeInput

The FakeInput request.

GetVersion

The GetVersion request.

GetVersionCookie

Cookie type for GetVersion.

GetVersionCookieUnchecked

Unchecked cookie type for GetVersion.

GetVersionReply

Reply type for GetVersion.

GrabControl

The GrabControl request.

Enums

Cursor

Constants

MAJOR_VERSION

The major version of the Test extension.

MINOR_VERSION

The minor version of the Test extension.

VERSION_STRING

The version string of the Test extension.

XNAME

The official identifier for the Test extension.

Functions

get_extension_data

Fetch server runtime info data of the Test extension.

prefetch_extension_data

Prefetch server runtime info data of the Test extension.