Module xcb::xv

The XVideo X extension.

Accessible with the xv cargo feature.

Structs

AdaptorInfo

AdaptorInfoBuf

AdaptorInfoIterator

AttributeFlag

AttributeInfo

AttributeInfoBuf

AttributeInfoIterator

BadControlError

The BadControlError error.

BadEncodingError

The BadEncodingError error.

BadPortError

The BadPortError error.

Encoding

EncodingInfo

EncodingInfoBuf

EncodingInfoIterator

Format

GetPortAttribute

The GetPortAttribute request.

GetPortAttributeCookie

Cookie type for GetPortAttribute.

GetPortAttributeCookieUnchecked

Unchecked cookie type for GetPortAttribute.

GetPortAttributeReply

Reply type for GetPortAttribute.

GetStill

The GetStill request.

GetVideo

The GetVideo request.

GrabPort

The GrabPort request.

GrabPortCookie

Cookie type for GrabPort.

GrabPortCookieUnchecked

Unchecked cookie type for GrabPort.

GrabPortReply

Reply type for GrabPort.

Image

ImageBuf

ImageFormatInfo

ImageIterator

ListImageFormats

The ListImageFormats request.

ListImageFormatsCookie

Cookie type for ListImageFormats.

ListImageFormatsCookieUnchecked

Unchecked cookie type for ListImageFormats.

ListImageFormatsReply

Reply type for ListImageFormats.

Port

PortNotifyEvent

The PortNotifyEvent event.

PutImage

The PutImage request.

PutStill

The PutStill request.

PutVideo

The PutVideo request.

QueryAdaptors

The QueryAdaptors request.

QueryAdaptorsCookie

Cookie type for QueryAdaptors.

QueryAdaptorsCookieUnchecked

Unchecked cookie type for QueryAdaptors.

QueryAdaptorsReply

Reply type for QueryAdaptors.

QueryBestSize

The QueryBestSize request.

QueryBestSizeCookie

Cookie type for QueryBestSize.

QueryBestSizeCookieUnchecked

Unchecked cookie type for QueryBestSize.

QueryBestSizeReply

Reply type for QueryBestSize.

QueryEncodings

The QueryEncodings request.

QueryEncodingsCookie

Cookie type for QueryEncodings.

QueryEncodingsCookieUnchecked

Unchecked cookie type for QueryEncodings.

QueryEncodingsReply

Reply type for QueryEncodings.

QueryExtension

The QueryExtension request.

QueryExtensionCookie

Cookie type for QueryExtension.

QueryExtensionCookieUnchecked

Unchecked cookie type for QueryExtension.

QueryExtensionReply

Reply type for QueryExtension.

QueryImageAttributes

The QueryImageAttributes request.

QueryImageAttributesCookie

Cookie type for QueryImageAttributes.

QueryImageAttributesCookieUnchecked

Unchecked cookie type for QueryImageAttributes.

QueryImageAttributesReply

Reply type for QueryImageAttributes.

QueryPortAttributes

The QueryPortAttributes request.

QueryPortAttributesCookie

Cookie type for QueryPortAttributes.

QueryPortAttributesCookieUnchecked

Unchecked cookie type for QueryPortAttributes.

QueryPortAttributesReply

Reply type for QueryPortAttributes.

Rational

SelectPortNotify

The SelectPortNotify request.

SelectVideoNotify

The SelectVideoNotify request.

SetPortAttribute

The SetPortAttribute request.

ShmPutImage

The ShmPutImage request.

StopVideo

The StopVideo request.

Type

UngrabPort

The UngrabPort request.

VideoNotifyEvent

The VideoNotifyEvent event.

Enums

Error

Unified error type for the Xv extension

Event

Unified event type for the Xv extension

GrabPortStatus

ImageFormatInfoFormat

ImageFormatInfoType

ScanlineOrder

VideoNotifyReason

Constants

MAJOR_VERSION

The major version of the Xv extension.

MINOR_VERSION

The minor version of the Xv extension.

VERSION_STRING

The version string of the Xv extension.

XNAME

The official identifier for the Xv extension.

Functions

get_extension_data

Fetch server runtime info data of the Xv extension.

prefetch_extension_data

Prefetch server runtime info data of the Xv extension.